ExperimentRegistry Constructor

makeExperimentRegistry constructs a special Registry which is suitable for the definition of large scale computer experiments.

Each experiments consists of a Problem and an Algorithm. These can be parametrized with addExperiments to actually define computational jobs.

makeExperimentRegistry(file.dir = "registry", work.dir = getwd(),
  conf.file = findConfFile(), packages = character(0L),
  namespaces = character(0L), source = character(0L),
  load = character(0L), seed = NULL, make.default = TRUE)

Arguments

file.dir

[character(1)] Path where all files of the registry are saved. Default is directory “registry” in the current working directory. The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well. If you pass NA, a temporary directory will be used. This way, you can create disposable registries for btlapply or examples. By default, the temporary directory tempdir() will be used. If you want to use another directory, e.g. a directory which is shared between nodes, you can set it in your configuration file by setting the variable temp.dir.

work.dir

[character(1)] Working directory for R process for running jobs. Defaults to the working directory currently set during Registry construction (see getwd). loadRegistry uses the stored work.dir, but you may also explicitly overwrite it, e.g., after switching to another system. The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well.

conf.file

[character(1)] Path to a configuration file which is sourced while the registry is created. For example, you can set cluster functions or default resources in it. The script is executed inside the environment of the registry after the defaults for all variables are set, thus you can set and overwrite slots, e.g. default.resources = list(walltime = 3600) to set default resources. The file lookup defaults to a heuristic which first tries to read “batchtools.conf.R” in the current working directory. If not found, it looks for a configuration file “config.R” in the OS dependent user configuration directory as reported by via rappdirs::user_config_dir("batchtools", expand = FALSE) (e.g., on linux this usually resolves to “~/.config/batchtools/config.R”). If this file is also not found, the heuristic finally tries to read the file “.batchtools.conf.R” in the home directory (“~”). Set to character(0) if you want to disable this heuristic.

packages

[character] Packages that will always be loaded on each node. Uses require internally. Default is character(0).

namespaces

[character] Same as packages, but the packages will not be attached. Uses requireNamespace internally. Default is character(0).

source

[character] Files which should be sourced on the slaves prior to executing a job. Calls sys.source using the .GlobalEnv.

load

[character] Files which should be loaded on the slaves prior to executing a job. Calls load using the .GlobalEnv.

seed

[integer(1)] Start seed for jobs. Each job uses the (seed + job.id) as seed. Default is a random number in the range [1, .Machine$integer.max/2].

make.default

[logical(1)] If set to TRUE, the created registry is saved inside the package namespace and acts as default registry. You might want to switch this off if you work with multiple registries simultaneously. Default is TRUE.

Value

[ExperimentRegistry].

See also

Other Experiment: addExperiments, removeExperiments, summarizeExperiments

Examples

tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

# Definde one problem, two algorithms and add them with some parameters:
addProblem(reg = tmp, "p1",
  fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd))
addAlgorithm(reg = tmp, "a1", fun = function(job, data, instance, ...) mean(instance))
addAlgorithm(reg = tmp, "a2", fun = function(job, data, instance, ...) median(instance))
ids = addExperiments(reg = tmp, list(p1 = CJ(n = c(50, 100), mean = -2:2, sd = 1:4)))

Adding 40 experiments ('p1'[40] x 'a1'[1] x repls[1]) ...

Adding 40 experiments ('p1'[40] x 'a2'[1] x repls[1]) ...

# Overview over defined experiments:
getProblemIds(reg = tmp)

#> [1] "p1"
#> 

getAlgorithmIds(reg = tmp)

#> [1] "a1" "a2"
#> 

summarizeExperiments(reg = tmp)

#>    problem algorithm .count
#> 1:      p1        a1     40
#> 2:      p1        a2     40
#> 

summarizeExperiments(reg = tmp, by = c("problem", "algorithm", "n"))

#>    problem algorithm   n .count
#> 1:      p1        a1  50     20
#> 2:      p1        a1 100     20
#> 3:      p1        a2  50     20
#> 4:      p1        a2 100     20
#> 

ids = findExperiments(prob.pars = (n == 50), reg = tmp)
getJobPars(ids, reg = tmp)

#>     job.id problem algorithm  n mean sd
#>  1:      1      p1        a1 50   -2  1
#>  2:      2      p1        a1 50   -2  2
#>  3:      3      p1        a1 50   -2  3
#>  4:      4      p1        a1 50   -2  4
#>  5:      5      p1        a1 50   -1  1
#>  6:      6      p1        a1 50   -1  2
#>  7:      7      p1        a1 50   -1  3
#>  8:      8      p1        a1 50   -1  4
#>  9:      9      p1        a1 50    0  1
#> 10:     10      p1        a1 50    0  2
#> 11:     11      p1        a1 50    0  3
#> 12:     12      p1        a1 50    0  4
#> 13:     13      p1        a1 50    1  1
#> 14:     14      p1        a1 50    1  2
#> 15:     15      p1        a1 50    1  3
#> 16:     16      p1        a1 50    1  4
#> 17:     17      p1        a1 50    2  1
#> 18:     18      p1        a1 50    2  2
#> 19:     19      p1        a1 50    2  3
#> 20:     20      p1        a1 50    2  4
#> 21:     41      p1        a2 50   -2  1
#> 22:     42      p1        a2 50   -2  2
#> 23:     43      p1        a2 50   -2  3
#> 24:     44      p1        a2 50   -2  4
#> 25:     45      p1        a2 50   -1  1
#> 26:     46      p1        a2 50   -1  2
#> 27:     47      p1        a2 50   -1  3
#> 28:     48      p1        a2 50   -1  4
#> 29:     49      p1        a2 50    0  1
#> 30:     50      p1        a2 50    0  2
#> 31:     51      p1        a2 50    0  3
#> 32:     52      p1        a2 50    0  4
#> 33:     53      p1        a2 50    1  1
#> 34:     54      p1        a2 50    1  2
#> 35:     55      p1        a2 50    1  3
#> 36:     56      p1        a2 50    1  4
#> 37:     57      p1        a2 50    2  1
#> 38:     58      p1        a2 50    2  2
#> 39:     59      p1        a2 50    2  3
#> 40:     60      p1        a2 50    2  4
#>     job.id problem algorithm  n mean sd
#> 

# Chunk jobs per algorithm and submit them:
ids = chunkIds(getJobPars(reg = tmp), group.by = "algorithm", reg = tmp)
submitJobs(ids, reg = tmp)

Submitting 80 jobs in 2 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

Syncing 2 files ...

#> [1] TRUE
#> 

# Reduce the results of algorithm a1
ids.mean = findExperiments(algo.name = "a1", reg = tmp)
reduceResults(ids.mean, fun = function(aggr, res, ...) c(aggr, res), reg = tmp)

#>  [1] -1.91433603 -1.51182725 -1.98742936 -2.36910443 -0.98787171 -1.27022246
#>  [7] -1.25992955 -1.34513502 -0.15700077 -0.23088698  0.07027429  0.87552783
#> [13]  0.96372348  1.01272758  1.21870563  1.09210651  2.32384233  2.21704556
#> [19]  2.89319252  2.39510729 -1.97107549 -1.60247254 -1.72943243 -1.13280855
#> [25] -0.96364419 -1.02812339 -0.77289340 -1.46904223  0.02351158  0.17083372
#> [31]  0.18285818  0.08181464  1.03563515  0.74558157  0.85903262  0.90236768
#> [37]  2.06103503  1.88740822  1.88288169  1.96829931
#> 

# Join info table with all results and calculate mean of results
# grouped by n and algorithm
ids = findDone(reg = tmp)
pars = getJobPars(ids, reg = tmp)
results = reduceResultsDataTable(ids, fun = function(res) list(res = res), reg = tmp)
tab = ljoin(pars, results)
tab[, list(mres = mean(res)), by = c("n", "algorithm")]

#>      n algorithm       mres
#> 1:  50        a1 0.10142547
#> 2: 100        a1 0.05658836
#> 3:  50        a2 0.13367759
#> 4: 100        a2 0.03343343
#>