Add Experiments to the Registry

Adds experiments (parametrized combinations of problems with algorithms) to the registry and thereby defines batch jobs.

If multiple problem designs or algorithm designs are provided, they are combined via the Cartesian product. E.g., if you have two problems p1 and p2 and three algorithms a1, a2 and a3, addExperiments creates experiments for all parameters for the combinations (p1, a1), (p1, a2), (p1, a3), (p2, a1), (p2, a2) and (p2, a3).

addExperiments(prob.designs = NULL, algo.designs = NULL, repls = 1L,
  combine = "crossprod", reg = getDefaultRegistry())

Arguments

prob.designs

[named list of data.frame] Named list of data frames (or data.table). The name must match the problem name while the column names correspond to parameters of the problem. If NULL, experiments for all defined problems without any parameters are added.

algo.designs

[named list of data.table or data.frame] Named list of data frames (or data.table). The name must match the algorithm name while the column names correspond to parameters of the algorithm. If NULL, experiments for all defined algorithms without any parameters are added.

repls

[integer(1)] Number of replications for each experiment.

combine

[character(1)] How to combine the rows of a single problem design with the rows of a single algorithm design? Default is “crossprod” which combines each row of the problem design which each row of the algorithm design in a cross-product fashion. Set to “bind” to just cbind the tables of problem and algorithm designs where the shorter table is repeated if necessary.

reg

[ExperimentRegistry] Registry. If not explicitly passed, uses the last created registry.

Value

[data.table] with ids of added jobs stored in column “job.id”.

See also

Other Experiment: ExperimentRegistry, removeExperiments, summarizeExperiments

Examples

tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Creating directories in '/var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmphrwhkr/registry142dc1582b042'

# add first problem
fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd)
addProblem("p1", fun = fun, reg = tmp)

# add second problem
fun = function(job, data, n, lamba, ...) rexp(n, lambda = lambda)
addProblem("p2", fun = fun, reg = tmp)

# add first algorithm
fun = function(instance, method, ...) if (method == "mean") mean(instance) else median(instance)
addAlgorithm("a1", fun = fun, reg = tmp)

# add second algorithm
fun = function(instance, ...) se(instance)
addAlgorithm("a2", reg = tmp)

# define problem and algorithm designs
prob.designs = algo.designs = list()
prob.designs$p1 = expand.grid(n = 100, mean = -1:1, sd = 1:5)
prob.designs$p2 = data.table(lambda = 1:5)
algo.designs$a1 = data.table(method = c("mean", "median"))
algo.designs$a2 = data.table()

# add experiments
addExperiments(prob.designs, algo.designs, reg = tmp)

#> Adding 30 experiments ('p1'[15] x 'a1'[2] x repls[1]) ...

#> Adding 15 experiments ('p1'[15] x 'a2'[1] x repls[1]) ...

#> Adding 10 experiments ('p2'[5] x 'a1'[2] x repls[1]) ...

#> Adding 5 experiments ('p2'[5] x 'a2'[1] x repls[1]) ...

# check what has been created
summarizeExperiments(reg = tmp)

#>    problem algorithm .count
#> 1:      p1        a1     30
#> 2:      p1        a2     15
#> 3:      p2        a1     10
#> 4:      p2        a2      5

getJobPars(reg = tmp)

#>     job.id problem algorithm   n mean sd method lambda
#>  1:      1      p1        a1 100   -1  1   mean     NA
#>  2:      2      p1        a1 100   -1  1 median     NA
#>  3:      3      p1        a1 100    0  1   mean     NA
#>  4:      4      p1        a1 100    0  1 median     NA
#>  5:      5      p1        a1 100    1  1   mean     NA
#>  6:      6      p1        a1 100    1  1 median     NA
#>  7:      7      p1        a1 100   -1  2   mean     NA
#>  8:      8      p1        a1 100   -1  2 median     NA
#>  9:      9      p1        a1 100    0  2   mean     NA
#> 10:     10      p1        a1 100    0  2 median     NA
#> 11:     11      p1        a1 100    1  2   mean     NA
#> 12:     12      p1        a1 100    1  2 median     NA
#> 13:     13      p1        a1 100   -1  3   mean     NA
#> 14:     14      p1        a1 100   -1  3 median     NA
#> 15:     15      p1        a1 100    0  3   mean     NA
#> 16:     16      p1        a1 100    0  3 median     NA
#> 17:     17      p1        a1 100    1  3   mean     NA
#> 18:     18      p1        a1 100    1  3 median     NA
#> 19:     19      p1        a1 100   -1  4   mean     NA
#> 20:     20      p1        a1 100   -1  4 median     NA
#> 21:     21      p1        a1 100    0  4   mean     NA
#> 22:     22      p1        a1 100    0  4 median     NA
#> 23:     23      p1        a1 100    1  4   mean     NA
#> 24:     24      p1        a1 100    1  4 median     NA
#> 25:     25      p1        a1 100   -1  5   mean     NA
#> 26:     26      p1        a1 100   -1  5 median     NA
#> 27:     27      p1        a1 100    0  5   mean     NA
#> 28:     28      p1        a1 100    0  5 median     NA
#> 29:     29      p1        a1 100    1  5   mean     NA
#> 30:     30      p1        a1 100    1  5 median     NA
#> 31:     31      p1        a2 100   -1  1     NA     NA
#> 32:     32      p1        a2 100    0  1     NA     NA
#> 33:     33      p1        a2 100    1  1     NA     NA
#> 34:     34      p1        a2 100   -1  2     NA     NA
#> 35:     35      p1        a2 100    0  2     NA     NA
#> 36:     36      p1        a2 100    1  2     NA     NA
#> 37:     37      p1        a2 100   -1  3     NA     NA
#> 38:     38      p1        a2 100    0  3     NA     NA
#> 39:     39      p1        a2 100    1  3     NA     NA
#> 40:     40      p1        a2 100   -1  4     NA     NA
#> 41:     41      p1        a2 100    0  4     NA     NA
#> 42:     42      p1        a2 100    1  4     NA     NA
#> 43:     43      p1        a2 100   -1  5     NA     NA
#> 44:     44      p1        a2 100    0  5     NA     NA
#> 45:     45      p1        a2 100    1  5     NA     NA
#> 46:     46      p2        a1  NA   NA NA   mean      1
#> 47:     47      p2        a1  NA   NA NA median      1
#> 48:     48      p2        a1  NA   NA NA   mean      2
#> 49:     49      p2        a1  NA   NA NA median      2
#> 50:     50      p2        a1  NA   NA NA   mean      3
#> 51:     51      p2        a1  NA   NA NA median      3
#> 52:     52      p2        a1  NA   NA NA   mean      4
#> 53:     53      p2        a1  NA   NA NA median      4
#> 54:     54      p2        a1  NA   NA NA   mean      5
#> 55:     55      p2        a1  NA   NA NA median      5
#> 56:     56      p2        a2  NA   NA NA     NA      1
#> 57:     57      p2        a2  NA   NA NA     NA      2
#> 58:     58      p2        a2  NA   NA NA     NA      3
#> 59:     59      p2        a2  NA   NA NA     NA      4
#> 60:     60      p2        a2  NA   NA NA     NA      5
#>     job.id problem algorithm   n mean sd method lambda