Export Objects to the Slaves

Objects are saved in subdirectory “exports” of the “file.dir” of reg. They are automatically loaded and placed in the global environment each time the registry is loaded or a job collection is executed.

batchExport(export = list(), unexport = character(0L),
  reg = getDefaultRegistry())

Arguments

export

[list] Named list of objects to export.

unexport

[character] Vector of object names to un-export.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] with name and uri to the exported objects.

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

# list exports
exports = batchExport(reg = tmp)
print(exports)

#> Empty data.table (0 rows) of 2 cols: name,uri
#> 

#> NULL
#> 

# add a job and required exports
batchMap(function(x) x^2 + y + z, x = 1:3, reg = tmp)

Adding 3 jobs ...

exports = batchExport(export = list(y = 99, z = 1), reg = tmp)

Exporting new objects: 'y','z'

print(exports)

#>    name                                         uri
#> 1:    y /scratch/registry15b26c4076ea/exports/y.rds
#> 2:    z /scratch/registry15b26c4076ea/exports/z.rds
#> 

submitJobs(reg = tmp)

Submitting 3 jobs in 3 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

Syncing 3 files ...

#> [1] TRUE
#> 

stopifnot(loadResult(1, reg = tmp) == 101)

# Un-export z
exports = batchExport(unexport = "z", reg = tmp)

Un-exporting exported objects: 'z'

print(exports)

#>    name                                         uri
#> 1:    y /scratch/registry15b26c4076ea/exports/y.rds
#>