Execute Jobs of a JobCollection

Executes every job in a JobCollection. This function is intended to be called on the slave.

doJobCollection(jc, output = NULL)

Arguments

jc

[JobCollection] Either an object of class “JobCollection” as returned by makeJobCollection or a string with the path to file containing a “JobCollection” as RDS file (as stored by submitJobs).

output

[character(1)] Path to a file to write the output to. Defaults to NULL which means that output is written to the active sink.

Value

[character(1)]: Hash of the JobCollection executed.

See also

Other JobCollection: makeJobCollection

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

batchMap(identity, 1:2, reg = tmp)

Adding 2 jobs ...

jc = makeJobCollection(1:2, reg = tmp)
doJobCollection(jc)

#> ### [bt 2016-10-09 23:15:29]: Starting calculation of 2 jobs
#> ### [bt 2016-10-09 23:15:29]: Setting working directory to '/home/michel/Projekte/batchtools/docs/reference'
#> ### [bt 2016-10-09 23:15:29]: Memory measurement disabled
#> ### [bt 2016-10-09 23:15:29]: Starting job [batchtools job.id=1]
#> 
#> ### [bt 2016-10-09 23:15:29]: Job terminated successfully [batchtools job.id=1]
#> ### [bt 2016-10-09 23:15:29]: Starting job [batchtools job.id=2]
#> 
#> ### [bt 2016-10-09 23:15:29]: Job terminated successfully [batchtools job.id=2]
#> ### [bt 2016-10-09 23:15:29]: Calculation finished!
#>