Find and Filter Jobs

These functions are used to find and filter jobs, depending on either their parameters (findJobs and findExperiments), their tags (findTagged), or their computational status (all other functions).

For a summarizing overview over the status, see getStatus. Note that findOnSystem and findExpired are somewhat heuristic and may report misleading results, depending on the state of the system and the ClusterFunctions implementation.

findJobs(expr, ids = NULL, reg = getDefaultRegistry())

findExperiments(prob.name = NULL, algo.name = NULL, prob.pars, algo.pars,
  repls = NULL, ids = NULL, reg = getDefaultRegistry())

findSubmitted(ids = NULL, reg = getDefaultRegistry())

findNotSubmitted(ids = NULL, reg = getDefaultRegistry())

findStarted(ids = NULL, reg = getDefaultRegistry())

findNotStarted(ids = NULL, reg = getDefaultRegistry())

findDone(ids = NULL, reg = getDefaultRegistry())

findNotDone(ids = NULL, reg = getDefaultRegistry())

findErrors(ids = NULL, reg = getDefaultRegistry())

findOnSystem(ids = NULL, reg = getDefaultRegistry())

findRunning(ids = NULL, reg = getDefaultRegistry())

findQueued(ids = NULL, reg = getDefaultRegistry())

findExpired(ids = NULL, reg = getDefaultRegistry())

findTagged(tags = character(0L), ids = NULL, reg = getDefaultRegistry())

Arguments

expr

[expression] Predicate expression evaluated in the job parameters. Jobs for which expr evaluates to TRUE are returned.

ids

[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to all jobs.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

prob.name

[character] Whitelist of problem names. If the string starts with “~”, it is treated as regular expression. If not provided, all problems are matched.

algo.name

[character] Whitelist of algorithm names. If the string starts with “~”, it is treated as regular expression. If not provided, all algorithms are matched.

prob.pars

[expression] Predicate expression evaluated in the problem parameters.

algo.pars

[expression] Predicate expression evaluated in the algorithm parameters.

repls

[integer] Whitelist of replication numbers. If not provided, all replications are matched.

tags

[character] Return jobs which are tagged with any of the tags provided.

Value

[data.table] with column “job.id” containing matched jobs.

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Creating directories in '/var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T//Rtmphrwhkr/registry142dc6cc15681'

batchMap(identity, i = 1:3, reg = tmp)

#> Adding 3 jobs ...

ids = findNotSubmitted(reg = tmp)

# get all jobs:
findJobs(reg = tmp)

#>    job.id
#> 1:      1
#> 2:      2
#> 3:      3

# filter for jobs with parameter i >= 2
findJobs(i >= 2, reg = tmp)

#>    job.id
#> 1:      2
#> 2:      3

# filter on the computational status
findSubmitted(reg = tmp)

#> Empty data.table (0 rows) of 1 col: job.id

findNotDone(reg = tmp)

#>    job.id
#> 1:      1
#> 2:      2
#> 3:      3

# filter on tags
addJobTags(2:3, "my_tag", reg = tmp)
findTagged(tags = "my_tag", reg = tmp)

#>    job.id
#> 1:      2
#> 2:      3

# combine filter functions using joins
# -> jobs which are not done and not tagged (using an anti-join):
ajoin(findNotDone(reg = tmp), findTagged("my_tag", reg = tmp))

#>    job.id
#> 1:      1