Query Job Information

getJobStatus returns the internal table which stores information about the computational status of jobs, getJobPars a table with the job parameters, getJobResources a table with the resources which were set to submit the jobs, and getJobTags the tags of the jobs (see Tags).

getJobTable returns all these tables joined.

getJobTable(ids = NULL, flatten = NULL, prefix = FALSE,
  reg = getDefaultRegistry())

getJobStatus(ids = NULL, reg = getDefaultRegistry())

getJobResources(ids = NULL, flatten = NULL, prefix = FALSE,
  reg = getDefaultRegistry())

getJobPars(ids = NULL, flatten = NULL, prefix = FALSE,
  reg = getDefaultRegistry())

getJobTags(ids = NULL, reg = getDefaultRegistry())

Arguments

ids

[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to all jobs.

flatten

[logical(1)] Transform the job parameters and/or resource specifications to data frame columns? Defaults to TRUE if all parameters or resources are atomics, FALSE otherwise where each row of the column will hold a named list. New columns will be named using prefix.

prefix

[logical(1)] If set to TRUE, the prefix “par.” is used to name column names of parameters for a Registry and prefixes “prob.par.” and “algo.par.” are used to name the columns of a ExperimentRegistry. Resources are prefixed with “res.”.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] with the following columns (not necessarily in this order):

job.id

Unique Job ID as integer.

submitted

Time the job was submitted to the batch system as POSIXct.

started

Time the job was started on the batch system as POSIXct.

done

Time the job terminated (successfully or with an error) as POSIXct.

error

Either NA if the job terminated successfully or the error message.

memory

Estimate of the memory usage.

batch.id

Batch ID as reported by the scheduler.

job.hash

Unique string identifying the job or chunk.

time.queued

Time in seconds (as difftime) the job was queued.

time.running

Time in seconds (as difftime) the job was running.

pars

List of parameters/arguments for this job. Possibly expanded to separate columns (see prefix.

resources

List of computational resources set for this job. Possibly expanded to separate columns (see prefix.

tags

Tags as joined string, delimited by “,”.

problem

Only for ExperimentRegistry: the problem identifier.

algorithm

Only for ExperimentRegistry: the algorithm identifier.

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

f = function(x) if (x < 0) stop("x must be > 0") else sqrt(x)
batchMap(f, x = c(-1, 0, 1), reg = tmp)

#> Adding 3 jobs ...

submitJobs(reg = tmp)

#> Submitting 3 jobs in 3 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 3 files ...

#> [1] FALSE

# Complete table:
getJobTable(reg = tmp, flatten = FALSE)

#>    job.id           submitted             started                done
#> 1:      1 2016-11-08 00:06:10 2016-11-08 00:06:10 2016-11-08 00:06:10
#> 2:      2 2016-11-08 00:06:10 2016-11-08 00:06:10 2016-11-08 00:06:10
#> 3:      3 2016-11-08 00:06:10 2016-11-08 00:06:10 2016-11-08 00:06:10
#>                                      error memory      batch.id
#> 1: Error in (function (x)  : x must be > 0     NA cfInteractive
#> 2:                                      NA     NA cfInteractive
#> 3:                                      NA     NA cfInteractive
#>                            job.hash time.queued time.running   pars resources
#> 1: e9b7879399a67d80628e531d4785ce52      0 secs       0 secs <list>    <list>
#> 2: e23b523716b4d8c1852a35b47e4bd0a5      0 secs       0 secs <list>    <list>
#> 3: 32cf8233132db57799eebcfc7336cd36      0 secs       0 secs <list>    <list>
#>    tags
#> 1:   NA
#> 2:   NA
#> 3:   NA

# Table with tags:
addJobTags(1:2, "tag1", reg = tmp)
addJobTags(2, "tag2", reg = tmp)
getJobTags(reg = tmp)

#>    job.id      tags
#> 1:      1      tag1
#> 2:      2 tag1,tag2
#> 3:      3        NA

# Job parameters with tags right-joined:
rjoin(getJobPars(reg = tmp), getJobTags(reg = tmp))

#>    job.id  x      tags
#> 1:      1 -1      tag1
#> 2:      2  0 tag1,tag2
#> 3:      3  1        NA