Summarize the Computational Status

This function gives an encompassing overview over the computational status on your system.

getStatus(ids = NULL, reg = getDefaultRegistry())

Arguments

ids

[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to all jobs.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] (with class “Status” for printing).

See also

Other debug: getErrorMessages, grepLogs, killJobs, resetJobs, showLog, testJob

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

fun = function(i) if (i == 3) stop(i) else i
ids = batchMap(fun, i = 1:5, reg = tmp)

Adding 5 jobs ...

submitJobs(ids = 1:4, reg = tmp)

Submitting 4 jobs in 4 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

Syncing 4 files ...

#> [1] FALSE
#> 

tab = getStatus(reg = tmp)
print(tab)

#> Status for 5 jobs:
#>   Submitted : 4 ( 80.0%)
#>   Queued    : 0 (  0.0%)
#>   Started   : 4 ( 80.0%)
#>   Running   : 0 (  0.0%)
#>   Done      : 3 ( 60.0%)
#>   Error     : 1 ( 20.0%)
#>   Expired   : 0 (  0.0%)
#> 

str(tab)

#> Classes ‘Status’, ‘data.table’ and 'data.frame':	1 obs. of  9 variables:
#>  $ defined  : int 5
#>  $ submitted: int 4
#>  $ started  : int 4
#>  $ done     : int 3
#>  $ error    : int 1
#>  $ queued   : int 0
#>  $ running  : int 0
#>  $ expired  : int 0
#>  $ system   : int 0
#>  - attr(*, ".internal.selfref")=<externalptr> 
#>