ClusterFunctions Constructor

This is the constructor used to create custom cluster functions. Note that some standard implementations for Torque, Slurm, LSF, SGE, etc. ship with the package.

makeClusterFunctions(name, submitJob, killJob = NULL, listJobsQueued = NULL,
  listJobsRunning = NULL, array.var = NA_character_, store.job = TRUE,
  hooks = list())

Arguments

name

[character(1)] Name of cluster functions.

submitJob

[function(reg, jc, ...)] Function to submit new jobs. Must return a SubmitJobResult object. The arguments are reg (Registry) and jobs (JobCollection).

killJob

[function(reg, batch.id)] Function to kill a job on the batch system. Make sure that you definitely kill the job! Return value is currently ignored. Must have the arguments reg (Registry) and batch.id (character(1) as returned by submitJob). Note that there is a helper function cfKillJob to repeatedly try to kill jobs. Set killJob to NULL if killing jobs cannot be supported.

listJobsQueued

[function(reg)] List all queued jobs on the batch system for the current user and registry. Must return an character vector of batch ids, same format as they are returned by submitJob. It does not matter if you return a few job ids too many (e.g., all for the current user instead of all for the current registry), but you have to include all relevant ones. Must have the argument are reg (Registry). Set listJobsQueued to NULL if listing of queued jobs is not supported.

listJobsRunning

[function(reg)] List all running jobs on the batch system for the current user and registry. This includes running, held, idle, etc. jobs. Must return an character vector of batch ids, same format as they are returned by submitJob. It does not matter if you return a few job ids too many (e.g. all for the current user instead of all for the current registry), but you have to include all relevant ones. Must have the argument are reg (Registry). Set listJobsRunning to NULL if listing of running jobs is not supported.

array.var

[character(1)] Name of the environment variable set by the scheduler to identify IDs of job arrays. Default is NA for no array support.

store.job

[logical(1)] Store the job on the file system before submitting? Default is TRUE.

hooks

[list] Named list of functions which will we called on certain events like “pre.submit” or “post.sync”. See Hooks.

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSSH, makeClusterFunctionsSlurm, makeClusterFunctionsSocket, makeClusterFunctionsTorque

Other ClusterFunctionsHelper: cfBrewTemplate, cfHandleUnknownSubmitError, cfKillJob, cfReadBrewTemplate, makeSubmitJobResult, runOSCommand