ClusterFunctions for Remote SSH Execution

Jobs are spawned by starting multiple R sessions via Rscript over SSH. If the hostname of the Worker equals “localhost”, Rscript is called directly so that you do not need to have an SSH client installed.

makeClusterFunctionsSSH(workers)

Arguments

workers

[list of Worker] List of Workers as constructed with Worker.

Value

[ClusterFunctions].

Note

If you use a custom “.ssh/config” file, make sure your ProxyCommand passes ‘-q’ to ssh, otherwise each output will end with the message “Killed by signal 1” and this will break the communication with the nodes.

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSlurm, makeClusterFunctionsSocket, makeClusterFunctionsTorque, makeClusterFunctions

Examples

## Not run: ------------------------------------
# # cluster functions for multicore execution on the local machine
# makeClusterFunctionsSSH(list(Worker$new("localhost", ncpus = 2)))
## ---------------------------------------------