ClusterFunctions for Slurm Systems

Cluster functions for Slurm (http://slurm.schedmd.com/).

Job files are created based on the brew template template.file. This file is processed with brew and then submitted to the queue using the sbatch command. Jobs are killed using the scancel command and the list of running jobs is retrieved using squeue. The user must have the appropriate privileges to submit, delete and list jobs on the cluster (this is usually the case).

The template file can access all resources passed to submitJobs as well as all variables stored in the JobCollection. It is the template file's job to choose a queue for the job and handle the desired resource allocations.

Note that you might have to specify the cluster name here if you do not want to use the default, otherwise the commands for listing and killing jobs will not work.

makeClusterFunctionsSlurm(template = "slurm", clusters = NULL,
  array.jobs = TRUE, nodename = "localhost", scheduler.latency = 1,
  fs.latency = 65)

Arguments

template
[character(1)] Either a path to a brew template file (with extension “tmpl”), or a short descriptive name enabling the following heuristic for the file lookup:

“batchtools.[template].tmpl” in the path specified by the environment variable “R_BATCHTOOLS_SEARCH_PATH”.

“batchtools.[template].tmpl” in the current working directory.

“[template].tmpl” in the user config directory (see user_config_dir); on linux this is usually “~/.config/batchtools/[template].tmpl”.

“.batchtools.[template].tmpl” in the home directory.

“[template].tmpl” in the package installation directory in the subfolder “templates”.

Here, the default for template is “slurm”. Alternatively, the template itself can be provided directly as a string which is then passed to brew. The string must include at least on newline (“\n”) in this case.

clusters
[character(1)] If multiple clusters are managed by one Slurm system, the name of one cluster has to be specified. If only one cluster is present, this argument may be omitted. Note that you should not select the cluster in your template file via #SBATCH --clusters.

array.jobs
[logical(1)] If array jobs are disabled on the computing site, set to FALSE.

nodename
[character(1)] Nodename of the master. All commands are send via SSH to this host. Only works iff

Passwordless authentication (e.g., via SSH public key authentication) is set up.

The file directory is shared across machines, e.g. mounted via SSHFS.

The absolute path to the file.dir are identical on the machines, or paths are provided relative to the home directory.

scheduler.latency
[numeric(1)] Time to sleep after important interactions with the scheduler to ensure a sane state. Currently only triggered after calling submitJobs.

fs.latency
[numeric(1)] Expected maximum latency of the file system, in seconds. Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to access files and directories. Usually safe to set to NA to disable the heuristic, e.g. if you are working on a local file system.

Value

[ClusterFunctions].

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSSH, makeClusterFunctionsSocket, makeClusterFunctionsTORQUE, makeClusterFunctions