ClusterFunctions for Parallel Socket Execution

Jobs are spawned asynchronously using the package snow.

makeClusterFunctionsSocket(ncpus = NA_integer_, fs.latency = 65)

Arguments

ncpus
[integer(1)] Number of CPUs. Default is to use all logical cores. The total number of cores "available" can be set via the option mc.cores and defaults to the heuristic implemented in detectCores.

fs.latency
[numeric(1)] Expected maximum latency of the file system, in seconds. Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to access files and directories. Usually safe to set to NA to disable the heuristic, e.g. if you are working on a local file system.

Value

[ClusterFunctions].

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSSH, makeClusterFunctionsSlurm, makeClusterFunctionsTORQUE, makeClusterFunctions