ClusterFunctions for OpenPBS/Torque Systems

Cluster functions for Torque/PBS (http://www.adaptivecomputing.com/products/open-source/torque/).

Job files are created based on the brew template template.file. This file is processed with brew and then submitted to the queue using the qsub command. Jobs are killed using the qdel command and the list of running jobs is retrieved using qselect. The user must have the appropriate privileges to submit, delete and list jobs on the cluster (this is usually the case).

The template file can access all resources passed to submitJobs as well as all variables stored in the JobCollection. It is the template file's job to choose a queue for the job and handle the desired resource allocations.

makeClusterFunctionsTorque(template = findTemplateFile("torque"))

Arguments

template

[character(1)] Path to a brew template file that can be used for the job description. Alternatively, the template as a single string (including at least one newline “\n”). Defaults to a heuristic which looks for a template file in the following locations and picks the first one found:

“batchtools.torque.tmpl” in the current working directory.

“torque.tmpl” in the user config directory (see user_config_dir); on linux is is usually “~/.config/batchtools/torque.tmpl”.

“.batchtools.torque.tmpl” in the home directory.

“torque.default.tmpl” in the package installation directory in the subfolder “templates” (this probably does not exist for your cluster system or needs adaptation).

Value

[ClusterFunctions].

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSSH, makeClusterFunctionsSlurm, makeClusterFunctionsSocket, makeClusterFunctions