Apply Functions on Results

Applies a function on the results of your finished jobs and thereby collects them in a list or data.table. The later requires the provided function to return a list (or data.frame) of scalar values. See rbindlist for features and limitations of the aggregation.

If not all jobs are terminated, the respective result will be NULL.

reduceResultsList(ids = NULL, fun = NULL, ..., reg = getDefaultRegistry())

reduceResultsDataTable(ids = NULL, fun = NULL, ..., fill = FALSE,
  reg = getDefaultRegistry())

Arguments

ids

[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to the return value of findDone.

fun

[function] Function to apply to each result. The result is passed unnamed as first argument. If NULL, the identity is used. If the function has the formal argument “job”, the Job/Experiment is also passed to the function.

...

[ANY] Additional arguments passed to to function fun.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

fill

[logical(1)] In reduceResultsDataTable: This flag is passed down to rbindlist which is used to convert the results to a data.table.

Value

reduceResultsList returns a list of the results in the same order as the provided ids. reduceResultsDataTable returns a data.table with columns “job.id” and additional result columns created via rbindlist, sorted by “job.id”.

Note

If you have thousands of jobs, disabling the progress bar (options(batchtools.progress = FALSE)) can significantly increase the performance.

See also

reduceResults.

Other Results: batchMapResults, loadResult, reduceResults

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

batchMap(function(x) x^2, x = 1:10, reg = tmp)

#> Adding 10 jobs ...

submitJobs(reg = tmp)

#> Submitting 10 jobs in 10 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 10 files ...

#> [1] TRUE

reduceResultsList(fun = sqrt, reg = tmp)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] 6
#> 
#> [[7]]
#> [1] 7
#> 
#> [[8]]
#> [1] 8
#> 
#> [[9]]
#> [1] 9
#> 
#> [[10]]
#> [1] 10
#>