Remove a Registry from the File System

All files will be erased from the file system, including all results. If you wish to remove only intermediate files, use sweepRegistry.

removeRegistry(wait = 5, reg = getDefaultRegistry())

Arguments

wait
[numeric(1)] Seconds to wait before proceeding. This is a safety measure to not accidentally remove your precious files. Set to 0 in non-interactive scripts to disable this precaution.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[character(1)]: Path of the deleted file directory.

See also

Other Registry: clearRegistry, getDefaultRegistry, loadRegistry, makeRegistry, saveRegistry, sweepRegistry, syncRegistry

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

removeRegistry(0, tmp)

#> Recursively removing files in '/tmp/batchtools-example/reg' ...