Run OS Commands on Local or Remote Machines

This is a helper function to run arbitrary OS commands on local or remote machines. The interface is similar to system2, but it always returns the exit status and the output.

Usage

runOSCommand(
  sys.cmd,
  sys.args = character(0L),
  stdin = "",
  nodename = "localhost"
)

Arguments

sys.cmd

[character(1)]
Command to run.

sys.args

[character]
Arguments for sys.cmd.

stdin

[character(1)]
Argument passed to system2.

nodename

[character(1)]
Name of the SSH node to run the command on. If set to “localhost” (default), the command is not piped through SSH.

Value

[named list] with “sys.cmd”, “sys.args”, “exit.code” (integer), “output” (character).

See also

Examples

if (FALSE) { # \dontrun{
runOSCommand("ls")
runOSCommand("ls", "-al")
runOSCommand("notfound")
} # }