Inspect Log Files

showLog opens the log in the pager. For customization, see file.show. getLog returns the log as character vector.

showLog(id, reg = getDefaultRegistry())

getLog(id, reg = getDefaultRegistry())

Arguments

id

[integer(1) or data.table] Single integer to specify the job or a data.table with column job.id and exactly one row.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

Nothing.

See also

Other debug: getErrorMessages, getStatus, grepLogs, killJobs, resetJobs, testJob

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

# Create some dummy jobs
fun = function(i) {
  if (i == 3) stop(i)
  if (i %% 2 == 1) warning("That's odd.")
}
ids = batchMap(fun, i = 1:5, reg = tmp)

Adding 5 jobs ...

submitJobs(reg = tmp)

Submitting 5 jobs in 5 chunks using cluster functions 'Interactive' ...

Warning message:
That's odd.

Warning message:
That's odd.

waitForJobs(reg = tmp)

Syncing 5 files ...

#> [1] FALSE
#> 

getStatus(reg = tmp)

#> Status for 5 jobs:
#>   Submitted : 5 (100.0%)
#>   Queued    : 0 (  0.0%)
#>   Started   : 5 (100.0%)
#>   Running   : 0 (  0.0%)
#>   Done      : 4 ( 80.0%)
#>   Error     : 1 ( 20.0%)
#>   Expired   : 0 (  0.0%)
#> 

writeLines(getLog(ids[1], reg = tmp))

#> ### [bt 2016-10-09 23:15:32]: Starting calculation of 1 jobs
#> ### [bt 2016-10-09 23:15:32]: Setting working directory to '/home/michel/Projekte/batchtools/docs/reference'
#> ### [bt 2016-10-09 23:15:32]: Memory measurement disabled
#> ### [bt 2016-10-09 23:15:32]: Starting job [batchtools job.id=1]
#> 
#> ### [bt 2016-10-09 23:15:32]: Job terminated successfully [batchtools job.id=1]
#> ### [bt 2016-10-09 23:15:32]: Calculation finished!
#> 

## Not run: ------------------------------------
# showLog(ids[1], reg = tmp)
## ---------------------------------------------

grepLogs(pattern = "warning", ignore.case = TRUE, reg = tmp)

#> Empty data.table (0 rows) of 2 cols: job.id,matches
#>