Submit Jobs to the Batch Systems

Submits defined jobs to the batch system.

If an additional column “chunk” is found in the table ids, jobs will be grouped accordingly to be executed sequentially on the same slave. The utility functions chunk, binpack and lpt can assist in grouping jobs. Jobs are submitted in the order of chunks, i.e. jobs which have chunk number unique(ids$chunk)[1] first, then jobs with chunk number unique(ids$chunk)[2] and so on. If no chunks are provided, jobs are submitted in the order of ids$job.id.

After submitting the jobs, you can use waitForJobs to wait for the termination of jobs or call reduceResultsList/reduceResults to collect partial results. The progress can be monitored with getStatus.

submitJobs(ids = NULL, resources = list(), sleep = NULL,
  reg = getDefaultRegistry())

Arguments

ids
[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to the return value of findNotSubmitted. Invalid ids are ignored.

resources
[named list] Computational resources for the batch jobs. The elements of this list (e.g. something like “walltime” or “nodes”) depend on your template file. See notes for reserved special resource names. Defaults can be stored in the configuration file by providing the named list default.resources. Settings in resources overwrite those in default.resources.

sleep
[function(i) | numeric(1)] Parameter to control the duration to sleep between temporary errors. You can pass an absolute numeric value in seconds or a function(i) which returns the number of seconds to sleep in the i-th iteration between temporary errors. If not provided (NULL), tries to read the value (number/function) from the configuration file (stored in reg$sleep) or defaults to a function with exponential backoff between 5 and 120 seconds.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] with columns “job.id” and “chunk”.

Note

Setting the resource measure.memory to TRUE turns on memory measurement: gc is called directly before and after the job and the difference is stored in the internal database. Note that this is just a rough estimate and does neither work reliably for external code like C/C++ nor in combination with threading.

If your cluster supports array jobs, you can set the resource chunks.as.arrayjobs to TRUE in order to execute chunks as job arrays. To do so, the job must be repeated nrow(jobs) times via the cluster functions template. The function doJobCollection (which is called on the slave) now retrieves the repetition number from the environment and restricts the computation to the respective job in the JobCollection.

Furthermore, the package provides support for inner parallelization using threading, sockets or MPI via the package parallelMap. If you set the resource “pm.backend” to “multicore”, “socket” or “mpi”, parallelStart is called on the slave before the first job in the chunk is started and parallelStop is called after the last job terminated. This way, the used resources for inner parallelization are set in the same place as the resources for the outer parallelization done by batchtools and all resources get stored together in the Registry. The user function just has to call parallelMap to start parallelization using the preconfigured backend.

Note that you should set the resource ncpus to control the number of CPUs to use in parallelMap. ncpus defaults to the number of available CPUs (as reported by (see detectCores)) on the executing machine for multicore and socket mode and defaults to the return value of mpi.universe.size-1 for MPI. Your template must be set up to handle the parallelization, e.g. start R with mpirun or request the right number of CPUs. You may pass further options like level to parallelStart via the named list “pm.opts”.

Also note that if you have thousands of jobs, disabling the progress bar (options(batchtools.progress = FALSE)) can significantly increase the performance of submitJobs.

Examples

### Example 1: Using memory measurement
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry38041485ab7e' using cluster functions 'Interactive'

# Toy function which creates a large matrix and returns the column sums
fun = function(n, p) colMeans(matrix(runif(n*p), n, p))

# Arguments to fun:
args = CJ(n = c(1e4, 1e5), p = c(10, 50)) # like expand.grid()
print(args)

#>        n     p
#>    <num> <num>
#> 1: 1e+04    10
#> 2: 1e+04    50
#> 3: 1e+05    10
#> 4: 1e+05    50

# Map function to create jobs
ids = batchMap(fun, args = args, reg = tmp)

#> Adding 4 jobs ...

# Set resources: enable memory measurement
res = list(measure.memory = TRUE)

# Submit jobs using the currently configured cluster functions
submitJobs(ids, resources = res, reg = tmp)

#> Submitting 4 jobs in 4 chunks using cluster functions 'Interactive' ...

# Retrive information about memory, combine with parameters
info = ijoin(getJobStatus(reg = tmp)[, .(job.id, memory)], getJobPars(reg = tmp))
print(info)

#>    job.id memory     n     p
#>     <int>  <num> <num> <num>
#> 1:      1  111.5 1e+04    10
#> 2:      2  116.4 1e+04    50
#> 3:      3  123.7 1e+05    10
#> 4:      4  184.7 1e+05    50

# Combine job info with results -> each job is aggregated using mean()
ijoin(info, reduceResultsDataTable(fun = function(res) list(res = mean(res)), reg = tmp))

#>    job.id memory     n     p       res
#>     <int>  <num> <num> <num>     <num>
#> 1:      1  111.5 1e+04    10 0.5006977
#> 2:      2  116.4 1e+04    50 0.4992960
#> 3:      3  123.7 1e+05    10 0.5002269
#> 4:      4  184.7 1e+05    50 0.5001242

### Example 2: Multicore execution on the slave
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry3804623222c9' using cluster functions 'Interactive'

# Function which sleeps 10 seconds, i-times
f = function(i) {
  parallelMap::parallelMap(Sys.sleep, rep(10, i))
}

# Create one job with parameter i=4
ids = batchMap(f, i = 4, reg = tmp)

#> Adding 1 jobs ...

# Set resources: Use parallelMap in multicore mode with 4 CPUs
# batchtools internally loads the namespace of parallelMap and then
# calls parallelStart() before the job and parallelStop() right
# after the job last job in the chunk terminated.
res = list(pm.backend = "multicore", ncpus = 4)
# NOT RUN {
# Submit both jobs and wait for them
submitJobs(resources = res, reg = tmp)
waitForJobs(reg = tmp)

# If successfull, the running time should be ~10s
getJobTable(reg = tmp)[, .(job.id, time.running)]

# There should also be a note in the log:
grepLogs(pattern = "parallelMap", reg = tmp)
# }