Define Algorithms for Experiments

Algorithms are functions which get the codedata part as well as the problem instance (the return value of the function defined in Problem) and return an arbitrary R object.

This function serializes all components to the file system and registers the algorithm in the ExperimentRegistry. removeAlgorithm removes all jobs from the registry which depend on the specific algorithm. getAlgorithmIds can be used to retrieve the IDs of already algorithms.

addAlgorithm(name, fun = NULL, reg = getDefaultRegistry())

removeAlgorithms(name, reg = getDefaultRegistry())

getAlgorithmIds(reg = getDefaultRegistry())

Arguments

name
[character(1)] Unique identifier for the algorithm.

fun
[function] The algorithm function. The static problem part is passed as “data”, the generated problem instance is passed as “instance” and the Job/Experiment as “job”. Therefore, your function must have the formal arguments “job”, “data” and “instance” (or dots ...).

If you do not provide a function, it defaults to a function which just returns the instance.

reg
[ExperimentRegistry] Registry. If not explicitly passed, uses the last created registry.

Value

[Algorithm]. Object of class “Algorithm”.

Examples

tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry7c583ed2de55' using cluster functions 'Interactive'

addProblem("p1", fun = function(job, data) data, reg = tmp)

#> Adding problem 'p1'

addProblem("p2", fun = function(job, data) job, reg = tmp)

#> Adding problem 'p2'

getProblemIds(reg = tmp)

#> [1] "p1" "p2"

addAlgorithm("a1", fun = function(job, data, instance) instance, reg = tmp)

#> Adding algorithm 'a1'

getAlgorithmIds(reg = tmp)

#> [1] "a1"

removeAlgorithms("a1", reg = tmp)

#> Removing Algorithm 'a1' and 0 corresponding jobs ...

#> Removing 0 obsolete result files ...

#> Removing 0 obsolete log files ...

#> Removing 0 obsolete job collection files ...

#> Removing 0 job description files ...

#> Removing 0 external directories of unsubmitted jobs ...

#> Removing 0 resource specifications ...

#> Removing 0 tags ...

getAlgorithmIds(reg = tmp)

#> character(0)