batchtools: Tools for Computation on Batch Systems

For bug reports and feature requests please use the tracker: https://github.com/mllg/batchtools.

Package options

batchtools.verbose

Verbosity. Set to FALSE to suppress info messages and progress bars.

batchtools.progress

Progress bars. Set to FALSE to disable them.

batchtools.timestamps

Add time stamps to log output. Set to FALSE to disable them.

Furthermore, you may enable a debug mode using the debugme package by setting the environment variable “DEBUGME” to “batchtools” before loading batchtools.

See also

Useful links:

https://github.com/mllg/batchtools

Report bugs at https://github.com/mllg/batchtools/issues