Chunk Jobs for Sequential Execution

Partition jobs into “chunks” which will be executed together on the nodes.

Chunks are submitted via submitJobs by simply providing a data frame with columns “job.id” and corresponding “chunk” (both integer). All jobs with the same chunk number will be grouped together on one node as a single computational job.

If neither n.chunks nor chunk.size are provided, each job will be assigned to its own chunk.

chunkIds(ids = NULL, n.chunks = NULL, chunk.size = NULL,
  group.by = character(0L), reg = getDefaultRegistry())

Arguments

ids

[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to all jobs.

n.chunks

[integer(1)] Requested number of chunks. If more chunks than elements in ids are requested, empty chunks are ignored. Mutually exclusive with chunks.size.

chunk.size

[integer(1)] Requested number of elements in each chunk. If ids cannot be chunked evenly, some chunks will have less elements than others. Mutually exclusive with with n.chunks.

group.by

[character(0)] If ids is a data.frame with additional columns (in addition to the required column “job.id”), then the chunking is performed using subgroups defined by the columns set in group.by. See example.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] with columns “job.id” and “chunk”.

Examples

# chunking for Registry
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

batchMap(identity, 1:25, reg = tmp)

#> Adding 25 jobs ...

ids = chunkIds(chunk.size = 10, reg = tmp)
print(ids)

#>     job.id chunk
#>  1:      1     2
#>  2:      2     3
#>  3:      3     1
#>  4:      4     3
#>  5:      5     3
#>  6:      6     3
#>  7:      7     2
#>  8:      8     1
#>  9:      9     2
#> 10:     10     1
#> 11:     11     2
#> 12:     12     1
#> 13:     13     2
#> 14:     14     1
#> 15:     15     3
#> 16:     16     2
#> 17:     17     1
#> 18:     18     3
#> 19:     19     2
#> 20:     20     2
#> 21:     21     3
#> 22:     22     1
#> 23:     23     3
#> 24:     24     1
#> 25:     25     1
#>     job.id chunk

print(table(ids$chunk))

#> 
#> 1 2 3 
#> 9 8 8 

# Creating chunks for an ExperimentRegistry
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

prob = addProblem(reg = tmp, "prob1", data = iris, fun = function(job, data) nrow(data))
prob = addProblem(reg = tmp, "prob2", data = Titanic, fun = function(job, data) nrow(data))
algo = addAlgorithm(reg = tmp, "algo", fun = function(job, data, instance, i, ...) problem)
prob.designs = list(prob1 = data.table(), prob2 = data.table(x = 1:2))
algo.designs = list(algo = data.table(i = 1:3))
addExperiments(prob.designs, algo.designs, repls = 3, reg = tmp)

#> Adding 9 experiments ('prob1'[1] x 'algo'[3] x repls[3]) ...

#> Adding 18 experiments ('prob2'[2] x 'algo'[3] x repls[3]) ...

# group into chunks of 5 jobs, but do not put multiple problems into a single chunk
# -> only one problem has to be loaded per chunk, and only once because it is then cached.
ids = getJobTable(reg = tmp)[, .(job.id, problem, algorithm)]
chunked = chunkIds(ids, chunk.size = 5, group.by = "problem", reg = tmp)
print(chunked)

#>     job.id chunk
#>  1:      1     1
#>  2:      2     1
#>  3:      3     2
#>  4:      4     2
#>  5:      5     1
#>  6:      6     1
#>  7:      7     2
#>  8:      8     1
#>  9:      9     2
#> 10:     10     3
#> 11:     11     4
#> 12:     12     5
#> 13:     13     3
#> 14:     14     6
#> 15:     15     5
#> 16:     16     6
#> 17:     17     3
#> 18:     18     4
#> 19:     19     6
#> 20:     20     4
#> 21:     21     5
#> 22:     22     6
#> 23:     23     5
#> 24:     24     4
#> 25:     25     4
#> 26:     26     3
#> 27:     27     6
#>     job.id chunk

dcast(ijoin(ids, chunked), chunk ~ problem)

#> Using 'chunk' as value column. Use 'value.var' to override

#> Aggregate function missing, defaulting to 'length'

#>    chunk prob1 prob2
#> 1:     1     5     0
#> 2:     2     4     0
#> 3:     3     0     4
#> 4:     4     0     5
#> 5:     5     0     4
#> 6:     6     0     5