Execute Jobs of a JobCollection

Executes every job in a JobCollection. This function is intended to be called on the slave.

doJobCollection(jc, output = NULL)

Arguments

jc
[JobCollection] Either an object of class “JobCollection” as returned by makeJobCollection or a string with the path to file containing a “JobCollection” as RDS file (as stored by submitJobs).

output
[character(1)] Path to a file to write the output to. Defaults to NULL which means that output is written to the active sink. Do not set this if your scheduler redirects output to a log file.

Value

[character(1)]: Hash of the JobCollection executed.

See also

Other JobCollection: makeJobCollection

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

batchMap(identity, 1:2, reg = tmp)

#> Adding 2 jobs ...

jc = makeJobCollection(1:2, reg = tmp)
doJobCollection(jc)

#> ### [bt]: This is batchtools v0.9.8
#> ### [bt]: Starting calculation of 2 jobs
#> ### [bt]: Setting working directory to '/home/lang/Projekte/batchtools/docs/reference'
#> ### [bt]: Memory measurement disabled
#> ### [bt]: Starting job [batchtools job.id=1]

#> ### [bt]: Setting seed to 1 ...

#> 
#> ### [bt]: Job terminated successfully [batchtools job.id=1]
#> ### [bt]: Starting job [batchtools job.id=2]

#> ### [bt]: Setting seed to 2 ...

#> 
#> ### [bt]: Job terminated successfully [batchtools job.id=2]
#> ### [bt]: Calculation finished!