Estimate Remaining Runtimes
Source: R/estimateRuntimes.R

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate]
Object to print.

n
[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> No configuration file found

#> Created registry in '/var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA     NA       NA          NA
#>   2:      2      2         NA         NA         NA     NA       NA          NA
#>   3:      3      3         NA         NA         NA     NA       NA          NA
#>   4:      4      4 1521796108 1521796108 1521796208     NA       NA           1
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA     NA       NA          NA
#> 497:    497    497 1521796110 1521796110 1521797610     NA       NA           1
#> 498:    498    498         NA         NA         NA     NA       NA          NA
#> 499:    499    499         NA         NA         NA     NA       NA          NA
#> 500:    500    500         NA         NA         NA     NA       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:            NA       NA                                  NA       NA     1
#>   2:            NA       NA                                  NA       NA     1
#>   3:            NA       NA                                  NA       NA     1
#>   4: cfInteractive       NA job3dd73de2a199d4d2e7af96e104183530       NA     1
#>  ---                                                                          
#> 496:            NA       NA                                  NA       NA     1
#> 497: cfInteractive       NA jobd8ea1d46e825a5997e01b9f9da8326bc       NA     1
#> 498:            NA       NA                                  NA       NA     1
#> 499:            NA       NA                                  NA       NA     1
#> 500:            NA       NA                                  NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0115 secs
#>  2:     16    iris      nrow     4      a 1100.0281 secs
#>  3:     34    iris      nrow     7      d  100.0109 secs
#>  4:     41    iris      nrow     9      a 1100.0070 secs
#>  5:     46    iris      nrow    10      a 1100.0110 secs
#>  6:     48    iris      nrow    10      c  100.0117 secs
#>  7:     69    iris      nrow    14      d  100.0067 secs
#>  8:     85    iris      nrow    17      e  100.0081 secs
#>  9:     89    iris      nrow    18      d  100.0070 secs
#> 10:     94    iris      nrow    19      d  100.0102 secs
#> 11:     97    iris      nrow    20      b 1100.0071 secs
#> 12:     98    iris      nrow    20      c  100.0120 secs
#> 13:    115    iris      nrow    23      e  100.0083 secs
#> 14:    117    iris      nrow    24      b 1100.0082 secs
#> 15:    120    iris      nrow    24      e  100.0090 secs
#> 16:    145    iris      nrow    29      e  100.0070 secs
#> 17:    148    iris      nrow    30      c  100.0099 secs
#> 18:    156    iris      nrow    32      a 1100.0076 secs
#> 19:    167    iris      nrow    34      b 1100.0083 secs
#> 20:    187    iris      nrow    38      b 1100.0082 secs
#> 21:    193    iris      nrow    39      c  100.0096 secs
#> 22:    220    iris      nrow    44      e  100.0069 secs
#> 23:    223    iris      nrow    45      c  100.0120 secs
#> 24:    227    iris      nrow    46      b 1100.0067 secs
#> 25:    246    iris      nrow    50      a 1100.0071 secs
#> 26:    258    iris      ncol     2      c  500.0090 secs
#> 27:    270    iris      ncol     4      e  500.0067 secs
#> 28:    271    iris      ncol     5      a 1500.0092 secs
#> 29:    275    iris      ncol     5      e  500.0103 secs
#> 30:    296    iris      ncol    10      a 1500.0078 secs
#> 31:    315    iris      ncol    13      e  500.0077 secs
#> 32:    325    iris      ncol    15      e  500.0086 secs
#> 33:    352    iris      ncol    21      b 1500.0089 secs
#> 34:    354    iris      ncol    21      d  500.0078 secs
#> 35:    358    iris      ncol    22      c  500.0097 secs
#> 36:    364    iris      ncol    23      d  500.0080 secs
#> 37:    370    iris      ncol    24      e  500.0092 secs
#> 38:    378    iris      ncol    26      c  500.0076 secs
#> 39:    388    iris      ncol    28      c  500.0101 secs
#> 40:    402    iris      ncol    31      b 1500.0081 secs
#> 41:    411    iris      ncol    33      a 1500.0100 secs
#> 42:    419    iris      ncol    34      d  500.0084 secs
#> 43:    432    iris      ncol    37      b 1500.0077 secs
#> 44:    434    iris      ncol    37      d  500.0082 secs
#> 45:    436    iris      ncol    38      a 1500.0092 secs
#> 46:    438    iris      ncol    38      c  500.0071 secs
#> 47:    447    iris      ncol    40      b 1500.0072 secs
#> 48:    451    iris      ncol    41      a 1500.0112 secs
#> 49:    454    iris      ncol    41      d  500.0073 secs
#> 50:    497    iris      ncol    50      b 1500.0070 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.5s
#>   Remaining: 3d 16h 23m 21.0s
#>   Total    : 4d 02h 06m 41.5s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1043.1353
#>   2:      2    iris      nrow     1      b estimated 1042.6340
#>   3:      3    iris      nrow     1      c estimated  190.8529
#>   4:      4    iris      nrow     1      d  observed  100.0115
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1396.1413
#> 497:    497    iris      ncol    50      b  observed 1500.0070
#> 498:    498    iris      ncol    50      c estimated  596.3900
#> 499:    499    iris      ncol    50      d estimated  572.4746
#> 500:    500    iris      ncol    50      e estimated  560.1503

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.5s
#>   Remaining: 3d 16h 23m 21.0s
#>   Parallel : 0d 08h 51m 28.8s
#>   Total    : 4d 02h 06m 41.5s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1433.6250
#>   2:    456 estimated 1424.6351
#>   3:    431 estimated 1421.7917
#>   4:    426 estimated 1421.5517
#>  ---                           
#> 446:     90 estimated  135.5263
#> 447:     99 estimated  131.5583
#> 448:     74 estimated  129.9317
#> 449:     79 estimated  128.2873
#> 450:     84 estimated  125.5806
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1353    75
#>   2:      2 estimated 1042.6340    75
#>   3:      3 estimated  190.8529    71
#>   4:      5 estimated  195.5074     9
#>  ---                                 
#> 446:    495 estimated  560.1503    35
#> 447:    496 estimated 1396.1413    12
#> 448:    498 estimated  596.3900     6
#> 449:    499 estimated  572.4746    21
#> 450:    500 estimated  560.1503    36

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1353    75
#>   2:      2 estimated 1042.6340    75
#>   3:      3 estimated  190.8529    71
#>   4:      5 estimated  195.5074     9
#>  ---                                 
#> 446:    495 estimated  560.1503    35
#> 447:    496 estimated 1396.1413    12
#> 448:    498 estimated  596.3900     6
#> 449:    499 estimated  572.4746    21
#> 450:    500 estimated  560.1503    36

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    75 3599.300
#>  2:    71 3532.549
#>  3:     9 3599.464
#>  4:    74 3599.323
#>  5:    70 3548.379
#>  6:     7 3599.944
#>  7:    44 3598.583
#>  8:    64 3596.812
#>  9:    61 3599.720
#> 10:    69 3557.232
#> 11:    54 3599.277
#> 12:    60 3596.218
#> 13:    56 3599.375
#> 14:    66 3580.744
#> 15:    51 3474.894
#> 16:    55 3599.865
#> 17:    58 3582.460
#> 18:    68 3571.211
#> 19:    49 3493.625
#> 20:    57 3592.257
#> 21:     8 3597.088
#> 22:    62 3599.961
#> 23:    72 3513.739
#> 24:    43 3592.660
#> 25:     5 3599.052
#> 26:    59 3579.823
#> 27:     4 3598.902
#> 28:    73 3482.661
#> 29:    78 3550.890
#> 30:     6 3597.690
#> 31:    77 3565.749
#> 32:    85 3514.238
#> 33:    53 3599.707
#> 34:    42 3599.734
#> 35:    41 3599.321
#> 36:    39 3599.329
#> 37:    82 3583.411
#> 38:     1 3599.048
#> 39:    80 3494.831
#> 40:    83 3563.280
#> 41:    37 3599.654
#> 42:    90 1271.974
#> 43:    52 3599.799
#> 44:    40 3599.960
#> 45:    81 3470.098
#> 46:    65 3586.757
#> 47:    67 3573.733
#> 48:     3 3598.243
#> 49:    76 3582.451
#> 50:     2 3599.591
#> 51:    79 3525.475
#> 52:    84 3547.442
#> 53:    63 3599.643
#> 54:    32 3488.498
#> 55:    13 3595.863
#> 56:    48 3501.245
#> 57:    50 3479.767
#> 58:    28 3513.630
#> 59:    30 3500.774
#> 60:    35 3481.031
#> 61:    38 3471.187
#> 62:    46 3517.254
#> 63:    33 3487.501
#> 64:    14 3599.622
#> 65:    45 3597.069
#> 66:    34 3485.333
#> 67:    36 3477.041
#> 68:    47 3506.486
#> 69:    12 3599.451
#> 70:    11 3598.856
#> 71:    10 3599.948
#> 72:    16 3599.690
#> 73:    17 3588.430
#> 74:    15 3598.620
#> 75:    29 3508.412
#> 76:    27 3519.659
#> 77:    31 3497.117
#> 78:    23 3555.930
#> 79:    20 3569.205
#> 80:    26 3539.969
#> 81:    21 3567.367
#> 82:    18 3585.755
#> 83:    24 3545.226
#> 84:    22 3557.271
#> 85:    19 3578.485
#> 86:    25 3543.665
#> 87:    86 3587.553
#> 88:    87 3568.405
#> 89:    88 3550.697
#> 90:    89 3531.835
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1353     8
#>   2:      2 estimated 1042.6340     6
#>   3:      3 estimated  190.8529     2
#>   4:      5 estimated  195.5074     8
#>  ---                                 
#> 446:    495 estimated  560.1503     4
#> 447:    496 estimated 1396.1413     9
#> 448:    498 estimated  596.3900     5
#> 449:    499 estimated  572.4746     7
#> 450:    500 estimated  560.1503     7

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     8 31755.20
#>  2:     6 31755.86
#>  3:     2 31880.62
#>  4:     4 31755.34
#>  5:    10 31884.81
#>  6:     3 31888.81
#>  7:     1 31885.55
#>  8:     5 31755.98
#>  9:     7 31883.20
#> 10:     9 31755.63