Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate]
Object to print.

n
[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

 batchtools:::example_push_temp(1) 
# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> No readable configuration file found

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done error mem.used resource.id
#>   1:      1      1         NA         NA         NA  <NA>       NA          NA
#>   2:      2      2         NA         NA         NA  <NA>       NA          NA
#>   3:      3      3         NA         NA         NA  <NA>       NA          NA
#>   4:      4      4         NA         NA         NA  <NA>       NA          NA
#>   5:      5      5         NA         NA         NA  <NA>       NA          NA
#>  ---                                                                          
#> 496:    496    496         NA         NA         NA  <NA>       NA          NA
#> 497:    497    497         NA         NA         NA  <NA>       NA          NA
#> 498:    498    498         NA         NA         NA  <NA>       NA          NA
#> 499:    499    499 1603197327 1603197327 1603197827  <NA>       NA           1
#> 500:    500    500         NA         NA         NA  <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name repl
#>   1:          <NA>     <NA>                                <NA>     <NA>    1
#>   2:          <NA>     <NA>                                <NA>     <NA>    1
#>   3:          <NA>     <NA>                                <NA>     <NA>    1
#>   4:          <NA>     <NA>                                <NA>     <NA>    1
#>   5:          <NA>     <NA>                                <NA>     <NA>    1
#>  ---                                                                         
#> 496:          <NA>     <NA>                                <NA>     <NA>    1
#> 497:          <NA>     <NA>                                <NA>     <NA>    1
#> 498:          <NA>     <NA>                                <NA>     <NA>    1
#> 499: cfInteractive     <NA> job7c4bbe21691c77d776b738c39a54fd28     <NA>    1
#> 500:          <NA>     <NA>                                <NA>     <NA>    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y   time.running
#>  1:     32    iris      nrow  7 b 1100.0049 secs
#>  2:     42    iris      nrow  9 b 1100.0046 secs
#>  3:     47    iris      nrow 10 b 1100.0046 secs
#>  4:     66    iris      nrow 14 a 1100.0055 secs
#>  5:     73    iris      nrow 15 c  100.0065 secs
#>  6:     75    iris      nrow 15 e  100.0099 secs
#>  7:     86    iris      nrow 18 a 1100.0058 secs
#>  8:    100    iris      nrow 20 e  100.0060 secs
#>  9:    101    iris      nrow 21 a 1100.0049 secs
#> 10:    103    iris      nrow 21 c  100.0046 secs
#> 11:    123    iris      nrow 25 c  100.0042 secs
#> 12:    125    iris      nrow 25 e  100.0059 secs
#> 13:    161    iris      nrow 33 a 1100.0059 secs
#> 14:    165    iris      nrow 33 e  100.0110 secs
#> 15:    169    iris      nrow 34 d  100.0049 secs
#> 16:    183    iris      nrow 37 c  100.0058 secs
#> 17:    184    iris      nrow 37 d  100.0058 secs
#> 18:    203    iris      nrow 41 c  100.0051 secs
#> 19:    207    iris      nrow 42 b 1100.0050 secs
#> 20:    209    iris      nrow 42 d  100.0049 secs
#> 21:    220    iris      nrow 44 e  100.0052 secs
#> 22:    227    iris      nrow 46 b 1100.0054 secs
#> 23:    229    iris      nrow 46 d  100.0068 secs
#> 24:    231    iris      nrow 47 a 1100.0059 secs
#> 25:    244    iris      nrow 49 d  100.0057 secs
#> 26:    260    iris      ncol  2 e  500.0053 secs
#> 27:    276    iris      ncol  6 a 1500.0050 secs
#> 28:    278    iris      ncol  6 c  500.0066 secs
#> 29:    279    iris      ncol  6 d  500.0054 secs
#> 30:    296    iris      ncol 10 a 1500.0066 secs
#> 31:    320    iris      ncol 14 e  500.0052 secs
#> 32:    340    iris      ncol 18 e  500.0046 secs
#> 33:    347    iris      ncol 20 b 1500.0046 secs
#> 34:    363    iris      ncol 23 c  500.0061 secs
#> 35:    369    iris      ncol 24 d  500.0062 secs
#> 36:    373    iris      ncol 25 c  500.0063 secs
#> 37:    387    iris      ncol 28 b 1500.0051 secs
#> 38:    410    iris      ncol 32 e  500.0055 secs
#> 39:    421    iris      ncol 35 a 1500.0070 secs
#> 40:    436    iris      ncol 38 a 1500.0054 secs
#> 41:    444    iris      ncol 39 d  500.0056 secs
#> 42:    448    iris      ncol 40 c  500.0055 secs
#> 43:    456    iris      ncol 42 a 1500.0059 secs
#> 44:    459    iris      ncol 42 d  500.0058 secs
#> 45:    467    iris      ncol 44 b 1500.0053 secs
#> 46:    468    iris      ncol 44 c  500.0057 secs
#> 47:    475    iris      ncol 45 e  500.0070 secs
#> 48:    482    iris      ncol 47 b 1500.0070 secs
#> 49:    492    iris      ncol 49 b 1500.0061 secs
#> 50:    499    iris      ncol 50 d  500.0059 secs
#>     job.id problem algorithm  x y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 17h 34m 34.9s
#>   Total    : 4d 03h 17m 55.2s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated 1103.4590
#>   2:      2    iris      nrow  1 b estimated 1087.2531
#>   3:      3    iris      nrow  1 c estimated  338.3327
#>   4:      4    iris      nrow  1 d estimated  318.0785
#>   5:      5    iris      nrow  1 e estimated  317.7231
#>  ---                                                  
#> 496:    496    iris      ncol 50 a estimated 1387.5999
#> 497:    497    iris      ncol 50 b estimated 1394.6896
#> 498:    498    iris      ncol 50 c estimated  618.0633
#> 499:    499    iris      ncol 50 d  observed  500.0059
#> 500:    500    iris      ncol 50 e estimated  578.6555

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 17h 34m 34.9s
#>   Parallel : 0d 08h 58m 10.3s
#>   Total    : 4d 03h 17m 55.2s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    466 estimated 1422.9803
#>   2:    461 estimated 1419.5870
#>   3:    472 estimated 1415.9535
#>   4:    462 estimated 1415.6001
#>   5:    457 estimated 1414.8001
#>  ---                           
#> 446:    204 estimated  132.6668
#> 447:    185 estimated  132.3737
#> 448:    189 estimated  131.9505
#> 449:    174 estimated  131.3331
#> 450:    179 estimated  130.1864

if (FALSE) {
submitJobs(ids, reg = tmp)
}

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1103.4590    48
#>   2:      2 estimated 1087.2531    52
#>   3:      3 estimated  338.3327    36
#>   4:      4 estimated  318.0785    33
#>   5:      5 estimated  317.7231    70
#>  ---                                 
#> 446:    495 estimated  584.3901    15
#> 447:    496 estimated 1387.5999    18
#> 448:    497 estimated 1394.6896    13
#> 449:    498 estimated  618.0633     4
#> 450:    500 estimated  578.6555    20

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1103.4590    48
#>   2:      2 estimated 1087.2531    52
#>   3:      3 estimated  338.3327    36
#>   4:      4 estimated  318.0785    33
#>   5:      5 estimated  317.7231    70
#>  ---                                 
#> 446:    495 estimated  584.3901    15
#> 447:    496 estimated 1387.5999    18
#> 448:    497 estimated 1394.6896    13
#> 449:    498 estimated  618.0633     4
#> 450:    500 estimated  578.6555    20

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    48 3485.338
#>  2:    52 3598.644
#>  3:    36 3599.459
#>  4:    33 3596.494
#>  5:    70 3490.925
#>  6:    49 3479.219
#>  7:    37 3595.168
#>  8:    51 3596.367
#>  9:    71 3488.611
#> 10:    53 3599.742
#> 11:    68 3496.038
#> 12:    72 3485.351
#> 13:    54 3589.155
#> 14:    69 3494.538
#> 15:    73 3478.069
#> 16:    46 3516.899
#> 17:    50 3599.762
#> 18:    42 3572.213
#> 19:    64 3516.755
#> 20:    66 3506.507
#> 21:    43 3569.239
#> 22:    65 3512.215
#> 23:    67 3505.174
#> 24:    41 3581.446
#> 25:    59 3537.144
#> 26:    60 3532.235
#> 27:    47 3493.115
#> 28:    38 3599.214
#> 29:    35 3599.986
#> 30:    40 3589.694
#> 31:    57 3561.085
#> 32:    58 3546.237
#> 33:    39 3594.115
#> 34:    55 3581.046
#> 35:    56 3570.053
#> 36:    44 3534.662
#> 37:    61 3522.996
#> 38:    34 3598.099
#> 39:    45 3533.418
#> 40:    63 3518.338
#> 41:    62 3521.086
#> 42:    27 3597.000
#> 43:    26 3586.504
#> 44:    24 3596.226
#> 45:    25 3597.570
#> 46:    23 3597.820
#> 47:    29 3567.933
#> 48:    75 3599.097
#> 49:    20 3522.443
#> 50:    74 3474.955
#> 51:     8 3591.306
#> 52:    21 3515.784
#> 53:    11 3565.927
#> 54:     7 3592.927
#> 55:    12 3563.426
#> 56:     5 3593.237
#> 57:     6 3598.126
#> 58:    32 3599.468
#> 59:    10 3579.486
#> 60:    80 3481.959
#> 61:    82 3598.299
#> 62:    83 3597.802
#> 63:     4 3598.686
#> 64:    79 3493.361
#> 65:    81 3474.061
#> 66:    76 3587.729
#> 67:    87 3564.202
#> 68:    89 3546.702
#> 69:    91 2158.021
#> 70:    78 3506.583
#> 71:    88 3554.982
#> 72:    77 3523.895
#> 73:    85 3580.521
#> 74:     3 3599.219
#> 75:    90 3522.918
#> 76:    84 3586.090
#> 77:    86 3572.917
#> 78:     2 3599.764
#> 79:     1 3599.763
#> 80:     9 3587.324
#> 81:    30 3560.138
#> 82:    13 3599.521
#> 83:    14 3598.014
#> 84:    22 3598.040
#> 85:    19 3574.641
#> 86:    17 3585.269
#> 87:    31 3546.246
#> 88:    18 3581.006
#> 89:    28 3574.646
#> 90:    16 3598.238
#> 91:    15 3599.249
#>     chunk  runtime

if (FALSE) {
submitJobs(ids, reg = tmp)
}

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1103.4590     7
#>   2:      2 estimated 1087.2531     5
#>   3:      3 estimated  338.3327     9
#>   4:      4 estimated  318.0785     3
#>   5:      5 estimated  317.7231     5
#>  ---                                 
#> 446:    495 estimated  584.3901    10
#> 447:    496 estimated 1387.5999     6
#> 448:    497 estimated 1394.6896     6
#> 449:    498 estimated  618.0633     8
#> 450:    500 estimated  578.6555     9

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:     7 32290.29
#>  2:     5 32288.78
#>  3:     9 32212.48
#>  4:     3 32215.25
#>  5:     8 32271.77
#>  6:     4 32212.70
#>  7:     2 32210.72
#>  8:     1 32289.79
#>  9:     6 32271.49
#> 10:    10 32211.62