Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Ignoring resource 'chunks.as.arrayjobs', not supported by cluster functions 'Interactive'

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started done error memory resource.id batch.id
#>   1:      1      1        NA      NA   NA    NA     NA          NA       NA
#>   2:      2      2        NA      NA   NA    NA     NA          NA       NA
#>   3:      3      3        NA      NA   NA    NA     NA          NA       NA
#>   4:      4      4        NA      NA   NA    NA     NA          NA       NA
#>   5:      5      5        NA      NA   NA    NA     NA          NA       NA
#>  ---                                                                       
#> 496:    496    496        NA      NA   NA    NA     NA          NA       NA
#> 497:    497    497        NA      NA   NA    NA     NA          NA       NA
#> 498:    498    498        NA      NA   NA    NA     NA          NA       NA
#> 499:    499    499        NA      NA   NA    NA     NA          NA       NA
#> 500:    500    500        NA      NA   NA    NA     NA          NA       NA
#>      log.file job.hash repl
#>   1:       NA       NA    1
#>   2:       NA       NA    1
#>   3:       NA       NA    1
#>   4:       NA       NA    1
#>   5:       NA       NA    1
#>  ---                       
#> 496:       NA       NA    1
#> 497:       NA       NA    1
#> 498:       NA       NA    1
#> 499:       NA       NA    1
#> 500:       NA       NA    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y time.running
#>  1:     12    iris      nrow  3 b    1100 secs
#>  2:     30    iris      nrow  6 e     100 secs
#>  3:     43    iris      nrow  9 c     101 secs
#>  4:     46    iris      nrow 10 a    1100 secs
#>  5:     53    iris      nrow 11 c     100 secs
#>  6:     55    iris      nrow 11 e     101 secs
#>  7:     65    iris      nrow 13 e     100 secs
#>  8:     66    iris      nrow 14 a    1100 secs
#>  9:     89    iris      nrow 18 d     101 secs
#> 10:     91    iris      nrow 19 a    1100 secs
#> 11:     93    iris      nrow 19 c     100 secs
#> 12:    124    iris      nrow 25 d     101 secs
#> 13:    136    iris      nrow 28 a    1100 secs
#> 14:    147    iris      nrow 30 b    1100 secs
#> 15:    157    iris      nrow 32 b    1101 secs
#> 16:    160    iris      nrow 32 e     100 secs
#> 17:    163    iris      nrow 33 c     100 secs
#> 18:    178    iris      nrow 36 c     101 secs
#> 19:    179    iris      nrow 36 d     100 secs
#> 20:    211    iris      nrow 43 a    1100 secs
#> 21:    222    iris      nrow 45 b    1101 secs
#> 22:    232    iris      nrow 47 b    1100 secs
#> 23:    235    iris      nrow 47 e     100 secs
#> 24:    239    iris      nrow 48 d     101 secs
#> 25:    244    iris      nrow 49 d     100 secs
#> 26:    251    iris      ncol  1 a    1500 secs
#> 27:    260    iris      ncol  2 e     501 secs
#> 28:    278    iris      ncol  6 c     500 secs
#> 29:    292    iris      ncol  9 b    1500 secs
#> 30:    326    iris      ncol 16 a    1501 secs
#> 31:    341    iris      ncol 19 a    1500 secs
#> 32:    343    iris      ncol 19 c     500 secs
#> 33:    346    iris      ncol 20 a    1501 secs
#> 34:    365    iris      ncol 23 e     500 secs
#> 35:    367    iris      ncol 24 b    1500 secs
#> 36:    372    iris      ncol 25 b    1501 secs
#> 37:    374    iris      ncol 25 d     500 secs
#> 38:    379    iris      ncol 26 d     500 secs
#> 39:    397    iris      ncol 30 b    1501 secs
#> 40:    409    iris      ncol 32 d     500 secs
#> 41:    410    iris      ncol 32 e     501 secs
#> 42:    418    iris      ncol 34 c     500 secs
#> 43:    423    iris      ncol 35 c     500 secs
#> 44:    425    iris      ncol 35 e     501 secs
#> 45:    439    iris      ncol 38 d     500 secs
#> 46:    442    iris      ncol 39 b    1500 secs
#> 47:    443    iris      ncol 39 c     501 secs
#> 48:    450    iris      ncol 40 e     500 secs
#> 49:    451    iris      ncol 41 a    1500 secs
#> 50:    459    iris      ncol 42 d     501 secs
#>     job.id problem algorithm  x y time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 37.0s
#>   Remaining: 3d 15h 20m 43.0s
#>   Total    : 4d 01h 04m 20.0s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated 1084.0589
#>   2:      2    iris      nrow  1 b estimated 1087.7987
#>   3:      3    iris      nrow  1 c estimated  296.8125
#>   4:      4    iris      nrow  1 d estimated  282.2683
#>   5:      5    iris      nrow  1 e estimated  284.8363
#>  ---                                                  
#> 496:    496    iris      ncol 50 a estimated 1259.9025
#> 497:    497    iris      ncol 50 b estimated 1258.8532
#> 498:    498    iris      ncol 50 c estimated  482.4642
#> 499:    499    iris      ncol 50 d estimated  468.9096
#> 500:    500    iris      ncol 50 e estimated  470.2686

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 37.0s
#>   Remaining: 3d 15h 20m 43.0s
#>   Parallel : 0d 08h 45m 14.2s
#>   Total    : 4d 01h 04m 20.0s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    347 estimated 1438.0114
#>   2:    352 estimated 1438.0098
#>   3:    362 estimated 1433.7575
#>   4:    357 estimated 1433.3002
#>   5:    342 estimated 1433.0618
#>  ---                           
#> 446:     64 estimated  149.3028
#> 447:    184 estimated  148.7860
#> 448:     54 estimated  147.3256
#> 449:     60 estimated  144.4702
#> 450:     59 estimated  142.7903

not_run({
  submitJobs(ids, reg = tmp)
})

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1084.0589    46
#>   2:      2 estimated 1087.7987    46
#>   3:      3 estimated  296.8125    28
#>   4:      4 estimated  282.2683    24
#>   5:      5 estimated  284.8363    25
#>  ---                                 
#> 446:    496 estimated 1259.9025    44
#> 447:    497 estimated 1258.8532    45
#> 448:    498 estimated  482.4642    80
#> 449:    499 estimated  468.9096    87
#> 450:    500 estimated  470.2686    87

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1084.0589    46
#>   2:      2 estimated 1087.7987    46
#>   3:      3 estimated  296.8125    28
#>   4:      4 estimated  282.2683    24
#>   5:      5 estimated  284.8363    25
#>  ---                                 
#> 446:    496 estimated 1259.9025    44
#> 447:    497 estimated 1258.8532    45
#> 448:    498 estimated  482.4642    80
#> 449:    499 estimated  468.9096    87
#> 450:    500 estimated  470.2686    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>  1:    46 3593.3036
#>  2:    28 3598.9423
#>  3:    24 3599.7305
#>  4:    25 3597.2095
#>  5:    57 3540.6777
#>  6:    54 3561.6132
#>  7:    17 3599.0298
#>  8:    20 3564.6247
#>  9:    14 3595.8755
#> 10:    60 3519.6925
#> 11:    18 3595.4964
#> 12:    21 3560.5547
#> 13:    16 3589.0886
#> 14:    61 3511.8717
#> 15:    59 3524.5950
#> 16:    15 3598.0058
#> 17:    12 3598.4456
#> 18:    27 3535.9845
#> 19:    66 3598.3998
#> 20:    63 3491.1326
#> 21:    41 3599.8948
#> 22:    55 3554.9255
#> 23:    56 3546.9064
#> 24:    47 3599.6214
#> 25:    67 3597.6202
#> 26:    62 3501.5089
#> 27:    49 3599.9906
#> 28:    88 3539.4935
#> 29:    87 3585.0084
#> 30:     4 3599.2349
#> 31:    68 3597.5919
#> 32:    65 3599.5787
#> 33:    69 3598.8179
#> 34:     2 3597.8309
#> 35:    89  766.7533
#> 36:    36 3598.8170
#> 37:    38 3594.4472
#> 38:    76 3525.9577
#> 39:     5 3599.7168
#> 40:    51 3583.8165
#> 41:    64 3482.7683
#> 42:    43 3599.8797
#> 43:    50 3591.9492
#> 44:    75 3539.8074
#> 45:    48 3598.5287
#> 46:    45 3594.9809
#> 47:    58 3532.8128
#> 48:    79 3475.1461
#> 49:    77 3509.8307
#> 50:    44 3598.4608
#> 51:    78 3494.1428
#> 52:    52 3577.5891
#> 53:    53 3569.3880
#> 54:     8 3599.6086
#> 55:     7 3599.1854
#> 56:    39 3599.8861
#> 57:    70 3599.6683
#> 58:    73 3599.9803
#> 59:    71 3599.9738
#> 60:    72 3599.9286
#> 61:     3 3599.9395
#> 62:    37 3599.7929
#> 63:    29 3508.6990
#> 64:    26 3539.8085
#> 65:    35 3460.7511
#> 66:    31 3496.3766
#> 67:    22 3557.1366
#> 68:    34 3463.9329
#> 69:    30 3504.4398
#> 70:    23 3551.7409
#> 71:     9 3599.5901
#> 72:    11 3591.7363
#> 73:    19 3574.5039
#> 74:    33 3468.6348
#> 75:    32 3486.4907
#> 76:    13 3596.1834
#> 77:    40 3599.1777
#> 78:    10 3597.1892
#> 79:    42 3595.3596
#> 80:    74 3523.1169
#> 81:     6 3599.9682
#> 82:     1 3467.1609
#> 83:    82 3593.0927
#> 84:    83 3577.3961
#> 85:    84 3558.2828
#> 86:    80 3599.3449
#> 87:    85 3543.0194
#> 88:    86 3471.3809
#> 89:    81 3593.4729
#>     chunk   runtime

not_run({
  submitJobs(ids, reg = tmp)
})

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1084.0589    10
#>   2:      2 estimated 1087.7987     1
#>   3:      3 estimated  296.8125     1
#>   4:      4 estimated  282.2683     3
#>   5:      5 estimated  284.8363     4
#>  ---                                 
#> 446:    496 estimated 1259.9025     6
#> 447:    497 estimated 1258.8532     5
#> 448:    498 estimated  482.4642    10
#> 449:    499 estimated  468.9096     7
#> 450:    500 estimated  470.2686    10

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    10 31512.61
#>  2:     1 31378.59
#>  3:     3 31380.24
#>  4:     4 31380.38
#>  5:     5 31372.80
#>  6:     8 31372.15
#>  7:     7 31508.25
#>  8:     9 31514.03
#>  9:     2 31509.77
#> 10:     6 31514.22