Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate]
Object to print.

n
[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/michel/.batchtools.conf.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 4 (seed = 5) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 16 (seed = 17) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 34 (seed = 35) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 41 (seed = 42) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 46 (seed = 47) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 48 (seed = 49) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 69 (seed = 70) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 85 (seed = 86) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 89 (seed = 90) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 94 (seed = 95) ...

#> 

#> Submitting [=========>-----------------------------------------]  20% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 97 (seed = 98) ...

#> 

#> Submitting [==========>----------------------------------------]  22% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 98 (seed = 99) ...

#> 

#> Submitting [===========>---------------------------------------]  24% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 115 (seed = 116) ...

#> 

#> Submitting [============>--------------------------------------]  26% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 117 (seed = 118) ...

#> 

#> Submitting [=============>-------------------------------------]  28% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 120 (seed = 121) ...

#> 

#> Submitting [==============>------------------------------------]  30% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 145 (seed = 146) ...

#> 

#> Submitting [===============>-----------------------------------]  32% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 148 (seed = 149) ...

#> 

#> Submitting [================>----------------------------------]  34% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 156 (seed = 157) ...

#> 

#> Submitting [=================>---------------------------------]  36% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 167 (seed = 168) ...

#> 

#> Submitting [==================>--------------------------------]  38% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 187 (seed = 188) ...

#> 

#> Submitting [===================>-------------------------------]  40% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 193 (seed = 194) ...

#> 

#> Submitting [====================>------------------------------]  42% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 220 (seed = 221) ...

#> 

#> Submitting [=====================>-----------------------------]  44% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 223 (seed = 224) ...

#> 

#> Submitting [======================>----------------------------]  46% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 227 (seed = 228) ...

#> 

#> Submitting [=======================>---------------------------]  48% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 246 (seed = 247) ...

#> 

#> Submitting [=========================>-------------------------]  50% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 258 (seed = 259) ...

#> 

#> Submitting [==========================>------------------------]  52% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 270 (seed = 271) ...

#> 

#> Submitting [===========================>-----------------------]  54% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 271 (seed = 272) ...

#> 

#> Submitting [============================>----------------------]  56% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 275 (seed = 276) ...

#> 

#> Submitting [=============================>---------------------]  58% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 296 (seed = 297) ...

#> 

#> Submitting [==============================>--------------------]  60% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 315 (seed = 316) ...

#> 

#> Submitting [===============================>-------------------]  62% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 325 (seed = 326) ...

#> 

#> Submitting [================================>------------------]  64% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 352 (seed = 353) ...

#> 

#> Submitting [=================================>-----------------]  66% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 354 (seed = 355) ...

#> 

#> Submitting [==================================>----------------]  68% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 358 (seed = 359) ...

#> 

#> Submitting [===================================>---------------]  70% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 364 (seed = 365) ...

#> 

#> Submitting [====================================>--------------]  72% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 370 (seed = 371) ...

#> 

#> Submitting [=====================================>-------------]  74% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 378 (seed = 379) ...

#> 

#> Submitting [======================================>------------]  76% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 388 (seed = 389) ...

#> 

#> Submitting [=======================================>-----------]  78% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 402 (seed = 403) ...

#> 

#> Submitting [========================================>----------]  80% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 411 (seed = 412) ...

#> 

#> Submitting [=========================================>---------]  82% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 419 (seed = 420) ...

#> 

#> Submitting [==========================================>--------]  84% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 432 (seed = 433) ...

#> 

#> Submitting [===========================================>-------]  86% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 434 (seed = 435) ...

#> 

#> Submitting [============================================>------]  88% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 436 (seed = 437) ...

#> 

#> Submitting [=============================================>-----]  90% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 438 (seed = 439) ...

#> 

#> Submitting [==============================================>----]  92% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 447 (seed = 448) ...

#> 

#> Submitting [===============================================>---]  94% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 451 (seed = 452) ...

#> 

#> Submitting [================================================>--]  96% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 454 (seed = 455) ...

#> 

#> Submitting [=================================================>-]  98% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 497 (seed = 498) ...

#> 

#> Submitting [===================================================] 100% eta:  0s

#>                                                                               

#> 

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4 1526760745 1526760745 1526760845   <NA>       NA           1
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497 1526760746 1526760746 1526762246   <NA>       NA           1
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499         NA         NA         NA   <NA>       NA          NA
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4: cfInteractive     <NA> jobf0815e13aa22fece603af2f3ce458998     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497: cfInteractive     <NA> jobc51056b4f43b88aad0ea8a6e79194e6b     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499:          <NA>     <NA>                                <NA>     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0034 secs
#>  2:     16    iris      nrow     4      a 1100.0095 secs
#>  3:     34    iris      nrow     7      d  100.0034 secs
#>  4:     41    iris      nrow     9      a 1100.0034 secs
#>  5:     46    iris      nrow    10      a 1100.0034 secs
#>  6:     48    iris      nrow    10      c  100.0034 secs
#>  7:     69    iris      nrow    14      d  100.0033 secs
#>  8:     85    iris      nrow    17      e  100.0034 secs
#>  9:     89    iris      nrow    18      d  100.0034 secs
#> 10:     94    iris      nrow    19      d  100.0034 secs
#> 11:     97    iris      nrow    20      b 1100.0034 secs
#> 12:     98    iris      nrow    20      c  100.0034 secs
#> 13:    115    iris      nrow    23      e  100.0034 secs
#> 14:    117    iris      nrow    24      b 1100.0033 secs
#> 15:    120    iris      nrow    24      e  100.0033 secs
#> 16:    145    iris      nrow    29      e  100.0034 secs
#> 17:    148    iris      nrow    30      c  100.0034 secs
#> 18:    156    iris      nrow    32      a 1100.0034 secs
#> 19:    167    iris      nrow    34      b 1100.0034 secs
#> 20:    187    iris      nrow    38      b 1100.0033 secs
#> 21:    193    iris      nrow    39      c  100.0033 secs
#> 22:    220    iris      nrow    44      e  100.0034 secs
#> 23:    223    iris      nrow    45      c  100.0034 secs
#> 24:    227    iris      nrow    46      b 1100.0039 secs
#> 25:    246    iris      nrow    50      a 1100.0034 secs
#> 26:    258    iris      ncol     2      c  500.0033 secs
#> 27:    270    iris      ncol     4      e  500.0034 secs
#> 28:    271    iris      ncol     5      a 1500.0035 secs
#> 29:    275    iris      ncol     5      e  500.0035 secs
#> 30:    296    iris      ncol    10      a 1500.0035 secs
#> 31:    315    iris      ncol    13      e  500.0035 secs
#> 32:    325    iris      ncol    15      e  500.0035 secs
#> 33:    352    iris      ncol    21      b 1500.0035 secs
#> 34:    354    iris      ncol    21      d  500.0034 secs
#> 35:    358    iris      ncol    22      c  500.0033 secs
#> 36:    364    iris      ncol    23      d  500.0034 secs
#> 37:    370    iris      ncol    24      e  500.0034 secs
#> 38:    378    iris      ncol    26      c  500.0034 secs
#> 39:    388    iris      ncol    28      c  500.0033 secs
#> 40:    402    iris      ncol    31      b 1500.0067 secs
#> 41:    411    iris      ncol    33      a 1500.0034 secs
#> 42:    419    iris      ncol    34      d  500.0034 secs
#> 43:    432    iris      ncol    37      b 1500.0034 secs
#> 44:    434    iris      ncol    37      d  500.0034 secs
#> 45:    436    iris      ncol    38      a 1500.0033 secs
#> 46:    438    iris      ncol    38      c  500.0034 secs
#> 47:    447    iris      ncol    40      b 1500.0039 secs
#> 48:    451    iris      ncol    41      a 1500.0033 secs
#> 49:    454    iris      ncol    41      d  500.0034 secs
#> 50:    497    iris      ncol    50      b 1500.0034 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 16h 24m 54.5s
#>   Total    : 4d 02h 08m 14.7s

rjoin(tab, est$runtimes)

#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1044.3833
#>   2:      2    iris      nrow     1      b estimated 1046.3668
#>   3:      3    iris      nrow     1      c estimated  208.1937
#>   4:      4    iris      nrow     1      d  observed  100.0034
#>   5:      5    iris      nrow     1      e estimated  219.5752
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1396.7774
#> 497:    497    iris      ncol    50      b  observed 1500.0034
#> 498:    498    iris      ncol    50      c estimated  597.5062
#> 499:    499    iris      ncol    50      d estimated  572.0760
#> 500:    500    iris      ncol    50      e estimated  566.7513

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 16h 24m 54.5s
#>   Parallel : 0d 08h 51m 27.3s
#>   Total    : 4d 02h 08m 14.7s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1428.9220
#>   2:    431 estimated 1426.1983
#>   3:    406 estimated 1425.8476
#>   4:    456 estimated 1424.2020
#>   5:    426 estimated 1423.5067
#>  ---                           
#> 446:     90 estimated  136.7803
#> 447:     95 estimated  136.7803
#> 448:     74 estimated  132.8907
#> 449:     79 estimated  132.0907
#> 450:     84 estimated  126.7584
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1044.3833    69
#>   2:      2 estimated 1046.3668    68
#>   3:      3 estimated  208.1937    60
#>   4:      5 estimated  219.5752    58
#>   5:      6 estimated 1050.9833    66
#>  ---                                 
#> 446:    495 estimated  566.7513    19
#> 447:    496 estimated 1396.7774    11
#> 448:    498 estimated  597.5062     4
#> 449:    499 estimated  572.0760    15
#> 450:    500 estimated  566.7513    20

print(ids)

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1044.3833    69
#>   2:      2 estimated 1046.3668    68
#>   3:      3 estimated  208.1937    60
#>   4:      5 estimated  219.5752    58
#>   5:      6 estimated 1050.9833    66
#>  ---                                 
#> 446:    495 estimated  566.7513    19
#> 447:    496 estimated 1396.7774    11
#> 448:    498 estimated  597.5062     4
#> 449:    499 estimated  572.0760    15
#> 450:    500 estimated  566.7513    20

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>     <int>     <num>
#>  1:    69 3526.5812
#>  2:    68 3535.8126
#>  3:    60 3599.7995
#>  4:    58 3598.2284
#>  5:    66 3563.3717
#>  6:    65 3574.2068
#>  7:    62 3589.1593
#>  8:    54 3591.7788
#>  9:    59 3599.6985
#> 10:    57 3598.1682
#> 11:    10 3599.8496
#> 12:    46 3500.1613
#> 13:    52 3597.1956
#> 14:    42 3597.5026
#> 15:    53 3596.6142
#> 16:    64 3579.0685
#> 17:    12 3599.9050
#> 18:    56 3598.2718
#> 19:    63 3583.7839
#> 20:    71 3516.3405
#> 21:    73 3468.9391
#> 22:     6 3594.2652
#> 23:    55 3598.4313
#> 24:     8 3599.1717
#> 25:    77 3551.7182
#> 26:    61 3595.7387
#> 27:    74 3598.5939
#> 28:    72 3502.1723
#> 29:    82 3590.5095
#> 30:     7 3599.3601
#> 31:    70 3521.5996
#> 32:    44 3598.4242
#> 33:    90  970.5616
#> 34:     5 3599.3209
#> 35:    43 3597.8709
#> 36:    81 3594.1642
#> 37:    89 3577.5041
#> 38:    50 3598.4437
#> 39:    83 3567.5372
#> 40:    80 3595.5570
#> 41:     2 3598.7527
#> 42:    75 3598.3410
#> 43:    51 3599.9663
#> 44:    76 3567.3972
#> 45:    67 3556.5666
#> 46:    79 3490.6798
#> 47:    45 3599.0170
#> 48:     4 3599.1229
#> 49:     1 3599.6114
#> 50:    78 3521.4281
#> 51:     3 3599.6362
#> 52:    41 3475.3081
#> 53:    47 3488.2508
#> 54:    18 3592.6048
#> 55:    37 3489.2869
#> 56:    48 3485.0520
#> 57:    28 3541.3406
#> 58:    35 3502.8685
#> 59:    13 3599.3233
#> 60:    31 3525.0521
#> 61:    38 3484.9237
#> 62:    32 3522.9662
#> 63:    39 3483.6839
#> 64:    17 3597.8528
#> 65:    40 3480.7739
#> 66:    49 3474.1252
#> 67:    16 3598.8011
#> 68:    34 3505.3774
#> 69:    11 3599.5959
#> 70:     9 3599.9779
#> 71:    29 3531.5211
#> 72:    15 3598.1800
#> 73:    14 3596.4544
#> 74:    30 3529.1093
#> 75:    33 3507.3109
#> 76:    36 3498.0792
#> 77:    26 3590.4385
#> 78:    20 3598.1479
#> 79:    27 3566.3473
#> 80:    24 3599.5365
#> 81:    21 3599.4573
#> 82:    19 3599.7217
#> 83:    25 3595.7357
#> 84:    22 3596.3357
#> 85:    23 3586.4197
#> 86:    84 3599.4421
#> 87:    87 3560.6517
#> 88:    86 3573.1026
#> 89:    85 3585.3276
#> 90:    88 3540.0906
#>     chunk   runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1044.3833    10
#>   2:      2 estimated 1046.3668    10
#>   3:      3 estimated  208.1937     2
#>   4:      5 estimated  219.5752     4
#>   5:      6 estimated 1050.9833     4
#>  ---                                 
#> 446:    495 estimated  566.7513     2
#> 447:    496 estimated 1396.7774     2
#> 448:    498 estimated  597.5062     6
#> 449:    499 estimated  572.0760     7
#> 450:    500 estimated  566.7513     6

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    10 31887.22
#>  2:     2 31773.96
#>  3:     4 31774.87
#>  4:     7 31776.13
#>  5:     8 31883.96
#>  6:     6 31884.41
#>  7:     1 31773.75
#>  8:     9 31878.74
#>  9:     3 31774.17
#> 10:     5 31887.26