Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry7fbc3ee316e6' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = flatten(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0019 secs
#>  2:     30    iris      nrow     6      e  100.0070 secs
#>  3:     43    iris      nrow     9      c  100.0020 secs
#>  4:     46    iris      nrow    10      a 1100.0020 secs
#>  5:     53    iris      nrow    11      c  100.0022 secs
#>  6:     55    iris      nrow    11      e  100.0020 secs
#>  7:     65    iris      nrow    13      e  100.0020 secs
#>  8:     66    iris      nrow    14      a 1100.0021 secs
#>  9:     89    iris      nrow    18      d  100.0034 secs
#> 10:     91    iris      nrow    19      a 1100.0020 secs
#> 11:     93    iris      nrow    19      c  100.0019 secs
#> 12:    124    iris      nrow    25      d  100.0019 secs
#> 13:    136    iris      nrow    28      a 1100.0026 secs
#> 14:    147    iris      nrow    30      b 1100.0020 secs
#> 15:    157    iris      nrow    32      b 1100.0019 secs
#> 16:    160    iris      nrow    32      e  100.0020 secs
#> 17:    163    iris      nrow    33      c  100.0018 secs
#> 18:    178    iris      nrow    36      c  100.0020 secs
#> 19:    179    iris      nrow    36      d  100.0020 secs
#> 20:    211    iris      nrow    43      a 1100.0039 secs
#> 21:    222    iris      nrow    45      b 1100.0091 secs
#> 22:    232    iris      nrow    47      b 1100.0040 secs
#> 23:    235    iris      nrow    47      e  100.0019 secs
#> 24:    239    iris      nrow    48      d  100.0054 secs
#> 25:    244    iris      nrow    49      d  100.0020 secs
#> 26:    251    iris      ncol     1      a 1500.0030 secs
#> 27:    260    iris      ncol     2      e  500.0028 secs
#> 28:    278    iris      ncol     6      c  500.0024 secs
#> 29:    292    iris      ncol     9      b 1500.0027 secs
#> 30:    326    iris      ncol    16      a 1500.0027 secs
#> 31:    341    iris      ncol    19      a 1500.0023 secs
#> 32:    343    iris      ncol    19      c  500.0021 secs
#> 33:    346    iris      ncol    20      a 1500.0021 secs
#> 34:    365    iris      ncol    23      e  500.0018 secs
#> 35:    367    iris      ncol    24      b 1500.0020 secs
#> 36:    372    iris      ncol    25      b 1500.0018 secs
#> 37:    374    iris      ncol    25      d  500.0021 secs
#> 38:    379    iris      ncol    26      d  500.0019 secs
#> 39:    397    iris      ncol    30      b 1500.0018 secs
#> 40:    409    iris      ncol    32      d  500.0019 secs
#> 41:    410    iris      ncol    32      e  500.0022 secs
#> 42:    418    iris      ncol    34      c  500.0019 secs
#> 43:    423    iris      ncol    35      c  500.0019 secs
#> 44:    425    iris      ncol    35      e  500.0019 secs
#> 45:    439    iris      ncol    38      d  500.0049 secs
#> 46:    442    iris      ncol    39      b 1500.0047 secs
#> 47:    443    iris      ncol    39      c  500.0044 secs
#> 48:    450    iris      ncol    40      e  500.0021 secs
#> 49:    451    iris      ncol    41      a 1500.0030 secs
#> 50:    459    iris      ncol    42      d  500.0019 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 19m 2.4s
#>   Total    : 4d 01h 02m 22.6s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1090.3899
#>   2:      2    iris      nrow     1      b estimated 1092.8477
#>   3:      3    iris      nrow     1      c estimated  297.8941
#>   4:      4    iris      nrow     1      d estimated  282.4756
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1262.4736
#> 497:    497    iris      ncol    50      b estimated 1260.4319
#> 498:    498    iris      ncol    50      c estimated  481.9342
#> 499:    499    iris      ncol    50      d estimated  467.3649
#> 500:    500    iris      ncol    50      e estimated  469.2935

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 19m 2.4s
#>   Parallel : 0d 08h 45m 3.6s
#>   Total    : 4d 01h 02m 22.6s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    347 estimated 1437.3593
#>   2:    352 estimated 1437.3593
#>   3:    362 estimated 1436.3453
#>   4:    357 estimated 1435.8786
#>  ---                           
#> 446:    184 estimated  148.8044
#> 447:     64 estimated  146.6495
#> 448:     54 estimated  145.1393
#> 449:     60 estimated  142.6300
#> 450:     59 estimated  140.6059
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3899    46
#>   2:      2 estimated 1092.8477    46
#>   3:      3 estimated  297.8941    27
#>   4:      4 estimated  282.4756    24
#>  ---                                 
#> 446:    496 estimated 1262.4736    44
#> 447:    497 estimated 1260.4319    45
#> 448:    498 estimated  481.9342    86
#> 449:    499 estimated  467.3649    87
#> 450:    500 estimated  469.2935    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3899    46
#>   2:      2 estimated 1092.8477    46
#>   3:      3 estimated  297.8941    27
#>   4:      4 estimated  282.4756    24
#>  ---                                 
#> 446:    496 estimated 1262.4736    44
#> 447:    497 estimated 1260.4319    45
#> 448:    498 estimated  481.9342    86
#> 449:    499 estimated  467.3649    87
#> 450:    500 estimated  469.2935    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    46 3598.472
#>  2:    27 3597.892
#>  3:    24 3598.210
#>  4:    25 3595.580
#>  5:    53 3566.468
#>  6:    51 3581.383
#>  7:    18 3598.080
#>  8:    22 3552.916
#>  9:    14 3596.998
#> 10:    58 3537.022
#> 11:    19 3587.076
#> 12:    23 3551.018
#> 13:    15 3594.180
#> 14:    60 3526.470
#> 15:    16 3599.534
#> 16:    28 3525.183
#> 17:    26 3536.835
#> 18:    64 3486.276
#> 19:    62 3512.665
#> 20:    39 3599.722
#> 21:    54 3562.122
#> 22:    65 3479.886
#> 23:    63 3499.862
#> 24:    40 3599.537
#> 25:    55 3556.947
#> 26:    66 3475.183
#> 27:    61 3520.309
#> 28:    47 3599.802
#> 29:    59 3533.411
#> 30:    67 3598.047
#> 31:    88 3507.823
#> 32:    87 3523.116
#> 33:    69 3598.695
#> 34:    89 1061.714
#> 35:    68 3599.922
#> 36:    37 3598.755
#> 37:     2 3598.168
#> 38:    38 3599.987
#> 39:     3 3599.771
#> 40:    56 3551.166
#> 41:    50 3587.215
#> 42:     4 3599.938
#> 43:    49 3598.012
#> 44:    76 3510.994
#> 45:    48 3599.970
#> 46:    57 3542.078
#> 47:    77 3496.781
#> 48:    52 3574.212
#> 49:    73 3595.755
#> 50:    45 3599.054
#> 51:    44 3599.685
#> 52:    75 3523.800
#> 53:    43 3599.717
#> 54:    41 3597.853
#> 55:     5 3599.910
#> 56:     6 3599.971
#> 57:    78 3477.635
#> 58:    70 3599.335
#> 59:    71 3599.731
#> 60:    72 3599.525
#> 61:    11 3595.842
#> 62:    21 3558.653
#> 63:    35 3468.892
#> 64:    31 3499.475
#> 65:    20 3562.518
#> 66:    34 3472.074
#> 67:    30 3501.791
#> 68:    13 3599.940
#> 69:    33 3478.121
#> 70:    10 3599.011
#> 71:    17 3586.920
#> 72:     9 3599.651
#> 73:    32 3491.352
#> 74:    12 3599.797
#> 75:     8 3599.958
#> 76:    36 3462.691
#> 77:    29 3505.673
#> 78:     7 3599.862
#> 79:    42 3597.280
#> 80:    74 3513.995
#> 81:     1 3467.822
#> 82:    82 3582.182
#> 83:    83 3566.580
#> 84:    84 3544.571
#> 85:    79 3597.577
#> 86:    85 3527.136
#> 87:    80 3599.849
#> 88:    81 3599.965
#> 89:    86 3455.874
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3899     7
#>   2:      2 estimated 1092.8477     2
#>   3:      3 estimated  297.8941     3
#>   4:      4 estimated  282.4756     5
#>  ---                                 
#> 446:    496 estimated 1262.4736     6
#> 447:    497 estimated 1260.4319     7
#> 448:    498 estimated  481.9342     8
#> 449:    499 estimated  467.3649     1
#> 450:    500 estimated  469.2935     8

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     7 31500.51
#>  2:     2 31501.67
#>  3:     3 31370.36
#>  4:     5 31371.58
#>  5:     9 31370.87
#>  6:    10 31364.82
#>  7:     6 31364.82
#>  8:     8 31498.11
#>  9:     1 31503.61
#> 10:     4 31496.10