Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab

[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[any] Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate] Object to print.

n

[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

lpt binpack

Examples

# Create a simple toy registry
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done error memory resource.id
#>   1:      1      1         NA         NA         NA    NA     NA          NA
#>   2:      2      2         NA         NA         NA    NA     NA          NA
#>   3:      3      3         NA         NA         NA    NA     NA          NA
#>   4:      4      4         NA         NA         NA    NA     NA          NA
#>   5:      5      5 1483361236 1483361236 1483361336    NA     NA           1
#>  ---                                                                        
#> 496:    496    496 1483361237 1483361237 1483362737    NA     NA           1
#> 497:    497    497         NA         NA         NA    NA     NA          NA
#> 498:    498    498         NA         NA         NA    NA     NA          NA
#> 499:    499    499         NA         NA         NA    NA     NA          NA
#> 500:    500    500 1483361237 1483361237 1483361737    NA     NA           1
#>           batch.id                            job.hash repl
#>   1:            NA                                  NA    1
#>   2:            NA                                  NA    1
#>   3:            NA                                  NA    1
#>   4:            NA                                  NA    1
#>   5: cfInteractive job58c68292c1c5f0e08c7f40f5cdcfec00    1
#>  ---                                                       
#> 496: cfInteractive job78553dfa3a2b2ac2793950ff5d5168c8    1
#> 497:            NA                                  NA    1
#> 498:            NA                                  NA    1
#> 499:            NA                                  NA    1
#> 500: cfInteractive jobd055601d788c400d19b2a54aa00652e8    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y time.running
#>  1:      5    iris      nrow  1 e     100 secs
#>  2:     25    iris      nrow  5 e     100 secs
#>  3:     33    iris      nrow  7 c     100 secs
#>  4:     40    iris      nrow  8 e     100 secs
#>  5:     53    iris      nrow 11 c     100 secs
#>  6:     62    iris      nrow 13 b    1100 secs
#>  7:     82    iris      nrow 17 b    1100 secs
#>  8:     84    iris      nrow 17 d     100 secs
#>  9:     91    iris      nrow 19 a    1100 secs
#> 10:    106    iris      nrow 22 a    1100 secs
#> 11:    108    iris      nrow 22 c     100 secs
#> 12:    110    iris      nrow 22 e     100 secs
#> 13:    147    iris      nrow 30 b    1100 secs
#> 14:    149    iris      nrow 30 d     100 secs
#> 15:    157    iris      nrow 32 b    1100 secs
#> 16:    158    iris      nrow 32 c     100 secs
#> 17:    170    iris      nrow 34 e     100 secs
#> 18:    174    iris      nrow 35 d     100 secs
#> 19:    191    iris      nrow 39 a    1100 secs
#> 20:    202    iris      nrow 41 b    1100 secs
#> 21:    216    iris      nrow 44 a    1100 secs
#> 22:    234    iris      nrow 47 d     100 secs
#> 23:    241    iris      nrow 49 a    1100 secs
#> 24:    248    iris      nrow 50 c     100 secs
#> 25:    249    iris      nrow 50 d     100 secs
#> 26:    258    iris      ncol  2 c     500 secs
#> 27:    268    iris      ncol  4 c     500 secs
#> 28:    292    iris      ncol  9 b    1500 secs
#> 29:    297    iris      ncol 10 b    1500 secs
#> 30:    305    iris      ncol 11 e     500 secs
#> 31:    308    iris      ncol 12 c     500 secs
#> 32:    357    iris      ncol 22 b    1500 secs
#> 33:    361    iris      ncol 23 a    1500 secs
#> 34:    366    iris      ncol 24 a    1500 secs
#> 35:    372    iris      ncol 25 b    1500 secs
#> 36:    374    iris      ncol 25 d     500 secs
#> 37:    375    iris      ncol 25 e     500 secs
#> 38:    384    iris      ncol 27 d     500 secs
#> 39:    388    iris      ncol 28 c     500 secs
#> 40:    394    iris      ncol 29 d     500 secs
#> 41:    395    iris      ncol 29 e     500 secs
#> 42:    398    iris      ncol 30 c     500 secs
#> 43:    406    iris      ncol 32 a    1500 secs
#> 44:    414    iris      ncol 33 d     500 secs
#> 45:    415    iris      ncol 33 e     500 secs
#> 46:    432    iris      ncol 37 b    1500 secs
#> 47:    479    iris      ncol 46 d     500 secs
#> 48:    486    iris      ncol 48 a    1500 secs
#> 49:    496    iris      ncol 50 a    1500 secs
#> 50:    500    iris      ncol 50 e     500 secs
#>     job.id problem algorithm  x y time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 17m 33.5s
#>   Total    : 4d 00h 00m 53.5s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated  911.4618
#>   2:      2    iris      nrow  1 b estimated  908.7240
#>   3:      3    iris      nrow  1 c estimated  163.1707
#>   4:      4    iris      nrow  1 d estimated  150.6731
#>   5:      5    iris      nrow  1 e  observed  100.0000
#>  ---                                                  
#> 496:    496    iris      ncol 50 a  observed 1500.0000
#> 497:    497    iris      ncol 50 b estimated 1352.8819
#> 498:    498    iris      ncol 50 c estimated  521.8169
#> 499:    499    iris      ncol 50 d estimated  526.8793
#> 500:    500    iris      ncol 50 e  observed  500.0000

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 17m 33.5s
#>   Parallel : 0d 08h 38m 32.4s
#>   Total    : 4d 00h 00m 53.5s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    362 estimated 1428.2699
#>   2:    371 estimated 1426.6383
#>   3:    367 estimated 1422.3524
#>   4:    401 estimated 1418.7957
#>   5:    426 estimated 1417.4987
#>  ---                           
#> 446:     30 estimated  136.9780
#> 447:     35 estimated  136.9780
#> 448:     29 estimated  136.6671
#> 449:     34 estimated  136.6671
#> 450:     24 estimated  136.5931

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated  911.4618    69
#>   2:      2 estimated  908.7240    71
#>   3:      3 estimated  163.1707    87
#>   4:      4 estimated  150.6731    88
#>   5:      6 estimated  912.2618    68
#>  ---                                 
#> 446:    494 estimated  531.0326    77
#> 447:    495 estimated  543.9307    18
#> 448:    497 estimated 1352.8819    37
#> 449:    498 estimated  521.8169    82
#> 450:    499 estimated  526.8793    78

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated  911.4618    69
#>   2:      2 estimated  908.7240    71
#>   3:      3 estimated  163.1707    87
#>   4:      4 estimated  150.6731    88
#>   5:      6 estimated  912.2618    68
#>  ---                                 
#> 446:    494 estimated  531.0326    77
#> 447:    495 estimated  543.9307    18
#> 448:    497 estimated 1352.8819    37
#> 449:    498 estimated  521.8169    82
#> 450:    499 estimated  526.8793    78

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    69 3599.796
#>  2:    71 3599.663
#>  3:    87 3551.269
#>  4:    88 2039.137
#>  5:    68 3515.269
#>  6:    70 3598.595
#>  7:    72 3599.747
#>  8:    37 3598.562
#>  9:     1 3599.539
#> 10:    81 3465.338
#> 11:    73 3587.105
#> 12:    36 3599.838
#> 13:    82 3597.560
#> 14:    58 3494.275
#> 15:    59 3485.254
#> 16:     5 3599.875
#> 17:    27 3503.862
#> 18:    14 3559.373
#> 19:    65 3599.622
#> 20:    66 3597.051
#> 21:    47 3561.410
#> 22:    51 3542.471
#> 23:     3 3599.589
#> 24:    38 3591.980
#> 25:    67 3558.681
#> 26:    74 3561.873
#> 27:    62 3474.844
#> 28:    60 3480.599
#> 29:    77 3522.602
#> 30:    75 3546.289
#> 31:    61 3478.783
#> 32:    64 3598.968
#> 33:    78 3502.763
#> 34:    79 3488.025
#> 35:    76 3531.082
#> 36:    80 3476.450
#> 37:    53 3527.845
#> 38:    86 3584.917
#> 39:    54 3523.706
#> 40:    49 3550.397
#> 41:    52 3532.577
#> 42:    40 3588.723
#> 43:    48 3553.999
#> 44:    56 3514.713
#> 45:    55 3518.802
#> 46:    39 3598.047
#> 47:    41 3571.999
#> 48:    50 3548.554
#> 49:    57 3511.701
#> 50:    46 3565.212
#> 51:    42 3553.065
#> 52:    44 3536.321
#> 53:    15 3557.485
#> 54:    20 3540.372
#> 55:     7 3593.573
#> 56:    43 3544.140
#> 57:    45 3527.397
#> 58:    16 3548.056
#> 59:    21 3538.301
#> 60:     8 3590.894
#> 61:    23 3528.413
#> 62:    32 3484.260
#> 63:     6 3599.875
#> 64:    24 3517.748
#> 65:    33 3480.287
#> 66:    10 3583.071
#> 67:    25 3511.069
#> 68:    34 3478.919
#> 69:    12 3576.443
#> 70:    26 3507.583
#> 71:    35 3464.091
#> 72:    13 3565.351
#> 73:    17 3547.506
#> 74:    18 3544.619
#> 75:    19 3543.160
#> 76:    63 3599.520
#> 77:     2 3599.998
#> 78:    28 3500.698
#> 79:    29 3500.236
#> 80:     4 3599.534
#> 81:    22 3534.803
#> 82:    31 3485.660
#> 83:     9 3587.959
#> 84:    30 3493.844
#> 85:    11 3579.292
#> 86:    85 3516.106
#> 87:    83 3598.066
#> 88:    84 3597.488
#>     chunk  runtime

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated  911.4618     2
#>   2:      2 estimated  908.7240     4
#>   3:      3 estimated  163.1707    10
#>   4:      4 estimated  150.6731    10
#>   5:      6 estimated  912.2618     8
#>  ---                                 
#> 446:    494 estimated  531.0326     7
#> 447:    495 estimated  543.9307     9
#> 448:    497 estimated 1352.8819     2
#> 449:    498 estimated  521.8169    10
#> 450:    499 estimated  526.8793     2

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:     2 31017.41
#>  2:     4 31111.42
#>  3:    10 31112.12
#>  4:     8 31112.39
#>  5:     3 31112.45
#>  6:     5 31112.37
#>  7:     7 31016.96
#>  8:     9 31020.73
#>  9:     1 31019.83
#> 10:     6 31017.85