Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab

[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate] Object to print.

n

[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Ignoring resource 'chunks.as.arrayjobs', not supported by cluster functions 'Multicore'

#> Submitting 50 jobs in 50 chunks using cluster functions 'Multicore' ...

waitForJobs(reg = tmp)

#> Syncing 49 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started done error memory resource.id batch.id
#>   1:      1      1        NA      NA   NA    NA     NA          NA       NA
#>   2:      2      2        NA      NA   NA    NA     NA          NA       NA
#>   3:      3      3        NA      NA   NA    NA     NA          NA       NA
#>   4:      4      4        NA      NA   NA    NA     NA          NA       NA
#>   5:      5      5        NA      NA   NA    NA     NA          NA       NA
#>  ---                                                                       
#> 496:    496    496        NA      NA   NA    NA     NA          NA       NA
#> 497:    497    497        NA      NA   NA    NA     NA          NA       NA
#> 498:    498    498        NA      NA   NA    NA     NA          NA       NA
#> 499:    499    499        NA      NA   NA    NA     NA          NA       NA
#> 500:    500    500        NA      NA   NA    NA     NA          NA       NA
#>      log.file job.hash repl
#>   1:       NA       NA    1
#>   2:       NA       NA    1
#>   3:       NA       NA    1
#>   4:       NA       NA    1
#>   5:       NA       NA    1
#>  ---                       
#> 496:       NA       NA    1
#> 497:       NA       NA    1
#> 498:       NA       NA    1
#> 499:       NA       NA    1
#> 500:       NA       NA    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y time.running
#>  1:     12    iris      nrow  3 b    1100 secs
#>  2:     30    iris      nrow  6 e     100 secs
#>  3:     43    iris      nrow  9 c     100 secs
#>  4:     46    iris      nrow 10 a    1100 secs
#>  5:     53    iris      nrow 11 c     100 secs
#>  6:     55    iris      nrow 11 e     100 secs
#>  7:     65    iris      nrow 13 e     100 secs
#>  8:     66    iris      nrow 14 a    1100 secs
#>  9:     89    iris      nrow 18 d     100 secs
#> 10:     91    iris      nrow 19 a    1100 secs
#> 11:     93    iris      nrow 19 c     100 secs
#> 12:    124    iris      nrow 25 d     100 secs
#> 13:    136    iris      nrow 28 a    1100 secs
#> 14:    147    iris      nrow 30 b    1100 secs
#> 15:    157    iris      nrow 32 b    1101 secs
#> 16:    160    iris      nrow 32 e     100 secs
#> 17:    163    iris      nrow 33 c     100 secs
#> 18:    178    iris      nrow 36 c     100 secs
#> 19:    179    iris      nrow 36 d     100 secs
#> 20:    211    iris      nrow 43 a    1100 secs
#> 21:    222    iris      nrow 45 b    1100 secs
#> 22:    232    iris      nrow 47 b    1100 secs
#> 23:    235    iris      nrow 47 e     100 secs
#> 24:    239    iris      nrow 48 d     100 secs
#> 25:    244    iris      nrow 49 d     100 secs
#> 26:    251    iris      ncol  1 a    1500 secs
#> 27:    260    iris      ncol  2 e     500 secs
#> 28:    278    iris      ncol  6 c     500 secs
#> 29:    292    iris      ncol  9 b    1500 secs
#> 30:    326    iris      ncol 16 a    1500 secs
#> 31:    341    iris      ncol 19 a    1500 secs
#> 32:    343    iris      ncol 19 c     500 secs
#> 33:    346    iris      ncol 20 a    1500 secs
#> 34:    365    iris      ncol 23 e     500 secs
#> 35:    367    iris      ncol 24 b    1500 secs
#> 36:    372    iris      ncol 25 b    1500 secs
#> 37:    374    iris      ncol 25 d     500 secs
#> 38:    379    iris      ncol 26 d     500 secs
#> 39:    397    iris      ncol 30 b    1500 secs
#> 40:    409    iris      ncol 32 d     500 secs
#> 41:    410    iris      ncol 32 e     500 secs
#> 42:    418    iris      ncol 34 c     500 secs
#> 43:    423    iris      ncol 35 c     500 secs
#> 44:    425    iris      ncol 35 e     500 secs
#> 45:    439    iris      ncol 38 d     500 secs
#> 46:    442    iris      ncol 39 b    1500 secs
#> 47:    443    iris      ncol 39 c     500 secs
#> 48:    450    iris      ncol 40 e     500 secs
#> 49:    451    iris      ncol 41 a    1500 secs
#> 50:    459    iris      ncol 42 d     500 secs
#>     job.id problem algorithm  x y time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.0s
#>   Remaining: 3d 14h 21m 35.1s
#>   Total    : 4d 00h 04m 56.1s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated 1078.7200
#>   2:      2    iris      nrow  1 b estimated 1061.7974
#>   3:      3    iris      nrow  1 c estimated  267.8001
#>   4:      4    iris      nrow  1 d estimated  257.2484
#>   5:      5    iris      nrow  1 e estimated  262.3271
#>  ---                                                  
#> 496:    496    iris      ncol 50 a estimated 1272.1558
#> 497:    497    iris      ncol 50 b estimated 1263.2513
#> 498:    498    iris      ncol 50 c estimated  461.6721
#> 499:    499    iris      ncol 50 d estimated  460.4041
#> 500:    500    iris      ncol 50 e estimated  461.1751

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.0s
#>   Remaining: 3d 14h 21m 35.1s
#>   Parallel : 0d 08h 39m 19.1s
#>   Total    : 4d 00h 04m 56.1s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    352 estimated 1441.3745
#>   2:    357 estimated 1441.3745
#>   3:    362 estimated 1441.3745
#>   4:    321 estimated 1437.9971
#>   5:    366 estimated 1437.9079
#>  ---                           
#> 446:     58 estimated  139.1190
#> 447:     63 estimated  139.1190
#> 448:     54 estimated  133.9673
#> 449:     59 estimated  133.6207
#> 450:     64 estimated  133.6207

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1078.7200    46
#>   2:      2 estimated 1061.7974    54
#>   3:      3 estimated  267.8001    20
#>   4:      4 estimated  257.2484    15
#>   5:      5 estimated  262.3271    19
#>  ---                                 
#> 446:    496 estimated 1272.1558    44
#> 447:    497 estimated 1263.2513    45
#> 448:    498 estimated  461.6721    85
#> 449:    499 estimated  460.4041    86
#> 450:    500 estimated  461.1751    85

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1078.7200    46
#>   2:      2 estimated 1061.7974    54
#>   3:      3 estimated  267.8001    20
#>   4:      4 estimated  257.2484    15
#>   5:      5 estimated  262.3271    19
#>  ---                                 
#> 446:    496 estimated 1272.1558    44
#> 447:    497 estimated 1263.2513    45
#> 448:    498 estimated  461.6721    85
#> 449:    499 estimated  460.4041    86
#> 450:    500 estimated  461.1751    85

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    46 3581.360
#>  2:    54 3511.746
#>  3:    20 3595.410
#>  4:    15 3599.939
#>  5:    19 3599.752
#>  6:    62 3599.937
#>  7:    64 3599.722
#>  8:    13 3598.186
#>  9:    14 3586.368
#> 10:    12 3597.986
#> 11:    11 3599.792
#> 12:    16 3548.919
#> 13:    10 3597.645
#> 14:    65 3599.514
#> 15:    71 3542.556
#> 16:    27 3484.192
#> 17:    53 3519.376
#> 18:    50 3544.053
#> 19:    28 3482.616
#> 20:    29 3481.057
#> 21:    63 3589.698
#> 22:    69 3563.573
#> 23:    76 3464.679
#> 24:    87 3498.169
#> 25:    60 3467.240
#> 26:    70 3549.477
#> 27:    85 3598.631
#> 28:    61 3463.615
#> 29:    56 3502.799
#> 30:    86 3579.476
#> 31:    72 3492.152
#> 32:     1 3597.024
#> 33:    66 3599.205
#> 34:    42 3595.667
#> 35:    88 1585.934
#> 36:    34 3598.744
#> 37:    36 3598.745
#> 38:    67 3599.912
#> 39:    35 3595.822
#> 40:    43 3599.396
#> 41:    57 3493.286
#> 42:    58 3474.817
#> 43:     3 3598.918
#> 44:     2 3598.127
#> 45:    68 3591.301
#> 46:    49 3548.098
#> 47:    37 3599.885
#> 48:    59 3471.996
#> 49:    84 3599.937
#> 50:    51 3534.475
#> 51:    75 3478.086
#> 52:    48 3555.151
#> 53:    55 3504.855
#> 54:    30 3476.982
#> 55:    39 3585.355
#> 56:     9 3574.392
#> 57:    32 3470.969
#> 58:    18 3518.424
#> 59:    33 3467.081
#> 60:    21 3506.188
#> 61:    47 3567.320
#> 62:    25 3488.579
#> 63:    26 3485.582
#> 64:    52 3526.385
#> 65:    73 3599.488
#> 66:    44 3599.965
#> 67:    38 3599.964
#> 68:    40 3564.292
#> 69:    31 3476.411
#> 70:    24 3490.020
#> 71:     4 3599.584
#> 72:    45 3598.475
#> 73:     6 3591.884
#> 74:    17 3519.049
#> 75:     5 3598.346
#> 76:     7 3589.661
#> 77:    23 3498.623
#> 78:     8 3580.533
#> 79:    22 3502.978
#> 80:    41 3598.839
#> 81:    74 3599.995
#> 82:    77 3599.125
#> 83:    79 3596.367
#> 84:    78 3599.934
#> 85:    80 3592.707
#> 86:    81 3586.140
#> 87:    82 3598.354
#> 88:    83 3588.061
#>     chunk  runtime

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1078.7200     7
#>   2:      2 estimated 1061.7974     9
#>   3:      3 estimated  267.8001     6
#>   4:      4 estimated  257.2484     3
#>   5:      5 estimated  262.3271     5
#>  ---                                 
#> 446:    496 estimated 1272.1558    10
#> 447:    497 estimated 1263.2513     6
#> 448:    498 estimated  461.6721    10
#> 449:    499 estimated  460.4041     7
#> 450:    500 estimated  461.1751     8

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:     7 31024.38
#>  2:     9 31154.48
#>  3:     6 31021.42
#>  4:     3 31159.07
#>  5:     5 31158.97
#>  6:     4 31021.70
#>  7:     1 31021.96
#>  8:     2 31154.71
#>  9:     8 31024.21
#> 10:    10 31154.18