Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab

[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[any] Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate] Object to print.

n

[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

lpt binpack

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started done error memory resource.id batch.id
#>   1:      1      1        NA      NA   NA    NA     NA          NA       NA
#>   2:      2      2        NA      NA   NA    NA     NA          NA       NA
#>   3:      3      3        NA      NA   NA    NA     NA          NA       NA
#>   4:      4      4        NA      NA   NA    NA     NA          NA       NA
#>   5:      5      5        NA      NA   NA    NA     NA          NA       NA
#>  ---                                                                       
#> 496:    496    496        NA      NA   NA    NA     NA          NA       NA
#> 497:    497    497        NA      NA   NA    NA     NA          NA       NA
#> 498:    498    498        NA      NA   NA    NA     NA          NA       NA
#> 499:    499    499        NA      NA   NA    NA     NA          NA       NA
#> 500:    500    500        NA      NA   NA    NA     NA          NA       NA
#>      job.hash repl
#>   1:       NA    1
#>   2:       NA    1
#>   3:       NA    1
#>   4:       NA    1
#>   5:       NA    1
#>  ---              
#> 496:       NA    1
#> 497:       NA    1
#> 498:       NA    1
#> 499:       NA    1
#> 500:       NA    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y time.running
#>  1:     12    iris      nrow  3 b    1100 secs
#>  2:     30    iris      nrow  6 e     100 secs
#>  3:     43    iris      nrow  9 c     100 secs
#>  4:     46    iris      nrow 10 a    1100 secs
#>  5:     53    iris      nrow 11 c     100 secs
#>  6:     55    iris      nrow 11 e     100 secs
#>  7:     65    iris      nrow 13 e     100 secs
#>  8:     66    iris      nrow 14 a    1100 secs
#>  9:     89    iris      nrow 18 d     100 secs
#> 10:     91    iris      nrow 19 a    1100 secs
#> 11:     93    iris      nrow 19 c     100 secs
#> 12:    124    iris      nrow 25 d     100 secs
#> 13:    136    iris      nrow 28 a    1100 secs
#> 14:    147    iris      nrow 30 b    1100 secs
#> 15:    157    iris      nrow 32 b    1100 secs
#> 16:    160    iris      nrow 32 e     100 secs
#> 17:    163    iris      nrow 33 c     100 secs
#> 18:    178    iris      nrow 36 c     100 secs
#> 19:    179    iris      nrow 36 d     100 secs
#> 20:    211    iris      nrow 43 a    1100 secs
#> 21:    222    iris      nrow 45 b    1100 secs
#> 22:    232    iris      nrow 47 b    1100 secs
#> 23:    235    iris      nrow 47 e     100 secs
#> 24:    239    iris      nrow 48 d     100 secs
#> 25:    244    iris      nrow 49 d     100 secs
#> 26:    251    iris      ncol  1 a    1500 secs
#> 27:    260    iris      ncol  2 e     500 secs
#> 28:    278    iris      ncol  6 c     500 secs
#> 29:    292    iris      ncol  9 b    1500 secs
#> 30:    326    iris      ncol 16 a    1500 secs
#> 31:    341    iris      ncol 19 a    1500 secs
#> 32:    343    iris      ncol 19 c     500 secs
#> 33:    346    iris      ncol 20 a    1500 secs
#> 34:    365    iris      ncol 23 e     500 secs
#> 35:    367    iris      ncol 24 b    1500 secs
#> 36:    372    iris      ncol 25 b    1500 secs
#> 37:    374    iris      ncol 25 d     500 secs
#> 38:    379    iris      ncol 26 d     500 secs
#> 39:    397    iris      ncol 30 b    1500 secs
#> 40:    409    iris      ncol 32 d     500 secs
#> 41:    410    iris      ncol 32 e     500 secs
#> 42:    418    iris      ncol 34 c     500 secs
#> 43:    423    iris      ncol 35 c     500 secs
#> 44:    425    iris      ncol 35 e     500 secs
#> 45:    439    iris      ncol 38 d     500 secs
#> 46:    442    iris      ncol 39 b    1500 secs
#> 47:    443    iris      ncol 39 c     500 secs
#> 48:    450    iris      ncol 40 e     500 secs
#> 49:    451    iris      ncol 41 a    1500 secs
#> 50:    459    iris      ncol 42 d     500 secs
#>     job.id problem algorithm  x y time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 10m 4.4s
#>   Total    : 3d 23h 53m 24.4s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated 1076.6090
#>   2:      2    iris      nrow  1 b estimated 1051.1828
#>   3:      3    iris      nrow  1 c estimated  286.8379
#>   4:      4    iris      nrow  1 d estimated  282.1404
#>   5:      5    iris      nrow  1 e estimated  286.3669
#>  ---                                                  
#> 496:    496    iris      ncol 50 a estimated 1218.3134
#> 497:    497    iris      ncol 50 b estimated 1203.0382
#> 498:    498    iris      ncol 50 c estimated  465.8779
#> 499:    499    iris      ncol 50 d estimated  463.9445
#> 500:    500    iris      ncol 50 e estimated  458.2146

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 10m 4.4s
#>   Parallel : 0d 08h 38m 10.3s
#>   Total    : 3d 23h 53m 24.4s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    347 estimated 1432.7301
#>   2:    352 estimated 1430.8378
#>   3:    357 estimated 1430.8378
#>   4:    362 estimated 1430.8378
#>   5:    366 estimated 1430.6995
#>  ---                           
#> 446:    248 estimated  152.3007
#> 447:     59 estimated  151.6543
#> 448:     64 estimated  151.6543
#> 449:     54 estimated  149.7972
#> 450:    249 estimated  148.7674

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1076.6090    42
#>   2:      2 estimated 1051.1828    53
#>   3:      3 estimated  286.8379    22
#>   4:      4 estimated  282.1404    21
#>   5:      5 estimated  286.3669    23
#>  ---                                 
#> 446:    496 estimated 1218.3134    44
#> 447:    497 estimated 1203.0382    45
#> 448:    498 estimated  465.8779    56
#> 449:    499 estimated  463.9445    78
#> 450:    500 estimated  458.2146    84

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1076.6090    42
#>   2:      2 estimated 1051.1828    53
#>   3:      3 estimated  286.8379    22
#>   4:      4 estimated  282.1404    21
#>   5:      5 estimated  286.3669    23
#>  ---                                 
#> 446:    496 estimated 1218.3134    44
#> 447:    497 estimated 1203.0382    45
#> 448:    498 estimated  465.8779    56
#> 449:    499 estimated  463.9445    78
#> 450:    500 estimated  458.2146    84

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    42 3593.125
#>  2:    53 3515.139
#>  3:    22 3593.486
#>  4:    21 3596.998
#>  5:    23 3587.126
#>  6:    58 3597.921
#>  7:    40 3596.722
#>  8:    11 3574.199
#>  9:    12 3567.282
#> 10:    10 3585.175
#> 11:    56 3584.918
#> 12:    13 3562.230
#> 13:    14 3552.285
#> 14:     9 3595.597
#> 15:    64 3598.805
#> 16:    70 3559.330
#> 17:    84 3599.450
#> 18:    85 3501.821
#> 19:    59 3599.562
#> 20:    67 3586.528
#> 21:    86 3502.318
#> 22:    68 3579.865
#> 23:    62 3591.948
#> 24:    69 3576.390
#> 25:    87 3597.310
#> 26:    60 3599.843
#> 27:    65 3595.627
#> 28:    27 3599.754
#> 29:    51 3540.898
#> 30:    54 3593.001
#> 31:    52 3531.773
#> 32:    41 3598.658
#> 33:    55 3586.728
#> 34:    44 3490.112
#> 35:    48 3568.466
#> 36:    73 3531.026
#> 37:    47 3574.836
#> 38:    45 3473.583
#> 39:    74 3518.260
#> 40:    46 3586.234
#> 41:    75 3513.325
#> 42:    49 3564.826
#> 43:    43 3506.334
#> 44:    50 3549.825
#> 45:    28 3599.874
#> 46:     5 3596.582
#> 47:    72 3563.845
#> 48:     4 3599.939
#> 49:    57 3580.347
#> 50:    66 3594.592
#> 51:    61 3597.601
#> 52:    63 3599.364
#> 53:    38 3536.969
#> 54:    37 3555.175
#> 55:    24 3503.664
#> 56:    36 3574.798
#> 57:     8 3598.803
#> 58:    16 3546.414
#> 59:    18 3543.414
#> 60:    15 3549.936
#> 61:    19 3538.494
#> 62:    20 3531.716
#> 63:    17 3545.308
#> 64:    25 3492.209
#> 65:     7 3597.882
#> 66:    35 3578.385
#> 67:     6 3598.407
#> 68:    26 3465.408
#> 69:    39 3593.656
#> 70:    71 3599.918
#> 71:    30 3595.154
#> 72:    29 3599.878
#> 73:    31 3593.154
#> 74:    33 3577.445
#> 75:    32 3585.780
#> 76:    34 3576.143
#> 77:     2 3460.171
#> 78:     1 3462.966
#> 79:     3 3455.224
#> 80:    80 3598.931
#> 81:    76 3597.158
#> 82:    78 3581.160
#> 83:    81 3595.470
#> 84:    79 3592.230
#> 85:    82 3584.956
#> 86:    83 3555.261
#> 87:    77 3591.946
#>     chunk  runtime

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1076.6090     5
#>   2:      2 estimated 1051.1828     4
#>   3:      3 estimated  286.8379     4
#>   4:      4 estimated  282.1404     3
#>   5:      5 estimated  286.3669     9
#>  ---                                 
#> 446:    496 estimated 1218.3134     8
#> 447:    497 estimated 1203.0382     6
#> 448:    498 estimated  465.8779    10
#> 449:    499 estimated  463.9445     5
#> 450:    500 estimated  458.2146     7

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:     5 31090.14
#>  2:     4 30959.47
#>  3:     3 30958.57
#>  4:     9 30958.12
#>  5:     6 31089.13
#>  6:     8 30940.60
#>  7:     1 31090.34
#>  8:     7 30939.91
#>  9:     2 31088.59
#> 10:    10 31089.51