Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

#> Created registry in '/tmp/RtmpFgx6O2/registry4fc14823892b' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0019 secs
#>  2:     30    iris      nrow     6      e  100.0061 secs
#>  3:     43    iris      nrow     9      c  100.0018 secs
#>  4:     46    iris      nrow    10      a 1100.0018 secs
#>  5:     53    iris      nrow    11      c  100.0018 secs
#>  6:     55    iris      nrow    11      e  100.0017 secs
#>  7:     65    iris      nrow    13      e  100.0017 secs
#>  8:     66    iris      nrow    14      a 1100.0017 secs
#>  9:     89    iris      nrow    18      d  100.0017 secs
#> 10:     91    iris      nrow    19      a 1100.0017 secs
#> 11:     93    iris      nrow    19      c  100.0017 secs
#> 12:    124    iris      nrow    25      d  100.0018 secs
#> 13:    136    iris      nrow    28      a 1100.0016 secs
#> 14:    147    iris      nrow    30      b 1100.0018 secs
#> 15:    157    iris      nrow    32      b 1100.0017 secs
#> 16:    160    iris      nrow    32      e  100.0017 secs
#> 17:    163    iris      nrow    33      c  100.0017 secs
#> 18:    178    iris      nrow    36      c  100.0016 secs
#> 19:    179    iris      nrow    36      d  100.0016 secs
#> 20:    211    iris      nrow    43      a 1100.0017 secs
#> 21:    222    iris      nrow    45      b 1100.0019 secs
#> 22:    232    iris      nrow    47      b 1100.0018 secs
#> 23:    235    iris      nrow    47      e  100.0019 secs
#> 24:    239    iris      nrow    48      d  100.0017 secs
#> 25:    244    iris      nrow    49      d  100.0017 secs
#> 26:    251    iris      ncol     1      a 1500.0017 secs
#> 27:    260    iris      ncol     2      e  500.0018 secs
#> 28:    278    iris      ncol     6      c  500.0018 secs
#> 29:    292    iris      ncol     9      b 1500.0017 secs
#> 30:    326    iris      ncol    16      a 1500.0017 secs
#> 31:    341    iris      ncol    19      a 1500.0021 secs
#> 32:    343    iris      ncol    19      c  500.0017 secs
#> 33:    346    iris      ncol    20      a 1500.0018 secs
#> 34:    365    iris      ncol    23      e  500.0017 secs
#> 35:    367    iris      ncol    24      b 1500.0018 secs
#> 36:    372    iris      ncol    25      b 1500.0020 secs
#> 37:    374    iris      ncol    25      d  500.0017 secs
#> 38:    379    iris      ncol    26      d  500.0017 secs
#> 39:    397    iris      ncol    30      b 1500.0017 secs
#> 40:    409    iris      ncol    32      d  500.0018 secs
#> 41:    410    iris      ncol    32      e  500.0021 secs
#> 42:    418    iris      ncol    34      c  500.0017 secs
#> 43:    423    iris      ncol    35      c  500.0017 secs
#> 44:    425    iris      ncol    35      e  500.0017 secs
#> 45:    439    iris      ncol    38      d  500.0017 secs
#> 46:    442    iris      ncol    39      b 1500.0020 secs
#> 47:    443    iris      ncol    39      c  500.0017 secs
#> 48:    450    iris      ncol    40      e  500.0017 secs
#> 49:    451    iris      ncol    41      a 1500.0021 secs
#> 50:    459    iris      ncol    42      d  500.0017 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 56.9s
#>   Total    : 4d 01h 02m 17.0s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1087.9892
#>   2:      2    iris      nrow     1      b estimated 1092.2870
#>   3:      3    iris      nrow     1      c estimated  295.8935
#>   4:      4    iris      nrow     1      d estimated  280.3948
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1259.7921
#> 497:    497    iris      ncol    50      b estimated 1257.4302
#> 498:    498    iris      ncol    50      c estimated  481.5329
#> 499:    499    iris      ncol    50      d estimated  467.6437
#> 500:    500    iris      ncol    50      e estimated  468.8258

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 56.9s
#>   Parallel : 0d 08h 45m 5.9s
#>   Total    : 4d 01h 02m 17.0s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    352 estimated 1438.5277
#>   2:    347 estimated 1438.5277
#>   3:    362 estimated 1435.7994
#>   4:    357 estimated 1435.3327
#>  ---                           
#> 446:    184 estimated  149.7682
#> 447:     64 estimated  146.5688
#> 448:     54 estimated  145.0586
#> 449:     60 estimated  143.5295
#> 450:     59 estimated  140.5253
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9892    46
#>   2:      2 estimated 1092.2870    46
#>   3:      3 estimated  295.8935    28
#>   4:      4 estimated  280.3948    23
#>  ---                                 
#> 446:    496 estimated 1259.7921    44
#> 447:    497 estimated 1257.4302    45
#> 448:    498 estimated  481.5329    86
#> 449:    499 estimated  467.6437    87
#> 450:    500 estimated  468.8258    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9892    46
#>   2:      2 estimated 1092.2870    46
#>   3:      3 estimated  295.8935    28
#>   4:      4 estimated  280.3948    23
#>  ---                                 
#> 446:    496 estimated 1259.7921    44
#> 447:    497 estimated 1257.4302    45
#> 448:    498 estimated  481.5329    86
#> 449:    499 estimated  467.6437    87
#> 450:    500 estimated  468.8258    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    46 3457.710
#>  2:    28 3597.566
#>  3:    23 3599.908
#>  4:    25 3598.565
#>  5:    56 3554.950
#>  6:    52 3581.786
#>  7:    18 3599.736
#>  8:    24 3548.488
#>  9:    14 3599.181
#> 10:    59 3534.320
#> 11:    19 3589.266
#> 12:    26 3539.320
#> 13:    15 3595.091
#> 14:    61 3521.548
#> 15:    58 3541.711
#> 16:    17 3597.034
#> 17:    27 3532.016
#> 18:    13 3599.516
#> 19:    64 3487.327
#> 20:    62 3514.179
#> 21:    40 3599.404
#> 22:    54 3566.609
#> 23:    66 3474.763
#> 24:    63 3504.385
#> 25:    47 3599.924
#> 26:    57 3548.896
#> 27:     8 3598.448
#> 28:    67 3599.697
#> 29:    88 3517.132
#> 30:    87 3526.277
#> 31:    68 3595.900
#> 32:    65 3479.876
#> 33:    89  899.377
#> 34:    69 3599.710
#> 35:     2 3599.280
#> 36:    36 3597.046
#> 37:    37 3598.848
#> 38:    75 3521.549
#> 39:    55 3559.886
#> 40:    51 3589.157
#> 41:    50 3598.957
#> 42:    53 3575.038
#> 43:    49 3599.005
#> 44:    76 3508.978
#> 45:    44 3599.250
#> 46:    48 3599.523
#> 47:    45 3594.526
#> 48:    77 3496.882
#> 49:    60 3528.149
#> 50:    41 3599.942
#> 51:    39 3598.753
#> 52:     7 3597.527
#> 53:    43 3599.958
#> 54:     6 3599.428
#> 55:    73 3587.005
#> 56:    70 3599.978
#> 57:    38 3599.742
#> 58:    71 3599.194
#> 59:    72 3597.478
#> 60:     4 3599.738
#> 61:     3 3599.667
#> 62:     5 3599.704
#> 63:    78 3481.318
#> 64:    11 3595.796
#> 65:    22 3556.923
#> 66:    35 3462.892
#> 67:    31 3499.108
#> 68:    20 3565.575
#> 69:    34 3465.679
#> 70:    10 3599.912
#> 71:    21 3560.619
#> 72:    33 3470.898
#> 73:    30 3502.803
#> 74:    16 3590.905
#> 75:     9 3599.523
#> 76:    32 3488.237
#> 77:    12 3599.460
#> 78:    29 3506.314
#> 79:    42 3591.956
#> 80:    74 3505.919
#> 81:    86 3599.600
#> 82:     1 3467.358
#> 83:    82 3575.486
#> 84:    83 3554.961
#> 85:    84 3537.863
#> 86:    79 3599.699
#> 87:    85 3520.828
#> 88:    81 3595.494
#> 89:    80 3598.013
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9892     5
#>   2:      2 estimated 1092.2870     6
#>   3:      3 estimated  295.8935     3
#>   4:      4 estimated  280.3948     9
#>  ---                                 
#> 446:    496 estimated 1259.7921    10
#> 447:    497 estimated 1257.4302     5
#> 448:    498 estimated  481.5329     8
#> 449:    499 estimated  467.6437     5
#> 450:    500 estimated  468.8258     8

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     5 31502.74
#>  2:     6 31505.94
#>  3:     3 31368.67
#>  4:     9 31369.20
#>  5:     2 31369.89
#>  6:    10 31499.98
#>  7:     1 31361.00
#>  8:     8 31497.24
#>  9:     7 31360.92
#> 10:     4 31501.37