Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry5af972151a32' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA     NA       NA          NA
#>   2:      2      2         NA         NA         NA     NA       NA          NA
#>   3:      3      3         NA         NA         NA     NA       NA          NA
#>   4:      4      4 1513694024 1513694024 1513694124     NA       NA           1
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA     NA       NA          NA
#> 497:    497    497 1513694025 1513694025 1513695525     NA       NA           1
#> 498:    498    498         NA         NA         NA     NA       NA          NA
#> 499:    499    499         NA         NA         NA     NA       NA          NA
#> 500:    500    500         NA         NA         NA     NA       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:            NA       NA                                  NA       NA     1
#>   2:            NA       NA                                  NA       NA     1
#>   3:            NA       NA                                  NA       NA     1
#>   4: cfInteractive       NA job18625181bbf2b281bf04299dce6faea8       NA     1
#>  ---                                                                          
#> 496:            NA       NA                                  NA       NA     1
#> 497: cfInteractive       NA job69d573dea06a724c0ff9d8bb19162c40       NA     1
#> 498:            NA       NA                                  NA       NA     1
#> 499:            NA       NA                                  NA       NA     1
#> 500:            NA       NA                                  NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0024 secs
#>  2:     16    iris      nrow     4      a 1100.0101 secs
#>  3:     34    iris      nrow     7      d  100.0023 secs
#>  4:     41    iris      nrow     9      a 1100.0022 secs
#>  5:     46    iris      nrow    10      a 1100.0022 secs
#>  6:     48    iris      nrow    10      c  100.0023 secs
#>  7:     69    iris      nrow    14      d  100.0022 secs
#>  8:     85    iris      nrow    17      e  100.0022 secs
#>  9:     89    iris      nrow    18      d  100.0023 secs
#> 10:     94    iris      nrow    19      d  100.0022 secs
#> 11:     97    iris      nrow    20      b 1100.0022 secs
#> 12:     98    iris      nrow    20      c  100.0021 secs
#> 13:    115    iris      nrow    23      e  100.0022 secs
#> 14:    117    iris      nrow    24      b 1100.0022 secs
#> 15:    120    iris      nrow    24      e  100.0022 secs
#> 16:    145    iris      nrow    29      e  100.0023 secs
#> 17:    148    iris      nrow    30      c  100.0030 secs
#> 18:    156    iris      nrow    32      a 1100.0022 secs
#> 19:    167    iris      nrow    34      b 1100.0023 secs
#> 20:    187    iris      nrow    38      b 1100.0030 secs
#> 21:    193    iris      nrow    39      c  100.0025 secs
#> 22:    220    iris      nrow    44      e  100.0031 secs
#> 23:    223    iris      nrow    45      c  100.0030 secs
#> 24:    227    iris      nrow    46      b 1100.0032 secs
#> 25:    246    iris      nrow    50      a 1100.0030 secs
#> 26:    258    iris      ncol     2      c  500.0026 secs
#> 27:    270    iris      ncol     4      e  500.0031 secs
#> 28:    271    iris      ncol     5      a 1500.0027 secs
#> 29:    275    iris      ncol     5      e  500.0030 secs
#> 30:    296    iris      ncol    10      a 1500.0030 secs
#> 31:    315    iris      ncol    13      e  500.0029 secs
#> 32:    325    iris      ncol    15      e  500.0029 secs
#> 33:    352    iris      ncol    21      b 1500.0028 secs
#> 34:    354    iris      ncol    21      d  500.0031 secs
#> 35:    358    iris      ncol    22      c  500.0025 secs
#> 36:    364    iris      ncol    23      d  500.0029 secs
#> 37:    370    iris      ncol    24      e  500.0024 secs
#> 38:    378    iris      ncol    26      c  500.0032 secs
#> 39:    388    iris      ncol    28      c  500.0026 secs
#> 40:    402    iris      ncol    31      b 1500.0035 secs
#> 41:    411    iris      ncol    33      a 1500.0029 secs
#> 42:    419    iris      ncol    34      d  500.0033 secs
#> 43:    432    iris      ncol    37      b 1500.0029 secs
#> 44:    434    iris      ncol    37      d  500.0033 secs
#> 45:    436    iris      ncol    38      a 1500.0032 secs
#> 46:    438    iris      ncol    38      c  500.0031 secs
#> 47:    447    iris      ncol    40      b 1500.0031 secs
#> 48:    451    iris      ncol    41      a 1500.0028 secs
#> 49:    454    iris      ncol    41      d  500.0034 secs
#> 50:    497    iris      ncol    50      b 1500.0033 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 16h 27m 1.7s
#>   Total    : 4d 02h 10m 21.8s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1038.8328
#>   2:      2    iris      nrow     1      b estimated 1040.6162
#>   3:      3    iris      nrow     1      c estimated  206.1627
#>   4:      4    iris      nrow     1      d  observed  100.0024
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1396.1770
#> 497:    497    iris      ncol    50      b  observed 1500.0033
#> 498:    498    iris      ncol    50      c estimated  597.7725
#> 499:    499    iris      ncol    50      d estimated  573.8424
#> 500:    500    iris      ncol    50      e estimated  568.2176

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 16h 27m 1.7s
#>   Parallel : 0d 08h 51m 40.7s
#>   Total    : 4d 02h 10m 21.8s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1427.8548
#>   2:    431 estimated 1426.0311
#>   3:    406 estimated 1425.3133
#>   4:    426 estimated 1423.8061
#>  ---                           
#> 446:     90 estimated  135.5625
#> 447:     95 estimated  135.5625
#> 448:     74 estimated  131.7729
#> 449:     79 estimated  130.9729
#> 450:     84 estimated  125.6406
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1038.8328    75
#>   2:      2 estimated 1040.6162    73
#>   3:      3 estimated  206.1627    66
#>   4:      5 estimated  218.8243    58
#>  ---                                 
#> 446:    495 estimated  568.2176    19
#> 447:    496 estimated 1396.1770    12
#> 448:    498 estimated  597.7725     4
#> 449:    499 estimated  573.8424    14
#> 450:    500 estimated  568.2176    20

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1038.8328    75
#>   2:      2 estimated 1040.6162    73
#>   3:      3 estimated  206.1627    66
#>   4:      5 estimated  218.8243    58
#>  ---                                 
#> 446:    495 estimated  568.2176    19
#> 447:    496 estimated 1396.1770    12
#> 448:    498 estimated  597.7725     4
#> 449:    499 estimated  573.8424    14
#> 450:    500 estimated  568.2176    20

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    75 3599.024
#>  2:    73 3476.825
#>  3:    66 3578.254
#>  4:    58 3599.583
#>  5:    69 3529.972
#>  6:    68 3541.137
#>  7:    63 3593.076
#>  8:    59 3596.391
#>  9:    55 3589.101
#> 10:    64 3589.341
#> 11:    62 3597.945
#> 12:    57 3598.882
#> 13:    44 3596.485
#> 14:    50 3472.457
#> 15:    10 3599.645
#> 16:    41 3474.886
#> 17:    54 3599.896
#> 18:    51 3598.253
#> 19:    42 3595.132
#> 20:    56 3598.770
#> 21:    61 3599.504
#> 22:    71 3518.943
#> 23:    60 3599.969
#> 24:    72 3505.582
#> 25:     8 3598.241
#> 26:    70 3523.931
#> 27:    77 3556.640
#> 28:     9 3598.958
#> 29:     7 3599.036
#> 30:    76 3571.077
#> 31:    78 3516.796
#> 32:    67 3555.532
#> 33:     2 3599.604
#> 34:    80 3598.556
#> 35:    82 3570.280
#> 36:    83 3548.374
#> 37:    90 1240.026
#> 38:    81 3594.314
#> 39:    74 3599.979
#> 40:    53 3599.851
#> 41:    52 3599.938
#> 42:    43 3599.835
#> 43:    79 3480.241
#> 44:    45 3595.607
#> 45:     4 3596.246
#> 46:     6 3596.077
#> 47:     3 3599.049
#> 48:     1 3597.732
#> 49:    65 3585.523
#> 50:     5 3598.197
#> 51:    49 3482.514
#> 52:    38 3481.405
#> 53:    48 3488.623
#> 54:    18 3596.032
#> 55:    37 3485.680
#> 56:    46 3504.981
#> 57:    28 3542.466
#> 58:    36 3497.543
#> 59:    13 3597.625
#> 60:    47 3493.423
#> 61:    32 3521.701
#> 62:    40 3478.424
#> 63:    31 3524.251
#> 64:    39 3479.166
#> 65:    16 3599.632
#> 66:    12 3599.701
#> 67:    17 3593.750
#> 68:    11 3598.522
#> 69:    34 3503.451
#> 70:    29 3531.680
#> 71:    15 3597.390
#> 72:    14 3597.686
#> 73:    30 3528.640
#> 74:    33 3505.519
#> 75:    35 3499.589
#> 76:    24 3597.614
#> 77:    20 3597.271
#> 78:    19 3597.773
#> 79:    25 3597.583
#> 80:    21 3597.772
#> 81:    23 3586.431
#> 82:    26 3593.986
#> 83:    22 3595.449
#> 84:    27 3572.345
#> 85:    84 3598.551
#> 86:    89 3506.857
#> 87:    87 3564.858
#> 88:    86 3578.182
#> 89:    85 3593.361
#> 90:    88 3545.563
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1038.8328     1
#>   2:      2 estimated 1040.6162     3
#>   3:      3 estimated  206.1627     1
#>   4:      5 estimated  218.8243     2
#>  ---                                 
#> 446:    495 estimated  568.2176     5
#> 447:    496 estimated 1396.1770     5
#> 448:    498 estimated  597.7725     3
#> 449:    499 estimated  573.8424     1
#> 450:    500 estimated  568.2176     3

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     1 31897.13
#>  2:     3 31787.58
#>  3:     2 31788.10
#>  4:     6 31897.01
#>  5:     8 31787.72
#>  6:     9 31788.19
#>  7:     7 31900.75
#>  8:    10 31892.09
#>  9:     4 31785.00
#> 10:     5 31898.10