Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

#> Created registry in '/tmp/RtmpP9I8YP/registry1f22324f4f9a' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0018 secs
#>  2:     30    iris      nrow     6      e  100.0058 secs
#>  3:     43    iris      nrow     9      c  100.0017 secs
#>  4:     46    iris      nrow    10      a 1100.0017 secs
#>  5:     53    iris      nrow    11      c  100.0017 secs
#>  6:     55    iris      nrow    11      e  100.0017 secs
#>  7:     65    iris      nrow    13      e  100.0017 secs
#>  8:     66    iris      nrow    14      a 1100.0017 secs
#>  9:     89    iris      nrow    18      d  100.0016 secs
#> 10:     91    iris      nrow    19      a 1100.0017 secs
#> 11:     93    iris      nrow    19      c  100.0017 secs
#> 12:    124    iris      nrow    25      d  100.0017 secs
#> 13:    136    iris      nrow    28      a 1100.0017 secs
#> 14:    147    iris      nrow    30      b 1100.0016 secs
#> 15:    157    iris      nrow    32      b 1100.0018 secs
#> 16:    160    iris      nrow    32      e  100.0017 secs
#> 17:    163    iris      nrow    33      c  100.0016 secs
#> 18:    178    iris      nrow    36      c  100.0018 secs
#> 19:    179    iris      nrow    36      d  100.0017 secs
#> 20:    211    iris      nrow    43      a 1100.0017 secs
#> 21:    222    iris      nrow    45      b 1100.0017 secs
#> 22:    232    iris      nrow    47      b 1100.0017 secs
#> 23:    235    iris      nrow    47      e  100.0016 secs
#> 24:    239    iris      nrow    48      d  100.0018 secs
#> 25:    244    iris      nrow    49      d  100.0017 secs
#> 26:    251    iris      ncol     1      a 1500.0017 secs
#> 27:    260    iris      ncol     2      e  500.0017 secs
#> 28:    278    iris      ncol     6      c  500.0017 secs
#> 29:    292    iris      ncol     9      b 1500.0016 secs
#> 30:    326    iris      ncol    16      a 1500.0017 secs
#> 31:    341    iris      ncol    19      a 1500.0017 secs
#> 32:    343    iris      ncol    19      c  500.0020 secs
#> 33:    346    iris      ncol    20      a 1500.0017 secs
#> 34:    365    iris      ncol    23      e  500.0017 secs
#> 35:    367    iris      ncol    24      b 1500.0017 secs
#> 36:    372    iris      ncol    25      b 1500.0018 secs
#> 37:    374    iris      ncol    25      d  500.0019 secs
#> 38:    379    iris      ncol    26      d  500.0017 secs
#> 39:    397    iris      ncol    30      b 1500.0017 secs
#> 40:    409    iris      ncol    32      d  500.0017 secs
#> 41:    410    iris      ncol    32      e  500.0017 secs
#> 42:    418    iris      ncol    34      c  500.0020 secs
#> 43:    423    iris      ncol    35      c  500.0017 secs
#> 44:    425    iris      ncol    35      e  500.0017 secs
#> 45:    439    iris      ncol    38      d  500.0017 secs
#> 46:    442    iris      ncol    39      b 1500.0020 secs
#> 47:    443    iris      ncol    39      c  500.0018 secs
#> 48:    450    iris      ncol    40      e  500.0017 secs
#> 49:    451    iris      ncol    41      a 1500.0016 secs
#> 50:    459    iris      ncol    42      d  500.0017 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 16m 48.2s
#>   Total    : 4d 01h 00m 8.3s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1089.1892
#>   2:      2    iris      nrow     1      b estimated 1092.2869
#>   3:      3    iris      nrow     1      c estimated  295.5734
#>   4:      4    iris      nrow     1      d estimated  279.9947
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1260.4719
#> 497:    497    iris      ncol    50      b estimated 1257.2568
#> 498:    498    iris      ncol    50      c estimated  480.7330
#> 499:    499    iris      ncol    50      d estimated  466.9104
#> 500:    500    iris      ncol    50      e estimated  467.9591

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 16m 48.2s
#>   Parallel : 0d 08h 44m 52.6s
#>   Total    : 4d 01h 00m 8.3s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    347 estimated 1437.4876
#>   2:    352 estimated 1437.4876
#>   3:    362 estimated 1434.7593
#>   4:    357 estimated 1434.2926
#>  ---                           
#> 446:    184 estimated  150.0704
#> 447:     64 estimated  146.1688
#> 448:     54 estimated  144.6585
#> 449:     60 estimated  143.5294
#> 450:     59 estimated  140.1252
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.1892    46
#>   2:      2 estimated 1092.2869    46
#>   3:      3 estimated  295.5734    27
#>   4:      4 estimated  279.9947    23
#>  ---                                 
#> 446:    496 estimated 1260.4719    44
#> 447:    497 estimated 1257.2568    45
#> 448:    498 estimated  480.7330    86
#> 449:    499 estimated  466.9104    87
#> 450:    500 estimated  467.9591    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.1892    46
#>   2:      2 estimated 1092.2869    46
#>   3:      3 estimated  295.5734    27
#>   4:      4 estimated  279.9947    23
#>  ---                                 
#> 446:    496 estimated 1260.4719    44
#> 447:    497 estimated 1257.2568    45
#> 448:    498 estimated  480.7330    86
#> 449:    499 estimated  466.9104    87
#> 450:    500 estimated  467.9591    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>     <int>     <num>
#>  1:    46 3598.5151
#>  2:    27 3597.3971
#>  3:    23 3597.3610
#>  4:    25 3594.8715
#>  5:    54 3561.3510
#>  6:    52 3575.7109
#>  7:    17 3599.2067
#>  8:    13 3598.4515
#>  9:    14 3595.2401
#> 10:    59 3533.3927
#> 11:    18 3595.5640
#> 12:    24 3545.2821
#> 13:    16 3589.7014
#> 14:    60 3525.6637
#> 15:    58 3537.1037
#> 16:    15 3598.4498
#> 17:    28 3523.8590
#> 18:    26 3535.9641
#> 19:    64 3484.9654
#> 20:    62 3508.1369
#> 21:    40 3599.9658
#> 22:    47 3599.7466
#> 23:    65 3477.1448
#> 24:    63 3498.1354
#> 25:    41 3598.7542
#> 26:    56 3552.0809
#> 27:    66 3599.3522
#> 28:    61 3518.7893
#> 29:    67 3598.8305
#> 30:     5 3599.1022
#> 31:    88 3510.8938
#> 32:    87 3553.1423
#> 33:    68 3598.8350
#> 34:    69 3599.6993
#> 35:     2 3598.9994
#> 36:    36 3597.1539
#> 37:    89  599.0802
#> 38:    37 3598.1193
#> 39:    39 3599.9947
#> 40:    75 3523.2632
#> 41:    43 3598.9441
#> 42:    51 3584.3727
#> 43:    50 3589.5141
#> 44:    73 3592.7971
#> 45:    57 3543.5673
#> 46:    53 3564.8109
#> 47:    49 3598.5528
#> 48:    77 3499.0944
#> 49:    48 3599.5612
#> 50:    44 3598.3692
#> 51:    45 3591.9389
#> 52:    78 3481.4106
#> 53:     4 3599.9618
#> 54:    76 3511.6215
#> 55:    55 3558.4482
#> 56:     6 3598.9540
#> 57:     7 3599.9103
#> 58:     3 3598.9846
#> 59:    70 3599.6905
#> 60:    71 3599.9354
#> 61:    86 3599.6512
#> 62:    72 3599.5298
#> 63:    38 3599.6326
#> 64:    10 3597.2472
#> 65:    22 3554.6264
#> 66:    35 3463.5211
#> 67:    31 3496.1201
#> 68:    20 3562.0011
#> 69:    33 3472.5582
#> 70:    30 3499.2691
#> 71:    21 3558.2151
#> 72:     9 3598.6363
#> 73:    29 3503.5728
#> 74:    19 3572.3237
#> 75:    34 3467.3527
#> 76:    32 3489.5746
#> 77:    12 3596.2731
#> 78:     8 3599.6689
#> 79:    11 3591.1601
#> 80:    42 3594.5420
#> 81:    74 3512.0546
#> 82:     1 3467.7575
#> 83:    82 3573.9948
#> 84:    83 3561.5626
#> 85:    84 3540.6387
#> 86:    79 3593.0428
#> 87:    85 3520.7156
#> 88:    80 3595.7278
#> 89:    81 3599.5480
#>     chunk   runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.1892     4
#>   2:      2 estimated 1092.2869    10
#>   3:      3 estimated  295.5734    10
#>   4:      4 estimated  279.9947     2
#>  ---                                 
#> 446:    496 estimated 1260.4719     5
#> 447:    497 estimated 1257.2568     4
#> 448:    498 estimated  480.7330     8
#> 449:    499 estimated  466.9104     1
#> 450:    500 estimated  467.9591     8

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     4 31489.26
#>  2:    10 31356.46
#>  3:     2 31357.10
#>  4:     5 31357.88
#>  5:     8 31486.72
#>  6:     7 31349.05
#>  7:     6 31347.82
#>  8:     9 31492.64
#>  9:     3 31487.83
#> 10:     1 31483.48