Estimate Remaining Runtimes
Source: R/estimateRuntimes.R

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate]
Object to print.

n
[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> No configuration file found

#> Created registry in '/var/folders/pj/dcslzytj3t30v_ylpyyfyyl00000gn/T/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 4 (seed = 5) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 16 (seed = 17) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 34 (seed = 35) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 41 (seed = 42) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 46 (seed = 47) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 48 (seed = 49) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 69 (seed = 70) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 85 (seed = 86) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 89 (seed = 90) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 94 (seed = 95) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 97 (seed = 98) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 98 (seed = 99) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 115 (seed = 116) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 117 (seed = 118) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 120 (seed = 121) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 145 (seed = 146) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 148 (seed = 149) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 156 (seed = 157) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 167 (seed = 168) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 187 (seed = 188) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 193 (seed = 194) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 220 (seed = 221) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 223 (seed = 224) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 227 (seed = 228) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 246 (seed = 247) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 258 (seed = 259) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 270 (seed = 271) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 271 (seed = 272) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 275 (seed = 276) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 296 (seed = 297) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 315 (seed = 316) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 325 (seed = 326) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 352 (seed = 353) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 354 (seed = 355) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 358 (seed = 359) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 364 (seed = 365) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 370 (seed = 371) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 378 (seed = 379) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 388 (seed = 389) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 402 (seed = 403) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 411 (seed = 412) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 419 (seed = 420) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 432 (seed = 433) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 434 (seed = 435) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 436 (seed = 437) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 438 (seed = 439) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 447 (seed = 448) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 451 (seed = 452) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 454 (seed = 455) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 497 (seed = 498) ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA     NA       NA          NA
#>   2:      2      2         NA         NA         NA     NA       NA          NA
#>   3:      3      3         NA         NA         NA     NA       NA          NA
#>   4:      4      4 1521642338 1521642338 1521642438     NA       NA           1
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA     NA       NA          NA
#> 497:    497    497 1521642340 1521642340 1521643840     NA       NA           1
#> 498:    498    498         NA         NA         NA     NA       NA          NA
#> 499:    499    499         NA         NA         NA     NA       NA          NA
#> 500:    500    500         NA         NA         NA     NA       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:            NA       NA                                  NA       NA     1
#>   2:            NA       NA                                  NA       NA     1
#>   3:            NA       NA                                  NA       NA     1
#>   4: cfInteractive       NA job6d0c221f37dee98263354ebec7c2d128       NA     1
#>  ---                                                                          
#> 496:            NA       NA                                  NA       NA     1
#> 497: cfInteractive       NA joba097ebb20ece21a66ab9b4b2dade603f       NA     1
#> 498:            NA       NA                                  NA       NA     1
#> 499:            NA       NA                                  NA       NA     1
#> 500:            NA       NA                                  NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0132 secs
#>  2:     16    iris      nrow     4      a 1100.0288 secs
#>  3:     34    iris      nrow     7      d  100.0128 secs
#>  4:     41    iris      nrow     9      a 1100.0080 secs
#>  5:     46    iris      nrow    10      a 1100.0084 secs
#>  6:     48    iris      nrow    10      c  100.0144 secs
#>  7:     69    iris      nrow    14      d  100.0083 secs
#>  8:     85    iris      nrow    17      e  100.0093 secs
#>  9:     89    iris      nrow    18      d  100.0084 secs
#> 10:     94    iris      nrow    19      d  100.0100 secs
#> 11:     97    iris      nrow    20      b 1100.0122 secs
#> 12:     98    iris      nrow    20      c  100.0081 secs
#> 13:    115    iris      nrow    23      e  100.0118 secs
#> 14:    117    iris      nrow    24      b 1100.0094 secs
#> 15:    120    iris      nrow    24      e  100.0114 secs
#> 16:    145    iris      nrow    29      e  100.0088 secs
#> 17:    148    iris      nrow    30      c  100.0103 secs
#> 18:    156    iris      nrow    32      a 1100.0152 secs
#> 19:    167    iris      nrow    34      b 1100.0096 secs
#> 20:    187    iris      nrow    38      b 1100.0094 secs
#> 21:    193    iris      nrow    39      c  100.0110 secs
#> 22:    220    iris      nrow    44      e  100.0082 secs
#> 23:    223    iris      nrow    45      c  100.0175 secs
#> 24:    227    iris      nrow    46      b 1100.0111 secs
#> 25:    246    iris      nrow    50      a 1100.0119 secs
#> 26:    258    iris      ncol     2      c  500.0101 secs
#> 27:    270    iris      ncol     4      e  500.0082 secs
#> 28:    271    iris      ncol     5      a 1500.0086 secs
#> 29:    275    iris      ncol     5      e  500.0085 secs
#> 30:    296    iris      ncol    10      a 1500.0083 secs
#> 31:    315    iris      ncol    13      e  500.0088 secs
#> 32:    325    iris      ncol    15      e  500.0119 secs
#> 33:    352    iris      ncol    21      b 1500.0099 secs
#> 34:    354    iris      ncol    21      d  500.0081 secs
#> 35:    358    iris      ncol    22      c  500.0091 secs
#> 36:    364    iris      ncol    23      d  500.0105 secs
#> 37:    370    iris      ncol    24      e  500.0117 secs
#> 38:    378    iris      ncol    26      c  500.0119 secs
#> 39:    388    iris      ncol    28      c  500.0084 secs
#> 40:    402    iris      ncol    31      b 1500.0081 secs
#> 41:    411    iris      ncol    33      a 1500.0082 secs
#> 42:    419    iris      ncol    34      d  500.0080 secs
#> 43:    432    iris      ncol    37      b 1500.0099 secs
#> 44:    434    iris      ncol    37      d  500.0104 secs
#> 45:    436    iris      ncol    38      a 1500.0112 secs
#> 46:    438    iris      ncol    38      c  500.0085 secs
#> 47:    447    iris      ncol    40      b 1500.0113 secs
#> 48:    451    iris      ncol    41      a 1500.0091 secs
#> 49:    454    iris      ncol    41      d  500.0107 secs
#> 50:    497    iris      ncol    50      b 1500.0119 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.5s
#>   Remaining: 3d 16h 23m 10.7s
#>   Total    : 4d 02h 06m 31.2s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1042.7630
#>   2:      2    iris      nrow     1      b estimated 1042.5288
#>   3:      3    iris      nrow     1      c estimated  190.2161
#>   4:      4    iris      nrow     1      d  observed  100.0132
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1398.7162
#> 497:    497    iris      ncol    50      b  observed 1500.0119
#> 498:    498    iris      ncol    50      c estimated  600.2321
#> 499:    499    iris      ncol    50      d estimated  577.6769
#> 500:    500    iris      ncol    50      e estimated  565.3525

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.5s
#>   Remaining: 3d 16h 23m 10.7s
#>   Parallel : 0d 08h 51m 28.0s
#>   Total    : 4d 02h 06m 31.2s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1433.0659
#>   2:    456 estimated 1425.0759
#>   3:    431 estimated 1420.8725
#>   4:    441 estimated 1420.6791
#>  ---                           
#> 446:     90 estimated  135.5278
#> 447:     99 estimated  131.9226
#> 448:     74 estimated  130.6965
#> 449:     79 estimated  129.0520
#> 450:     84 estimated  125.9453
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1042.7630    75
#>   2:      2 estimated 1042.5288    75
#>   3:      3 estimated  190.2161    71
#>   4:      5 estimated  194.7092    69
#>  ---                                 
#> 446:    495 estimated  565.3525    28
#> 447:    496 estimated 1398.7162    11
#> 448:    498 estimated  600.2321     5
#> 449:    499 estimated  577.6769    20
#> 450:    500 estimated  565.3525    29

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1042.7630    75
#>   2:      2 estimated 1042.5288    75
#>   3:      3 estimated  190.2161    71
#>   4:      5 estimated  194.7092    69
#>  ---                                 
#> 446:    495 estimated  565.3525    28
#> 447:    496 estimated 1398.7162    11
#> 448:    498 estimated  600.2321     5
#> 449:    499 estimated  577.6769    20
#> 450:    500 estimated  565.3525    29

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    75 3599.634
#>  2:    71 3526.596
#>  3:    69 3555.678
#>  4:    74 3598.490
#>  5:    70 3547.382
#>  6:    73 3481.064
#>  7:    44 3598.983
#>  8:    63 3598.639
#>  9:    64 3591.342
#> 10:    61 3599.632
#> 11:    54 3599.794
#> 12:    60 3598.456
#> 13:    11 3598.433
#> 14:    66 3575.416
#> 15:    12 3599.998
#> 16:    56 3599.357
#> 17:    58 3579.388
#> 18:    67 3572.117
#> 19:    48 3493.377
#> 20:    55 3599.610
#> 21:    65 3584.737
#> 22:    57 3583.530
#> 23:    72 3507.284
#> 24:     5 3599.728
#> 25:     9 3599.812
#> 26:    59 3599.546
#> 27:     8 3598.353
#> 28:     6 3599.869
#> 29:    42 3599.998
#> 30:    43 3595.108
#> 31:     2 3599.456
#> 32:     7 3595.887
#> 33:    76 3593.940
#> 34:    83 3559.835
#> 35:     4 3599.716
#> 36:    80 3487.133
#> 37:    81 3466.403
#> 38:    38 3599.211
#> 39:    82 3581.822
#> 40:    51 3596.527
#> 41:    41 3599.323
#> 42:    40 3599.856
#> 43:    37 3598.738
#> 44:    85 3511.181
#> 45:    52 3597.366
#> 46:     1 3598.435
#> 47:    79 3514.954
#> 48:    53 3599.106
#> 49:    68 3567.779
#> 50:     3 3599.286
#> 51:    39 3599.455
#> 52:    84 3547.831
#> 53:    78 3545.777
#> 54:    77 3560.818
#> 55:    36 3476.231
#> 56:    46 3520.416
#> 57:    50 3471.632
#> 58:    30 3503.415
#> 59:    16 3598.306
#> 60:    14 3599.257
#> 61:    49 3488.351
#> 62:    35 3479.562
#> 63:    45 3599.563
#> 64:    33 3486.125
#> 65:    34 3483.502
#> 66:    13 3598.647
#> 67:    47 3506.455
#> 68:    15 3599.217
#> 69:    62 3598.765
#> 70:    10 3599.863
#> 71:    17 3595.936
#> 72:    28 3520.109
#> 73:    31 3501.391
#> 74:    27 3521.868
#> 75:    29 3512.526
#> 76:    32 3495.982
#> 77:    20 3577.901
#> 78:    23 3561.676
#> 79:    26 3542.520
#> 80:    18 3588.465
#> 81:    21 3575.404
#> 82:    24 3551.578
#> 83:    19 3581.656
#> 84:    22 3564.861
#> 85:    25 3549.851
#> 86:    86 3584.037
#> 87:    90 1005.619
#> 88:    87 3565.369
#> 89:    88 3550.605
#> 90:    89 3532.826
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1042.7630     7
#>   2:      2 estimated 1042.5288    10
#>   3:      3 estimated  190.2161     9
#>   4:      5 estimated  194.7092     1
#>  ---                                 
#> 446:    495 estimated  565.3525     6
#> 447:    496 estimated 1398.7162    10
#> 448:    498 estimated  600.2321     4
#> 449:    499 estimated  577.6769     6
#> 450:    500 estimated  565.3525     9

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     7 31882.73
#>  2:    10 31884.67
#>  3:     9 31754.67
#>  4:     1 31879.63
#>  5:     8 31754.51
#>  6:     4 31887.96
#>  7:     2 31753.71
#>  8:     3 31884.28
#>  9:     6 31753.85
#> 10:     5 31754.66