Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry62e652230057' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash  repl
#>        <char>   <char> <int>
#>   1:       NA       NA     1
#>   2:       NA       NA     1
#>   3:       NA       NA     1
#>   4:       NA       NA     1
#>  ---                        
#> 496:       NA       NA     1
#> 497:       NA       NA     1
#> 498:       NA       NA     1
#> 499:       NA       NA     1
#> 500:       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0023 secs
#>  2:     30    iris      nrow     6      e  100.0085 secs
#>  3:     43    iris      nrow     9      c  100.0032 secs
#>  4:     46    iris      nrow    10      a 1100.0030 secs
#>  5:     53    iris      nrow    11      c  100.0026 secs
#>  6:     55    iris      nrow    11      e  100.0028 secs
#>  7:     65    iris      nrow    13      e  100.0033 secs
#>  8:     66    iris      nrow    14      a 1100.0028 secs
#>  9:     89    iris      nrow    18      d  100.0029 secs
#> 10:     91    iris      nrow    19      a 1100.0025 secs
#> 11:     93    iris      nrow    19      c  100.0035 secs
#> 12:    124    iris      nrow    25      d  100.0030 secs
#> 13:    136    iris      nrow    28      a 1100.0038 secs
#> 14:    147    iris      nrow    30      b 1100.0032 secs
#> 15:    157    iris      nrow    32      b 1100.0039 secs
#> 16:    160    iris      nrow    32      e  100.0033 secs
#> 17:    163    iris      nrow    33      c  100.0033 secs
#> 18:    178    iris      nrow    36      c  100.0035 secs
#> 19:    179    iris      nrow    36      d  100.0027 secs
#> 20:    211    iris      nrow    43      a 1100.0034 secs
#> 21:    222    iris      nrow    45      b 1100.0033 secs
#> 22:    232    iris      nrow    47      b 1100.0039 secs
#> 23:    235    iris      nrow    47      e  100.0037 secs
#> 24:    239    iris      nrow    48      d  100.0039 secs
#> 25:    244    iris      nrow    49      d  100.0032 secs
#> 26:    251    iris      ncol     1      a 1500.0036 secs
#> 27:    260    iris      ncol     2      e  500.0034 secs
#> 28:    278    iris      ncol     6      c  500.0033 secs
#> 29:    292    iris      ncol     9      b 1500.0041 secs
#> 30:    326    iris      ncol    16      a 1500.0035 secs
#> 31:    341    iris      ncol    19      a 1500.0033 secs
#> 32:    343    iris      ncol    19      c  500.0036 secs
#> 33:    346    iris      ncol    20      a 1500.0033 secs
#> 34:    365    iris      ncol    23      e  500.0035 secs
#> 35:    367    iris      ncol    24      b 1500.0038 secs
#> 36:    372    iris      ncol    25      b 1500.0033 secs
#> 37:    374    iris      ncol    25      d  500.0033 secs
#> 38:    379    iris      ncol    26      d  500.0035 secs
#> 39:    397    iris      ncol    30      b 1500.0028 secs
#> 40:    409    iris      ncol    32      d  500.0037 secs
#> 41:    410    iris      ncol    32      e  500.0036 secs
#> 42:    418    iris      ncol    34      c  500.0030 secs
#> 43:    423    iris      ncol    35      c  500.0034 secs
#> 44:    425    iris      ncol    35      e  500.0035 secs
#> 45:    439    iris      ncol    38      d  500.0035 secs
#> 46:    442    iris      ncol    39      b 1500.0041 secs
#> 47:    443    iris      ncol    39      c  500.0032 secs
#> 48:    450    iris      ncol    40      e  500.0038 secs
#> 49:    451    iris      ncol    41      a 1500.0032 secs
#> 50:    459    iris      ncol    42      d  500.0041 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 15h 16m 19.7s
#>   Total    : 4d 00h 59m 39.9s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1090.3905
#>   2:      2    iris      nrow     1      b estimated 1092.8482
#>   3:      3    iris      nrow     1      c estimated  294.7750
#>   4:      4    iris      nrow     1      d estimated  279.9963
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1261.2737
#> 497:    497    iris      ncol    50      b estimated 1259.1252
#> 498:    498    iris      ncol    50      c estimated  481.9347
#> 499:    499    iris      ncol    50      d estimated  468.5123
#> 500:    500    iris      ncol    50      e estimated  469.5611

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 15h 16m 19.7s
#>   Parallel : 0d 08h 44m 50.2s
#>   Total    : 4d 00h 59m 39.9s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    347 estimated 1437.3605
#>   2:    352 estimated 1437.3605
#>   3:    362 estimated 1434.8989
#>   4:    357 estimated 1434.4322
#>  ---                           
#> 446:    184 estimated  149.3252
#> 447:     64 estimated  145.8502
#> 448:     54 estimated  144.3399
#> 449:     60 estimated  143.5309
#> 450:     59 estimated  139.8066
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3905    46
#>   2:      2 estimated 1092.8482    46
#>   3:      3 estimated  294.7750    27
#>   4:      4 estimated  279.9963    22
#>  ---                                 
#> 446:    496 estimated 1261.2737    44
#> 447:    497 estimated 1259.1252    45
#> 448:    498 estimated  481.9347    86
#> 449:    499 estimated  468.5123    87
#> 450:    500 estimated  469.5611    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3905    46
#>   2:      2 estimated 1092.8482    46
#>   3:      3 estimated  294.7750    27
#>   4:      4 estimated  279.9963    22
#>  ---                                 
#> 446:    496 estimated 1261.2737    44
#> 447:    497 estimated 1259.1252    45
#> 448:    498 estimated  481.9347    86
#> 449:    499 estimated  468.5123    87
#> 450:    500 estimated  469.5611    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>     <int>     <num>
#>  1:    46 3599.2328
#>  2:    27 3593.4574
#>  3:    22 3599.3331
#>  4:    25 3594.4494
#>  5:    52 3571.5298
#>  6:    51 3585.5359
#>  7:    16 3599.0866
#>  8:    23 3547.2571
#>  9:    13 3598.9816
#> 10:    58 3537.1411
#> 11:    17 3597.7464
#> 12:    24 3545.5876
#> 13:    14 3595.1848
#> 14:    60 3524.6409
#> 15:    15 3596.5435
#> 16:    28 3521.3986
#> 17:    26 3535.7043
#> 18:    64 3485.9042
#> 19:    62 3513.4606
#> 20:    41 3599.4367
#> 21:    55 3557.2928
#> 22:    53 3563.8812
#> 23:    65 3478.9043
#> 24:    63 3500.6728
#> 25:     7 3599.0355
#> 26:    56 3551.7630
#> 27:    66 3473.1697
#> 28:    61 3517.8383
#> 29:    48 3599.3250
#> 30:    67 3596.7067
#> 31:    88 3506.7256
#> 32:    87 3528.2838
#> 33:    59 3532.2201
#> 34:    68 3596.6111
#> 35:    89  756.4566
#> 36:    36 3597.2318
#> 37:    37 3599.0557
#> 38:     2 3599.7579
#> 39:    50 3590.1906
#> 40:    49 3597.6894
#> 41:    43 3599.1331
#> 42:    54 3562.0891
#> 43:    77 3495.8279
#> 44:    45 3595.5484
#> 45:    57 3544.6045
#> 46:    47 3599.8675
#> 47:    44 3599.8453
#> 48:    78 3478.1622
#> 49:    73 3590.5278
#> 50:    40 3597.5775
#> 51:     3 3599.7475
#> 52:    76 3508.8096
#> 53:    75 3521.8978
#> 54:     5 3599.7072
#> 55:     6 3599.9608
#> 56:    39 3599.3068
#> 57:    69 3599.4841
#> 58:    70 3598.7686
#> 59:    71 3599.6841
#> 60:    72 3599.9577
#> 61:    86 3598.5098
#> 62:     4 3599.7902
#> 63:    11 3594.1493
#> 64:     8 3599.8776
#> 65:    20 3560.6566
#> 66:    35 3465.7659
#> 67:    31 3496.1678
#> 68:    19 3571.3215
#> 69:    34 3469.5853
#> 70:    30 3498.4430
#> 71:    21 3556.1970
#> 72:     9 3598.3857
#> 73:    29 3500.7084
#> 74:    18 3583.5274
#> 75:    33 3473.6857
#> 76:    32 3487.8816
#> 77:    12 3597.7197
#> 78:    38 3599.8980
#> 79:    10 3597.5876
#> 80:    42 3596.7140
#> 81:    74 3510.7331
#> 82:     1 3466.7049
#> 83:    81 3599.9559
#> 84:    83 3557.3393
#> 85:    82 3571.6814
#> 86:    84 3538.1496
#> 87:    85 3520.3145
#> 88:    80 3597.3409
#> 89:    79 3595.9922
#>     chunk   runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1090.3905     3
#>   2:      2 estimated 1092.8482     8
#>   3:      3 estimated  294.7750     2
#>   4:      4 estimated  279.9963     8
#>  ---                                 
#> 446:    496 estimated 1261.2737     2
#> 447:    497 estimated 1259.1252     3
#> 448:    498 estimated  481.9347     3
#> 449:    499 estimated  468.5123     9
#> 450:    500 estimated  469.5611     3

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 31485.85
#>  2:     8 31353.17
#>  3:     2 31354.26
#>  4:     1 31354.14
#>  5:     5 31490.23
#>  6:     4 31486.02
#>  7:    10 31343.06
#>  8:     6 31343.43
#>  9:     7 31482.38
#> 10:     9 31487.18