Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/michel/.config/batchtools/config.R' ...

#> Created registry in '/tmp/RtmpErAdYG/registry17a37bcd7e24' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = flatten(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0017 secs
#>  2:     30    iris      nrow     6      e  100.0059 secs
#>  3:     43    iris      nrow     9      c  100.0018 secs
#>  4:     46    iris      nrow    10      a 1100.0018 secs
#>  5:     53    iris      nrow    11      c  100.0017 secs
#>  6:     55    iris      nrow    11      e  100.0017 secs
#>  7:     65    iris      nrow    13      e  100.0017 secs
#>  8:     66    iris      nrow    14      a 1100.0017 secs
#>  9:     89    iris      nrow    18      d  100.0017 secs
#> 10:     91    iris      nrow    19      a 1100.0017 secs
#> 11:     93    iris      nrow    19      c  100.0017 secs
#> 12:    124    iris      nrow    25      d  100.0018 secs
#> 13:    136    iris      nrow    28      a 1100.0017 secs
#> 14:    147    iris      nrow    30      b 1100.0017 secs
#> 15:    157    iris      nrow    32      b 1100.0017 secs
#> 16:    160    iris      nrow    32      e  100.0017 secs
#> 17:    163    iris      nrow    33      c  100.0018 secs
#> 18:    178    iris      nrow    36      c  100.0017 secs
#> 19:    179    iris      nrow    36      d  100.0017 secs
#> 20:    211    iris      nrow    43      a 1100.0017 secs
#> 21:    222    iris      nrow    45      b 1100.0017 secs
#> 22:    232    iris      nrow    47      b 1100.0018 secs
#> 23:    235    iris      nrow    47      e  100.0018 secs
#> 24:    239    iris      nrow    48      d  100.0017 secs
#> 25:    244    iris      nrow    49      d  100.0017 secs
#> 26:    251    iris      ncol     1      a 1500.0017 secs
#> 27:    260    iris      ncol     2      e  500.0018 secs
#> 28:    278    iris      ncol     6      c  500.0019 secs
#> 29:    292    iris      ncol     9      b 1500.0021 secs
#> 30:    326    iris      ncol    16      a 1500.0017 secs
#> 31:    341    iris      ncol    19      a 1500.0017 secs
#> 32:    343    iris      ncol    19      c  500.0018 secs
#> 33:    346    iris      ncol    20      a 1500.0018 secs
#> 34:    365    iris      ncol    23      e  500.0017 secs
#> 35:    367    iris      ncol    24      b 1500.0017 secs
#> 36:    372    iris      ncol    25      b 1500.0018 secs
#> 37:    374    iris      ncol    25      d  500.0018 secs
#> 38:    379    iris      ncol    26      d  500.0017 secs
#> 39:    397    iris      ncol    30      b 1500.0018 secs
#> 40:    409    iris      ncol    32      d  500.0020 secs
#> 41:    410    iris      ncol    32      e  500.0018 secs
#> 42:    418    iris      ncol    34      c  500.0017 secs
#> 43:    423    iris      ncol    35      c  500.0018 secs
#> 44:    425    iris      ncol    35      e  500.0020 secs
#> 45:    439    iris      ncol    38      d  500.0017 secs
#> 46:    442    iris      ncol    39      b 1500.0020 secs
#> 47:    443    iris      ncol    39      c  500.0017 secs
#> 48:    450    iris      ncol    40      e  500.0019 secs
#> 49:    451    iris      ncol    41      a 1500.0018 secs
#> 50:    459    iris      ncol    42      d  500.0018 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 14h 57m 18.6s
#>   Total    : 4d 00h 40m 38.7s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1084.3783
#>   2:      2    iris      nrow     1      b estimated 1088.0360
#>   3:      3    iris      nrow     1      c estimated  280.6025
#>   4:      4    iris      nrow     1      d estimated  264.3304
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1238.2161
#> 497:    497    iris      ncol    50      b estimated 1236.5476
#> 498:    498    iris      ncol    50      c estimated  481.0237
#> 499:    499    iris      ncol    50      d estimated  466.5345
#> 500:    500    iris      ncol    50      e estimated  468.1833

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 14h 57m 18.6s
#>   Parallel : 0d 08h 42m 54.3s
#>   Total    : 4d 00h 40m 38.7s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    352 estimated 1439.2401
#>   2:    347 estimated 1439.2401
#>   3:    342 estimated 1437.7801
#>   4:    332 estimated 1436.6801
#>  ---                           
#> 446:     64 estimated  141.1123
#> 447:    184 estimated  140.4490
#> 448:     60 estimated  140.0194
#> 449:     54 estimated  139.6552
#> 450:     59 estimated  137.1218
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1084.3783    44
#>   2:      2 estimated 1088.0360    43
#>   3:      3 estimated  280.6025    20
#>   4:      4 estimated  264.3304    21
#>  ---                                 
#> 446:    496 estimated 1238.2161    44
#> 447:    497 estimated 1236.5476    45
#> 448:    498 estimated  481.0237    87
#> 449:    499 estimated  466.5345    87
#> 450:    500 estimated  468.1833    82

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1084.3783    44
#>   2:      2 estimated 1088.0360    43
#>   3:      3 estimated  280.6025    20
#>   4:      4 estimated  264.3304    21
#>  ---                                 
#> 446:    496 estimated 1238.2161    44
#> 447:    497 estimated 1236.5476    45
#> 448:    498 estimated  481.0237    87
#> 449:    499 estimated  466.5345    87
#> 450:    500 estimated  468.1833    82

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    44 3560.810
#>  2:    43 3582.400
#>  3:    20 3598.674
#>  4:    21 3575.820
#>  5:    19 3597.709
#>  6:    59 3520.006
#>  7:    56 3536.511
#>  8:    13 3598.056
#>  9:    23 3512.339
#> 10:    18 3558.067
#> 11:    61 3502.250
#> 12:    14 3587.949
#> 13:    24 3507.649
#> 14:    22 3527.092
#> 15:    62 3495.053
#> 16:    12 3598.305
#> 17:    26 3494.474
#> 18:    25 3503.323
#> 19:    67 3599.566
#> 20:    65 3473.678
#> 21:    32 3599.908
#> 22:    50 3565.141
#> 23:    49 3567.953
#> 24:    68 3598.779
#> 25:    34 3599.747
#> 26:     4 3599.721
#> 27:    72 3595.119
#> 28:    39 3560.927
#> 29:    58 3526.911
#> 30:    57 3533.779
#> 31:    71 3595.382
#> 32:    64 3478.772
#> 33:    88 3463.146
#> 34:    53 3547.246
#> 35:    79 3470.129
#> 36:    66 3466.952
#> 37:    31 3599.839
#> 38:    80 3599.194
#> 39:    60 3512.956
#> 40:    73 3599.986
#> 41:    87 3595.868
#> 42:    51 3551.166
#> 43:    47 3587.782
#> 44:    81 3599.990
#> 45:    52 3548.824
#> 46:    76 3512.505
#> 47:    48 3576.280
#> 48:     3 3599.923
#> 49:    75 3538.565
#> 50:    33 3599.843
#> 51:    55 3538.751
#> 52:    54 3542.386
#> 53:    63 3484.963
#> 54:    37 3599.386
#> 55:    35 3598.434
#> 56:    40 3550.578
#> 57:    36 3598.583
#> 58:    46 3599.349
#> 59:    69 3597.630
#> 60:    70 3598.387
#> 61:    78 3486.882
#> 62:     2 3599.220
#> 63:    77 3498.467
#> 64:     5 3599.744
#> 65:    27 3483.812
#> 66:     7 3595.803
#> 67:    45 3555.918
#> 68:    17 3565.942
#> 69:    29 3472.421
#> 70:    15 3575.673
#> 71:     8 3599.745
#> 72:    16 3572.226
#> 73:    28 3477.911
#> 74:    11 3599.766
#> 75:    30 3466.637
#> 76:    10 3597.867
#> 77:    38 3578.542
#> 78:     9 3598.560
#> 79:    42 3560.531
#> 80:     6 3598.732
#> 81:    41 3594.688
#> 82:    74 3598.705
#> 83:     1 3467.783
#> 84:    84 3593.183
#> 85:    85 3556.368
#> 86:    86 3532.801
#> 87:    82 3591.006
#> 88:    83 3589.182
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1084.3783     9
#>   2:      2 estimated 1088.0360     8
#>   3:      3 estimated  280.6025     7
#>   4:      4 estimated  264.3304    10
#>  ---                                 
#> 446:    496 estimated 1238.2161     5
#> 447:    497 estimated 1236.5476     7
#> 448:    498 estimated  481.0237     8
#> 449:    499 estimated  466.5345     8
#> 450:    500 estimated  468.1833     6

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     9 31373.60
#>  2:     8 31374.30
#>  3:     7 31234.73
#>  4:    10 31373.59
#>  5:     1 31234.47
#>  6:     2 31373.67
#>  7:     4 31234.37
#>  8:     3 31234.41
#>  9:     5 31371.18
#> 10:     6 31234.31