Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab

[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[any] Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate] Object to print.

n

[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

lpt binpack

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> Syncing 50 files ...

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started done error memory resource.id batch.id
#>   1:      1      1        NA      NA   NA    NA     NA          NA       NA
#>   2:      2      2        NA      NA   NA    NA     NA          NA       NA
#>   3:      3      3        NA      NA   NA    NA     NA          NA       NA
#>   4:      4      4        NA      NA   NA    NA     NA          NA       NA
#>   5:      5      5        NA      NA   NA    NA     NA          NA       NA
#>  ---                                                                       
#> 496:    496    496        NA      NA   NA    NA     NA          NA       NA
#> 497:    497    497        NA      NA   NA    NA     NA          NA       NA
#> 498:    498    498        NA      NA   NA    NA     NA          NA       NA
#> 499:    499    499        NA      NA   NA    NA     NA          NA       NA
#> 500:    500    500        NA      NA   NA    NA     NA          NA       NA
#>      job.hash repl
#>   1:       NA    1
#>   2:       NA    1
#>   3:       NA    1
#>   4:       NA    1
#>   5:       NA    1
#>  ---              
#> 496:       NA    1
#> 497:       NA    1
#> 498:       NA    1
#> 499:       NA    1
#> 500:       NA    1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm  x y time.running
#>  1:     12    iris      nrow  3 b    1100 secs
#>  2:     30    iris      nrow  6 e     100 secs
#>  3:     43    iris      nrow  9 c     100 secs
#>  4:     46    iris      nrow 10 a    1100 secs
#>  5:     53    iris      nrow 11 c     100 secs
#>  6:     55    iris      nrow 11 e     100 secs
#>  7:     65    iris      nrow 13 e     100 secs
#>  8:     66    iris      nrow 14 a    1100 secs
#>  9:     89    iris      nrow 18 d     100 secs
#> 10:     91    iris      nrow 19 a    1100 secs
#> 11:     93    iris      nrow 19 c     100 secs
#> 12:    124    iris      nrow 25 d     100 secs
#> 13:    136    iris      nrow 28 a    1100 secs
#> 14:    147    iris      nrow 30 b    1100 secs
#> 15:    157    iris      nrow 32 b    1100 secs
#> 16:    160    iris      nrow 32 e     100 secs
#> 17:    163    iris      nrow 33 c     100 secs
#> 18:    178    iris      nrow 36 c     100 secs
#> 19:    179    iris      nrow 36 d     100 secs
#> 20:    211    iris      nrow 43 a    1100 secs
#> 21:    222    iris      nrow 45 b    1100 secs
#> 22:    232    iris      nrow 47 b    1100 secs
#> 23:    235    iris      nrow 47 e     100 secs
#> 24:    239    iris      nrow 48 d     100 secs
#> 25:    244    iris      nrow 49 d     100 secs
#> 26:    251    iris      ncol  1 a    1500 secs
#> 27:    260    iris      ncol  2 e     500 secs
#> 28:    278    iris      ncol  6 c     500 secs
#> 29:    292    iris      ncol  9 b    1500 secs
#> 30:    326    iris      ncol 16 a    1500 secs
#> 31:    341    iris      ncol 19 a    1500 secs
#> 32:    343    iris      ncol 19 c     500 secs
#> 33:    346    iris      ncol 20 a    1500 secs
#> 34:    365    iris      ncol 23 e     500 secs
#> 35:    367    iris      ncol 24 b    1500 secs
#> 36:    372    iris      ncol 25 b    1500 secs
#> 37:    374    iris      ncol 25 d     500 secs
#> 38:    379    iris      ncol 26 d     500 secs
#> 39:    397    iris      ncol 30 b    1500 secs
#> 40:    409    iris      ncol 32 d     500 secs
#> 41:    410    iris      ncol 32 e     500 secs
#> 42:    418    iris      ncol 34 c     500 secs
#> 43:    423    iris      ncol 35 c     500 secs
#> 44:    425    iris      ncol 35 e     500 secs
#> 45:    439    iris      ncol 38 d     500 secs
#> 46:    442    iris      ncol 39 b    1500 secs
#> 47:    443    iris      ncol 39 c     500 secs
#> 48:    450    iris      ncol 40 e     500 secs
#> 49:    451    iris      ncol 41 a    1500 secs
#> 50:    459    iris      ncol 42 d     500 secs
#>     job.id problem algorithm  x y time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 14m 50.9s
#>   Total    : 3d 23h 58m 10.9s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm  x y      type   runtime
#>   1:      1    iris      nrow  1 a estimated 1080.3872
#>   2:      2    iris      nrow  1 b estimated 1070.9411
#>   3:      3    iris      nrow  1 c estimated  269.0999
#>   4:      4    iris      nrow  1 d estimated  266.0873
#>   5:      5    iris      nrow  1 e estimated  271.3281
#>  ---                                                  
#> 496:    496    iris      ncol 50 a estimated 1261.7432
#> 497:    497    iris      ncol 50 b estimated 1253.0704
#> 498:    498    iris      ncol 50 c estimated  457.2788
#> 499:    499    iris      ncol 50 d estimated  454.3188
#> 500:    500    iris      ncol 50 e estimated  451.5841

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.0s
#>   Remaining: 3d 14h 14m 50.9s
#>   Parallel : 0d 08h 38m 42.7s
#>   Total    : 3d 23h 58m 10.9s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>   1:    347 estimated 1454.5167
#>   2:    352 estimated 1448.0575
#>   3:    357 estimated 1448.0575
#>   4:    362 estimated 1448.0575
#>   5:    342 estimated 1445.4501
#>  ---                           
#> 446:     58 estimated  148.4124
#> 447:     63 estimated  148.4124
#> 448:     59 estimated  144.5199
#> 449:     64 estimated  144.5199
#> 450:     54 estimated  143.3465

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1080.3872    44
#>   2:      2 estimated 1070.9411    45
#>   3:      3 estimated  269.0999    23
#>   4:      4 estimated  266.0873    21
#>   5:      5 estimated  271.3281    22
#>  ---                                 
#> 446:    496 estimated 1261.7432    44
#> 447:    497 estimated 1253.0704    45
#> 448:    498 estimated  457.2788    84
#> 449:    499 estimated  454.3188    84
#> 450:    500 estimated  451.5841    85

print(ids)

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1080.3872    44
#>   2:      2 estimated 1070.9411    45
#>   3:      3 estimated  269.0999    23
#>   4:      4 estimated  266.0873    21
#>   5:      5 estimated  271.3281    22
#>  ---                                 
#> 446:    496 estimated 1261.7432    44
#> 447:    497 estimated 1253.0704    45
#> 448:    498 estimated  457.2788    84
#> 449:    499 estimated  454.3188    84
#> 450:    500 estimated  451.5841    85

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>  1:    44 3595.2008
#>  2:    45 3577.0818
#>  3:    23 3585.6985
#>  4:    21 3597.8974
#>  5:    22 3592.9727
#>  6:    57 3597.1451
#>  7:    60 3597.4774
#>  8:    11 3590.9735
#>  9:    12 3584.4269
#> 10:    10 3598.4518
#> 11:    13 3574.0377
#> 12:    16 3560.1903
#> 13:    14 3568.6160
#> 14:    67 3599.8291
#> 15:    69 3588.0863
#> 16:    56 3503.0277
#> 17:    85 3598.5772
#> 18:    84 3587.2656
#> 19:    70 3581.8430
#> 20:    73 3574.2120
#> 21:    83 3569.2681
#> 22:    63 3599.8691
#> 23:    68 3595.8994
#> 24:    86 3557.9900
#> 25:    87 3531.5799
#> 26:    64 3599.9429
#> 27:    32 3599.3305
#> 28:    71 3567.8356
#> 29:    88  143.3465
#> 30:    66 3599.7785
#> 31:    29 3597.2261
#> 32:    61 3598.2032
#> 33:    54 3522.7072
#> 34:    58 3589.7445
#> 35:    30 3599.9346
#> 36:    53 3524.8604
#> 37:    46 3580.9767
#> 38:    43 3594.5390
#> 39:    52 3532.9123
#> 40:    49 3560.8829
#> 41:    47 3578.2748
#> 42:    50 3554.9600
#> 43:    34 3599.7228
#> 44:    48 3566.1777
#> 45:     4 3599.4851
#> 46:    36 3599.5816
#> 47:     6 3599.6544
#> 48:    38 3549.1989
#> 49:     7 3599.4668
#> 50:     5 3599.4994
#> 51:    55 3513.9422
#> 52:    51 3541.5611
#> 53:    59 3585.8336
#> 54:    65 3599.9971
#> 55:    62 3598.4414
#> 56:    31 3599.9975
#> 57:    37 3570.8153
#> 58:    28 3458.9428
#> 59:    33 3599.9584
#> 60:    27 3482.1736
#> 61:    19 3550.6883
#> 62:    25 3496.3835
#> 63:    15 3563.8816
#> 64:    18 3555.1361
#> 65:    24 3513.7435
#> 66:     9 3599.6271
#> 67:    20 3546.3606
#> 68:    26 3492.0468
#> 69:    17 3558.6069
#> 70:     8 3598.0830
#> 71:    39 3544.6802
#> 72:    35 3599.2602
#> 73:    72 3549.0062
#> 74:    40 3597.7701
#> 75:    41 3579.8224
#> 76:    42 3558.4768
#> 77:     1 3477.6060
#> 78:     2 3469.7342
#> 79:     3 3461.3035
#> 80:    75 3599.4355
#> 81:    77 3581.9414
#> 82:    78 3572.2635
#> 83:    79 3560.7770
#> 84:    80 3548.2837
#> 85:    81 3527.6448
#> 86:    82 3481.2571
#> 87:    76 3590.5268
#> 88:    74 3599.0306
#>     chunk   runtime

## Not run: ------------------------------------
# submitJobs(ids, reg = tmp)
## ---------------------------------------------

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>   1:      1 estimated 1080.3872    10
#>   2:      2 estimated 1070.9411     1
#>   3:      3 estimated  269.0999     7
#>   4:      4 estimated  266.0873    10
#>   5:      5 estimated  271.3281     4
#>  ---                                 
#> 446:    496 estimated 1261.7432     5
#> 447:    497 estimated 1253.0704     9
#> 448:    498 estimated  457.2788     5
#> 449:    499 estimated  454.3188     8
#> 450:    500 estimated  451.5841     5

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>  1:    10 31119.81
#>  2:     1 30976.63
#>  3:     7 31120.96
#>  4:     4 30978.80
#>  5:     6 30976.89
#>  6:     2 31119.17
#>  7:     9 30976.62
#>  8:     3 31122.73
#>  9:     5 30976.72
#> 10:     8 31122.57