Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 4 (seed = 5) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 16 (seed = 17) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 34 (seed = 35) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 41 (seed = 42) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 46 (seed = 47) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 48 (seed = 49) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 69 (seed = 70) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 85 (seed = 86) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 89 (seed = 90) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 94 (seed = 95) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 97 (seed = 98) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 98 (seed = 99) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 115 (seed = 116) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 117 (seed = 118) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 120 (seed = 121) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 145 (seed = 146) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 148 (seed = 149) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 156 (seed = 157) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 167 (seed = 168) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 187 (seed = 188) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 193 (seed = 194) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 220 (seed = 221) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 223 (seed = 224) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 227 (seed = 228) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 246 (seed = 247) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 258 (seed = 259) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 270 (seed = 271) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 271 (seed = 272) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 275 (seed = 276) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 296 (seed = 297) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 315 (seed = 316) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 325 (seed = 326) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 352 (seed = 353) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 354 (seed = 355) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 358 (seed = 359) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 364 (seed = 365) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 370 (seed = 371) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 378 (seed = 379) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 388 (seed = 389) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 402 (seed = 403) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 411 (seed = 412) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 419 (seed = 420) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 432 (seed = 433) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 434 (seed = 435) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 436 (seed = 437) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 438 (seed = 439) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 447 (seed = 448) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 451 (seed = 452) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 454 (seed = 455) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 497 (seed = 498) ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA     NA       NA          NA
#>   2:      2      2         NA         NA         NA     NA       NA          NA
#>   3:      3      3         NA         NA         NA     NA       NA          NA
#>   4:      4      4 1517223274 1517223274 1517223374     NA       NA           1
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA     NA       NA          NA
#> 497:    497    497 1517223275 1517223275 1517224775     NA       NA           1
#> 498:    498    498         NA         NA         NA     NA       NA          NA
#> 499:    499    499         NA         NA         NA     NA       NA          NA
#> 500:    500    500         NA         NA         NA     NA       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:            NA       NA                                  NA       NA     1
#>   2:            NA       NA                                  NA       NA     1
#>   3:            NA       NA                                  NA       NA     1
#>   4: cfInteractive       NA jobe67784c016c09a5a110d540b7cf5e8ef       NA     1
#>  ---                                                                          
#> 496:            NA       NA                                  NA       NA     1
#> 497: cfInteractive       NA job267d741cbdad70926d3c65f6e19d9e91       NA     1
#> 498:            NA       NA                                  NA       NA     1
#> 499:            NA       NA                                  NA       NA     1
#> 500:            NA       NA                                  NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0048 secs
#>  2:     16    iris      nrow     4      a 1100.0106 secs
#>  3:     34    iris      nrow     7      d  100.0045 secs
#>  4:     41    iris      nrow     9      a 1100.0048 secs
#>  5:     46    iris      nrow    10      a 1100.0044 secs
#>  6:     48    iris      nrow    10      c  100.0044 secs
#>  7:     69    iris      nrow    14      d  100.0070 secs
#>  8:     85    iris      nrow    17      e  100.0044 secs
#>  9:     89    iris      nrow    18      d  100.0044 secs
#> 10:     94    iris      nrow    19      d  100.0044 secs
#> 11:     97    iris      nrow    20      b 1100.0055 secs
#> 12:     98    iris      nrow    20      c  100.0048 secs
#> 13:    115    iris      nrow    23      e  100.0053 secs
#> 14:    117    iris      nrow    24      b 1100.0048 secs
#> 15:    120    iris      nrow    24      e  100.0061 secs
#> 16:    145    iris      nrow    29      e  100.0049 secs
#> 17:    148    iris      nrow    30      c  100.0053 secs
#> 18:    156    iris      nrow    32      a 1100.0052 secs
#> 19:    167    iris      nrow    34      b 1100.0047 secs
#> 20:    187    iris      nrow    38      b 1100.0044 secs
#> 21:    193    iris      nrow    39      c  100.0059 secs
#> 22:    220    iris      nrow    44      e  100.0049 secs
#> 23:    223    iris      nrow    45      c  100.0048 secs
#> 24:    227    iris      nrow    46      b 1100.0065 secs
#> 25:    246    iris      nrow    50      a 1100.0058 secs
#> 26:    258    iris      ncol     2      c  500.0053 secs
#> 27:    270    iris      ncol     4      e  500.0053 secs
#> 28:    271    iris      ncol     5      a 1500.0065 secs
#> 29:    275    iris      ncol     5      e  500.0056 secs
#> 30:    296    iris      ncol    10      a 1500.0054 secs
#> 31:    315    iris      ncol    13      e  500.0051 secs
#> 32:    325    iris      ncol    15      e  500.0056 secs
#> 33:    352    iris      ncol    21      b 1500.0052 secs
#> 34:    354    iris      ncol    21      d  500.0061 secs
#> 35:    358    iris      ncol    22      c  500.0052 secs
#> 36:    364    iris      ncol    23      d  500.0062 secs
#> 37:    370    iris      ncol    24      e  500.0051 secs
#> 38:    378    iris      ncol    26      c  500.0048 secs
#> 39:    388    iris      ncol    28      c  500.0064 secs
#> 40:    402    iris      ncol    31      b 1500.0050 secs
#> 41:    411    iris      ncol    33      a 1500.0048 secs
#> 42:    419    iris      ncol    34      d  500.0061 secs
#> 43:    432    iris      ncol    37      b 1500.0051 secs
#> 44:    434    iris      ncol    37      d  500.0048 secs
#> 45:    436    iris      ncol    38      a 1500.0059 secs
#> 46:    438    iris      ncol    38      c  500.0075 secs
#> 47:    447    iris      ncol    40      b 1500.0046 secs
#> 48:    451    iris      ncol    41      a 1500.0053 secs
#> 49:    454    iris      ncol    41      d  500.0047 secs
#> 50:    497    iris      ncol    50      b 1500.0045 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 16h 24m 0.7s
#>   Total    : 4d 02h 07m 21.0s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1040.4347
#>   2:      2    iris      nrow     1      b estimated 1042.4183
#>   3:      3    iris      nrow     1      c estimated  205.8452
#>   4:      4    iris      nrow     1      d  observed  100.0048
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1395.5220
#> 497:    497    iris      ncol    50      b  observed 1500.0045
#> 498:    498    iris      ncol    50      c estimated  595.4513
#> 499:    499    iris      ncol    50      d estimated  570.3209
#> 500:    500    iris      ncol    50      e estimated  563.0961

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 16h 24m 0.7s
#>   Parallel : 0d 08h 51m 22.4s
#>   Total    : 4d 02h 07m 21.0s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1430.6666
#>   2:    431 estimated 1428.4230
#>   3:    406 estimated 1426.9385
#>   4:    456 estimated 1423.9466
#>  ---                           
#> 446:     90 estimated  137.5604
#> 447:     95 estimated  137.5604
#> 448:     74 estimated  133.1712
#> 449:     79 estimated  132.3710
#> 450:     84 estimated  127.0384
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.4347    71
#>   2:      2 estimated 1042.4183    70
#>   3:      3 estimated  205.8452    66
#>   4:      5 estimated  219.3068    58
#>  ---                                 
#> 446:    495 estimated  563.0961    20
#> 447:    496 estimated 1395.5220    12
#> 448:    498 estimated  595.4513     4
#> 449:    499 estimated  570.3209    15
#> 450:    500 estimated  563.0961    21

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.4347    71
#>   2:      2 estimated 1042.4183    70
#>   3:      3 estimated  205.8452    66
#>   4:      5 estimated  219.3068    58
#>  ---                                 
#> 446:    495 estimated  563.0961    20
#> 447:    496 estimated 1395.5220    12
#> 448:    498 estimated  595.4513     4
#> 449:    499 estimated  570.3209    15
#> 450:    500 estimated  563.0961    21

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>     <int>     <num>
#>  1:    71 3508.0912
#>  2:    70 3515.7668
#>  3:    66 3567.6190
#>  4:    58 3597.9586
#>  5:    67 3548.4963
#>  6:    63 3582.6485
#>  7:    68 3532.9486
#>  8:    55 3591.6104
#>  9:    59 3599.8052
#> 10:    57 3599.3832
#> 11:    65 3575.3639
#> 12:    44 3598.4777
#> 13:    54 3598.8282
#> 14:    51 3598.4533
#> 15:    10 3599.6305
#> 16:    11 3598.0222
#> 17:    53 3599.7059
#> 18:    52 3598.1638
#> 19:    64 3580.4533
#> 20:    41 3475.4828
#> 21:    56 3598.8397
#> 22:    61 3594.2572
#> 23:    69 3524.3898
#> 24:     8 3597.7859
#> 25:    73 3477.8906
#> 26:    78 3522.6317
#> 27:    60 3599.9888
#> 28:    77 3550.2357
#> 29:    62 3588.2110
#> 30:    76 3564.2691
#> 31:    80 3599.2793
#> 32:     7 3595.4760
#> 33:    82 3582.7404
#> 34:    90  982.6564
#> 35:     5 3597.0620
#> 36:    72 3496.7738
#> 37:    74 3598.1087
#> 38:     3 3599.4890
#> 39:    89 3576.1124
#> 40:    42 3599.7014
#> 41:    83 3556.3711
#> 42:    75 3597.8898
#> 43:     1 3598.6696
#> 44:    50 3595.6922
#> 45:     2 3599.8991
#> 46:     4 3598.2560
#> 47:    81 3598.2491
#> 48:    79 3487.0620
#> 49:    45 3597.3778
#> 50:     6 3597.4956
#> 51:    43 3599.5385
#> 52:    12 3599.8162
#> 53:    48 3486.0258
#> 54:    18 3599.3036
#> 55:    37 3488.6326
#> 56:    46 3504.5412
#> 57:    28 3546.0322
#> 58:    14 3596.0049
#> 59:    13 3594.7425
#> 60:    32 3523.6399
#> 61:    39 3482.3993
#> 62:    47 3489.5693
#> 63:    31 3525.6613
#> 64:    38 3483.3920
#> 65:    17 3597.5928
#> 66:    40 3481.5764
#> 67:    49 3478.5744
#> 68:    16 3597.8121
#> 69:    15 3594.4598
#> 70:     9 3599.2405
#> 71:    29 3534.5781
#> 72:    33 3508.6385
#> 73:    35 3501.5545
#> 74:    30 3532.4324
#> 75:    34 3506.5459
#> 76:    36 3500.9694
#> 77:    26 3588.9882
#> 78:    22 3588.5470
#> 79:    19 3598.6527
#> 80:    24 3599.9603
#> 81:    20 3598.7487
#> 82:    23 3584.4709
#> 83:    25 3595.8378
#> 84:    21 3595.7461
#> 85:    27 3570.1712
#> 86:    84 3599.5743
#> 87:    86 3570.8782
#> 88:    85 3578.7001
#> 89:    87 3552.1519
#> 90:    88 3525.2473
#>     chunk   runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.4347     3
#>   2:      2 estimated 1042.4183     5
#>   3:      3 estimated  205.8452     4
#>   4:      5 estimated  219.3068     3
#>  ---                                 
#> 446:    495 estimated  563.0961     7
#> 447:    496 estimated 1395.5220     2
#> 448:    498 estimated  595.4513     4
#> 449:    499 estimated  570.3209     9
#> 450:    500 estimated  563.0961     1

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 31769.28
#>  2:     5 31770.56
#>  3:     4 31878.09
#>  4:     9 31770.26
#>  5:     7 31770.83
#>  6:     6 31882.45
#>  7:    10 31767.98
#>  8:     2 31872.25
#>  9:     8 31881.50
#> 10:     1 31877.52