Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry38047b31e95a' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0024 secs
#>  2:     30    iris      nrow     6      e  100.0064 secs
#>  3:     43    iris      nrow     9      c  100.0021 secs
#>  4:     46    iris      nrow    10      a 1100.0020 secs
#>  5:     53    iris      nrow    11      c  100.0020 secs
#>  6:     55    iris      nrow    11      e  100.0021 secs
#>  7:     65    iris      nrow    13      e  100.0023 secs
#>  8:     66    iris      nrow    14      a 1100.0022 secs
#>  9:     89    iris      nrow    18      d  100.0023 secs
#> 10:     91    iris      nrow    19      a 1100.0024 secs
#> 11:     93    iris      nrow    19      c  100.0024 secs
#> 12:    124    iris      nrow    25      d  100.0023 secs
#> 13:    136    iris      nrow    28      a 1100.0021 secs
#> 14:    147    iris      nrow    30      b 1100.0020 secs
#> 15:    157    iris      nrow    32      b 1100.0023 secs
#> 16:    160    iris      nrow    32      e  100.0022 secs
#> 17:    163    iris      nrow    33      c  100.0023 secs
#> 18:    178    iris      nrow    36      c  100.0022 secs
#> 19:    179    iris      nrow    36      d  100.0024 secs
#> 20:    211    iris      nrow    43      a 1100.0021 secs
#> 21:    222    iris      nrow    45      b 1100.0018 secs
#> 22:    232    iris      nrow    47      b 1100.0018 secs
#> 23:    235    iris      nrow    47      e  100.0019 secs
#> 24:    239    iris      nrow    48      d  100.0020 secs
#> 25:    244    iris      nrow    49      d  100.0019 secs
#> 26:    251    iris      ncol     1      a 1500.0020 secs
#> 27:    260    iris      ncol     2      e  500.0021 secs
#> 28:    278    iris      ncol     6      c  500.0023 secs
#> 29:    292    iris      ncol     9      b 1500.0024 secs
#> 30:    326    iris      ncol    16      a 1500.0023 secs
#> 31:    341    iris      ncol    19      a 1500.0020 secs
#> 32:    343    iris      ncol    19      c  500.0022 secs
#> 33:    346    iris      ncol    20      a 1500.0022 secs
#> 34:    365    iris      ncol    23      e  500.0022 secs
#> 35:    367    iris      ncol    24      b 1500.0024 secs
#> 36:    372    iris      ncol    25      b 1500.0023 secs
#> 37:    374    iris      ncol    25      d  500.0024 secs
#> 38:    379    iris      ncol    26      d  500.0023 secs
#> 39:    397    iris      ncol    30      b 1500.0024 secs
#> 40:    409    iris      ncol    32      d  500.0023 secs
#> 41:    410    iris      ncol    32      e  500.0022 secs
#> 42:    418    iris      ncol    34      c  500.0022 secs
#> 43:    423    iris      ncol    35      c  500.0021 secs
#> 44:    425    iris      ncol    35      e  500.0025 secs
#> 45:    439    iris      ncol    38      d  500.0022 secs
#> 46:    442    iris      ncol    39      b 1500.0024 secs
#> 47:    443    iris      ncol    39      c  500.0022 secs
#> 48:    450    iris      ncol    40      e  500.0023 secs
#> 49:    451    iris      ncol    41      a 1500.0021 secs
#> 50:    459    iris      ncol    42      d  500.0022 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 13.6s
#>   Total    : 4d 01h 01m 33.7s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1087.9896
#>   2:      2    iris      nrow     1      b estimated 1092.2874
#>   3:      3    iris      nrow     1      c estimated  295.5739
#>   4:      4    iris      nrow     1      d estimated  280.8752
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1258.3923
#> 497:    497    iris      ncol    50      b estimated 1255.1239
#> 498:    498    iris      ncol    50      c estimated  480.7334
#> 499:    499    iris      ncol    50      d estimated  466.1642
#> 500:    500    iris      ncol    50      e estimated  468.0930

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 13.6s
#>   Parallel : 0d 08h 45m 0.3s
#>   Total    : 4d 01h 01m 33.7s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    347 estimated 1436.8882
#>   2:    352 estimated 1436.8882
#>   3:    362 estimated 1434.4266
#>   4:    357 estimated 1433.9599
#>  ---                           
#> 446:    184 estimated  150.5662
#> 447:     64 estimated  147.0493
#> 448:     54 estimated  145.5390
#> 449:     60 estimated  143.5299
#> 450:     59 estimated  141.0057
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9896    45
#>   2:      2 estimated 1092.2874    46
#>   3:      3 estimated  295.5739    29
#>   4:      4 estimated  280.8752    25
#>  ---                                 
#> 446:    496 estimated 1258.3923    44
#> 447:    497 estimated 1255.1239    45
#> 448:    498 estimated  480.7334    86
#> 449:    499 estimated  466.1642    87
#> 450:    500 estimated  468.0930    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9896    45
#>   2:      2 estimated 1092.2874    46
#>   3:      3 estimated  295.5739    29
#>   4:      4 estimated  280.8752    25
#>  ---                                 
#> 446:    496 estimated 1258.3923    44
#> 447:    497 estimated 1255.1239    45
#> 448:    498 estimated  480.7334    86
#> 449:    499 estimated  466.1642    87
#> 450:    500 estimated  468.0930    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk   runtime
#>     <int>     <num>
#>  1:    45 3598.2374
#>  2:    46 3599.8975
#>  3:    29 3590.3740
#>  4:    25 3592.8230
#>  5:    24 3598.7232
#>  6:    56 3551.7994
#>  7:    43 3599.8762
#>  8:    17 3599.4193
#>  9:    19 3574.9504
#> 10:    14 3596.7662
#> 11:    59 3531.4107
#> 12:    18 3597.1274
#> 13:    20 3563.4644
#> 14:    16 3589.0041
#> 15:    60 3522.8756
#> 16:    58 3536.9094
#> 17:    15 3599.0415
#> 18:    12 3597.5325
#> 19:    21 3558.9405
#> 20:    64 3483.7126
#> 21:    62 3507.7661
#> 22:    39 3599.2399
#> 23:    54 3561.2995
#> 24:    65 3477.5086
#> 25:    63 3496.1946
#> 26:    41 3595.5330
#> 27:    40 3599.4477
#> 28:    66 3599.2530
#> 29:    61 3518.1142
#> 30:    47 3599.5487
#> 31:    67 3598.9658
#> 32:    88 3509.9074
#> 33:    87 3560.1171
#> 34:    38 3599.8889
#> 35:    69 3597.7148
#> 36:    89  593.2558
#> 37:    68 3599.8451
#> 38:    37 3598.2585
#> 39:    35 3597.9774
#> 40:    36 3599.4484
#> 41:    76 3519.4516
#> 42:    75 3535.9716
#> 43:     2 3599.6912
#> 44:    51 3579.3298
#> 45:     3 3599.4393
#> 46:    55 3556.9373
#> 47:    50 3587.3274
#> 48:     4 3599.9551
#> 49:    49 3599.2661
#> 50:    77 3507.0941
#> 51:    48 3599.3592
#> 52:    78 3488.3884
#> 53:    57 3542.3784
#> 54:     5 3599.3284
#> 55:    44 3594.7705
#> 56:    52 3572.8903
#> 57:    73 3594.4807
#> 58:    53 3567.2230
#> 59:     6 3599.6452
#> 60:    70 3599.8098
#> 61:    71 3599.6558
#> 62:    72 3599.3266
#> 63:    10 3592.4739
#> 64:     7 3598.2744
#> 65:    26 3538.1632
#> 66:    31 3491.1537
#> 67:    22 3556.2980
#> 68:    33 3467.3216
#> 69:    30 3501.6119
#> 70:    23 3550.3983
#> 71:     9 3597.3603
#> 72:    27 3521.6758
#> 73:    13 3599.9023
#> 74:    34 3462.3205
#> 75:    32 3486.1309
#> 76:    11 3597.6657
#> 77:    28 3518.0052
#> 78:     8 3599.4329
#> 79:    42 3590.3499
#> 80:    74 3516.7927
#> 81:     1 3464.5591
#> 82:    81 3599.6370
#> 83:    82 3595.4912
#> 84:    84 3552.3305
#> 85:    79 3598.9762
#> 86:    85 3532.3762
#> 87:    80 3598.0414
#> 88:    83 3573.0596
#> 89:    86 3459.8820
#>     chunk   runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1087.9896     3
#>   2:      2 estimated 1092.2874     7
#>   3:      3 estimated  295.5739     6
#>   4:      4 estimated  280.8752     4
#>  ---                                 
#> 446:    496 estimated 1258.3923     9
#> 447:    497 estimated 1255.1239     3
#> 448:    498 estimated  480.7334     2
#> 449:    499 estimated  466.1642    10
#> 450:    500 estimated  468.0930     2

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 31359.71
#>  2:     7 31496.39
#>  3:     6 31365.58
#>  4:     4 31365.04
#>  5:     5 31363.53
#>  6:     8 31492.41
#>  7:     1 31500.27
#>  8:     9 31359.02
#>  9:    10 31494.51
#> 10:     2 31497.11