Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry7c585a9e6c55' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = getJobPars(reg = tmp)

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id submitted started  done  error memory resource.id batch.id
#>       <int>  <int>     <num>   <num> <num> <char>  <num>       <int>   <char>
#>   1:      1      1        NA      NA    NA     NA     NA          NA       NA
#>   2:      2      2        NA      NA    NA     NA     NA          NA       NA
#>   3:      3      3        NA      NA    NA     NA     NA          NA       NA
#>   4:      4      4        NA      NA    NA     NA     NA          NA       NA
#>  ---                                                                         
#> 496:    496    496        NA      NA    NA     NA     NA          NA       NA
#> 497:    497    497        NA      NA    NA     NA     NA          NA       NA
#> 498:    498    498        NA      NA    NA     NA     NA          NA       NA
#> 499:    499    499        NA      NA    NA     NA     NA          NA       NA
#> 500:    500    500        NA      NA    NA     NA     NA          NA       NA
#>      log.file job.hash job.name  repl
#>        <char>   <char>   <char> <int>
#>   1:       NA       NA       NA     1
#>   2:       NA       NA       NA     1
#>   3:       NA       NA       NA     1
#>   4:       NA       NA       NA     1
#>  ---                                 
#> 496:       NA       NA       NA     1
#> 497:       NA       NA       NA     1
#> 498:       NA       NA       NA     1
#> 499:       NA       NA       NA     1
#> 500:       NA       NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <fctr>    <fctr> <int> <char>     <difftime>
#>  1:     12    iris      nrow     3      b 1100.0019 secs
#>  2:     30    iris      nrow     6      e  100.0100 secs
#>  3:     43    iris      nrow     9      c  100.0021 secs
#>  4:     46    iris      nrow    10      a 1100.0021 secs
#>  5:     53    iris      nrow    11      c  100.0020 secs
#>  6:     55    iris      nrow    11      e  100.0019 secs
#>  7:     65    iris      nrow    13      e  100.0018 secs
#>  8:     66    iris      nrow    14      a 1100.0019 secs
#>  9:     89    iris      nrow    18      d  100.0019 secs
#> 10:     91    iris      nrow    19      a 1100.0022 secs
#> 11:     93    iris      nrow    19      c  100.0021 secs
#> 12:    124    iris      nrow    25      d  100.0020 secs
#> 13:    136    iris      nrow    28      a 1100.0020 secs
#> 14:    147    iris      nrow    30      b 1100.0020 secs
#> 15:    157    iris      nrow    32      b 1100.0019 secs
#> 16:    160    iris      nrow    32      e  100.0019 secs
#> 17:    163    iris      nrow    33      c  100.0018 secs
#> 18:    178    iris      nrow    36      c  100.0023 secs
#> 19:    179    iris      nrow    36      d  100.0022 secs
#> 20:    211    iris      nrow    43      a 1100.0020 secs
#> 21:    222    iris      nrow    45      b 1100.0021 secs
#> 22:    232    iris      nrow    47      b 1100.0025 secs
#> 23:    235    iris      nrow    47      e  100.0021 secs
#> 24:    239    iris      nrow    48      d  100.0021 secs
#> 25:    244    iris      nrow    49      d  100.0021 secs
#> 26:    251    iris      ncol     1      a 1500.0024 secs
#> 27:    260    iris      ncol     2      e  500.0022 secs
#> 28:    278    iris      ncol     6      c  500.0024 secs
#> 29:    292    iris      ncol     9      b 1500.0025 secs
#> 30:    326    iris      ncol    16      a 1500.0027 secs
#> 31:    341    iris      ncol    19      a 1500.0024 secs
#> 32:    343    iris      ncol    19      c  500.0024 secs
#> 33:    346    iris      ncol    20      a 1500.0024 secs
#> 34:    365    iris      ncol    23      e  500.0027 secs
#> 35:    367    iris      ncol    24      b 1500.0027 secs
#> 36:    372    iris      ncol    25      b 1500.0025 secs
#> 37:    374    iris      ncol    25      d  500.0026 secs
#> 38:    379    iris      ncol    26      d  500.0025 secs
#> 39:    397    iris      ncol    30      b 1500.0025 secs
#> 40:    409    iris      ncol    32      d  500.0024 secs
#> 41:    410    iris      ncol    32      e  500.0027 secs
#> 42:    418    iris      ncol    34      c  500.0025 secs
#> 43:    423    iris      ncol    35      c  500.0026 secs
#> 44:    425    iris      ncol    35      e  500.0029 secs
#> 45:    439    iris      ncol    38      d  500.0024 secs
#> 46:    442    iris      ncol    39      b 1500.0024 secs
#> 47:    443    iris      ncol    39      c  500.0025 secs
#> 48:    450    iris      ncol    40      e  500.0025 secs
#> 49:    451    iris      ncol    41      a 1500.0024 secs
#> 50:    459    iris      ncol    42      d  500.0026 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 5.0s
#>   Total    : 4d 01h 01m 25.1s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <fctr>    <fctr> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1089.0296
#>   2:      2    iris      nrow     1      b estimated 1092.4473
#>   3:      3    iris      nrow     1      c estimated  295.3459
#>   4:      4    iris      nrow     1      d estimated  279.9958
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1262.1926
#> 497:    497    iris      ncol    50      b estimated 1260.8974
#> 498:    498    iris      ncol    50      c estimated  482.5336
#> 499:    499    iris      ncol    50      d estimated  468.7111
#> 500:    500    iris      ncol    50      e estimated  469.7599

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.1s
#>   Remaining: 3d 15h 18m 5.0s
#>   Parallel : 0d 08h 45m 1.5s
#>   Total    : 4d 01h 01m 25.1s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    347 estimated 1436.9595
#>   2:    352 estimated 1436.9595
#>   3:    362 estimated 1434.4979
#>   4:    357 estimated 1434.0312
#>  ---                           
#> 446:    184 estimated  149.8042
#> 447:     64 estimated  145.8492
#> 448:     54 estimated  144.3390
#> 449:     60 estimated  142.2301
#> 450:     59 estimated  139.8056
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.0296    46
#>   2:      2 estimated 1092.4473    46
#>   3:      3 estimated  295.3459    27
#>   4:      4 estimated  279.9958    23
#>  ---                                 
#> 446:    496 estimated 1262.1926    44
#> 447:    497 estimated 1260.8974    45
#> 448:    498 estimated  482.5336    86
#> 449:    499 estimated  468.7111    87
#> 450:    500 estimated  469.7599    87

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.0296    46
#>   2:      2 estimated 1092.4473    46
#>   3:      3 estimated  295.3459    27
#>   4:      4 estimated  279.9958    23
#>  ---                                 
#> 446:    496 estimated 1262.1926    44
#> 447:    497 estimated 1260.8974    45
#> 448:    498 estimated  482.5336    86
#> 449:    499 estimated  468.7111    87
#> 450:    500 estimated  469.7599    87

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    46 3599.627
#>  2:    27 3597.372
#>  3:    23 3599.097
#>  4:    24 3598.121
#>  5:    55 3558.268
#>  6:    51 3584.678
#>  7:    16 3597.911
#>  8:    12 3599.795
#>  9:    14 3593.245
#> 10:    58 3538.552
#> 11:    17 3596.597
#> 12:    22 3553.100
#> 13:    18 3582.621
#> 14:    60 3525.636
#> 15:    15 3595.381
#> 16:    11 3599.859
#> 17:    25 3539.420
#> 18:    64 3487.956
#> 19:    62 3513.042
#> 20:    56 3553.497
#> 21:    65 3480.073
#> 22:    63 3502.121
#> 23:    47 3598.365
#> 24:    57 3543.388
#> 25:    66 3473.973
#> 26:    61 3521.218
#> 27:    48 3597.557
#> 28:     5 3599.600
#> 29:    88 3510.899
#> 30:    67 3596.156
#> 31:    59 3534.114
#> 32:    87 3536.545
#> 33:    68 3596.007
#> 34:    37 3594.570
#> 35:    89  900.714
#> 36:    36 3599.663
#> 37:    53 3569.449
#> 38:    75 3530.292
#> 39:    38 3599.629
#> 40:    50 3596.051
#> 41:    54 3561.561
#> 42:    49 3599.806
#> 43:     2 3598.923
#> 44:    77 3498.955
#> 45:    45 3599.492
#> 46:    78 3481.916
#> 47:    40 3598.916
#> 48:     3 3599.923
#> 49:    52 3576.352
#> 50:    44 3598.589
#> 51:    76 3515.829
#> 52:    41 3598.151
#> 53:     7 3599.957
#> 54:     8 3599.597
#> 55:     6 3599.034
#> 56:    39 3599.814
#> 57:    69 3599.219
#> 58:    70 3599.737
#> 59:    71 3599.490
#> 60:    72 3599.647
#> 61:     4 3599.330
#> 62:    73 3592.844
#> 63:    43 3599.998
#> 64:    10 3595.683
#> 65:    21 3557.981
#> 66:    35 3461.082
#> 67:    31 3492.224
#> 68:    19 3574.728
#> 69:    33 3468.581
#> 70:    30 3498.597
#> 71:    13 3599.361
#> 72:     9 3598.941
#> 73:    29 3505.331
#> 74:    20 3564.962
#> 75:    34 3464.507
#> 76:    32 3485.847
#> 77:    26 3532.201
#> 78:    28 3514.831
#> 79:    42 3599.324
#> 80:    74 3515.666
#> 81:     1 3462.073
#> 82:    82 3584.635
#> 83:    81 3599.860
#> 84:    84 3545.832
#> 85:    79 3599.553
#> 86:    85 3528.878
#> 87:    80 3591.323
#> 88:    83 3563.615
#> 89:    86 3468.126
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1089.0296     6
#>   2:      2 estimated 1092.4473     3
#>   3:      3 estimated  295.3459     1
#>   4:      4 estimated  279.9958     2
#>  ---                                 
#> 446:    496 estimated 1262.1926    10
#> 447:    497 estimated 1260.8974     8
#> 448:    498 estimated  482.5336     7
#> 449:    499 estimated  468.7111    10
#> 450:    500 estimated  469.7599     7

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     6 31362.97
#>  2:     3 31493.68
#>  3:     1 31362.96
#>  4:     2 31362.64
#>  5:     4 31355.57
#>  6:     7 31497.22
#>  7:     8 31354.62
#>  8:    10 31501.48
#>  9:     9 31495.23
#> 10:     5 31498.61