Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate]
Object to print.

n
[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 4 (seed = 5) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 16 (seed = 17) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 34 (seed = 35) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 41 (seed = 42) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 46 (seed = 47) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 48 (seed = 49) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 69 (seed = 70) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 85 (seed = 86) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 89 (seed = 90) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 94 (seed = 95) ...

#> 

#> Submitting [=========>-----------------------------------------]  20% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 97 (seed = 98) ...

#> 

#> Submitting [==========>----------------------------------------]  22% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 98 (seed = 99) ...

#> 

#> Submitting [===========>---------------------------------------]  24% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 115 (seed = 116) ...

#> 

#> Submitting [============>--------------------------------------]  26% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 117 (seed = 118) ...

#> 

#> Submitting [=============>-------------------------------------]  28% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 120 (seed = 121) ...

#> 

#> Submitting [==============>------------------------------------]  30% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 145 (seed = 146) ...

#> 

#> Submitting [===============>-----------------------------------]  32% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 148 (seed = 149) ...

#> 

#> Submitting [================>----------------------------------]  34% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 156 (seed = 157) ...

#> 

#> Submitting [=================>---------------------------------]  36% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 167 (seed = 168) ...

#> 

#> Submitting [==================>--------------------------------]  38% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 187 (seed = 188) ...

#> 

#> Submitting [===================>-------------------------------]  40% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 193 (seed = 194) ...

#> 

#> Submitting [====================>------------------------------]  42% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 220 (seed = 221) ...

#> 

#> Submitting [=====================>-----------------------------]  44% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 223 (seed = 224) ...

#> 

#> Submitting [======================>----------------------------]  46% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 227 (seed = 228) ...

#> 

#> Submitting [=======================>---------------------------]  48% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 246 (seed = 247) ...

#> 

#> Submitting [=========================>-------------------------]  50% eta:  1s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 258 (seed = 259) ...

#> 

#> Submitting [==========================>------------------------]  52% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 270 (seed = 271) ...

#> 

#> Submitting [===========================>-----------------------]  54% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 271 (seed = 272) ...

#> 

#> Submitting [============================>----------------------]  56% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 275 (seed = 276) ...

#> 

#> Submitting [=============================>---------------------]  58% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 296 (seed = 297) ...

#> 

#> Submitting [==============================>--------------------]  60% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 315 (seed = 316) ...

#> 

#> Submitting [===============================>-------------------]  62% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 325 (seed = 326) ...

#> 

#> Submitting [================================>------------------]  64% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 352 (seed = 353) ...

#> 

#> Submitting [=================================>-----------------]  66% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 354 (seed = 355) ...

#> 

#> Submitting [==================================>----------------]  68% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 358 (seed = 359) ...

#> 

#> Submitting [===================================>---------------]  70% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 364 (seed = 365) ...

#> 

#> Submitting [====================================>--------------]  72% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 370 (seed = 371) ...

#> 

#> Submitting [=====================================>-------------]  74% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 378 (seed = 379) ...

#> 

#> Submitting [======================================>------------]  76% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 388 (seed = 389) ...

#> 

#> Submitting [=======================================>-----------]  78% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 402 (seed = 403) ...

#> 

#> Submitting [========================================>----------]  80% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 411 (seed = 412) ...

#> 

#> Submitting [=========================================>---------]  82% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 419 (seed = 420) ...

#> 

#> Submitting [==========================================>--------]  84% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 432 (seed = 433) ...

#> 

#> Submitting [===========================================>-------]  86% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 434 (seed = 435) ...

#> 

#> Submitting [============================================>------]  88% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 436 (seed = 437) ...

#> 

#> Submitting [=============================================>-----]  90% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 438 (seed = 439) ...

#> 

#> Submitting [==============================================>----]  92% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 447 (seed = 448) ...

#> 

#> Submitting [===============================================>---]  94% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 451 (seed = 452) ...

#> 

#> Submitting [================================================>--]  96% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 454 (seed = 455) ...

#> 

#> Submitting [=================================================>-]  98% eta:  0s

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 497 (seed = 498) ...

#> 

#> Submitting [===================================================] 100% eta:  0s

#>                                                                               

#> 

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4 1526769090 1526769090 1526769190   <NA>       NA           1
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497 1526769091 1526769091 1526770591   <NA>       NA           1
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499         NA         NA         NA   <NA>       NA          NA
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4: cfInteractive     <NA> job5640f604acbbbdc1376c0b5c271e631b     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497: cfInteractive     <NA> jobebf9af537010298df61c74e074c6cff7     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499:          <NA>     <NA>                                <NA>     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0034 secs
#>  2:     16    iris      nrow     4      a 1100.0094 secs
#>  3:     34    iris      nrow     7      d  100.0034 secs
#>  4:     41    iris      nrow     9      a 1100.0034 secs
#>  5:     46    iris      nrow    10      a 1100.0034 secs
#>  6:     48    iris      nrow    10      c  100.0034 secs
#>  7:     69    iris      nrow    14      d  100.0034 secs
#>  8:     85    iris      nrow    17      e  100.0035 secs
#>  9:     89    iris      nrow    18      d  100.0034 secs
#> 10:     94    iris      nrow    19      d  100.0034 secs
#> 11:     97    iris      nrow    20      b 1100.0034 secs
#> 12:     98    iris      nrow    20      c  100.0034 secs
#> 13:    115    iris      nrow    23      e  100.0035 secs
#> 14:    117    iris      nrow    24      b 1100.0034 secs
#> 15:    120    iris      nrow    24      e  100.0034 secs
#> 16:    145    iris      nrow    29      e  100.0034 secs
#> 17:    148    iris      nrow    30      c  100.0034 secs
#> 18:    156    iris      nrow    32      a 1100.0035 secs
#> 19:    167    iris      nrow    34      b 1100.0034 secs
#> 20:    187    iris      nrow    38      b 1100.0033 secs
#> 21:    193    iris      nrow    39      c  100.0034 secs
#> 22:    220    iris      nrow    44      e  100.0034 secs
#> 23:    223    iris      nrow    45      c  100.0035 secs
#> 24:    227    iris      nrow    46      b 1100.0038 secs
#> 25:    246    iris      nrow    50      a 1100.0034 secs
#> 26:    258    iris      ncol     2      c  500.0033 secs
#> 27:    270    iris      ncol     4      e  500.0034 secs
#> 28:    271    iris      ncol     5      a 1500.0034 secs
#> 29:    275    iris      ncol     5      e  500.0035 secs
#> 30:    296    iris      ncol    10      a 1500.0034 secs
#> 31:    315    iris      ncol    13      e  500.0034 secs
#> 32:    325    iris      ncol    15      e  500.0035 secs
#> 33:    352    iris      ncol    21      b 1500.0067 secs
#> 34:    354    iris      ncol    21      d  500.0035 secs
#> 35:    358    iris      ncol    22      c  500.0035 secs
#> 36:    364    iris      ncol    23      d  500.0035 secs
#> 37:    370    iris      ncol    24      e  500.0034 secs
#> 38:    378    iris      ncol    26      c  500.0034 secs
#> 39:    388    iris      ncol    28      c  500.0034 secs
#> 40:    402    iris      ncol    31      b 1500.0040 secs
#> 41:    411    iris      ncol    33      a 1500.0034 secs
#> 42:    419    iris      ncol    34      d  500.0034 secs
#> 43:    432    iris      ncol    37      b 1500.0034 secs
#> 44:    434    iris      ncol    37      d  500.0034 secs
#> 45:    436    iris      ncol    38      a 1500.0035 secs
#> 46:    438    iris      ncol    38      c  500.0035 secs
#> 47:    447    iris      ncol    40      b 1500.0037 secs
#> 48:    451    iris      ncol    41      a 1500.0035 secs
#> 49:    454    iris      ncol    41      d  500.0035 secs
#> 50:    497    iris      ncol    50      b 1500.0034 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 16h 25m 19.7s
#>   Total    : 4d 02h 08m 39.8s

rjoin(tab, est$runtimes)

#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1040.8332
#>   2:      2    iris      nrow     1      b estimated 1042.6168
#>   3:      3    iris      nrow     1      c estimated  205.8437
#>   4:      4    iris      nrow     1      d  observed  100.0034
#>   5:      5    iris      nrow     1      e estimated  219.3053
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1395.9552
#> 497:    497    iris      ncol    50      b  observed 1500.0034
#> 498:    498    iris      ncol    50      c estimated  596.3063
#> 499:    499    iris      ncol    50      d estimated  571.5761
#> 500:    500    iris      ncol    50      e estimated  564.3513

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.2s
#>   Remaining: 3d 16h 25m 19.7s
#>   Parallel : 0d 08h 51m 30.2s
#>   Total    : 4d 02h 08m 39.8s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1430.2998
#>   2:    431 estimated 1427.1761
#>   3:    456 estimated 1425.5798
#>   4:    406 estimated 1424.9140
#>   5:    426 estimated 1422.6845
#>  ---                           
#> 446:     90 estimated  136.3303
#> 447:     95 estimated  136.3303
#> 448:     74 estimated  132.5407
#> 449:     79 estimated  131.7407
#> 450:     84 estimated  126.4084
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.8332    72
#>   2:      2 estimated 1042.6168    70
#>   3:      3 estimated  205.8437    65
#>   4:      5 estimated  219.3053    57
#>   5:      6 estimated 1047.4332    68
#>  ---                                 
#> 446:    495 estimated  564.3513    20
#> 447:    496 estimated 1395.9552    12
#> 448:    498 estimated  596.3063     4
#> 449:    499 estimated  571.5761    15
#> 450:    500 estimated  564.3513    21

print(ids)

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.8332    72
#>   2:      2 estimated 1042.6168    70
#>   3:      3 estimated  205.8437    65
#>   4:      5 estimated  219.3053    57
#>   5:      6 estimated 1047.4332    68
#>  ---                                 
#> 446:    495 estimated  564.3513    20
#> 447:    496 estimated 1395.9552    12
#> 448:    498 estimated  596.3063     4
#> 449:    499 estimated  571.5761    15
#> 450:    500 estimated  564.3513    21

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    72 3501.024
#>  2:    70 3517.799
#>  3:    65 3580.150
#>  4:    57 3599.801
#>  5:    68 3536.609
#>  6:    67 3547.271
#>  7:    64 3586.212
#>  8:    53 3599.887
#>  9:    63 3589.181
#> 10:    60 3599.647
#> 11:    66 3573.903
#> 12:    44 3598.942
#> 13:    55 3589.054
#> 14:    51 3597.819
#> 15:    11 3599.986
#> 16:    54 3595.094
#> 17:    52 3599.436
#> 18:    10 3599.877
#> 19:    41 3476.171
#> 20:    62 3592.930
#> 21:    59 3598.796
#> 22:    71 3508.057
#> 23:    69 3527.701
#> 24:     8 3598.028
#> 25:    73 3468.299
#> 26:     9 3599.672
#> 27:    77 3545.143
#> 28:     7 3599.268
#> 29:    76 3567.093
#> 30:    75 3599.634
#> 31:    80 3599.155
#> 32:     6 3598.621
#> 33:    90 1116.284
#> 34:    78 3505.705
#> 35:    82 3566.850
#> 36:    79 3480.599
#> 37:    83 3551.099
#> 38:    42 3596.398
#> 39:    81 3598.709
#> 40:    74 3599.656
#> 41:    50 3596.778
#> 42:     2 3598.812
#> 43:    58 3599.495
#> 44:    56 3599.143
#> 45:    61 3597.112
#> 46:     3 3599.882
#> 47:    45 3597.544
#> 48:     4 3599.651
#> 49:    43 3599.827
#> 50:     1 3599.604
#> 51:     5 3597.826
#> 52:    49 3478.057
#> 53:    38 3486.022
#> 54:    48 3485.043
#> 55:    30 3530.976
#> 56:    13 3597.656
#> 57:    46 3503.213
#> 58:    28 3544.402
#> 59:    36 3500.474
#> 60:    37 3487.621
#> 61:    47 3490.626
#> 62:    32 3522.485
#> 63:    12 3599.545
#> 64:    31 3525.592
#> 65:    39 3483.598
#> 66:    16 3598.463
#> 67:    40 3480.588
#> 68:    15 3599.893
#> 69:    34 3506.134
#> 70:    18 3598.136
#> 71:    17 3589.288
#> 72:    14 3597.971
#> 73:    29 3535.143
#> 74:    33 3508.845
#> 75:    35 3502.003
#> 76:    26 3589.319
#> 77:    20 3598.115
#> 78:    19 3598.418
#> 79:    24 3598.925
#> 80:    21 3596.375
#> 81:    23 3585.448
#> 82:    25 3593.676
#> 83:    22 3592.235
#> 84:    27 3569.856
#> 85:    84 3599.955
#> 86:    87 3558.943
#> 87:    86 3576.588
#> 88:    89 3496.612
#> 89:    88 3535.634
#> 90:    85 3586.550
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1040.8332     4
#>   2:      2 estimated 1042.6168     7
#>   3:      3 estimated  205.8437     5
#>   4:      5 estimated  219.3053     9
#>   5:      6 estimated 1047.4332     5
#>  ---                                 
#> 446:    495 estimated  564.3513     2
#> 447:    496 estimated 1395.9552     5
#> 448:    498 estimated  596.3063     5
#> 449:    499 estimated  571.5761     7
#> 450:    500 estimated  564.3513     4

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     4 31777.63
#>  2:     7 31886.27
#>  3:     5 31889.80
#>  4:     9 31777.60
#>  5:     2 31890.21
#>  6:     8 31777.55
#>  7:     3 31881.39
#>  8:     6 31775.06
#>  9:    10 31886.50
#> 10:     1 31777.67