Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for RuntimeEstimate
print(x, n = 1L, ...)

Arguments

tab
[data.table] Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...
[ANY] Additional parameters passed to ranger. Ignored for the print method.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x
[RuntimeEstimate] Object to print.

n
[integer(1)] Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

binpack and lpt to chunk jobs according to their estimated runtimes.

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)

#> Adding problem 'iris'

addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)

#> Adding algorithm 'nrow'

addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)

#> Adding algorithm 'ncol'

addExperiments(algo.designs = list(nrow = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...

addExperiments(algo.designs = list(ncol = CJ(x = 1:50, y = letters[1:5])), reg = tmp)

#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)

#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 4 (seed = 5) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 16 (seed = 17) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 34 (seed = 35) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 41 (seed = 42) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 46 (seed = 47) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 48 (seed = 49) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 69 (seed = 70) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 85 (seed = 86) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 89 (seed = 90) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 94 (seed = 95) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 97 (seed = 98) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 98 (seed = 99) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 115 (seed = 116) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 117 (seed = 118) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 120 (seed = 121) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 145 (seed = 146) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 148 (seed = 149) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 156 (seed = 157) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 167 (seed = 168) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 187 (seed = 188) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 193 (seed = 194) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 220 (seed = 221) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 223 (seed = 224) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 227 (seed = 228) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'nrow' on problem 'iris' for job 246 (seed = 247) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 258 (seed = 259) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 270 (seed = 271) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 271 (seed = 272) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 275 (seed = 276) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 296 (seed = 297) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 315 (seed = 316) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 325 (seed = 326) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 352 (seed = 353) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 354 (seed = 355) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 358 (seed = 359) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 364 (seed = 365) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 370 (seed = 371) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 378 (seed = 379) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 388 (seed = 389) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 402 (seed = 403) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 411 (seed = 412) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 419 (seed = 420) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 432 (seed = 433) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 434 (seed = 435) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 436 (seed = 437) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 438 (seed = 439) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 447 (seed = 448) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 451 (seed = 452) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 454 (seed = 455) ...

#> ### [bt]: Generating problem instance for problem 'iris' ...

#> ### [bt]: Applying algorithm 'ncol' on problem 'iris' for job 497 (seed = 498) ...

waitForJobs(reg = tmp)

#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]

#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA     NA       NA          NA
#>   2:      2      2         NA         NA         NA     NA       NA          NA
#>   3:      3      3         NA         NA         NA     NA       NA          NA
#>   4:      4      4 1517237169 1517237169 1517237269     NA       NA           1
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA     NA       NA          NA
#> 497:    497    497 1517237170 1517237170 1517238670     NA       NA           1
#> 498:    498    498         NA         NA         NA     NA       NA          NA
#> 499:    499    499         NA         NA         NA     NA       NA          NA
#> 500:    500    500         NA         NA         NA     NA       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:            NA       NA                                  NA       NA     1
#>   2:            NA       NA                                  NA       NA     1
#>   3:            NA       NA                                  NA       NA     1
#>   4: cfInteractive       NA job1d842e38174b87617bebee1ae5615783       NA     1
#>  ---                                                                          
#> 496:            NA       NA                                  NA       NA     1
#> 497: cfInteractive       NA job51fb9ebdd67e453b14228432fe877c46       NA     1
#> 498:            NA       NA                                  NA       NA     1
#> 499:            NA       NA                                  NA       NA     1
#> 500:            NA       NA                                  NA       NA     1

rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])

#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:      4    iris      nrow     1      d  100.0061 secs
#>  2:     16    iris      nrow     4      a 1100.0167 secs
#>  3:     34    iris      nrow     7      d  100.0054 secs
#>  4:     41    iris      nrow     9      a 1100.0056 secs
#>  5:     46    iris      nrow    10      a 1100.0059 secs
#>  6:     48    iris      nrow    10      c  100.0048 secs
#>  7:     69    iris      nrow    14      d  100.0055 secs
#>  8:     85    iris      nrow    17      e  100.0051 secs
#>  9:     89    iris      nrow    18      d  100.0054 secs
#> 10:     94    iris      nrow    19      d  100.0047 secs
#> 11:     97    iris      nrow    20      b 1100.0056 secs
#> 12:     98    iris      nrow    20      c  100.0050 secs
#> 13:    115    iris      nrow    23      e  100.0048 secs
#> 14:    117    iris      nrow    24      b 1100.0046 secs
#> 15:    120    iris      nrow    24      e  100.0050 secs
#> 16:    145    iris      nrow    29      e  100.0047 secs
#> 17:    148    iris      nrow    30      c  100.0056 secs
#> 18:    156    iris      nrow    32      a 1100.0050 secs
#> 19:    167    iris      nrow    34      b 1100.0050 secs
#> 20:    187    iris      nrow    38      b 1100.0050 secs
#> 21:    193    iris      nrow    39      c  100.0084 secs
#> 22:    220    iris      nrow    44      e  100.0049 secs
#> 23:    223    iris      nrow    45      c  100.0048 secs
#> 24:    227    iris      nrow    46      b 1100.0057 secs
#> 25:    246    iris      nrow    50      a 1100.0052 secs
#> 26:    258    iris      ncol     2      c  500.0049 secs
#> 27:    270    iris      ncol     4      e  500.0056 secs
#> 28:    271    iris      ncol     5      a 1500.0077 secs
#> 29:    275    iris      ncol     5      e  500.0067 secs
#> 30:    296    iris      ncol    10      a 1500.0068 secs
#> 31:    315    iris      ncol    13      e  500.0056 secs
#> 32:    325    iris      ncol    15      e  500.0055 secs
#> 33:    352    iris      ncol    21      b 1500.0074 secs
#> 34:    354    iris      ncol    21      d  500.0062 secs
#> 35:    358    iris      ncol    22      c  500.0058 secs
#> 36:    364    iris      ncol    23      d  500.0056 secs
#> 37:    370    iris      ncol    24      e  500.0074 secs
#> 38:    378    iris      ncol    26      c  500.0075 secs
#> 39:    388    iris      ncol    28      c  500.0054 secs
#> 40:    402    iris      ncol    31      b 1500.0063 secs
#> 41:    411    iris      ncol    33      a 1500.0082 secs
#> 42:    419    iris      ncol    34      d  500.0074 secs
#> 43:    432    iris      ncol    37      b 1500.0068 secs
#> 44:    434    iris      ncol    37      d  500.0069 secs
#> 45:    436    iris      ncol    38      a 1500.0067 secs
#> 46:    438    iris      ncol    38      c  500.0065 secs
#> 47:    447    iris      ncol    40      b 1500.0061 secs
#> 48:    451    iris      ncol    41      a 1500.0054 secs
#> 49:    454    iris      ncol    41      d  500.0061 secs
#> 50:    497    iris      ncol    50      b 1500.0047 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)

#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 16h 24m 53.3s
#>   Total    : 4d 02h 08m 13.6s

rjoin(tab, est$runtimes)

#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1043.1873
#>   2:      2    iris      nrow     1      b estimated 1045.1703
#>   3:      3    iris      nrow     1      c estimated  207.8759
#>   4:      4    iris      nrow     1      d  observed  100.0061
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1396.7798
#> 497:    497    iris      ncol    50      b  observed 1500.0047
#> 498:    498    iris      ncol    50      c estimated  596.3090
#> 499:    499    iris      ncol    50      d estimated  570.8789
#> 500:    500    iris      ncol    50      e estimated  565.5541

print(est, n = 10)

#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 20.3s
#>   Remaining: 3d 16h 24m 53.3s
#>   Parallel : 0d 08h 51m 28.0s
#>   Total    : 4d 02h 08m 13.6s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)

#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    446 estimated 1429.7248
#>   2:    431 estimated 1427.1015
#>   3:    406 estimated 1425.3172
#>   4:    456 estimated 1425.0047
#>  ---                           
#> 446:     90 estimated  137.2754
#> 447:     95 estimated  137.2753
#> 448:     74 estimated  134.3860
#> 449:     79 estimated  133.5860
#> 450:     84 estimated  128.2536
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1873    70
#>   2:      2 estimated 1045.1703    70
#>   3:      3 estimated  207.8759    10
#>   4:      5 estimated  220.0574    57
#>  ---                                 
#> 446:    495 estimated  565.5541    20
#> 447:    496 estimated 1396.7798    12
#> 448:    498 estimated  596.3090     4
#> 449:    499 estimated  570.8789    15
#> 450:    500 estimated  565.5541    21

print(ids)

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1873    70
#>   2:      2 estimated 1045.1703    70
#>   3:      3 estimated  207.8759    10
#>   4:      5 estimated  220.0574    57
#>  ---                                 
#> 446:    495 estimated  565.5541    20
#> 447:    496 estimated 1396.7798    12
#> 448:    498 estimated  596.3090     4
#> 449:    499 estimated  570.8789    15
#> 450:    500 estimated  565.5541    21

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:    70 3521.994
#>  2:    10 3599.727
#>  3:    57 3599.741
#>  4:    67 3554.303
#>  5:    63 3589.511
#>  6:    52 3598.344
#>  7:    60 3599.555
#>  8:    59 3598.384
#>  9:    65 3580.438
#> 10:    44 3597.586
#> 11:    55 3599.645
#> 12:    53 3599.553
#> 13:    61 3597.755
#> 14:    40 3478.151
#> 15:    54 3594.136
#> 16:    64 3586.005
#> 17:    12 3599.194
#> 18:    56 3598.776
#> 19:    58 3599.504
#> 20:     9 3599.403
#> 21:    72 3501.549
#> 22:    69 3529.859
#> 23:    73 3470.943
#> 24:     6 3594.881
#> 25:    68 3535.948
#> 26:     8 3598.753
#> 27:    75 3599.709
#> 28:    71 3513.146
#> 29:    74 3598.288
#> 30:     4 3598.667
#> 31:     7 3596.146
#> 32:    81 3596.326
#> 33:    90 1114.946
#> 34:    77 3547.574
#> 35:    82 3566.873
#> 36:    43 3599.235
#> 37:    42 3598.022
#> 38:     1 3599.592
#> 39:    50 3595.945
#> 40:    51 3599.263
#> 41:    80 3599.627
#> 42:     2 3599.642
#> 43:    76 3564.050
#> 44:    83 3549.674
#> 45:    79 3483.552
#> 46:    62 3592.904
#> 47:    45 3596.377
#> 48:    78 3514.043
#> 49:     3 3599.087
#> 50:    66 3575.886
#> 51:     5 3599.403
#> 52:    49 3477.843
#> 53:    41 3474.108
#> 54:    48 3487.330
#> 55:    30 3525.032
#> 56:    37 3486.864
#> 57:    46 3502.921
#> 58:    18 3599.892
#> 59:    34 3498.724
#> 60:    13 3598.866
#> 61:    31 3521.730
#> 62:    39 3484.233
#> 63:    47 3490.203
#> 64:    17 3599.920
#> 65:    38 3484.477
#> 66:    16 3597.647
#> 67:    11 3599.946
#> 68:    15 3599.540
#> 69:    14 3598.906
#> 70:    28 3531.822
#> 71:    32 3511.090
#> 72:    35 3496.443
#> 73:    29 3529.894
#> 74:    33 3507.079
#> 75:    36 3495.442
#> 76:    24 3598.014
#> 77:    19 3599.871
#> 78:    27 3565.807
#> 79:    25 3596.414
#> 80:    20 3598.819
#> 81:    22 3586.246
#> 82:    26 3590.131
#> 83:    21 3596.303
#> 84:    23 3583.926
#> 85:    89 3496.078
#> 86:    87 3560.350
#> 87:    86 3574.087
#> 88:    84 3599.767
#> 89:    85 3585.333
#> 90:    88 3540.585
#>     chunk  runtime
# NOT RUN {
submitJobs(ids, reg = tmp)
# }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]

#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1043.1873     3
#>   2:      2 estimated 1045.1703     6
#>   3:      3 estimated  207.8759     4
#>   4:      5 estimated  220.0574     4
#>  ---                                 
#> 446:    495 estimated  565.5541     1
#> 447:    496 estimated 1396.7798     5
#> 448:    498 estimated  596.3090     2
#> 449:    499 estimated  570.8789     9
#> 450:    500 estimated  565.5541     6

print(ids[, list(runtime = sum(runtime)), by = chunk])

#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 31887.16
#>  2:     6 31887.95
#>  3:     4 31772.93
#>  4:     2 31773.28
#>  5:     1 31775.77
#>  6:    10 31879.57
#>  7:     7 31884.51
#>  8:     9 31885.22
#>  9:     5 31773.95
#> 10:     8 31772.96