Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1755630524 1755630524 1755631024   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job9d54a570e7f44c67132cc2984ee779bc     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0345 secs
#>  2:     42    iris      nrow     9      b 1100.0386 secs
#>  3:     47    iris      nrow    10      b 1100.0344 secs
#>  4:     66    iris      nrow    14      a 1100.0352 secs
#>  5:     73    iris      nrow    15      c  100.0347 secs
#>  6:     75    iris      nrow    15      e  100.0340 secs
#>  7:     86    iris      nrow    18      a 1100.0338 secs
#>  8:    100    iris      nrow    20      e  100.0378 secs
#>  9:    101    iris      nrow    21      a 1100.0348 secs
#> 10:    103    iris      nrow    21      c  100.0360 secs
#> 11:    123    iris      nrow    25      c  100.0343 secs
#> 12:    125    iris      nrow    25      e  100.0346 secs
#> 13:    161    iris      nrow    33      a 1100.0334 secs
#> 14:    165    iris      nrow    33      e  100.0338 secs
#> 15:    169    iris      nrow    34      d  100.0351 secs
#> 16:    183    iris      nrow    37      c  100.0344 secs
#> 17:    184    iris      nrow    37      d  100.0335 secs
#> 18:    203    iris      nrow    41      c  100.0334 secs
#> 19:    207    iris      nrow    42      b 1100.0337 secs
#> 20:    209    iris      nrow    42      d  100.0348 secs
#> 21:    220    iris      nrow    44      e  100.0341 secs
#> 22:    227    iris      nrow    46      b 1100.0336 secs
#> 23:    229    iris      nrow    46      d  100.0336 secs
#> 24:    231    iris      nrow    47      a 1100.0385 secs
#> 25:    244    iris      nrow    49      d  100.0350 secs
#> 26:    260    iris      ncol     2      e  500.0353 secs
#> 27:    276    iris      ncol     6      a 1500.0341 secs
#> 28:    278    iris      ncol     6      c  500.0342 secs
#> 29:    279    iris      ncol     6      d  500.0372 secs
#> 30:    296    iris      ncol    10      a 1500.0346 secs
#> 31:    320    iris      ncol    14      e  500.0354 secs
#> 32:    340    iris      ncol    18      e  500.0337 secs
#> 33:    347    iris      ncol    20      b 1500.0340 secs
#> 34:    363    iris      ncol    23      c  500.0378 secs
#> 35:    369    iris      ncol    24      d  500.0354 secs
#> 36:    373    iris      ncol    25      c  500.0349 secs
#> 37:    387    iris      ncol    28      b 1500.0339 secs
#> 38:    410    iris      ncol    32      e  500.0341 secs
#> 39:    421    iris      ncol    35      a 1500.0601 secs
#> 40:    436    iris      ncol    38      a 1500.0355 secs
#> 41:    444    iris      ncol    39      d  500.0346 secs
#> 42:    448    iris      ncol    40      c  500.0343 secs
#> 43:    456    iris      ncol    42      a 1500.0335 secs
#> 44:    459    iris      ncol    42      d  500.0339 secs
#> 45:    467    iris      ncol    44      b 1500.0401 secs
#> 46:    468    iris      ncol    44      c  500.0350 secs
#> 47:    475    iris      ncol    45      e  500.0339 secs
#> 48:    482    iris      ncol    47      b 1500.0336 secs
#> 49:    492    iris      ncol    49      b 1500.0331 secs
#> 50:    499    iris      ncol    50      d  500.0332 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 38m 1.5s
#>   Total    : 4d 03h 21m 23.2s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.8491
#>   2:      2    iris      nrow     1      b estimated 1090.6435
#>   3:      3    iris      nrow     1      c estimated  337.8018
#>   4:      4    iris      nrow     1      d estimated  319.8676
#>   5:      5    iris      nrow     1      e estimated  318.0449
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1381.8762
#> 497:    497    iris      ncol    50      b estimated 1388.7640
#> 498:    498    iris      ncol    50      c estimated  613.6569
#> 499:    499    iris      ncol    50      d  observed  500.0332
#> 500:    500    iris      ncol    50      e estimated  574.9420
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 38m 1.5s
#>   Parallel : 0d 08h 58m 29.2s
#>   Total    : 4d 03h 21m 23.2s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1421.2544
#>   2:    461 estimated 1419.8610
#>   3:    462 estimated 1416.9581
#>   4:    457 estimated 1416.1582
#>   5:    487 estimated 1415.0829
#>  ---                           
#> 446:    164 estimated  133.4720
#> 447:    185 estimated  132.9347
#> 448:    204 estimated  131.5203
#> 449:    174 estimated  131.1220
#> 450:    179 estimated  129.9751
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8491    47
#>   2:      2 estimated 1090.6435    51
#>   3:      3 estimated  337.8018    55
#>   4:      4 estimated  319.8676    34
#>   5:      5 estimated  318.0449    33
#>  ---                                 
#> 446:    495 estimated  581.8767    16
#> 447:    496 estimated 1381.8762    21
#> 448:    497 estimated 1388.7640    16
#> 449:    498 estimated  613.6569     4
#> 450:    500 estimated  574.9420    26
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8491    47
#>   2:      2 estimated 1090.6435    51
#>   3:      3 estimated  337.8018    55
#>   4:      4 estimated  319.8676    34
#>   5:      5 estimated  318.0449    33
#>  ---                                 
#> 446:    495 estimated  581.8767    16
#> 447:    496 estimated 1381.8762    21
#> 448:    497 estimated 1388.7640    16
#> 449:    498 estimated  613.6569     4
#> 450:    500 estimated  574.9420    26
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3493.312
#>  2:    51 3596.540
#>  3:    55 3584.576
#>  4:    34 3594.436
#>  5:    33 3598.252
#>  6:    48 3490.904
#>  7:    56 3574.207
#>  8:    71 3487.864
#>  9:    52 3599.463
#> 10:    57 3565.257
#> 11:    69 3494.571
#> 12:    72 3485.476
#> 13:    58 3554.753
#> 14:    70 3492.514
#> 15:    73 3478.622
#> 16:    46 3519.772
#> 17:    50 3599.851
#> 18:    38 3598.696
#> 19:    63 3520.780
#> 20:    66 3507.926
#> 21:    43 3570.175
#> 22:    64 3519.307
#> 23:    67 3506.992
#> 24:    42 3580.285
#> 25:    61 3532.805
#> 26:    53 3599.770
#> 27:    40 3596.503
#> 28:    54 3589.567
#> 29:    60 3535.634
#> 30:    49 3485.262
#> 31:    41 3588.589
#> 32:    37 3599.514
#> 33:    36 3599.655
#> 34:    39 3599.546
#> 35:    59 3546.839
#> 36:    62 3529.696
#> 37:    44 3540.208
#> 38:    35 3595.720
#> 39:    65 3513.453
#> 40:    45 3539.408
#> 41:    68 3503.588
#> 42:    28 3593.599
#> 43:    25 3597.946
#> 44:    26 3590.745
#> 45:    27 3597.255
#> 46:    24 3599.789
#> 47:    29 3570.102
#> 48:    74 3599.690
#> 49:    75 3599.941
#> 50:     6 3593.711
#> 51:     8 3597.804
#> 52:    20 3522.721
#> 53:    21 3518.053
#> 54:     7 3599.683
#> 55:     5 3597.657
#> 56:    12 3563.500
#> 57:    11 3573.811
#> 58:    10 3576.179
#> 59:    32 3600.000
#> 60:    80 3484.012
#> 61:    81 3476.545
#> 62:    82 3598.538
#> 63:     2 3599.976
#> 64:    79 3497.754
#> 65:    76 3592.179
#> 66:    85 3583.820
#> 67:    89 3550.035
#> 68:    91 2033.909
#> 69:     3 3599.774
#> 70:    87 3569.242
#> 71:    88 3559.163
#> 72:    77 3522.083
#> 73:    86 3577.397
#> 74:    83 3598.331
#> 75:    78 3506.258
#> 76:    90 3528.717
#> 77:    84 3592.199
#> 78:     1 3599.203
#> 79:     4 3599.161
#> 80:     9 3589.643
#> 81:    22 3598.935
#> 82:    18 3576.438
#> 83:    13 3598.686
#> 84:    23 3593.751
#> 85:    19 3571.758
#> 86:    16 3587.553
#> 87:    31 3550.106
#> 88:    17 3582.124
#> 89:    30 3566.185
#> 90:    15 3595.651
#> 91:    14 3599.874
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8491     6
#>   2:      2 estimated 1090.6435     1
#>   3:      3 estimated  337.8018     9
#>   4:      4 estimated  319.8676     3
#>   5:      5 estimated  318.0449     4
#>  ---                                 
#> 446:    495 estimated  581.8767     2
#> 447:    496 estimated 1381.8762     6
#> 448:    497 estimated 1388.7640     4
#> 449:    498 estimated  613.6569     4
#> 450:    500 estimated  574.9420     6
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     6 32295.12
#>  2:     1 32231.78
#>  3:     9 32232.37
#>  4:     3 32237.05
#>  5:     4 32295.07
#>  6:     7 32309.22
#>  7:     8 32231.51
#>  8:     5 32308.89
#>  9:     2 32308.52
#> 10:    10 32231.94