Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1755527376 1755527376 1755527876   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job8b22d15da4bc0599a8609172eb948d03     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0340 secs
#>  2:     42    iris      nrow     9      b 1100.0403 secs
#>  3:     47    iris      nrow    10      b 1100.0345 secs
#>  4:     66    iris      nrow    14      a 1100.0352 secs
#>  5:     73    iris      nrow    15      c  100.0345 secs
#>  6:     75    iris      nrow    15      e  100.0343 secs
#>  7:     86    iris      nrow    18      a 1100.0342 secs
#>  8:    100    iris      nrow    20      e  100.0384 secs
#>  9:    101    iris      nrow    21      a 1100.0351 secs
#> 10:    103    iris      nrow    21      c  100.0354 secs
#> 11:    123    iris      nrow    25      c  100.0345 secs
#> 12:    125    iris      nrow    25      e  100.0340 secs
#> 13:    161    iris      nrow    33      a 1100.0338 secs
#> 14:    165    iris      nrow    33      e  100.0344 secs
#> 15:    169    iris      nrow    34      d  100.0353 secs
#> 16:    183    iris      nrow    37      c  100.0345 secs
#> 17:    184    iris      nrow    37      d  100.0341 secs
#> 18:    203    iris      nrow    41      c  100.0340 secs
#> 19:    207    iris      nrow    42      b 1100.0343 secs
#> 20:    209    iris      nrow    42      d  100.0354 secs
#> 21:    220    iris      nrow    44      e  100.0347 secs
#> 22:    227    iris      nrow    46      b 1100.0341 secs
#> 23:    229    iris      nrow    46      d  100.0341 secs
#> 24:    231    iris      nrow    47      a 1100.0387 secs
#> 25:    244    iris      nrow    49      d  100.0351 secs
#> 26:    260    iris      ncol     2      e  500.0353 secs
#> 27:    276    iris      ncol     6      a 1500.0342 secs
#> 28:    278    iris      ncol     6      c  500.0342 secs
#> 29:    279    iris      ncol     6      d  500.0378 secs
#> 30:    296    iris      ncol    10      a 1500.0353 secs
#> 31:    320    iris      ncol    14      e  500.0356 secs
#> 32:    340    iris      ncol    18      e  500.0347 secs
#> 33:    347    iris      ncol    20      b 1500.0353 secs
#> 34:    363    iris      ncol    23      c  500.0387 secs
#> 35:    369    iris      ncol    24      d  500.0392 secs
#> 36:    373    iris      ncol    25      c  500.0346 secs
#> 37:    387    iris      ncol    28      b 1500.0338 secs
#> 38:    410    iris      ncol    32      e  500.0338 secs
#> 39:    421    iris      ncol    35      a 1500.0600 secs
#> 40:    436    iris      ncol    38      a 1500.0352 secs
#> 41:    444    iris      ncol    39      d  500.0344 secs
#> 42:    448    iris      ncol    40      c  500.0343 secs
#> 43:    456    iris      ncol    42      a 1500.0342 secs
#> 44:    459    iris      ncol    42      d  500.0341 secs
#> 45:    467    iris      ncol    44      b 1500.0422 secs
#> 46:    468    iris      ncol    44      c  500.0365 secs
#> 47:    475    iris      ncol    45      e  500.0353 secs
#> 48:    482    iris      ncol    47      b 1500.0350 secs
#> 49:    492    iris      ncol    49      b 1500.0339 secs
#> 50:    499    iris      ncol    50      d  500.0338 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 37m 28.7s
#>   Total    : 4d 03h 20m 50.5s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.0495
#>   2:      2    iris      nrow     1      b estimated 1089.8440
#>   3:      3    iris      nrow     1      c estimated  337.8019
#>   4:      4    iris      nrow     1      d estimated  319.8678
#>   5:      5    iris      nrow     1      e estimated  318.8451
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1381.8768
#> 497:    497    iris      ncol    50      b estimated 1388.9651
#> 498:    498    iris      ncol    50      c estimated  613.6575
#> 499:    499    iris      ncol    50      d  observed  500.0338
#> 500:    500    iris      ncol    50      e estimated  575.2094
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 37m 28.7s
#>   Parallel : 0d 08h 58m 26.7s
#>   Total    : 4d 03h 20m 50.5s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1423.0550
#>   2:    461 estimated 1419.6615
#>   3:    462 estimated 1416.9591
#>   4:    472 estimated 1416.3126
#>   5:    457 estimated 1416.1591
#>  ---                           
#> 446:    164 estimated  133.4722
#> 447:    185 estimated  133.2017
#> 448:    204 estimated  131.5207
#> 449:    174 estimated  131.1223
#> 450:    179 estimated  129.9755
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0495    47
#>   2:      2 estimated 1089.8440    52
#>   3:      3 estimated  337.8019    54
#>   4:      4 estimated  319.8678    33
#>   5:      5 estimated  318.8451    32
#>  ---                                 
#> 446:    495 estimated  582.1440    16
#> 447:    496 estimated 1381.8768    21
#> 448:    497 estimated 1388.9651    16
#> 449:    498 estimated  613.6575     4
#> 450:    500 estimated  575.2094    26
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0495    47
#>   2:      2 estimated 1089.8440    52
#>   3:      3 estimated  337.8019    54
#>   4:      4 estimated  319.8678    33
#>   5:      5 estimated  318.8451    32
#>  ---                                 
#> 446:    495 estimated  582.1440    16
#> 447:    496 estimated 1381.8768    21
#> 448:    497 estimated 1388.9651    16
#> 449:    498 estimated  613.6575     4
#> 450:    500 estimated  575.2094    26
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3493.848
#>  2:    52 3598.460
#>  3:    54 3589.379
#>  4:    33 3595.219
#>  5:    32 3599.220
#>  6:    48 3488.772
#>  7:    55 3584.978
#>  8:    51 3597.206
#>  9:    70 3491.160
#> 10:    56 3573.808
#> 11:    68 3496.795
#> 12:    71 3488.598
#> 13:    53 3595.993
#> 14:    57 3565.259
#> 15:    69 3494.572
#> 16:    72 3486.077
#> 17:    46 3518.174
#> 18:    50 3599.812
#> 19:    38 3596.697
#> 20:    65 3513.898
#> 21:    43 3568.239
#> 22:    66 3508.167
#> 23:    42 3578.262
#> 24:    36 3599.996
#> 25:    62 3529.490
#> 26:    39 3598.473
#> 27:    37 3597.973
#> 28:    58 3554.520
#> 29:    49 3487.151
#> 30:    41 3585.448
#> 31:    60 3535.658
#> 32:    63 3526.823
#> 33:    40 3593.007
#> 34:    59 3546.241
#> 35:    61 3532.057
#> 36:    44 3537.638
#> 37:    35 3594.072
#> 38:    67 3507.376
#> 39:    45 3536.838
#> 40:    64 3518.331
#> 41:    34 3599.298
#> 42:    27 3598.754
#> 43:    25 3595.129
#> 44:    28 3573.821
#> 45:    26 3599.991
#> 46:    29 3566.255
#> 47:    24 3599.171
#> 48:    74 3599.779
#> 49:    75 3599.903
#> 50:     6 3593.313
#> 51:     7 3599.863
#> 52:    20 3521.789
#> 53:    31 3599.605
#> 54:     9 3590.712
#> 55:     5 3598.531
#> 56:    12 3563.769
#> 57:    11 3573.229
#> 58:    10 3577.507
#> 59:     4 3596.886
#> 60:    79 3496.144
#> 61:    81 3476.664
#> 62:     2 3599.999
#> 63:    73 3486.061
#> 64:    82 3599.100
#> 65:    80 3484.815
#> 66:    76 3583.519
#> 67:    86 3579.370
#> 68:    89 3552.208
#> 69:    91 1904.716
#> 70:     3 3599.344
#> 71:    88 3561.844
#> 72:    77 3521.902
#> 73:    84 3593.036
#> 74:    83 3599.497
#> 75:    78 3504.427
#> 76:    85 3586.573
#> 77:    87 3570.065
#> 78:    90 3532.810
#> 79:     1 3598.854
#> 80:     8 3598.705
#> 81:    22 3598.243
#> 82:    18 3576.100
#> 83:    13 3599.222
#> 84:    23 3593.760
#> 85:    19 3570.826
#> 86:    14 3594.498
#> 87:    21 3599.750
#> 88:    17 3582.198
#> 89:    30 3563.844
#> 90:    16 3593.865
#> 91:    15 3595.721
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0495     3
#>   2:      2 estimated 1089.8440     5
#>   3:      3 estimated  337.8019     8
#>   4:      4 estimated  319.8678    10
#>   5:      5 estimated  318.8451     7
#>  ---                                 
#> 446:    495 estimated  582.1440     7
#> 447:    496 estimated 1381.8768     6
#> 448:    497 estimated 1388.9651     3
#> 449:    498 estimated  613.6575     3
#> 450:    500 estimated  575.2094     2
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 32228.00
#>  2:     5 32227.71
#>  3:     8 32227.98
#>  4:    10 32234.40
#>  5:     7 32291.88
#>  6:     2 32292.13
#>  7:     4 32227.85
#>  8:     1 32306.67
#>  9:     9 32305.56
#> 10:     6 32306.49