Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1755526932 1755526932 1755527433   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job150870333fa3423b9bbf20d643aef569     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0349 secs
#>  2:     42    iris      nrow     9      b 1100.0401 secs
#>  3:     47    iris      nrow    10      b 1100.0353 secs
#>  4:     66    iris      nrow    14      a 1100.0359 secs
#>  5:     73    iris      nrow    15      c  100.0357 secs
#>  6:     75    iris      nrow    15      e  100.0427 secs
#>  7:     86    iris      nrow    18      a 1100.0345 secs
#>  8:    100    iris      nrow    20      e  100.0397 secs
#>  9:    101    iris      nrow    21      a 1100.0355 secs
#> 10:    103    iris      nrow    21      c  100.0362 secs
#> 11:    123    iris      nrow    25      c  100.0350 secs
#> 12:    125    iris      nrow    25      e  100.0352 secs
#> 13:    161    iris      nrow    33      a 1100.0342 secs
#> 14:    165    iris      nrow    33      e  100.0352 secs
#> 15:    169    iris      nrow    34      d  100.0360 secs
#> 16:    183    iris      nrow    37      c  100.0354 secs
#> 17:    184    iris      nrow    37      d  100.0350 secs
#> 18:    203    iris      nrow    41      c  100.0343 secs
#> 19:    207    iris      nrow    42      b 1100.0353 secs
#> 20:    209    iris      nrow    42      d  100.0368 secs
#> 21:    220    iris      nrow    44      e  100.0360 secs
#> 22:    227    iris      nrow    46      b 1100.0352 secs
#> 23:    229    iris      nrow    46      d  100.0358 secs
#> 24:    231    iris      nrow    47      a 1100.0400 secs
#> 25:    244    iris      nrow    49      d  100.0363 secs
#> 26:    260    iris      ncol     2      e  500.0367 secs
#> 27:    276    iris      ncol     6      a 1500.0356 secs
#> 28:    278    iris      ncol     6      c  500.0356 secs
#> 29:    279    iris      ncol     6      d  500.0396 secs
#> 30:    296    iris      ncol    10      a 1500.0366 secs
#> 31:    320    iris      ncol    14      e  500.0376 secs
#> 32:    340    iris      ncol    18      e  500.0359 secs
#> 33:    347    iris      ncol    20      b 1500.0356 secs
#> 34:    363    iris      ncol    23      c  500.0402 secs
#> 35:    369    iris      ncol    24      d  500.0368 secs
#> 36:    373    iris      ncol    25      c  500.0360 secs
#> 37:    387    iris      ncol    28      b 1500.0351 secs
#> 38:    410    iris      ncol    32      e  500.0368 secs
#> 39:    421    iris      ncol    35      a 1500.0629 secs
#> 40:    436    iris      ncol    38      a 1500.0365 secs
#> 41:    444    iris      ncol    39      d  500.0358 secs
#> 42:    448    iris      ncol    40      c  500.0358 secs
#> 43:    456    iris      ncol    42      a 1500.0351 secs
#> 44:    459    iris      ncol    42      d  500.0353 secs
#> 45:    467    iris      ncol    44      b 1500.0429 secs
#> 46:    468    iris      ncol    44      c  500.0362 secs
#> 47:    475    iris      ncol    45      e  500.0357 secs
#> 48:    482    iris      ncol    47      b 1500.0356 secs
#> 49:    492    iris      ncol    49      b 1500.0353 secs
#> 50:    499    iris      ncol    50      d  500.0352 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 37m 55.3s
#>   Total    : 4d 03h 21m 17.1s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.8502
#>   2:      2    iris      nrow     1      b estimated 1090.6446
#>   3:      3    iris      nrow     1      c estimated  337.9638
#>   4:      4    iris      nrow     1      d estimated  319.8699
#>   5:      5    iris      nrow     1      e estimated  318.1543
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1381.8781
#> 497:    497    iris      ncol    50      b estimated 1388.9261
#> 498:    498    iris      ncol    50      c estimated  613.0186
#> 499:    499    iris      ncol    50      d  observed  500.0352
#> 500:    500    iris      ncol    50      e estimated  573.6106
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 37m 55.3s
#>   Parallel : 0d 08h 58m 28.8s
#>   Total    : 4d 03h 21m 17.1s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1421.2563
#>   2:    461 estimated 1419.8628
#>   3:    462 estimated 1416.9601
#>   4:    457 estimated 1416.1601
#>   5:    487 estimated 1415.2450
#>  ---                           
#> 446:    164 estimated  133.4732
#> 447:    185 estimated  132.4032
#> 448:    174 estimated  131.1232
#> 449:    204 estimated  130.9618
#> 450:    179 estimated  129.9765
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8502    47
#>   2:      2 estimated 1090.6446    51
#>   3:      3 estimated  337.9638    55
#>   4:      4 estimated  319.8699    33
#>   5:      5 estimated  318.1543    70
#>  ---                                 
#> 446:    495 estimated  580.5452    19
#> 447:    496 estimated 1381.8781    21
#> 448:    497 estimated 1388.9261    16
#> 449:    498 estimated  613.0186     4
#> 450:    500 estimated  573.6106    28
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8502    47
#>   2:      2 estimated 1090.6446    51
#>   3:      3 estimated  337.9638    55
#>   4:      4 estimated  319.8699    33
#>   5:      5 estimated  318.1543    70
#>  ---                                 
#> 446:    495 estimated  580.5452    19
#> 447:    496 estimated 1381.8781    21
#> 448:    497 estimated 1388.9261    16
#> 449:    498 estimated  613.0186     4
#> 450:    500 estimated  573.6106    28
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3494.092
#>  2:    51 3596.706
#>  3:    55 3584.742
#>  4:    33 3599.442
#>  5:    70 3490.785
#>  6:    48 3491.684
#>  7:    56 3574.533
#>  8:    34 3593.804
#>  9:    71 3487.657
#> 10:    52 3599.629
#> 11:    57 3565.423
#> 12:    72 3485.428
#> 13:    58 3554.758
#> 14:    69 3494.116
#> 15:    73 3478.734
#> 16:    46 3519.778
#> 17:    50 3599.529
#> 18:    38 3598.063
#> 19:    53 3594.439
#> 20:    67 3506.876
#> 21:    43 3570.342
#> 22:    64 3518.993
#> 23:    68 3503.933
#> 24:    42 3580.542
#> 25:    61 3532.811
#> 26:    36 3599.599
#> 27:    39 3599.713
#> 28:    54 3589.573
#> 29:    60 3535.747
#> 30:    49 3484.734
#> 31:    41 3589.280
#> 32:    37 3599.629
#> 33:    63 3527.014
#> 34:    40 3595.389
#> 35:    59 3546.845
#> 36:    62 3530.156
#> 37:    44 3540.375
#> 38:    35 3595.567
#> 39:    65 3513.913
#> 40:    45 3539.575
#> 41:    66 3508.062
#> 42:    28 3593.781
#> 43:    25 3597.723
#> 44:    26 3590.241
#> 45:    27 3597.306
#> 46:    24 3599.726
#> 47:    29 3570.883
#> 48:    74 3599.890
#> 49:    75 3598.702
#> 50:    12 3563.909
#> 51:     9 3589.538
#> 52:    21 3517.954
#> 53:    20 3522.465
#> 54:     8 3597.598
#> 55:     6 3593.212
#> 56:     5 3597.265
#> 57:    10 3576.346
#> 58:    32 3599.755
#> 59:    11 3573.179
#> 60:    79 3498.775
#> 61:     2 3599.823
#> 62:    83 3599.246
#> 63:    81 3475.113
#> 64:    76 3593.553
#> 65:    86 3577.215
#> 66:    89 3548.516
#> 67:    91 2024.896
#> 68:     3 3599.941
#> 69:    87 3568.383
#> 70:    77 3523.135
#> 71:    88 3560.178
#> 72:    84 3590.425
#> 73:    78 3505.736
#> 74:    80 3488.394
#> 75:    82 3599.869
#> 76:    85 3583.637
#> 77:     1 3599.766
#> 78:    90 3531.093
#> 79:     4 3599.488
#> 80:     7 3599.536
#> 81:    23 3598.944
#> 82:    18 3575.590
#> 83:    13 3597.827
#> 84:    22 3596.723
#> 85:    19 3570.154
#> 86:    16 3587.278
#> 87:    31 3550.300
#> 88:    17 3581.598
#> 89:    30 3567.070
#> 90:    15 3593.897
#> 91:    14 3597.707
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8502     6
#>   2:      2 estimated 1090.6446     4
#>   3:      3 estimated  337.9638     7
#>   4:      4 estimated  319.8699     8
#>   5:      5 estimated  318.1543     6
#>  ---                                 
#> 446:    495 estimated  580.5452     1
#> 447:    496 estimated 1381.8781     6
#> 448:    497 estimated 1388.9261     2
#> 449:    498 estimated  613.0186     2
#> 450:    500 estimated  573.6106     6
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     6 32294.52
#>  2:     4 32308.40
#>  3:     7 32232.80
#>  4:     8 32236.73
#>  5:     3 32295.27
#>  6:     1 32230.02
#>  7:     9 32308.85
#>  8:     2 32308.81
#>  9:     5 32230.33
#> 10:    10 32229.57