Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1747928917 1747928917 1747929417   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job951a87685b71d8195f88203cff814b3f     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0354 secs
#>  2:     42    iris      nrow     9      b 1100.0333 secs
#>  3:     47    iris      nrow    10      b 1100.0337 secs
#>  4:     66    iris      nrow    14      a 1100.0387 secs
#>  5:     73    iris      nrow    15      c  100.0342 secs
#>  6:     75    iris      nrow    15      e  100.0336 secs
#>  7:     86    iris      nrow    18      a 1100.0329 secs
#>  8:    100    iris      nrow    20      e  100.0333 secs
#>  9:    101    iris      nrow    21      a 1100.0331 secs
#> 10:    103    iris      nrow    21      c  100.0380 secs
#> 11:    123    iris      nrow    25      c  100.0336 secs
#> 12:    125    iris      nrow    25      e  100.0330 secs
#> 13:    161    iris      nrow    33      a 1100.0332 secs
#> 14:    165    iris      nrow    33      e  100.0328 secs
#> 15:    169    iris      nrow    34      d  100.0327 secs
#> 16:    183    iris      nrow    37      c  100.0344 secs
#> 17:    184    iris      nrow    37      d  100.0342 secs
#> 18:    203    iris      nrow    41      c  100.0332 secs
#> 19:    207    iris      nrow    42      b 1100.0330 secs
#> 20:    209    iris      nrow    42      d  100.0331 secs
#> 21:    220    iris      nrow    44      e  100.0332 secs
#> 22:    227    iris      nrow    46      b 1100.0344 secs
#> 23:    229    iris      nrow    46      d  100.0335 secs
#> 24:    231    iris      nrow    47      a 1100.0330 secs
#> 25:    244    iris      nrow    49      d  100.0329 secs
#> 26:    260    iris      ncol     2      e  500.0329 secs
#> 27:    276    iris      ncol     6      a 1500.0373 secs
#> 28:    278    iris      ncol     6      c  500.0340 secs
#> 29:    279    iris      ncol     6      d  500.0332 secs
#> 30:    296    iris      ncol    10      a 1500.0330 secs
#> 31:    320    iris      ncol    14      e  500.0335 secs
#> 32:    340    iris      ncol    18      e  500.0356 secs
#> 33:    347    iris      ncol    20      b 1500.0350 secs
#> 34:    363    iris      ncol    23      c  500.0341 secs
#> 35:    369    iris      ncol    24      d  500.0333 secs
#> 36:    373    iris      ncol    25      c  500.0332 secs
#> 37:    387    iris      ncol    28      b 1500.0345 secs
#> 38:    410    iris      ncol    32      e  500.0397 secs
#> 39:    421    iris      ncol    35      a 1500.0358 secs
#> 40:    436    iris      ncol    38      a 1500.0344 secs
#> 41:    444    iris      ncol    39      d  500.0333 secs
#> 42:    448    iris      ncol    40      c  500.0330 secs
#> 43:    456    iris      ncol    42      a 1500.0329 secs
#> 44:    459    iris      ncol    42      d  500.0341 secs
#> 45:    467    iris      ncol    44      b 1500.0336 secs
#> 46:    468    iris      ncol    44      c  500.0331 secs
#> 47:    475    iris      ncol    45      e  500.0331 secs
#> 48:    482    iris      ncol    47      b 1500.0331 secs
#> 49:    492    iris      ncol    49      b 1500.0342 secs
#> 50:    499    iris      ncol    50      d  500.0333 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 36m 44.4s
#>   Total    : 4d 03h 20m 6.1s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1105.4888
#>   2:      2    iris      nrow     1      b estimated 1089.4428
#>   3:      3    iris      nrow     1      c estimated  338.4811
#>   4:      4    iris      nrow     1      d estimated  317.7063
#>   5:      5    iris      nrow     1      e estimated  317.3769
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1386.5528
#> 497:    497    iris      ncol    50      b estimated 1393.2825
#> 498:    498    iris      ncol    50      c estimated  616.0559
#> 499:    499    iris      ncol    50      d  observed  500.0333
#> 500:    500    iris      ncol    50      e estimated  577.2086
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 36m 44.4s
#>   Parallel : 0d 08h 58m 18.9s
#>   Total    : 4d 03h 20m 6.1s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1423.3334
#>   2:    461 estimated 1421.9401
#>   3:    457 estimated 1418.9934
#>   4:    462 estimated 1418.9933
#>   5:    472 estimated 1416.1467
#>  ---                           
#> 446:    164 estimated  133.9708
#> 447:    185 estimated  133.4605
#> 448:    204 estimated  131.8194
#> 449:    174 estimated  131.6208
#> 450:    179 estimated  130.4743
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1105.4888    48
#>   2:      2 estimated 1089.4428    51
#>   3:      3 estimated  338.4811    38
#>   4:      4 estimated  317.7063    34
#>   5:      5 estimated  317.3769    71
#>  ---                                 
#> 446:    495 estimated  582.1432    15
#> 447:    496 estimated 1386.5528    19
#> 448:    497 estimated 1393.2825    14
#> 449:    498 estimated  616.0559     4
#> 450:    500 estimated  577.2086    22
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1105.4888    48
#>   2:      2 estimated 1089.4428    51
#>   3:      3 estimated  338.4811    38
#>   4:      4 estimated  317.7063    34
#>   5:      5 estimated  317.3769    71
#>  ---                                 
#> 446:    495 estimated  582.1432    15
#> 447:    496 estimated 1386.5528    19
#> 448:    497 estimated 1393.2825    14
#> 449:    498 estimated  616.0559     4
#> 450:    500 estimated  577.2086    22
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    48 3484.913
#>  2:    51 3598.783
#>  3:    38 3593.102
#>  4:    34 3595.497
#>  5:    71 3491.750
#>  6:    49 3476.698
#>  7:    55 3584.071
#>  8:    68 3500.145
#>  9:    72 3488.432
#> 10:    52 3593.853
#> 11:    56 3574.637
#> 12:    69 3496.757
#> 13:    73 3482.072
#> 14:    57 3568.717
#> 15:    70 3493.656
#> 16:    74 3474.399
#> 17:    46 3519.930
#> 18:    50 3599.212
#> 19:    43 3567.890
#> 20:    64 3516.088
#> 21:    66 3512.282
#> 22:    44 3533.450
#> 23:    65 3515.292
#> 24:    67 3509.240
#> 25:    39 3598.642
#> 26:    37 3599.298
#> 27:    63 3526.087
#> 28:    47 3493.974
#> 29:    53 3598.537
#> 30:    40 3586.379
#> 31:    59 3547.232
#> 32:    42 3579.433
#> 33:    61 3537.082
#> 34:    62 3531.756
#> 35:    41 3584.106
#> 36:    58 3554.779
#> 37:    60 3540.276
#> 38:    54 3596.246
#> 39:    45 3531.543
#> 40:    36 3594.233
#> 41:    35 3598.056
#> 42:    28 3595.929
#> 43:    26 3593.732
#> 44:    25 3597.452
#> 45:    27 3597.697
#> 46:    30 3569.959
#> 47:    29 3575.673
#> 48:    75 3599.401
#> 49:    20 3522.400
#> 50:     8 3594.584
#> 51:    31 3599.933
#> 52:     6 3595.630
#> 53:     7 3599.756
#> 54:    12 3563.299
#> 55:     5 3599.159
#> 56:    11 3575.426
#> 57:    33 3599.394
#> 58:    10 3578.377
#> 59:    32 3599.955
#> 60:    82 3599.552
#> 61:    83 3597.562
#> 62:     4 3598.579
#> 63:    79 3487.425
#> 64:    80 3476.107
#> 65:     3 3599.441
#> 66:    85 3579.059
#> 67:    91 2034.999
#> 68:    81 3471.034
#> 69:    77 3519.667
#> 70:    88 3556.815
#> 71:    87 3565.184
#> 72:    78 3503.314
#> 73:    89 3548.820
#> 74:    84 3587.370
#> 75:    90 3532.212
#> 76:    86 3570.203
#> 77:     2 3599.797
#> 78:     1 3599.108
#> 79:    76 3584.637
#> 80:     9 3584.215
#> 81:    22 3596.042
#> 82:    18 3572.756
#> 83:    13 3598.577
#> 84:    21 3599.206
#> 85:    19 3568.695
#> 86:    16 3586.323
#> 87:    23 3588.038
#> 88:    17 3578.963
#> 89:    24 3596.707
#> 90:    15 3598.113
#> 91:    14 3599.616
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1105.4888     3
#>   2:      2 estimated 1089.4428     4
#>   3:      3 estimated  338.4811     4
#>   4:      4 estimated  317.7063     8
#>   5:      5 estimated  317.3769     3
#>  ---                                 
#> 446:    495 estimated  582.1432     3
#> 447:    496 estimated 1386.5528     5
#> 448:    497 estimated 1393.2825     9
#> 449:    498 estimated  616.0559     1
#> 450:    500 estimated  577.2086     6
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     3 32291.13
#>  2:     4 32224.38
#>  3:     8 32229.12
#>  4:     2 32297.83
#>  5:    10 32224.59
#>  6:     1 32224.44
#>  7:     7 32298.11
#>  8:     5 32291.41
#>  9:     9 32298.94
#> 10:     6 32224.48