Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1747924457 1747924457 1747924957   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job692f3f52cbf30c093a55bee6a40e1a9c     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0351 secs
#>  2:     42    iris      nrow     9      b 1100.0356 secs
#>  3:     47    iris      nrow    10      b 1100.0357 secs
#>  4:     66    iris      nrow    14      a 1100.0356 secs
#>  5:     73    iris      nrow    15      c  100.0406 secs
#>  6:     75    iris      nrow    15      e  100.0357 secs
#>  7:     86    iris      nrow    18      a 1100.0369 secs
#>  8:    100    iris      nrow    20      e  100.0360 secs
#>  9:    101    iris      nrow    21      a 1100.0363 secs
#> 10:    103    iris      nrow    21      c  100.0357 secs
#> 11:    123    iris      nrow    25      c  100.0408 secs
#> 12:    125    iris      nrow    25      e  100.0357 secs
#> 13:    161    iris      nrow    33      a 1100.0389 secs
#> 14:    165    iris      nrow    33      e  100.0360 secs
#> 15:    169    iris      nrow    34      d  100.0358 secs
#> 16:    183    iris      nrow    37      c  100.0355 secs
#> 17:    184    iris      nrow    37      d  100.0359 secs
#> 18:    203    iris      nrow    41      c  100.0364 secs
#> 19:    207    iris      nrow    42      b 1100.0361 secs
#> 20:    209    iris      nrow    42      d  100.0367 secs
#> 21:    220    iris      nrow    44      e  100.0358 secs
#> 22:    227    iris      nrow    46      b 1100.0416 secs
#> 23:    229    iris      nrow    46      d  100.0366 secs
#> 24:    231    iris      nrow    47      a 1100.0362 secs
#> 25:    244    iris      nrow    49      d  100.0355 secs
#> 26:    260    iris      ncol     2      e  500.0355 secs
#> 27:    276    iris      ncol     6      a 1500.0403 secs
#> 28:    278    iris      ncol     6      c  500.0355 secs
#> 29:    279    iris      ncol     6      d  500.0359 secs
#> 30:    296    iris      ncol    10      a 1500.0352 secs
#> 31:    320    iris      ncol    14      e  500.0355 secs
#> 32:    340    iris      ncol    18      e  500.0403 secs
#> 33:    347    iris      ncol    20      b 1500.0355 secs
#> 34:    363    iris      ncol    23      c  500.0357 secs
#> 35:    369    iris      ncol    24      d  500.0354 secs
#> 36:    373    iris      ncol    25      c  500.0347 secs
#> 37:    387    iris      ncol    28      b 1500.0407 secs
#> 38:    410    iris      ncol    32      e  500.0354 secs
#> 39:    421    iris      ncol    35      a 1500.0350 secs
#> 40:    436    iris      ncol    38      a 1500.0356 secs
#> 41:    444    iris      ncol    39      d  500.0405 secs
#> 42:    448    iris      ncol    40      c  500.0353 secs
#> 43:    456    iris      ncol    42      a 1500.0356 secs
#> 44:    459    iris      ncol    42      d  500.0359 secs
#> 45:    467    iris      ncol    44      b 1500.0606 secs
#> 46:    468    iris      ncol    44      c  500.0363 secs
#> 47:    475    iris      ncol    45      e  500.0358 secs
#> 48:    482    iris      ncol    47      b 1500.0348 secs
#> 49:    492    iris      ncol    49      b 1500.0359 secs
#> 50:    499    iris      ncol    50      d  500.0347 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 36m 3.6s
#>   Total    : 4d 03h 19m 25.5s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.2903
#>   2:      2    iris      nrow     1      b estimated 1090.0845
#>   3:      3    iris      nrow     1      c estimated  335.4038
#>   4:      4    iris      nrow     1      d estimated  317.2023
#>   5:      5    iris      nrow     1      e estimated  315.5128
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1381.8772
#> 497:    497    iris      ncol    50      b estimated 1388.7688
#> 498:    498    iris      ncol    50      c estimated  613.6583
#> 499:    499    iris      ncol    50      d  observed  500.0347
#> 500:    500    iris      ncol    50      e estimated  574.0106
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 36m 3.6s
#>   Parallel : 0d 08h 58m 20.7s
#>   Total    : 4d 03h 19m 25.5s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1423.2558
#>   2:    461 estimated 1419.8606
#>   3:    472 estimated 1417.2316
#>   4:    462 estimated 1416.8765
#>   5:    457 estimated 1416.0764
#>  ---                           
#> 446:    164 estimated  133.7336
#> 447:    204 estimated  133.6223
#> 448:    185 estimated  133.4631
#> 449:    174 estimated  131.3836
#> 450:    179 estimated  130.2369
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2903    47
#>   2:      2 estimated 1090.0845    51
#>   3:      3 estimated  335.4038    37
#>   4:      4 estimated  317.2023    32
#>   5:      5 estimated  315.5128    69
#>  ---                                 
#> 446:    495 estimated  581.7453    15
#> 447:    496 estimated 1381.8772    21
#> 448:    497 estimated 1388.7688    16
#> 449:    498 estimated  613.6583     4
#> 450:    500 estimated  574.0106    26
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2903    47
#>   2:      2 estimated 1090.0845    51
#>   3:      3 estimated  335.4038    37
#>   4:      4 estimated  317.2023    32
#>   5:      5 estimated  315.5128    69
#>  ---                                 
#> 446:    495 estimated  581.7453    15
#> 447:    496 estimated 1381.8772    21
#> 448:    497 estimated 1388.7688    16
#> 449:    498 estimated  613.6583     4
#> 450:    500 estimated  574.0106    26
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3493.664
#>  2:    51 3597.153
#>  3:    37 3599.038
#>  4:    32 3597.139
#>  5:    69 3491.234
#>  6:    48 3490.963
#>  7:    53 3594.416
#>  8:    66 3499.562
#>  9:    70 3489.574
#> 10:    52 3599.559
#> 11:    54 3586.280
#> 12:    67 3497.708
#> 13:    71 3487.341
#> 14:    55 3582.518
#> 15:    68 3494.875
#> 16:    72 3483.703
#> 17:    46 3518.631
#> 18:    50 3599.927
#> 19:    38 3595.125
#> 20:    64 3510.931
#> 21:    43 3560.416
#> 22:    61 3522.246
#> 23:    65 3508.021
#> 24:    42 3570.216
#> 25:    60 3533.405
#> 26:    39 3594.748
#> 27:    36 3599.001
#> 28:    57 3562.802
#> 29:    49 3487.261
#> 30:    41 3583.759
#> 31:    56 3569.275
#> 32:    59 3543.538
#> 33:    40 3590.103
#> 34:    35 3599.535
#> 35:    58 3552.517
#> 36:    44 3530.658
#> 37:    34 3599.684
#> 38:    63 3514.552
#> 39:    45 3529.858
#> 40:    62 3518.878
#> 41:    33 3599.451
#> 42:    27 3599.387
#> 43:    25 3592.519
#> 44:    24 3599.205
#> 45:    26 3599.742
#> 46:    23 3599.264
#> 47:    28 3572.966
#> 48:    75 3599.512
#> 49:    12 3562.256
#> 50:    74 3472.329
#> 51:    10 3584.482
#> 52:    21 3516.590
#> 53:    20 3522.987
#> 54:     9 3593.021
#> 55:     6 3594.737
#> 56:     5 3597.873
#> 57:     7 3597.032
#> 58:    11 3574.966
#> 59:     4 3598.488
#> 60:    80 3485.147
#> 61:    82 3471.173
#> 62:    83 3599.923
#> 63:    81 3479.934
#> 64:    73 3489.530
#> 65:    77 3524.367
#> 66:    79 3500.049
#> 67:    76 3584.074
#> 68:    85 3587.398
#> 69:    86 3577.929
#> 70:    91 2313.168
#> 71:    78 3508.232
#> 72:    90 3536.674
#> 73:    88 3562.655
#> 74:    89 3553.812
#> 75:    84 3594.190
#> 76:    87 3568.958
#> 77:     3 3599.407
#> 78:     1 3599.917
#> 79:     2 3599.990
#> 80:     8 3599.563
#> 81:    22 3598.431
#> 82:    18 3575.364
#> 83:    13 3597.534
#> 84:    30 3557.287
#> 85:    19 3571.127
#> 86:    16 3585.491
#> 87:    31 3546.633
#> 88:    17 3581.746
#> 89:    29 3566.583
#> 90:    15 3592.721
#> 91:    14 3597.989
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2903     1
#>   2:      2 estimated 1090.0845    10
#>   3:      3 estimated  335.4038     8
#>   4:      4 estimated  317.2023    10
#>   5:      5 estimated  315.5128     5
#>  ---                                 
#> 446:    495 estimated  581.7453     8
#> 447:    496 estimated 1381.8772     6
#> 448:    497 estimated 1388.7688     3
#> 449:    498 estimated  613.6583     3
#> 450:    500 estimated  574.0106    10
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     1 32224.67
#>  2:    10 32217.46
#>  3:     8 32218.91
#>  4:     5 32284.56
#>  5:     3 32299.03
#>  6:     2 32218.96
#>  7:     4 32284.88
#>  8:     6 32300.68
#>  9:     7 32298.01
#> 10:     9 32216.45