Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1747928522 1747928522 1747929022   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job452505811eea36cdb2e44efb146082d3     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0341 secs
#>  2:     42    iris      nrow     9      b 1100.0330 secs
#>  3:     47    iris      nrow    10      b 1100.0329 secs
#>  4:     66    iris      nrow    14      a 1100.0371 secs
#>  5:     73    iris      nrow    15      c  100.0344 secs
#>  6:     75    iris      nrow    15      e  100.0337 secs
#>  7:     86    iris      nrow    18      a 1100.0331 secs
#>  8:    100    iris      nrow    20      e  100.0328 secs
#>  9:    101    iris      nrow    21      a 1100.0333 secs
#> 10:    103    iris      nrow    21      c  100.0382 secs
#> 11:    123    iris      nrow    25      c  100.0337 secs
#> 12:    125    iris      nrow    25      e  100.0330 secs
#> 13:    161    iris      nrow    33      a 1100.0330 secs
#> 14:    165    iris      nrow    33      e  100.0331 secs
#> 15:    169    iris      nrow    34      d  100.0331 secs
#> 16:    183    iris      nrow    37      c  100.0346 secs
#> 17:    184    iris      nrow    37      d  100.0335 secs
#> 18:    203    iris      nrow    41      c  100.0332 secs
#> 19:    207    iris      nrow    42      b 1100.0337 secs
#> 20:    209    iris      nrow    42      d  100.0330 secs
#> 21:    220    iris      nrow    44      e  100.0382 secs
#> 22:    227    iris      nrow    46      b 1100.0350 secs
#> 23:    229    iris      nrow    46      d  100.0361 secs
#> 24:    231    iris      nrow    47      a 1100.0338 secs
#> 25:    244    iris      nrow    49      d  100.0339 secs
#> 26:    260    iris      ncol     2      e  500.0404 secs
#> 27:    276    iris      ncol     6      a 1500.0361 secs
#> 28:    278    iris      ncol     6      c  500.0338 secs
#> 29:    279    iris      ncol     6      d  500.0332 secs
#> 30:    296    iris      ncol    10      a 1500.0330 secs
#> 31:    320    iris      ncol    14      e  500.0374 secs
#> 32:    340    iris      ncol    18      e  500.0340 secs
#> 33:    347    iris      ncol    20      b 1500.0330 secs
#> 34:    363    iris      ncol    23      c  500.0328 secs
#> 35:    369    iris      ncol    24      d  500.0329 secs
#> 36:    373    iris      ncol    25      c  500.0380 secs
#> 37:    387    iris      ncol    28      b 1500.0336 secs
#> 38:    410    iris      ncol    32      e  500.0330 secs
#> 39:    421    iris      ncol    35      a 1500.0329 secs
#> 40:    436    iris      ncol    38      a 1500.0377 secs
#> 41:    444    iris      ncol    39      d  500.0338 secs
#> 42:    448    iris      ncol    40      c  500.0330 secs
#> 43:    456    iris      ncol    42      a 1500.0333 secs
#> 44:    459    iris      ncol    42      d  500.0371 secs
#> 45:    467    iris      ncol    44      b 1500.0339 secs
#> 46:    468    iris      ncol    44      c  500.0332 secs
#> 47:    475    iris      ncol    45      e  500.0329 secs
#> 48:    482    iris      ncol    47      b 1500.0375 secs
#> 49:    492    iris      ncol    49      b 1500.0335 secs
#> 50:    499    iris      ncol    50      d  500.0331 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 38m 48.6s
#>   Total    : 4d 03h 22m 10.3s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.2882
#>   2:      2    iris      nrow     1      b estimated 1090.0822
#>   3:      3    iris      nrow     1      c estimated  338.3615
#>   4:      4    iris      nrow     1      d estimated  318.4003
#>   5:      5    iris      nrow     1      e estimated  318.6844
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1383.3536
#> 497:    497    iris      ncol    50      b estimated 1390.8033
#> 498:    498    iris      ncol    50      c estimated  617.6164
#> 499:    499    iris      ncol    50      d  observed  500.0331
#> 500:    500    iris      ncol    50      e estimated  577.8087
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 38m 48.6s
#>   Parallel : 0d 08h 58m 31.4s
#>   Total    : 4d 03h 22m 10.3s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1422.2506
#>   2:    461 estimated 1420.8573
#>   3:    462 estimated 1417.8705
#>   4:    457 estimated 1417.8705
#>   5:    472 estimated 1416.3840
#>  ---                           
#> 446:    164 estimated  133.2311
#> 447:    185 estimated  132.9609
#> 448:    174 estimated  130.8811
#> 449:    204 estimated  130.6533
#> 450:    179 estimated  129.7345
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2882    47
#>   2:      2 estimated 1090.0822    52
#>   3:      3 estimated  338.3615    54
#>   4:      4 estimated  318.4003    71
#>   5:      5 estimated  318.6844    34
#>  ---                                 
#> 446:    495 estimated  583.9433    15
#> 447:    496 estimated 1383.3536    20
#> 448:    497 estimated 1390.8033    14
#> 449:    498 estimated  617.6164     4
#> 450:    500 estimated  577.8087    22
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2882    47
#>   2:      2 estimated 1090.0822    52
#>   3:      3 estimated  338.3615    54
#>   4:      4 estimated  318.4003    71
#>   5:      5 estimated  318.6844    34
#>  ---                                 
#> 446:    495 estimated  583.9433    15
#> 447:    496 estimated 1383.3536    20
#> 448:    497 estimated 1390.8033    14
#> 449:    498 estimated  617.6164     4
#> 450:    500 estimated  577.8087    22
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3492.589
#>  2:    52 3597.707
#>  3:    54 3589.347
#>  4:    71 3488.824
#>  5:    34 3595.435
#>  6:    48 3487.468
#>  7:    55 3588.145
#>  8:    72 3485.742
#>  9:    51 3596.939
#> 10:    56 3578.037
#> 11:    73 3479.173
#> 12:    69 3493.230
#> 13:    53 3593.612
#> 14:    57 3568.131
#> 15:    74 3599.924
#> 16:    70 3491.205
#> 17:    46 3517.555
#> 18:    50 3599.758
#> 19:    38 3598.215
#> 20:    35 3596.151
#> 21:    43 3568.880
#> 22:    36 3594.844
#> 23:    42 3578.680
#> 24:    65 3521.878
#> 25:    63 3528.095
#> 26:    39 3599.567
#> 27:    59 3548.155
#> 28:    58 3557.114
#> 29:    49 3476.398
#> 30:    41 3590.040
#> 31:    61 3532.998
#> 32:    64 3524.495
#> 33:    40 3597.914
#> 34:    60 3536.382
#> 35:    62 3530.253
#> 36:    44 3539.378
#> 37:    37 3593.295
#> 38:    66 3509.244
#> 39:    45 3538.578
#> 40:    67 3506.707
#> 41:    68 3504.220
#> 42:    28 3595.972
#> 43:    25 3594.089
#> 44:    26 3592.283
#> 45:    27 3598.297
#> 46:    24 3597.532
#> 47:    29 3574.208
#> 48:    75 3599.421
#> 49:    20 3521.951
#> 50:     9 3592.459
#> 51:    21 3517.969
#> 52:     6 3595.924
#> 53:     8 3598.863
#> 54:    12 3564.068
#> 55:     5 3599.861
#> 56:    11 3575.886
#> 57:    33 3599.640
#> 58:     7 3597.520
#> 59:    32 3598.720
#> 60:    82 3470.950
#> 61:    84 3599.752
#> 62:    76 3597.586
#> 63:    83 3599.673
#> 64:    81 3480.846
#> 65:    80 3490.038
#> 66:    77 3526.342
#> 67:    87 3566.312
#> 68:    91 2291.496
#> 69:    78 3513.018
#> 70:    89 3546.162
#> 71:    88 3557.259
#> 72:    90 3526.446
#> 73:    79 3499.730
#> 74:    85 3584.622
#> 75:    86 3577.126
#> 76:     2 3599.863
#> 77:     4 3599.576
#> 78:     3 3599.998
#> 79:    10 3584.064
#> 80:    22 3598.547
#> 81:    18 3574.047
#> 82:    13 3599.054
#> 83:    30 3560.042
#> 84:    19 3569.077
#> 85:    14 3594.458
#> 86:    31 3550.907
#> 87:    17 3580.538
#> 88:    23 3598.489
#> 89:    15 3597.707
#> 90:    16 3594.202
#> 91:     1 3471.683
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.2882     6
#>   2:      2 estimated 1090.0822    10
#>   3:      3 estimated  338.3615     1
#>   4:      4 estimated  318.4003     2
#>   5:      5 estimated  318.6844    10
#>  ---                                 
#> 446:    495 estimated  583.9433     1
#> 447:    496 estimated 1383.3536     2
#> 448:    497 estimated 1390.8033     7
#> 449:    498 estimated  617.6164     3
#> 450:    500 estimated  577.8087     2
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     6 32310.61
#>  2:    10 32241.78
#>  3:     1 32237.39
#>  4:     2 32300.58
#>  5:     7 32237.49
#>  6:     8 32311.43
#>  7:     4 32238.64
#>  8:     5 32300.62
#>  9:     9 32311.21
#> 10:     3 32238.82