Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1747924755 1747924755 1747925255   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job7e7c20b6dfc51fbb713547bfecdcf52d     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0336 secs
#>  2:     42    iris      nrow     9      b 1100.0336 secs
#>  3:     47    iris      nrow    10      b 1100.0333 secs
#>  4:     66    iris      nrow    14      a 1100.0335 secs
#>  5:     73    iris      nrow    15      c  100.0339 secs
#>  6:     75    iris      nrow    15      e  100.0350 secs
#>  7:     86    iris      nrow    18      a 1100.0329 secs
#>  8:    100    iris      nrow    20      e  100.0328 secs
#>  9:    101    iris      nrow    21      a 1100.0329 secs
#> 10:    103    iris      nrow    21      c  100.0335 secs
#> 11:    123    iris      nrow    25      c  100.0335 secs
#> 12:    125    iris      nrow    25      e  100.0336 secs
#> 13:    161    iris      nrow    33      a 1100.0330 secs
#> 14:    165    iris      nrow    33      e  100.0328 secs
#> 15:    169    iris      nrow    34      d  100.0329 secs
#> 16:    183    iris      nrow    37      c  100.0371 secs
#> 17:    184    iris      nrow    37      d  100.0333 secs
#> 18:    203    iris      nrow    41      c  100.0337 secs
#> 19:    207    iris      nrow    42      b 1100.0329 secs
#> 20:    209    iris      nrow    42      d  100.0330 secs
#> 21:    220    iris      nrow    44      e  100.0328 secs
#> 22:    227    iris      nrow    46      b 1100.0335 secs
#> 23:    229    iris      nrow    46      d  100.0333 secs
#> 24:    231    iris      nrow    47      a 1100.0338 secs
#> 25:    244    iris      nrow    49      d  100.0329 secs
#> 26:    260    iris      ncol     2      e  500.0351 secs
#> 27:    276    iris      ncol     6      a 1500.0328 secs
#> 28:    278    iris      ncol     6      c  500.0334 secs
#> 29:    279    iris      ncol     6      d  500.0334 secs
#> 30:    296    iris      ncol    10      a 1500.0336 secs
#> 31:    320    iris      ncol    14      e  500.0330 secs
#> 32:    340    iris      ncol    18      e  500.0326 secs
#> 33:    347    iris      ncol    20      b 1500.0328 secs
#> 34:    363    iris      ncol    23      c  500.0334 secs
#> 35:    369    iris      ncol    24      d  500.0334 secs
#> 36:    373    iris      ncol    25      c  500.0337 secs
#> 37:    387    iris      ncol    28      b 1500.0330 secs
#> 38:    410    iris      ncol    32      e  500.0327 secs
#> 39:    421    iris      ncol    35      a 1500.0330 secs
#> 40:    436    iris      ncol    38      a 1500.0338 secs
#> 41:    444    iris      ncol    39      d  500.0332 secs
#> 42:    448    iris      ncol    40      c  500.0336 secs
#> 43:    456    iris      ncol    42      a 1500.0330 secs
#> 44:    459    iris      ncol    42      d  500.0329 secs
#> 45:    467    iris      ncol    44      b 1500.0365 secs
#> 46:    468    iris      ncol    44      c  500.0335 secs
#> 47:    475    iris      ncol    45      e  500.0334 secs
#> 48:    482    iris      ncol    47      b 1500.0338 secs
#> 49:    492    iris      ncol    49      b 1500.0328 secs
#> 50:    499    iris      ncol    50      d  500.0329 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 37m 39.8s
#>   Total    : 4d 03h 21m 1.4s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.0471
#>   2:      2    iris      nrow     1      b estimated 1090.0015
#>   3:      3    iris      nrow     1      c estimated  337.9606
#>   4:      4    iris      nrow     1      d estimated  319.5995
#>   5:      5    iris      nrow     1      e estimated  319.0836
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1382.7474
#> 497:    497    iris      ncol    50      b estimated 1389.9974
#> 498:    498    iris      ncol    50      c estimated  615.8907
#> 499:    499    iris      ncol    50      d  observed  500.0329
#> 500:    500    iris      ncol    50      e estimated  576.8826
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 37m 39.8s
#>   Parallel : 0d 08h 58m 26.8s
#>   Total    : 4d 03h 21m 1.4s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1421.9248
#>   2:    461 estimated 1418.5312
#>   3:    472 estimated 1416.2585
#>   4:    462 estimated 1415.7448
#>   5:    487 estimated 1415.1163
#>  ---                           
#> 446:    164 estimated  133.7307
#> 447:    185 estimated  133.4604
#> 448:    174 estimated  131.3807
#> 449:    204 estimated  131.2193
#> 450:    179 estimated  130.2342
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0471    47
#>   2:      2 estimated 1090.0015    51
#>   3:      3 estimated  337.9606    55
#>   4:      4 estimated  319.5995    33
#>   5:      5 estimated  319.0836    70
#>  ---                                 
#> 446:    495 estimated  584.6172    15
#> 447:    496 estimated 1382.7474    19
#> 448:    497 estimated 1389.9974    14
#> 449:    498 estimated  615.8907     4
#> 450:    500 estimated  576.8826    22
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0471    47
#>   2:      2 estimated 1090.0015    51
#>   3:      3 estimated  337.9606    55
#>   4:      4 estimated  319.5995    33
#>   5:      5 estimated  319.0836    70
#>  ---                                 
#> 446:    495 estimated  584.6172    15
#> 447:    496 estimated 1382.7474    19
#> 448:    497 estimated 1389.9974    14
#> 449:    498 estimated  615.8907     4
#> 450:    500 estimated  576.8826    22
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3493.812
#>  2:    51 3593.480
#>  3:    55 3585.769
#>  4:    33 3597.824
#>  5:    70 3491.432
#>  6:    48 3489.803
#>  7:    56 3575.456
#>  8:    34 3593.823
#>  9:    71 3488.682
#> 10:    52 3597.187
#> 11:    57 3568.154
#> 12:    72 3484.530
#> 13:    58 3558.669
#> 14:    69 3493.458
#> 15:    73 3479.013
#> 16:    46 3518.298
#> 17:    50 3599.657
#> 18:    38 3596.206
#> 19:    64 3516.009
#> 20:    39 3595.294
#> 21:    63 3519.243
#> 22:    65 3513.059
#> 23:    43 3578.941
#> 24:    36 3599.703
#> 25:    61 3533.437
#> 26:    40 3598.365
#> 27:    37 3595.139
#> 28:    54 3589.486
#> 29:    49 3483.679
#> 30:    42 3584.057
#> 31:    59 3549.053
#> 32:    62 3529.134
#> 33:    41 3587.238
#> 34:    53 3599.601
#> 35:    60 3539.266
#> 36:    44 3546.013
#> 37:    35 3596.856
#> 38:    67 3505.365
#> 39:    45 3545.213
#> 40:    66 3508.753
#> 41:    68 3502.879
#> 42:    27 3597.159
#> 43:    24 3596.792
#> 44:    26 3585.269
#> 45:    25 3599.237
#> 46:    28 3571.554
#> 47:    23 3599.921
#> 48:    74 3473.566
#> 49:    75 3599.812
#> 50:    12 3562.623
#> 51:     8 3593.418
#> 52:    31 3599.629
#> 53:    20 3520.097
#> 54:     7 3596.481
#> 55:    11 3564.663
#> 56:     5 3593.564
#> 57:     6 3596.080
#> 58:    32 3599.959
#> 59:    10 3578.636
#> 60:     4 3597.817
#> 61:    81 3482.714
#> 62:     3 3599.966
#> 63:    83 3599.696
#> 64:    77 3527.579
#> 65:    79 3505.632
#> 66:    76 3591.969
#> 67:    85 3591.238
#> 68:    91 2163.075
#> 69:    82 3470.857
#> 70:    78 3513.407
#> 71:    89 3551.761
#> 72:    88 3560.164
#> 73:    87 3569.921
#> 74:     2 3599.757
#> 75:    80 3493.775
#> 76:    86 3579.504
#> 77:     1 3599.486
#> 78:    84 3599.447
#> 79:    90 3530.690
#> 80:     9 3589.667
#> 81:    22 3597.230
#> 82:    18 3566.949
#> 83:    15 3588.359
#> 84:    21 3599.328
#> 85:    19 3564.289
#> 86:    16 3586.779
#> 87:    30 3552.530
#> 88:    17 3574.283
#> 89:    29 3565.012
#> 90:    13 3599.801
#> 91:    14 3597.568
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.0471     5
#>   2:      2 estimated 1090.0015     8
#>   3:      3 estimated  337.9606     2
#>   4:      4 estimated  319.5995    10
#>   5:      5 estimated  319.0836     1
#>  ---                                 
#> 446:    495 estimated  584.6172     5
#> 447:    496 estimated 1382.7474     9
#> 448:    497 estimated 1389.9974     4
#> 449:    498 estimated  615.8907     1
#> 450:    500 estimated  576.8826     5
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     5 32306.40
#>  2:     8 32230.15
#>  3:     2 32306.83
#>  4:    10 32234.45
#>  5:     1 32295.78
#>  6:     4 32227.75
#>  7:     9 32295.66
#>  8:     7 32228.63
#>  9:     3 32306.41
#> 10:     6 32227.70