Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1747925064 1747925064 1747925564   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> jobf2f11644470be2973909b7bc0bd61e8d     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0343 secs
#>  2:     42    iris      nrow     9      b 1100.0343 secs
#>  3:     47    iris      nrow    10      b 1100.0338 secs
#>  4:     66    iris      nrow    14      a 1100.0338 secs
#>  5:     73    iris      nrow    15      c  100.0336 secs
#>  6:     75    iris      nrow    15      e  100.0338 secs
#>  7:     86    iris      nrow    18      a 1100.0333 secs
#>  8:    100    iris      nrow    20      e  100.0330 secs
#>  9:    101    iris      nrow    21      a 1100.0330 secs
#> 10:    103    iris      nrow    21      c  100.0343 secs
#> 11:    123    iris      nrow    25      c  100.0344 secs
#> 12:    125    iris      nrow    25      e  100.0344 secs
#> 13:    161    iris      nrow    33      a 1100.0342 secs
#> 14:    165    iris      nrow    33      e  100.0340 secs
#> 15:    169    iris      nrow    34      d  100.0346 secs
#> 16:    183    iris      nrow    37      c  100.0370 secs
#> 17:    184    iris      nrow    37      d  100.0371 secs
#> 18:    203    iris      nrow    41      c  100.0368 secs
#> 19:    207    iris      nrow    42      b 1100.0343 secs
#> 20:    209    iris      nrow    42      d  100.0341 secs
#> 21:    220    iris      nrow    44      e  100.0347 secs
#> 22:    227    iris      nrow    46      b 1100.0351 secs
#> 23:    229    iris      nrow    46      d  100.0355 secs
#> 24:    231    iris      nrow    47      a 1100.0357 secs
#> 25:    244    iris      nrow    49      d  100.0362 secs
#> 26:    260    iris      ncol     2      e  500.0346 secs
#> 27:    276    iris      ncol     6      a 1500.0343 secs
#> 28:    278    iris      ncol     6      c  500.0352 secs
#> 29:    279    iris      ncol     6      d  500.0345 secs
#> 30:    296    iris      ncol    10      a 1500.0370 secs
#> 31:    320    iris      ncol    14      e  500.0335 secs
#> 32:    340    iris      ncol    18      e  500.0333 secs
#> 33:    347    iris      ncol    20      b 1500.0332 secs
#> 34:    363    iris      ncol    23      c  500.0340 secs
#> 35:    369    iris      ncol    24      d  500.0338 secs
#> 36:    373    iris      ncol    25      c  500.0340 secs
#> 37:    387    iris      ncol    28      b 1500.0339 secs
#> 38:    410    iris      ncol    32      e  500.0337 secs
#> 39:    421    iris      ncol    35      a 1500.0342 secs
#> 40:    436    iris      ncol    38      a 1500.0346 secs
#> 41:    444    iris      ncol    39      d  500.0348 secs
#> 42:    448    iris      ncol    40      c  500.0347 secs
#> 43:    456    iris      ncol    42      a 1500.0344 secs
#> 44:    459    iris      ncol    42      d  500.0347 secs
#> 45:    467    iris      ncol    44      b 1500.0404 secs
#> 46:    468    iris      ncol    44      c  500.0344 secs
#> 47:    475    iris      ncol    45      e  500.0345 secs
#> 48:    482    iris      ncol    47      b 1500.0344 secs
#> 49:    492    iris      ncol    49      b 1500.0337 secs
#> 50:    499    iris      ncol    50      d  500.0334 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 36m 56.7s
#>   Total    : 4d 03h 20m 18.5s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1106.8479
#>   2:      2    iris      nrow     1      b estimated 1090.8023
#>   3:      3    iris      nrow     1      c estimated  338.4810
#>   4:      4    iris      nrow     1      d estimated  319.0667
#>   5:      5    iris      nrow     1      e estimated  318.3105
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1382.7488
#> 497:    497    iris      ncol    50      b estimated 1389.7987
#> 498:    498    iris      ncol    50      c estimated  614.5316
#> 499:    499    iris      ncol    50      d  observed  500.0334
#> 500:    500    iris      ncol    50      e estimated  576.0835
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.7s
#>   Remaining: 3d 17h 36m 56.7s
#>   Parallel : 0d 08h 58m 21.8s
#>   Total    : 4d 03h 20m 18.5s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1422.1264
#>   2:    461 estimated 1418.7326
#>   3:    462 estimated 1415.7463
#>   4:    472 estimated 1415.4601
#>   5:    457 estimated 1414.9463
#>  ---                           
#> 446:    185 estimated  133.3621
#> 447:    164 estimated  133.2322
#> 448:    204 estimated  131.1216
#> 449:    174 estimated  130.8823
#> 450:    179 estimated  129.7365
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8479    47
#>   2:      2 estimated 1090.8023    51
#>   3:      3 estimated  338.4810    36
#>   4:      4 estimated  319.0667    33
#>   5:      5 estimated  318.3105    32
#>  ---                                 
#> 446:    495 estimated  583.0181    15
#> 447:    496 estimated 1382.7488    20
#> 448:    497 estimated 1389.7987    14
#> 449:    498 estimated  614.5316     4
#> 450:    500 estimated  576.0835    26
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8479    47
#>   2:      2 estimated 1090.8023    51
#>   3:      3 estimated  338.4810    36
#>   4:      4 estimated  319.0667    33
#>   5:      5 estimated  318.3105    32
#>  ---                                 
#> 446:    495 estimated  583.0181    15
#> 447:    496 estimated 1382.7488    20
#> 448:    497 estimated 1389.7987    14
#> 449:    498 estimated  614.5316     4
#> 450:    500 estimated  576.0835    26
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    47 3494.242
#>  2:    51 3596.217
#>  3:    36 3599.568
#>  4:    33 3595.148
#>  5:    32 3599.795
#>  6:    48 3491.433
#>  7:    37 3592.701
#>  8:    70 3490.530
#>  9:    52 3599.940
#> 10:    54 3589.663
#> 11:    68 3495.524
#> 12:    71 3487.476
#> 13:    55 3587.560
#> 14:    69 3493.050
#> 15:    72 3485.333
#> 16:    46 3519.368
#> 17:    50 3599.563
#> 18:    38 3595.917
#> 19:    53 3593.789
#> 20:    34 3599.452
#> 21:    43 3571.988
#> 22:    63 3519.814
#> 23:    64 3517.256
#> 24:    39 3599.637
#> 25:    60 3537.747
#> 26:    62 3530.926
#> 27:    40 3596.901
#> 28:    56 3576.556
#> 29:    57 3566.696
#> 30:    49 3487.412
#> 31:    42 3582.514
#> 32:    59 3549.423
#> 33:    35 3599.971
#> 34:    41 3586.451
#> 35:    58 3558.297
#> 36:    61 3533.269
#> 37:    44 3542.048
#> 38:    66 3507.811
#> 39:    45 3541.248
#> 40:    65 3513.049
#> 41:    67 3505.518
#> 42:    27 3595.010
#> 43:    24 3596.566
#> 44:    26 3583.357
#> 45:    25 3598.827
#> 46:    23 3598.635
#> 47:    28 3570.452
#> 48:    74 3599.664
#> 49:    75 3599.282
#> 50:    12 3562.026
#> 51:     8 3595.369
#> 52:    31 3599.841
#> 53:    20 3521.727
#> 54:     7 3597.432
#> 55:    11 3563.774
#> 56:     5 3595.981
#> 57:    10 3576.682
#> 58:     6 3599.231
#> 59:    73 3479.882
#> 60:    79 3501.322
#> 61:    82 3599.274
#> 62:    81 3478.050
#> 63:    78 3509.132
#> 64:     3 3599.570
#> 65:    76 3589.144
#> 66:    85 3587.717
#> 67:    86 3576.300
#> 68:    91 1895.919
#> 69:    80 3487.113
#> 70:    77 3527.761
#> 71:    88 3563.170
#> 72:    87 3571.890
#> 73:    84 3597.201
#> 74:     2 3599.714
#> 75:    89 3555.498
#> 76:     1 3599.712
#> 77:    83 3599.588
#> 78:    90 3529.868
#> 79:     4 3597.500
#> 80:     9 3587.491
#> 81:    22 3595.002
#> 82:    19 3568.139
#> 83:    16 3586.117
#> 84:    30 3551.776
#> 85:    18 3572.269
#> 86:    13 3595.898
#> 87:    21 3597.905
#> 88:    17 3579.888
#> 89:    29 3561.114
#> 90:    15 3593.040
#> 91:    14 3598.108
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1106.8479     5
#>   2:      2 estimated 1090.8023     5
#>   3:      3 estimated  338.4810     9
#>   4:      4 estimated  319.0667     1
#>   5:      5 estimated  318.3105    10
#>  ---                                 
#> 446:    495 estimated  583.0181     4
#> 447:    496 estimated 1382.7488     9
#> 448:    497 estimated 1389.7987     5
#> 449:    498 estimated  614.5316     6
#> 450:    500 estimated  576.0835     2
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:     5 32301.74
#>  2:     9 32225.05
#>  3:     1 32230.81
#>  4:    10 32289.77
#>  5:     3 32289.60
#>  6:     2 32225.21
#>  7:     4 32225.12
#>  8:     7 32301.26
#>  9:     6 32301.79
#> 10:     8 32226.38