Estimate Remaining Runtimes

Estimates the runtimes of jobs using the random forest implemented in ranger. Observed runtimes are retrieved from the Registry and runtimes are predicted for unfinished jobs.

The estimated remaining time is calculated in the print method. You may also pass n here to determine the number of parallel jobs which is then used in a simple Longest Processing Time (LPT) algorithm to give an estimate for the parallel runtime.

Usage

estimateRuntimes(tab, ..., reg = getDefaultRegistry())

# S3 method for class 'RuntimeEstimate'
print(x, n = 1L, ...)

Arguments

tab

[data.table]
Table with column “job.id” and additional columns to predict the runtime. Observed runtimes will be looked up in the registry and serve as dependent variable. All columns in tab except “job.id” will be passed to ranger as independent variables to fit the model.

...

[ANY]
Additional parameters passed to ranger. Ignored for the print method.

reg

[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

x

[RuntimeEstimate]
Object to print.

n

[integer(1)]
Number of parallel jobs to assume for runtime estimation.

Value

[RuntimeEstimate] which is a list with two named elements: “runtimes” is a data.table with columns “job.id”, “runtime” (in seconds) and “type” (“estimated” if runtime is estimated, “observed” if runtime was observed). The other element of the list named “model”] contains the fitted random forest object.

See also

Examples

# Create a simple toy registry
set.seed(1)
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE, seed = 1)
#> No readable configuration file found
#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'
addProblem(name = "iris", data = iris, fun = function(data, ...) nrow(data), reg = tmp)
#> Adding problem 'iris'
addAlgorithm(name = "nrow", function(instance, ...) nrow(instance), reg = tmp)
#> Adding algorithm 'nrow'
addAlgorithm(name = "ncol", function(instance, ...) ncol(instance), reg = tmp)
#> Adding algorithm 'ncol'
addExperiments(algo.designs = list(nrow = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'nrow'[250] x repls[1]) ...
addExperiments(algo.designs = list(ncol = data.table::CJ(x = 1:50, y = letters[1:5])), reg = tmp)
#> Adding 250 experiments ('iris'[1] x 'ncol'[250] x repls[1]) ...

# We use the job parameters to predict runtimes
tab = unwrap(getJobPars(reg = tmp))

# First we need to submit some jobs so that the forest can train on some data.
# Thus, we just sample some jobs from the registry while grouping by factor variables.
library(data.table)
ids = tab[, .SD[sample(nrow(.SD), 5)], by = c("problem", "algorithm", "y")]
setkeyv(ids, "job.id")
submitJobs(ids, reg = tmp)
#> Submitting 50 jobs in 50 chunks using cluster functions 'Interactive' ...
waitForJobs(reg = tmp)
#> [1] TRUE

# We "simulate" some more realistic runtimes here to demonstrate the functionality:
# - Algorithm "ncol" is 5 times more expensive than "nrow"
# - x has no effect on the runtime
# - If y is "a" or "b", the runtimes are really high
runtime = function(algorithm, x, y) {
  ifelse(algorithm == "nrow", 100L, 500L) + 1000L * (y %in% letters[1:2])
}
tmp$status[ids, done := done + tab[ids, runtime(algorithm, x, y)]]
#> Key: <job.id>
#>      job.id def.id  submitted    started       done  error mem.used resource.id
#>       <int>  <int>      <num>      <num>      <num> <char>    <num>       <int>
#>   1:      1      1         NA         NA         NA   <NA>       NA          NA
#>   2:      2      2         NA         NA         NA   <NA>       NA          NA
#>   3:      3      3         NA         NA         NA   <NA>       NA          NA
#>   4:      4      4         NA         NA         NA   <NA>       NA          NA
#>   5:      5      5         NA         NA         NA   <NA>       NA          NA
#>  ---                                                                           
#> 496:    496    496         NA         NA         NA   <NA>       NA          NA
#> 497:    497    497         NA         NA         NA   <NA>       NA          NA
#> 498:    498    498         NA         NA         NA   <NA>       NA          NA
#> 499:    499    499 1755115230 1755115230 1755115730   <NA>       NA           1
#> 500:    500    500         NA         NA         NA   <NA>       NA          NA
#>           batch.id log.file                            job.hash job.name  repl
#>             <char>   <char>                              <char>   <char> <int>
#>   1:          <NA>     <NA>                                <NA>     <NA>     1
#>   2:          <NA>     <NA>                                <NA>     <NA>     1
#>   3:          <NA>     <NA>                                <NA>     <NA>     1
#>   4:          <NA>     <NA>                                <NA>     <NA>     1
#>   5:          <NA>     <NA>                                <NA>     <NA>     1
#>  ---                                                                          
#> 496:          <NA>     <NA>                                <NA>     <NA>     1
#> 497:          <NA>     <NA>                                <NA>     <NA>     1
#> 498:          <NA>     <NA>                                <NA>     <NA>     1
#> 499: cfInteractive     <NA> job653db15edcd3cad8ed900c3e1ba2c25d     <NA>     1
#> 500:          <NA>     <NA>                                <NA>     <NA>     1
rjoin(sjoin(tab, ids), getJobStatus(ids, reg = tmp)[, c("job.id", "time.running")])
#> Key: <job.id>
#>     job.id problem algorithm     x      y   time.running
#>      <int>  <char>    <char> <int> <char>     <difftime>
#>  1:     32    iris      nrow     7      b 1100.0365 secs
#>  2:     42    iris      nrow     9      b 1100.0369 secs
#>  3:     47    iris      nrow    10      b 1100.0358 secs
#>  4:     66    iris      nrow    14      a 1100.0356 secs
#>  5:     73    iris      nrow    15      c  100.0345 secs
#>  6:     75    iris      nrow    15      e  100.0378 secs
#>  7:     86    iris      nrow    18      a 1100.0372 secs
#>  8:    100    iris      nrow    20      e  100.0396 secs
#>  9:    101    iris      nrow    21      a 1100.0352 secs
#> 10:    103    iris      nrow    21      c  100.0361 secs
#> 11:    123    iris      nrow    25      c  100.0348 secs
#> 12:    125    iris      nrow    25      e  100.0357 secs
#> 13:    161    iris      nrow    33      a 1100.0387 secs
#> 14:    165    iris      nrow    33      e  100.0362 secs
#> 15:    169    iris      nrow    34      d  100.0352 secs
#> 16:    183    iris      nrow    37      c  100.0352 secs
#> 17:    184    iris      nrow    37      d  100.0344 secs
#> 18:    203    iris      nrow    41      c  100.0362 secs
#> 19:    207    iris      nrow    42      b 1100.0375 secs
#> 20:    209    iris      nrow    42      d  100.0358 secs
#> 21:    220    iris      nrow    44      e  100.0343 secs
#> 22:    227    iris      nrow    46      b 1100.0346 secs
#> 23:    229    iris      nrow    46      d  100.0395 secs
#> 24:    231    iris      nrow    47      a 1100.0379 secs
#> 25:    244    iris      nrow    49      d  100.0371 secs
#> 26:    260    iris      ncol     2      e  500.0342 secs
#> 27:    276    iris      ncol     6      a 1500.0347 secs
#> 28:    278    iris      ncol     6      c  500.0336 secs
#> 29:    279    iris      ncol     6      d  500.0355 secs
#> 30:    296    iris      ncol    10      a 1500.0369 secs
#> 31:    320    iris      ncol    14      e  500.0344 secs
#> 32:    340    iris      ncol    18      e  500.0339 secs
#> 33:    347    iris      ncol    20      b 1500.0342 secs
#> 34:    363    iris      ncol    23      c  500.0346 secs
#> 35:    369    iris      ncol    24      d  500.0373 secs
#> 36:    373    iris      ncol    25      c  500.0351 secs
#> 37:    387    iris      ncol    28      b 1500.0344 secs
#> 38:    410    iris      ncol    32      e  500.0340 secs
#> 39:    421    iris      ncol    35      a 1500.0376 secs
#> 40:    436    iris      ncol    38      a 1500.0367 secs
#> 41:    444    iris      ncol    39      d  500.0361 secs
#> 42:    448    iris      ncol    40      c  500.0339 secs
#> 43:    456    iris      ncol    42      a 1500.0343 secs
#> 44:    459    iris      ncol    42      d  500.0341 secs
#> 45:    467    iris      ncol    44      b 1500.0368 secs
#> 46:    468    iris      ncol    44      c  500.0361 secs
#> 47:    475    iris      ncol    45      e  500.0350 secs
#> 48:    482    iris      ncol    47      b 1500.0344 secs
#> 49:    492    iris      ncol    49      b 1500.0344 secs
#> 50:    499    iris      ncol    50      d  500.0367 secs
#>     job.id problem algorithm     x      y   time.running

# Estimate runtimes:
est = estimateRuntimes(tab, reg = tmp)
print(est)
#> Runtime Estimate for 500 jobs with 1 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 34m 18.0s
#>   Total    : 4d 03h 17m 39.8s
rjoin(tab, est$runtimes)
#> Key: <job.id>
#>      job.id problem algorithm     x      y      type   runtime
#>       <int>  <char>    <char> <int> <char>    <fctr>     <num>
#>   1:      1    iris      nrow     1      a estimated 1104.8500
#>   2:      2    iris      nrow     1      b estimated 1088.8042
#>   3:      3    iris      nrow     1      c estimated  337.8423
#>   4:      4    iris      nrow     1      d estimated  317.7083
#>   5:      5    iris      nrow     1      e estimated  317.6193
#>  ---                                                          
#> 496:    496    iris      ncol    50      a estimated 1385.9494
#> 497:    497    iris      ncol    50      b estimated 1392.8387
#> 498:    498    iris      ncol    50      c estimated  615.6928
#> 499:    499    iris      ncol    50      d  observed  500.0367
#> 500:    500    iris      ncol    50      e estimated  575.4848
print(est, n = 10)
#> Runtime Estimate for 500 jobs with 10 CPUs
#>   Done     : 0d 09h 43m 21.8s
#>   Remaining: 3d 17h 34m 18.0s
#>   Parallel : 0d 08h 58m 6.8s
#>   Total    : 4d 03h 17m 39.8s

# Submit jobs with longest runtime first:
ids = est$runtimes[type == "estimated"][order(runtime, decreasing = TRUE)]
print(ids)
#>      job.id      type   runtime
#>       <int>    <fctr>     <num>
#>   1:    466 estimated 1423.2101
#>   2:    461 estimated 1419.8167
#>   3:    462 estimated 1416.7555
#>   4:    472 estimated 1416.2688
#>   5:    457 estimated 1415.9555
#>  ---                           
#> 446:    199 estimated  131.9122
#> 447:    194 estimated  131.7520
#> 448:    174 estimated  131.3631
#> 449:    204 estimated  131.0019
#> 450:    179 estimated  130.2164
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into chunks with runtime < 1h
ids = est$runtimes[type == "estimated"]
ids[, chunk := binpack(runtime, 3600)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1104.8500    48
#>   2:      2 estimated 1088.8042    52
#>   3:      3 estimated  337.8423    45
#>   4:      4 estimated  317.7083    31
#>   5:      5 estimated  317.6193    68
#>  ---                                 
#> 446:    495 estimated  580.4194    18
#> 447:    496 estimated 1385.9494    18
#> 448:    497 estimated 1392.8387    13
#> 449:    498 estimated  615.6928     4
#> 450:    500 estimated  575.4848    26
print(ids)
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1104.8500    48
#>   2:      2 estimated 1088.8042    52
#>   3:      3 estimated  337.8423    45
#>   4:      4 estimated  317.7083    31
#>   5:      5 estimated  317.6193    68
#>  ---                                 
#> 446:    495 estimated  580.4194    18
#> 447:    496 estimated 1385.9494    18
#> 448:    497 estimated 1392.8387    13
#> 449:    498 estimated  615.6928     4
#> 450:    500 estimated  575.4848    26
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    48 3487.877
#>  2:    52 3599.612
#>  3:    45 3521.512
#>  4:    31 3599.119
#>  5:    68 3495.551
#>  6:    49 3483.815
#>  7:    54 3589.042
#>  8:    32 3596.968
#>  9:    69 3493.787
#> 10:    55 3585.721
#> 11:    33 3592.511
#> 12:    70 3491.079
#> 13:    53 3592.177
#> 14:    56 3575.052
#> 15:    51 3595.649
#> 16:    71 3488.287
#> 17:    46 3524.016
#> 18:    50 3599.126
#> 19:    38 3594.692
#> 20:    64 3517.482
#> 21:    66 3507.411
#> 22:    43 3566.310
#> 23:    65 3512.957
#> 24:    67 3506.666
#> 25:    42 3576.510
#> 26:    35 3599.326
#> 27:    47 3494.808
#> 28:    39 3593.590
#> 29:    36 3599.060
#> 30:    57 3564.483
#> 31:    41 3582.135
#> 32:    60 3535.143
#> 33:    61 3532.756
#> 34:    40 3585.258
#> 35:    58 3553.027
#> 36:    59 3544.596
#> 37:    44 3530.196
#> 38:    34 3598.772
#> 39:    37 3599.895
#> 40:    63 3519.450
#> 41:    62 3522.291
#> 42:    27 3594.418
#> 43:    25 3589.788
#> 44:    24 3598.533
#> 45:    26 3597.139
#> 46:    23 3599.735
#> 47:    28 3568.166
#> 48:    75 3591.715
#> 49:    20 3520.245
#> 50:    74 3599.941
#> 51:     8 3593.904
#> 52:     6 3594.201
#> 53:    72 3497.314
#> 54:     7 3599.762
#> 55:    12 3560.954
#> 56:     5 3597.605
#> 57:    11 3572.588
#> 58:    10 3575.449
#> 59:     4 3598.606
#> 60:    80 3482.491
#> 61:    81 3473.349
#> 62:    82 3599.157
#> 63:    73 3479.907
#> 64:    78 3504.541
#> 65:     2 3599.694
#> 66:    86 3574.787
#> 67:    89 3552.306
#> 68:    91 1880.895
#> 69:    79 3492.477
#> 70:    76 3577.062
#> 71:    88 3560.715
#> 72:    87 3568.311
#> 73:     3 3598.409
#> 74:    85 3581.902
#> 75:    83 3597.804
#> 76:    77 3522.940
#> 77:    90 3525.247
#> 78:    84 3590.032
#> 79:     1 3599.192
#> 80:     9 3584.969
#> 81:    22 3597.366
#> 82:    18 3574.235
#> 83:    13 3597.753
#> 84:    21 3598.348
#> 85:    19 3570.430
#> 86:    14 3593.560
#> 87:    30 3548.993
#> 88:    17 3579.983
#> 89:    29 3564.044
#> 90:    15 3597.080
#> 91:    16 3592.268
#>     chunk  runtime
if (FALSE) { # \dontrun{
submitJobs(ids, reg = tmp)
} # }

# Group jobs into 10 chunks with similar runtime
ids = est$runtimes[type == "estimated"]
ids[, chunk := lpt(runtime, 10)]
#> Key: <job.id>
#>      job.id      type   runtime chunk
#>       <int>    <fctr>     <num> <int>
#>   1:      1 estimated 1104.8500    10
#>   2:      2 estimated 1088.8042     9
#>   3:      3 estimated  337.8423     1
#>   4:      4 estimated  317.7083     7
#>   5:      5 estimated  317.6193     6
#>  ---                                 
#> 446:    495 estimated  580.4194     4
#> 447:    496 estimated 1385.9494     9
#> 448:    497 estimated 1392.8387    10
#> 449:    498 estimated  615.6928     5
#> 450:    500 estimated  575.4848    10
print(ids[, list(runtime = sum(runtime)), by = chunk])
#>     chunk  runtime
#>     <int>    <num>
#>  1:    10 32271.89
#>  2:     9 32286.41
#>  3:     1 32210.66
#>  4:     7 32213.56
#>  5:     6 32286.54
#>  6:     5 32210.57
#>  7:     4 32210.29
#>  8:     3 32271.90
#>  9:     2 32286.81
#> 10:     8 32209.41