Summarize the Computational Status

This function gives an encompassing overview over the computational status on your system. The status can be one or many of the following:

“defined”: Jobs which are defined via batchMap or addExperiments, but are not yet submitted.

“submitted”: Jobs which are submitted to the batch system via submitJobs, scheduled for execution.

“started”: Jobs which have been started.

“done”: Jobs which terminated successfully.

“error”: Jobs which terminated with an exception.

“running”: Jobs which are listed by the cluster functions to be running on the live system. Not supported for all cluster functions.

“queued”: Jobs which are listed by the cluster functions to be queued on the live system. Not supported for all cluster functions.

“system”: Jobs which are listed by the cluster functions to be queued or running. Not supported for all cluster functions.

“expired”: Jobs which have been submitted, but vanished from the live system. Note that this is determined heuristically and may include some false positives.

Here, a job which terminated successfully counts towards the jobs which are submitted, started and done. To retrieve the corresponding job ids, see findJobs.

getStatus(ids = NULL, reg = getDefaultRegistry())

Arguments

ids
[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to all jobs. Invalid ids are ignored.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

[data.table] (with class “Status” for printing).

See also

findJobs<`3`>

Other debug: getErrorMessages, grepLogs, killJobs, resetJobs, showLog, testJob

Examples

tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

fun = function(i) if (i == 3) stop(i) else i
ids = batchMap(fun, i = 1:5, reg = tmp)

#> Adding 5 jobs ...

submitJobs(ids = 1:4, reg = tmp)

#> Submitting 4 jobs in 4 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Setting seed to 1 ...

#> ### [bt]: Setting seed to 2 ...

#> ### [bt]: Setting seed to 3 ...

#> ### [bt]: Setting seed to 4 ...

waitForJobs(reg = tmp)

#> [1] FALSE

tab = getStatus(reg = tmp)
print(tab)

#> Status for 5 jobs:
#>   Submitted    : 4 ( 80.0%)
#>   -- Queued    : 0 (  0.0%)
#>   -- Started   : 4 ( 80.0%)
#>   ---- Running : 0 (  0.0%)
#>   ---- Done    : 3 ( 60.0%)
#>   ---- Error   : 1 ( 20.0%)
#>   ---- Expired : 0 (  0.0%)

str(tab)

#> Classes 'Status', 'data.table' and 'data.frame':	1 obs. of  9 variables:
#>  $ defined  : int 5
#>  $ submitted: int 4
#>  $ started  : int 4
#>  $ done     : int 3
#>  $ error    : int 1
#>  $ queued   : int 0
#>  $ running  : int 0
#>  $ expired  : int 0
#>  $ system   : int 0
#>  - attr(*, ".internal.selfref")=<externalptr>