Load a Registry from the File System

Loads a registry from its file.dir.

Multiple R sessions accessing the same registry simultaneously can lead to database inconsistencies. Here, it does not matter if the sessions run on the same system or different systems via a file system mount.

If you just need to check on the status or peek into some preliminary results, you can load the registry in a read-only mode by setting writeable to FALSE. All operations that need to change the registry will raise an exception in this mode. Files communicated back by the computational nodes are parsed to update the registry in memory, but remain on the file system in order to be read and cleaned up by an R session with read-write access.

A heuristic tries to detect if the registry has been altered in the background. However, you should not completely rely on it. Thus, set to writeable to TRUE if and only if you are absolutely sure other R processes are terminated.

loadRegistry(file.dir, work.dir = NULL, conf.file = findConfFile(),
  make.default = TRUE, writeable = FALSE)

Arguments

file.dir
[character(1)] Path where all files of the registry are saved. Default is directory “registry” in the current working directory. The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well.

If you pass NA, a temporary directory will be used. This way, you can create disposable registries for btlapply or examples. By default, the temporary directory tempdir() will be used. If you want to use another directory, e.g. a directory which is shared between nodes, you can set it in your configuration file by setting the variable temp.dir.

work.dir
[character(1)] Working directory for R process for running jobs. Defaults to the working directory currently set during Registry construction (see getwd). loadRegistry uses the stored work.dir, but you may also explicitly overwrite it, e.g., after switching to another system.

The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well.

conf.file
[character(1)] Path to a configuration file which is sourced while the registry is created. In the configuration file you can define how batchtools interacts with the system via ClusterFunctions. Separating the configuration of the underlying host system from the R code allows to easily move computation to another site.

The file lookup is implemented in findConfFile which returns the first file found of the following candidates:

File “batchtools.conf.R” in the path specified by the environment variable “R_BATCHTOOLS_SEARCH_PATH”.

File “batchtools.conf.R” in the current working directory.

File “config.R” in the OS dependent user configuration directory as reported by via rappdirs::user_config_dir("batchtools", expand = FALSE) (e.g., on linux this usually resolves to “~/.config/batchtools/config.R”).

“.batchtools.conf.R” in the home directory (“~”).

Set to character(0) if you want to suppress reading any configuration file. If a configuration file is found, it gets sourced inside the environment of the registry after the defaults for all variables are set. Therefore you can set and overwrite slots, e.g. default.resources = list(walltime = 3600) to set default resources or “max.concurrent.jobs” to limit the number of jobs allowed to run simultaneously on the system.

make.default
[logical(1)] If set to TRUE, the created registry is saved inside the package namespace and acts as default registry. You might want to switch this off if you work with multiple registries simultaneously. Default is TRUE.

writeable
[logical(1)] Loads the registry in read-write mode. Default is FALSE.

Value

[Registry].

See also

Other Registry: clearRegistry, getDefaultRegistry, makeRegistry, removeRegistry, saveRegistry, sweepRegistry, syncRegistry