ClusterFunctions for Remote SSH Execution

Jobs are spawned by starting multiple R sessions via Rscript over SSH. If the hostname of the Worker equals “localhost”, Rscript is called directly so that you do not need to have an SSH client installed.

makeClusterFunctionsSSH(workers, fs.latency = 65)

Arguments

workers
[list of Worker] List of Workers as constructed with Worker.

fs.latency
[numeric(1)] Expected maximum latency of the file system, in seconds. Set to a positive number for network file systems like NFS which enables more robust (but also more expensive) mechanisms to access files and directories. Usually safe to set to NA to disable the heuristic, e.g. if you are working on a local file system.

Value

[ClusterFunctions].

Note

If you use a custom “.ssh/config” file, make sure your ProxyCommand passes ‘-q’ to ssh, otherwise each output will end with the message “Killed by signal 1” and this will break the communication with the nodes.

See also

Other ClusterFunctions: makeClusterFunctionsDocker, makeClusterFunctionsInteractive, makeClusterFunctionsLSF, makeClusterFunctionsMulticore, makeClusterFunctionsOpenLava, makeClusterFunctionsSGE, makeClusterFunctionsSlurm, makeClusterFunctionsSocket, makeClusterFunctionsTORQUE, makeClusterFunctions

Examples

# NOT RUN {
# cluster functions for multicore execution on the local machine
makeClusterFunctionsSSH(list(Worker$new("localhost", ncpus = 2)))
# }