ExperimentRegistry Constructor

makeExperimentRegistry constructs a special Registry which is suitable for the definition of large scale computer experiments.

Each experiments consists of a Problem and an Algorithm. These can be parametrized with addExperiments to actually define computational jobs.

makeExperimentRegistry(file.dir = "registry", work.dir = getwd(),
  conf.file = findConfFile(), packages = character(0L),
  namespaces = character(0L), source = character(0L),
  load = character(0L), seed = NULL, make.default = TRUE)

Arguments

file.dir
[character(1)]
Path where all files of the registry are saved. Default is directory “registry” in the current working directory. The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well.

If you pass NA, a temporary directory will be used. This way, you can create disposable registries for btlapply or examples. By default, the temporary directory tempdir() will be used. If you want to use another directory, e.g. a directory which is shared between nodes, you can set it in your configuration file by setting the variable temp.dir.

work.dir
[character(1)]
Working directory for R process for running jobs. Defaults to the working directory currently set during Registry construction (see getwd). loadRegistry uses the stored work.dir, but you may also explicitly overwrite it, e.g., after switching to another system.

The provided path will get normalized unless it is given relative to the home directory (i.e., starting with “~”). Note that some templates do not handle relative paths well.

conf.file
[character(1)]
Path to a configuration file which is sourced while the registry is created. In the configuration file you can define how batchtools interacts with the system via ClusterFunctions. Separating the configuration of the underlying host system from the R code allows to easily move computation to another site.

The file lookup is implemented in the internal (but exported) function findConfFile which returns the first file found of the following candidates:

File “batchtools.conf.R” in the path specified by the environment variable “R_BATCHTOOLS_SEARCH_PATH”.

File “batchtools.conf.R” in the current working directory.

File “config.R” in the user configuration directory as reported by rappdirs::user_config_dir("batchtools", expand = FALSE) (depending on OS, e.g., on linux this usually resolves to “~/.config/batchtools/config.R”).

“.batchtools.conf.R” in the home directory (“~”).

“config.R” in the site config directory as reported by rappdirs::site_config_dir("batchtools") (depending on OS). This file can be used for admins to set sane defaults for a computation site.

Set to NA if you want to suppress reading any configuration file. If a configuration file is found, it gets sourced inside the environment of the registry after the defaults for all variables are set. Therefore you can set and overwrite slots, e.g. default.resources = list(walltime = 3600) to set default resources or “max.concurrent.jobs” to limit the number of jobs allowed to run simultaneously on the system.

packages
[character]
Packages that will always be loaded on each node. Uses require internally. Default is character(0).

namespaces
[character]
Same as packages, but the packages will not be attached. Uses requireNamespace internally. Default is character(0).

source
[character]
Files which should be sourced on the slaves prior to executing a job. Calls sys.source using the .GlobalEnv.

load
[character]
Files which should be loaded on the slaves prior to executing a job. Calls load using the .GlobalEnv.

seed
[integer(1)]
Start seed for jobs. Each job uses the (seed + job.id) as seed. Default is a random number in the range [1, .Machine$integer.max/2].

make.default
[logical(1)]
If set to TRUE, the created registry is saved inside the package namespace and acts as default registry. You might want to switch this off if you work with multiple registries simultaneously. Default is TRUE.

Value

[ExperimentRegistry].

Examples

tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Created registry in '/tmp/batchtools-example/reg' using cluster functions 'Interactive'

# Definde one problem, two algorithms and add them with some parameters:
addProblem(reg = tmp, "p1",
  fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd))

#> Adding problem 'p1'

addAlgorithm(reg = tmp, "a1", fun = function(job, data, instance, ...) mean(instance))

#> Adding algorithm 'a1'

addAlgorithm(reg = tmp, "a2", fun = function(job, data, instance, ...) median(instance))

#> Adding algorithm 'a2'

ids = addExperiments(reg = tmp, list(p1 = CJ(n = c(50, 100), mean = -2:2, sd = 1:4)))

#> Adding 40 experiments ('p1'[40] x 'a1'[1] x repls[1]) ...

#> Adding 40 experiments ('p1'[40] x 'a2'[1] x repls[1]) ...

# Overview over defined experiments:
tmp$problems

#> [1] "p1"

tmp$algorithms

#> [1] "a1" "a2"

summarizeExperiments(reg = tmp)

#> Key: <problem, algorithm>
#>    problem algorithm .count
#>     <char>    <char>  <int>
#> 1:      p1        a1     40
#> 2:      p1        a2     40

summarizeExperiments(reg = tmp, by = c("problem", "algorithm", "n"))

#> Key: <problem, algorithm, n>
#>    problem algorithm     n .count
#>     <char>    <char> <num>  <int>
#> 1:      p1        a1    50     20
#> 2:      p1        a1   100     20
#> 3:      p1        a2    50     20
#> 4:      p1        a2   100     20

ids = findExperiments(prob.pars = (n == 50), reg = tmp)
print(unwrap(getJobPars(ids, reg = tmp)))

#> Key: <job.id>
#>     job.id problem algorithm     n  mean    sd
#>      <int>  <char>    <char> <num> <int> <int>
#>  1:      1      p1        a1    50    -2     1
#>  2:      2      p1        a1    50    -2     2
#>  3:      3      p1        a1    50    -2     3
#>  4:      4      p1        a1    50    -2     4
#>  5:      5      p1        a1    50    -1     1
#>  6:      6      p1        a1    50    -1     2
#>  7:      7      p1        a1    50    -1     3
#>  8:      8      p1        a1    50    -1     4
#>  9:      9      p1        a1    50     0     1
#> 10:     10      p1        a1    50     0     2
#> 11:     11      p1        a1    50     0     3
#> 12:     12      p1        a1    50     0     4
#> 13:     13      p1        a1    50     1     1
#> 14:     14      p1        a1    50     1     2
#> 15:     15      p1        a1    50     1     3
#> 16:     16      p1        a1    50     1     4
#> 17:     17      p1        a1    50     2     1
#> 18:     18      p1        a1    50     2     2
#> 19:     19      p1        a1    50     2     3
#> 20:     20      p1        a1    50     2     4
#> 21:     41      p1        a2    50    -2     1
#> 22:     42      p1        a2    50    -2     2
#> 23:     43      p1        a2    50    -2     3
#> 24:     44      p1        a2    50    -2     4
#> 25:     45      p1        a2    50    -1     1
#> 26:     46      p1        a2    50    -1     2
#> 27:     47      p1        a2    50    -1     3
#> 28:     48      p1        a2    50    -1     4
#> 29:     49      p1        a2    50     0     1
#> 30:     50      p1        a2    50     0     2
#> 31:     51      p1        a2    50     0     3
#> 32:     52      p1        a2    50     0     4
#> 33:     53      p1        a2    50     1     1
#> 34:     54      p1        a2    50     1     2
#> 35:     55      p1        a2    50     1     3
#> 36:     56      p1        a2    50     1     4
#> 37:     57      p1        a2    50     2     1
#> 38:     58      p1        a2    50     2     2
#> 39:     59      p1        a2    50     2     3
#> 40:     60      p1        a2    50     2     4
#>     job.id problem algorithm     n  mean    sd

# Submit jobs
submitJobs(reg = tmp)

#> Submitting 80 jobs in 80 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 1 (seed = 31063) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 2 (seed = 31064) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 3 (seed = 31065) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 4 (seed = 31066) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 5 (seed = 31067) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 6 (seed = 31068) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 7 (seed = 31069) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 8 (seed = 31070) ...

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 9 (seed = 31071) ...

#> 

#> Submitting [=====>---------------------------------------------]  11% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 10 (seed = 31072) ...

#> 

#> Submitting [=====>---------------------------------------------]  12% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 11 (seed = 31073) ...

#> 

#> Submitting [======>--------------------------------------------]  14% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 12 (seed = 31074) ...

#> 

#> Submitting [=======>-------------------------------------------]  15% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 13 (seed = 31075) ...

#> 

#> Submitting [=======>-------------------------------------------]  16% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 14 (seed = 31076) ...

#> 

#> Submitting [========>------------------------------------------]  18% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 15 (seed = 31077) ...

#> 

#> Submitting [=========>-----------------------------------------]  19% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 16 (seed = 31078) ...

#> 

#> Submitting [=========>-----------------------------------------]  20% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 17 (seed = 31079) ...

#> 

#> Submitting [==========>----------------------------------------]  21% eta:  2s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 18 (seed = 31080) ...

#> 

#> Submitting [==========>----------------------------------------]  22% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 19 (seed = 31081) ...

#> 

#> Submitting [===========>---------------------------------------]  24% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 20 (seed = 31082) ...

#> 

#> Submitting [============>--------------------------------------]  25% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 21 (seed = 31083) ...

#> 

#> Submitting [============>--------------------------------------]  26% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 22 (seed = 31084) ...

#> 

#> Submitting [=============>-------------------------------------]  28% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 23 (seed = 31085) ...

#> 

#> Submitting [==============>------------------------------------]  29% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 24 (seed = 31086) ...

#> 

#> Submitting [==============>------------------------------------]  30% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 25 (seed = 31087) ...

#> 

#> Submitting [===============>-----------------------------------]  31% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 26 (seed = 31088) ...

#> 

#> Submitting [================>----------------------------------]  32% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 27 (seed = 31089) ...

#> 

#> Submitting [================>----------------------------------]  34% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 28 (seed = 31090) ...

#> 

#> Submitting [=================>---------------------------------]  35% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 29 (seed = 31091) ...

#> 

#> Submitting [=================>---------------------------------]  36% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 30 (seed = 31092) ...

#> 

#> Submitting [==================>--------------------------------]  38% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 31 (seed = 31093) ...

#> 

#> Submitting [===================>-------------------------------]  39% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 32 (seed = 31094) ...

#> 

#> Submitting [===================>-------------------------------]  40% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 33 (seed = 31095) ...

#> 

#> Submitting [====================>------------------------------]  41% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 34 (seed = 31096) ...

#> 

#> Submitting [=====================>-----------------------------]  42% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 35 (seed = 31097) ...

#> 

#> Submitting [=====================>-----------------------------]  44% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 36 (seed = 31098) ...

#> 

#> Submitting [======================>----------------------------]  45% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 37 (seed = 31099) ...

#> 

#> Submitting [=======================>---------------------------]  46% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 38 (seed = 31100) ...

#> 

#> Submitting [=======================>---------------------------]  48% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 39 (seed = 31101) ...

#> 

#> Submitting [========================>--------------------------]  49% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a1' on problem 'p1' for job 40 (seed = 31102) ...

#> 

#> Submitting [=========================>-------------------------]  50% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 41 (seed = 31103) ...

#> 

#> Submitting [=========================>-------------------------]  51% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 42 (seed = 31104) ...

#> 

#> Submitting [==========================>------------------------]  52% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 43 (seed = 31105) ...

#> 

#> Submitting [==========================>------------------------]  54% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 44 (seed = 31106) ...

#> 

#> Submitting [===========================>-----------------------]  55% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 45 (seed = 31107) ...

#> 

#> Submitting [============================>----------------------]  56% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 46 (seed = 31108) ...

#> 

#> Submitting [============================>----------------------]  57% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 47 (seed = 31109) ...

#> 

#> Submitting [=============================>---------------------]  59% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 48 (seed = 31110) ...

#> 

#> Submitting [==============================>--------------------]  60% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 49 (seed = 31111) ...

#> 

#> Submitting [==============================>--------------------]  61% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 50 (seed = 31112) ...

#> 

#> Submitting [===============================>-------------------]  62% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 51 (seed = 31113) ...

#> 

#> Submitting [================================>------------------]  64% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 52 (seed = 31114) ...

#> 

#> Submitting [================================>------------------]  65% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 53 (seed = 31115) ...

#> 

#> Submitting [=================================>-----------------]  66% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 54 (seed = 31116) ...

#> 

#> Submitting [=================================>-----------------]  68% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 55 (seed = 31117) ...

#> 

#> Submitting [==================================>----------------]  69% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 56 (seed = 31118) ...

#> 

#> Submitting [===================================>---------------]  70% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 57 (seed = 31119) ...

#> 

#> Submitting [===================================>---------------]  71% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 58 (seed = 31120) ...

#> 

#> Submitting [====================================>--------------]  72% eta:  1s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 59 (seed = 31121) ...

#> 

#> Submitting [=====================================>-------------]  74% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 60 (seed = 31122) ...

#> 

#> Submitting [=====================================>-------------]  75% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 61 (seed = 31123) ...

#> 

#> Submitting [======================================>------------]  76% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 62 (seed = 31124) ...

#> 

#> Submitting [=======================================>-----------]  78% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 63 (seed = 31125) ...

#> 

#> Submitting [=======================================>-----------]  79% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 64 (seed = 31126) ...

#> 

#> Submitting [========================================>----------]  80% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 65 (seed = 31127) ...

#> 

#> Submitting [========================================>----------]  81% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 66 (seed = 31128) ...

#> 

#> Submitting [=========================================>---------]  82% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 67 (seed = 31129) ...

#> 

#> Submitting [==========================================>--------]  84% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 68 (seed = 31130) ...

#> 

#> Submitting [==========================================>--------]  85% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 69 (seed = 31131) ...

#> 

#> Submitting [===========================================>-------]  86% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 70 (seed = 31132) ...

#> 

#> Submitting [============================================>------]  88% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 71 (seed = 31133) ...

#> 

#> Submitting [============================================>------]  89% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 72 (seed = 31134) ...

#> 

#> Submitting [=============================================>-----]  90% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 73 (seed = 31135) ...

#> 

#> Submitting [==============================================>----]  91% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 74 (seed = 31136) ...

#> 

#> Submitting [==============================================>----]  92% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 75 (seed = 31137) ...

#> 

#> Submitting [===============================================>---]  94% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 76 (seed = 31138) ...

#> 

#> Submitting [===============================================>---]  95% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 77 (seed = 31139) ...

#> 

#> Submitting [================================================>--]  96% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 78 (seed = 31140) ...

#> 

#> Submitting [=================================================>-]  98% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 79 (seed = 31141) ...

#> 

#> Submitting [=================================================>-]  99% eta:  0s

#> ### [bt]: Generating problem instance for problem 'p1' ...

#> ### [bt]: Applying algorithm 'a2' on problem 'p1' for job 80 (seed = 31142) ...

#> 

#> Submitting [===================================================] 100% eta:  0s

#>                                                                               

#> 

waitForJobs(reg = tmp)

#> [1] TRUE

# Reduce the results of algorithm a1
ids.mean = findExperiments(algo.name = "a1", reg = tmp)
reduceResults(ids.mean, fun = function(aggr, res, ...) c(aggr, res), reg = tmp)

#>  [1] -1.91851189 -2.13263714 -1.15482212 -1.42420576 -0.87479570 -1.12460273
#>  [7] -1.31687864 -0.62168084 -0.09295716  0.38459832  0.72874542  0.01556041
#> [13]  1.00853645  1.10392756  0.64841617  1.97504271  1.86944683  2.12704529
#> [19]  2.13826884  2.59386422 -1.99038250 -1.85429185 -1.94528527 -2.45781568
#> [25] -0.97365152 -0.75482372 -1.12424658 -1.71008408  0.01493960  0.01634059
#> [31]  0.18217789  0.35560917  1.06718880  0.98909517  1.18238431  0.75007631
#> [37]  2.02479556  2.22462125  1.83573521  1.66398732

# Join info table with all results and calculate mean of results
# grouped by n and algorithm
ids = findDone(reg = tmp)
pars = unwrap(getJobPars(ids, reg = tmp))
results = unwrap(reduceResultsDataTable(ids, fun = function(res) list(res = res), reg = tmp))
tab = ljoin(pars, results)
tab[, list(mres = mean(res)), by = c("n", "algorithm")]

#>        n algorithm        mres
#>    <num>    <char>       <num>
#> 1:    50        a1  0.19661801
#> 2:   100        a1 -0.02518150
#> 3:    50        a2 -0.07334272
#> 4:   100        a2  0.06760491