Create a SubmitJobResult

This function is only intended for use in your own cluster functions implementation.

Use this function in your implementation of makeClusterFunctions to create a return value for the submitJob function.

makeSubmitJobResult(status, batch.id, log.file = NA_character_,
  msg = NA_character_)

Arguments

status
[integer(1)] Launch status of job. 0 means success, codes between 1 and 100 are temporary errors and any error greater than 100 is a permanent failure.

batch.id
[character()] Unique id of this job on batch system, as given by the batch system. Must be globally unique so that the job can be terminated using just this information. For array jobs, this may be a vector of length equal to the number of jobs in the array.

log.file
[character()] Log file. If NA, defaults to [job.hash].log. Some cluster functions set this for array jobs.

msg
[character(1)] Optional error message in case status is not equal to 0. Default is “OK”, “TEMPERROR”, “ERROR”, depending on status.

Value

[SubmitJobResult]. A list, containing status, batch.id and msg.

See also

Other ClusterFunctionsHelper: cfBrewTemplate, cfHandleUnknownSubmitError, cfKillJob, cfReadBrewTemplate, makeClusterFunctions, runOSCommand