Apply Functions on Results

Applies a function on the results of your finished jobs and thereby collects them in a list or data.table. The later requires the provided function to return a list (or data.frame) of scalar values. See rbindlist for features and limitations of the aggregation.

If not all jobs are terminated, the respective result will be NULL.

reduceResultsList(ids = NULL, fun = NULL, ..., missing.val,
  reg = getDefaultRegistry())

reduceResultsDataTable(ids = NULL, fun = NULL, ..., missing.val,
  reg = getDefaultRegistry())

Arguments

ids
[data.frame or integer]
A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to the return value of findDone. Invalid ids are ignored.

fun
[function]
Function to apply to each result. The result is passed unnamed as first argument. If NULL, the identity is used. If the function has the formal argument “job”, the Job/Experiment is also passed to the function.

...
[ANY]
Additional arguments passed to to function fun.

missing.val
[ANY]
Value to impute as result for a job which is not finished. If not provided and a result is missing, an exception is raised.

reg
[Registry]
Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

reduceResultsList returns a list of the results in the same order as the provided ids. reduceResultsDataTable returns a data.table with columns “job.id” and additional result columns created via rbindlist, sorted by “job.id”.

Note

If you have thousands of jobs, disabling the progress bar (options(batchtools.progress = FALSE)) can significantly increase the performance.

See also

Examples

### Example 1 - reduceResultsList
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Created registry in '/tmp/batchtools-example/reg1' using cluster functions 'Interactive'

batchMap(function(x) x^2, x = 1:10, reg = tmp)

#> Adding 10 jobs ...

submitJobs(reg = tmp)

#> Submitting 10 jobs in 10 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Setting seed to 1 ...

#> ### [bt]: Setting seed to 2 ...

#> ### [bt]: Setting seed to 3 ...

#> ### [bt]: Setting seed to 4 ...

#> ### [bt]: Setting seed to 5 ...

#> ### [bt]: Setting seed to 6 ...

#> ### [bt]: Setting seed to 7 ...

#> ### [bt]: Setting seed to 8 ...

#> ### [bt]: Setting seed to 9 ...

#> 

#> Submitting [=============================================>-----]  90% eta:  0s

#> ### [bt]: Setting seed to 10 ...

#> 

#> Submitting [===================================================] 100% eta:  0s

#>                                                                               

#> 

waitForJobs(reg = tmp)

#> [1] TRUE

reduceResultsList(fun = sqrt, reg = tmp)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] 6
#> 
#> [[7]]
#> [1] 7
#> 
#> [[8]]
#> [1] 8
#> 
#> [[9]]
#> [1] 9
#> 
#> [[10]]
#> [1] 10
#> 

### Example 2 - reduceResultsDataTable
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '~/.batchtools.conf.R' ...

#> Created registry in '/tmp/batchtools-example/reg2' using cluster functions 'Interactive'

# add first problem
fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd)
addProblem("rnorm", fun = fun, reg = tmp)

#> Adding problem 'rnorm'

# add second problem
fun = function(job, data, n, lambda, ...) rexp(n, rate = lambda)
addProblem("rexp", fun = fun, reg = tmp)

#> Adding problem 'rexp'

# add first algorithm
fun = function(instance, method, ...) if (method == "mean") mean(instance) else median(instance)
addAlgorithm("average", fun = fun, reg = tmp)

#> Adding algorithm 'average'

# add second algorithm
fun = function(instance, ...) sd(instance)
addAlgorithm("deviation", fun = fun, reg = tmp)

#> Adding algorithm 'deviation'

# define problem and algorithm designs
prob.designs = algo.designs = list()
prob.designs$rnorm = CJ(n = 100, mean = -1:1, sd = 1:5)
prob.designs$rexp = data.table(n = 100, lambda = 1:5)
algo.designs$average = data.table(method = c("mean", "median"))
algo.designs$deviation = data.table()

# add experiments and submit
addExperiments(prob.designs, algo.designs, reg = tmp)

#> Adding 30 experiments ('rnorm'[15] x 'average'[2] x repls[1]) ...

#> Adding 15 experiments ('rnorm'[15] x 'deviation'[1] x repls[1]) ...

#> Adding 10 experiments ('rexp'[5] x 'average'[2] x repls[1]) ...

#> Adding 5 experiments ('rexp'[5] x 'deviation'[1] x repls[1]) ...

submitJobs(reg = tmp)

#> Submitting 60 jobs in 60 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 1 (seed = 13297) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 2 (seed = 13298) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 3 (seed = 13299) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 4 (seed = 13300) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 5 (seed = 13301) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 6 (seed = 13302) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 7 (seed = 13303) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 8 (seed = 13304) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 9 (seed = 13305) ...

#> 

#> Submitting [=======>-------------------------------------------]  15% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 10 (seed = 13306) ...

#> 

#> Submitting [=======>-------------------------------------------]  17% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 11 (seed = 13307) ...

#> 

#> Submitting [========>------------------------------------------]  18% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 12 (seed = 13308) ...

#> 

#> Submitting [=========>-----------------------------------------]  20% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 13 (seed = 13309) ...

#> 

#> Submitting [==========>----------------------------------------]  22% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 14 (seed = 13310) ...

#> 

#> Submitting [===========>---------------------------------------]  23% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 15 (seed = 13311) ...

#> 

#> Submitting [============>--------------------------------------]  25% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 16 (seed = 13312) ...

#> 

#> Submitting [=============>-------------------------------------]  27% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 17 (seed = 13313) ...

#> 

#> Submitting [=============>-------------------------------------]  28% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 18 (seed = 13314) ...

#> 

#> Submitting [==============>------------------------------------]  30% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 19 (seed = 13315) ...

#> 

#> Submitting [===============>-----------------------------------]  32% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 20 (seed = 13316) ...

#> 

#> Submitting [================>----------------------------------]  33% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 21 (seed = 13317) ...

#> 

#> Submitting [=================>---------------------------------]  35% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 22 (seed = 13318) ...

#> 

#> Submitting [==================>--------------------------------]  37% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 23 (seed = 13319) ...

#> 

#> Submitting [===================>-------------------------------]  38% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 24 (seed = 13320) ...

#> 

#> Submitting [===================>-------------------------------]  40% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 25 (seed = 13321) ...

#> 

#> Submitting [====================>------------------------------]  42% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 26 (seed = 13322) ...

#> 

#> Submitting [=====================>-----------------------------]  43% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 27 (seed = 13323) ...

#> 

#> Submitting [======================>----------------------------]  45% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 28 (seed = 13324) ...

#> 

#> Submitting [=======================>---------------------------]  47% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 29 (seed = 13325) ...

#> 

#> Submitting [========================>--------------------------]  48% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 30 (seed = 13326) ...

#> 

#> Submitting [=========================>-------------------------]  50% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 31 (seed = 13327) ...

#> 

#> Submitting [=========================>-------------------------]  52% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 32 (seed = 13328) ...

#> 

#> Submitting [==========================>------------------------]  53% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 33 (seed = 13329) ...

#> 

#> Submitting [===========================>-----------------------]  55% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 34 (seed = 13330) ...

#> 

#> Submitting [============================>----------------------]  57% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 35 (seed = 13331) ...

#> 

#> Submitting [=============================>---------------------]  58% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 36 (seed = 13332) ...

#> 

#> Submitting [==============================>--------------------]  60% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 37 (seed = 13333) ...

#> 

#> Submitting [==============================>--------------------]  62% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 38 (seed = 13334) ...

#> 

#> Submitting [===============================>-------------------]  63% eta:  1s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 39 (seed = 13335) ...

#> 

#> Submitting [================================>------------------]  65% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 40 (seed = 13336) ...

#> 

#> Submitting [=================================>-----------------]  67% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 41 (seed = 13337) ...

#> 

#> Submitting [==================================>----------------]  68% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 42 (seed = 13338) ...

#> 

#> Submitting [===================================>---------------]  70% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 43 (seed = 13339) ...

#> 

#> Submitting [====================================>--------------]  72% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 44 (seed = 13340) ...

#> 

#> Submitting [====================================>--------------]  73% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 45 (seed = 13341) ...

#> 

#> Submitting [=====================================>-------------]  75% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 46 (seed = 13342) ...

#> 

#> Submitting [======================================>------------]  77% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 47 (seed = 13343) ...

#> 

#> Submitting [=======================================>-----------]  78% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 48 (seed = 13344) ...

#> 

#> Submitting [========================================>----------]  80% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 49 (seed = 13345) ...

#> 

#> Submitting [=========================================>---------]  82% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 50 (seed = 13346) ...

#> 

#> Submitting [=========================================>---------]  83% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 51 (seed = 13347) ...

#> 

#> Submitting [==========================================>--------]  85% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 52 (seed = 13348) ...

#> 

#> Submitting [===========================================>-------]  87% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 53 (seed = 13349) ...

#> 

#> Submitting [============================================>------]  88% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 54 (seed = 13350) ...

#> 

#> Submitting [=============================================>-----]  90% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 55 (seed = 13351) ...

#> 

#> Submitting [==============================================>----]  92% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 56 (seed = 13352) ...

#> 

#> Submitting [===============================================>---]  93% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 57 (seed = 13353) ...

#> 

#> Submitting [===============================================>---]  95% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 58 (seed = 13354) ...

#> 

#> Submitting [================================================>--]  97% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 59 (seed = 13355) ...

#> 

#> Submitting [=================================================>-]  98% eta:  0s

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 60 (seed = 13356) ...

#> 

#> Submitting [===================================================] 100% eta:  0s

#>                                                                               

#> 

# collect results and join them with problem and algorithm paramters
res = ijoin(
  getJobPars(reg = tmp),
  reduceResultsDataTable(reg = tmp, fun = function(x) list(res = x))
)
unwrap(res, sep = ".")

#> Key: <job.id>
#>     job.id problem algorithm prob.pars.n prob.pars.mean prob.pars.sd
#>      <int>  <char>    <char>       <num>          <int>        <int>
#>  1:      1   rnorm   average         100             -1            1
#>  2:      2   rnorm   average         100             -1            1
#>  3:      3   rnorm   average         100             -1            2
#>  4:      4   rnorm   average         100             -1            2
#>  5:      5   rnorm   average         100             -1            3
#>  6:      6   rnorm   average         100             -1            3
#>  7:      7   rnorm   average         100             -1            4
#>  8:      8   rnorm   average         100             -1            4
#>  9:      9   rnorm   average         100             -1            5
#> 10:     10   rnorm   average         100             -1            5
#> 11:     11   rnorm   average         100              0            1
#> 12:     12   rnorm   average         100              0            1
#> 13:     13   rnorm   average         100              0            2
#> 14:     14   rnorm   average         100              0            2
#> 15:     15   rnorm   average         100              0            3
#> 16:     16   rnorm   average         100              0            3
#> 17:     17   rnorm   average         100              0            4
#> 18:     18   rnorm   average         100              0            4
#> 19:     19   rnorm   average         100              0            5
#> 20:     20   rnorm   average         100              0            5
#> 21:     21   rnorm   average         100              1            1
#> 22:     22   rnorm   average         100              1            1
#> 23:     23   rnorm   average         100              1            2
#> 24:     24   rnorm   average         100              1            2
#> 25:     25   rnorm   average         100              1            3
#> 26:     26   rnorm   average         100              1            3
#> 27:     27   rnorm   average         100              1            4
#> 28:     28   rnorm   average         100              1            4
#> 29:     29   rnorm   average         100              1            5
#> 30:     30   rnorm   average         100              1            5
#> 31:     31   rnorm deviation         100             -1            1
#> 32:     32   rnorm deviation         100             -1            2
#> 33:     33   rnorm deviation         100             -1            3
#> 34:     34   rnorm deviation         100             -1            4
#> 35:     35   rnorm deviation         100             -1            5
#> 36:     36   rnorm deviation         100              0            1
#> 37:     37   rnorm deviation         100              0            2
#> 38:     38   rnorm deviation         100              0            3
#> 39:     39   rnorm deviation         100              0            4
#> 40:     40   rnorm deviation         100              0            5
#> 41:     41   rnorm deviation         100              1            1
#> 42:     42   rnorm deviation         100              1            2
#> 43:     43   rnorm deviation         100              1            3
#> 44:     44   rnorm deviation         100              1            4
#> 45:     45   rnorm deviation         100              1            5
#> 46:     46    rexp   average         100             NA           NA
#> 47:     47    rexp   average         100             NA           NA
#> 48:     48    rexp   average         100             NA           NA
#> 49:     49    rexp   average         100             NA           NA
#> 50:     50    rexp   average         100             NA           NA
#> 51:     51    rexp   average         100             NA           NA
#> 52:     52    rexp   average         100             NA           NA
#> 53:     53    rexp   average         100             NA           NA
#> 54:     54    rexp   average         100             NA           NA
#> 55:     55    rexp   average         100             NA           NA
#> 56:     56    rexp deviation         100             NA           NA
#> 57:     57    rexp deviation         100             NA           NA
#> 58:     58    rexp deviation         100             NA           NA
#> 59:     59    rexp deviation         100             NA           NA
#> 60:     60    rexp deviation         100             NA           NA
#>     job.id problem algorithm prob.pars.n prob.pars.mean prob.pars.sd
#>     prob.pars.lambda algo.pars.method   result.res
#>                <int>           <char>        <num>
#>  1:               NA             mean -0.842949204
#>  2:               NA           median -0.915451498
#>  3:               NA             mean -1.013829715
#>  4:               NA           median -1.026744582
#>  5:               NA             mean -0.544082757
#>  6:               NA           median -1.228882756
#>  7:               NA             mean -1.758195152
#>  8:               NA           median -1.017889502
#>  9:               NA             mean -0.721865242
#> 10:               NA           median -1.009905112
#> 11:               NA             mean -0.066968266
#> 12:               NA           median -0.001324423
#> 13:               NA             mean  0.116369676
#> 14:               NA           median -0.239593821
#> 15:               NA             mean -0.275914994
#> 16:               NA           median -0.219673804
#> 17:               NA             mean -0.390600924
#> 18:               NA           median -0.431722873
#> 19:               NA             mean -0.862369986
#> 20:               NA           median  0.310564934
#> 21:               NA             mean  0.919685582
#> 22:               NA           median  1.137638493
#> 23:               NA             mean  0.717251280
#> 24:               NA           median  1.000696628
#> 25:               NA             mean  1.079885160
#> 26:               NA           median  1.256486092
#> 27:               NA             mean  0.056569707
#> 28:               NA           median  1.511908035
#> 29:               NA             mean  1.772567494
#> 30:               NA           median  0.576546942
#> 31:               NA             <NA>  1.066915639
#> 32:               NA             <NA>  1.956072085
#> 33:               NA             <NA>  3.129048220
#> 34:               NA             <NA>  4.091646305
#> 35:               NA             <NA>  5.567688298
#> 36:               NA             <NA>  1.033162728
#> 37:               NA             <NA>  2.290784363
#> 38:               NA             <NA>  3.011799409
#> 39:               NA             <NA>  3.724917848
#> 40:               NA             <NA>  5.164125697
#> 41:               NA             <NA>  0.913834076
#> 42:               NA             <NA>  1.840190154
#> 43:               NA             <NA>  3.225940976
#> 44:               NA             <NA>  4.250388672
#> 45:               NA             <NA>  4.820613875
#> 46:                1             mean  0.976800412
#> 47:                1           median  0.561309444
#> 48:                2             mean  0.512651915
#> 49:                2           median  0.332483764
#> 50:                3             mean  0.363899709
#> 51:                3           median  0.282966778
#> 52:                4             mean  0.225608448
#> 53:                4           median  0.167649514
#> 54:                5             mean  0.186122531
#> 55:                5           median  0.128949969
#> 56:                1             <NA>  0.938268024
#> 57:                2             <NA>  0.488269167
#> 58:                3             <NA>  0.318698244
#> 59:                4             <NA>  0.223306907
#> 60:                5             <NA>  0.175971064
#>     prob.pars.lambda algo.pars.method   result.res