Apply Functions on Results

Applies a function on the results of your finished jobs and thereby collects them in a list or data.table. The later requires the provided function to return a list (or data.frame) of scalar values. See rbindlist for features and limitations of the aggregation.

If not all jobs are terminated, the respective result will be NULL.

reduceResultsList(ids = NULL, fun = NULL, ..., missing.val,
  reg = getDefaultRegistry())

reduceResultsDataTable(ids = NULL, fun = NULL, ..., flatten = NULL,
  missing.val, reg = getDefaultRegistry())

Arguments

ids
[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to the return value of findDone. Invalid ids are ignored.

fun
[function] Function to apply to each result. The result is passed unnamed as first argument. If NULL, the identity is used. If the function has the formal argument “job”, the Job/Experiment is also passed to the function.

...
[ANY] Additional arguments passed to to function fun.

missing.val
[ANY] Value to impute as result for a job which is not finished. If not provided and a result is missing, an exception is raised.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

flatten
[logical(1)] Transform results into separate data.table columns. Defaults to TRUE if all results are (a list of) scalar atomics, Otherwise each row of the column will hold a named list.

Value

reduceResultsList returns a list of the results in the same order as the provided ids. reduceResultsDataTable returns a data.table with columns “job.id” and additional result columns created via rbindlist, sorted by “job.id”.

Note

If you have thousands of jobs, disabling the progress bar (options(batchtools.progress = FALSE)) can significantly increase the performance.

See also

reduceResults

Other Results: batchMapResults, loadResult, reduceResults

Examples

### Example 1 - reduceResultsList
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry7c5879ea2f45' using cluster functions 'Interactive'

batchMap(function(x) x^2, x = 1:10, reg = tmp)

#> Adding 10 jobs ...

submitJobs(reg = tmp)

#> Submitting 10 jobs in 10 chunks using cluster functions 'Interactive' ...

waitForJobs(reg = tmp)

#> [1] TRUE

reduceResultsList(fun = sqrt, reg = tmp)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] 6
#> 
#> [[7]]
#> [1] 7
#> 
#> [[8]]
#> [1] 8
#> 
#> [[9]]
#> [1] 9
#> 
#> [[10]]
#> [1] 10
#> 

### Example 2 - reduceResultsDataTable
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/scratch/registry7c583ba50d49' using cluster functions 'Interactive'

# add first problem
fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd)
addProblem("rnorm", fun = fun, reg = tmp)

#> Adding problem 'rnorm'

# add second problem
fun = function(job, data, n, lambda, ...) rexp(n, rate = lambda)
addProblem("rexp", fun = fun, reg = tmp)

#> Adding problem 'rexp'

# add first algorithm
fun = function(instance, method, ...) if (method == "mean") mean(instance) else median(instance)
addAlgorithm("average", fun = fun, reg = tmp)

#> Adding algorithm 'average'

# add second algorithm
fun = function(instance, ...) sd(instance)
addAlgorithm("deviation", fun = fun, reg = tmp)

#> Adding algorithm 'deviation'

# define problem and algorithm designs
prob.designs = algo.designs = list()
prob.designs$rnorm = CJ(n = 100, mean = -1:1, sd = 1:5)
prob.designs$rexp = data.table(n = 100, lambda = 1:5)
algo.designs$average = data.table(method = c("mean", "median"))
algo.designs$deviation = data.table()

# add experiments and submit
addExperiments(prob.designs, algo.designs, reg = tmp)

#> Adding 30 experiments ('rnorm'[15] x 'average'[2] x repls[1]) ...

#> Adding 15 experiments ('rnorm'[15] x 'deviation'[1] x repls[1]) ...

#> Adding 10 experiments ('rexp'[5] x 'average'[2] x repls[1]) ...

#> Adding 5 experiments ('rexp'[5] x 'deviation'[1] x repls[1]) ...

submitJobs(reg = tmp)

#> Submitting 60 jobs in 60 chunks using cluster functions 'Interactive' ...

# collect results and join them # with problem and algorithm paramters
ijoin(
  getJobPars(reg = tmp),
  reduceResultsDataTable(reg = tmp, fun = function(x) list(res = x))
)

#>     job.id problem algorithm     n  mean    sd method lambda         res
#>      <int>  <fctr>    <fctr> <num> <int> <int> <char>  <int>       <num>
#>  1:      1   rnorm   average   100    -1     1   mean     NA -1.03092335
#>  2:      2   rnorm   average   100    -1     1 median     NA -1.09193236
#>  3:      3   rnorm   average   100    -1     2   mean     NA -0.67885778
#>  4:      4   rnorm   average   100    -1     2 median     NA -0.79856997
#>  5:      5   rnorm   average   100    -1     3   mean     NA -1.67194910
#>  6:      6   rnorm   average   100    -1     3 median     NA -1.54841149
#>  7:      7   rnorm   average   100    -1     4   mean     NA -1.87850580
#>  8:      8   rnorm   average   100    -1     4 median     NA -0.62072520
#>  9:      9   rnorm   average   100    -1     5   mean     NA -1.06936614
#> 10:     10   rnorm   average   100    -1     5 median     NA -0.51983799
#> 11:     11   rnorm   average   100     0     1   mean     NA  0.01717658
#> 12:     12   rnorm   average   100     0     1 median     NA  0.07116721
#> 13:     13   rnorm   average   100     0     2   mean     NA  0.05934450
#> 14:     14   rnorm   average   100     0     2 median     NA  0.09531172
#> 15:     15   rnorm   average   100     0     3   mean     NA -0.08971524
#> 16:     16   rnorm   average   100     0     3 median     NA  0.65120875
#> 17:     17   rnorm   average   100     0     4   mean     NA  0.58022370
#> 18:     18   rnorm   average   100     0     4 median     NA  1.04515507
#> 19:     19   rnorm   average   100     0     5   mean     NA  0.34010298
#> 20:     20   rnorm   average   100     0     5 median     NA  0.77314379
#> 21:     21   rnorm   average   100     1     1   mean     NA  1.09241448
#> 22:     22   rnorm   average   100     1     1 median     NA  1.10664442
#> 23:     23   rnorm   average   100     1     2   mean     NA  0.83024516
#> 24:     24   rnorm   average   100     1     2 median     NA  1.13635050
#> 25:     25   rnorm   average   100     1     3   mean     NA  0.57841953
#> 26:     26   rnorm   average   100     1     3 median     NA  1.14485380
#> 27:     27   rnorm   average   100     1     4   mean     NA  0.97827546
#> 28:     28   rnorm   average   100     1     4 median     NA  0.23115031
#> 29:     29   rnorm   average   100     1     5   mean     NA  0.86009224
#> 30:     30   rnorm   average   100     1     5 median     NA  1.52628409
#> 31:     31   rnorm deviation   100    -1     1     NA     NA  1.00789728
#> 32:     32   rnorm deviation   100    -1     2     NA     NA  1.93576611
#> 33:     33   rnorm deviation   100    -1     3     NA     NA  3.06922268
#> 34:     34   rnorm deviation   100    -1     4     NA     NA  4.52038823
#> 35:     35   rnorm deviation   100    -1     5     NA     NA  4.72436529
#> 36:     36   rnorm deviation   100     0     1     NA     NA  0.98484606
#> 37:     37   rnorm deviation   100     0     2     NA     NA  2.16470621
#> 38:     38   rnorm deviation   100     0     3     NA     NA  2.90804627
#> 39:     39   rnorm deviation   100     0     4     NA     NA  4.51527817
#> 40:     40   rnorm deviation   100     0     5     NA     NA  5.10206112
#> 41:     41   rnorm deviation   100     1     1     NA     NA  0.98106938
#> 42:     42   rnorm deviation   100     1     2     NA     NA  2.07716612
#> 43:     43   rnorm deviation   100     1     3     NA     NA  2.83850959
#> 44:     44   rnorm deviation   100     1     4     NA     NA  4.01032896
#> 45:     45   rnorm deviation   100     1     5     NA     NA  4.69268499
#> 46:     46    rexp   average   100    NA    NA   mean      1  1.18753506
#> 47:     47    rexp   average   100    NA    NA median      1  0.66870853
#> 48:     48    rexp   average   100    NA    NA   mean      2  0.36403045
#> 49:     49    rexp   average   100    NA    NA median      2  0.35749457
#> 50:     50    rexp   average   100    NA    NA   mean      3  0.35583467
#> 51:     51    rexp   average   100    NA    NA median      3  0.23735029
#> 52:     52    rexp   average   100    NA    NA   mean      4  0.30408752
#> 53:     53    rexp   average   100    NA    NA median      4  0.16366624
#> 54:     54    rexp   average   100    NA    NA   mean      5  0.21408259
#> 55:     55    rexp   average   100    NA    NA median      5  0.13066213
#> 56:     56    rexp deviation   100    NA    NA     NA      1  1.03778337
#> 57:     57    rexp deviation   100    NA    NA     NA      2  0.50034279
#> 58:     58    rexp deviation   100    NA    NA     NA      3  0.40798912
#> 59:     59    rexp deviation   100    NA    NA     NA      4  0.23243882
#> 60:     60    rexp deviation   100    NA    NA     NA      5  0.23601934
#>     job.id problem algorithm     n  mean    sd method lambda         res