Apply Functions on Results

Applies a function on the results of your finished jobs and thereby collects them in a list or data.table. The later requires the provided function to return a list (or data.frame) of scalar values. See rbindlist for features and limitations of the aggregation.

If not all jobs are terminated, the respective result will be NULL.

reduceResultsList(ids = NULL, fun = NULL, ..., missing.val,
  reg = getDefaultRegistry())

reduceResultsDataTable(ids = NULL, fun = NULL, ..., missing.val,
  reg = getDefaultRegistry())

Arguments

ids
[data.frame or integer] A data.frame (or data.table) with a column named “job.id”. Alternatively, you may also pass a vector of integerish job ids. If not set, defaults to the return value of findDone. Invalid ids are ignored.

fun
[function] Function to apply to each result. The result is passed unnamed as first argument. If NULL, the identity is used. If the function has the formal argument “job”, the Job/Experiment is also passed to the function.

...
[ANY] Additional arguments passed to to function fun.

missing.val
[ANY] Value to impute as result for a job which is not finished. If not provided and a result is missing, an exception is raised.

reg
[Registry] Registry. If not explicitly passed, uses the default registry (see setDefaultRegistry).

Value

reduceResultsList returns a list of the results in the same order as the provided ids. reduceResultsDataTable returns a data.table with columns “job.id” and additional result columns created via rbindlist, sorted by “job.id”.

Note

If you have thousands of jobs, disabling the progress bar (options(batchtools.progress = FALSE)) can significantly increase the performance.

See also

reduceResults

Other Results: batchMapResults, loadResult, reduceResults

Examples

### Example 1 - reduceResultsList
tmp = makeRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg1' using cluster functions 'Interactive'

batchMap(function(x) x^2, x = 1:10, reg = tmp)

#> Adding 10 jobs ...

submitJobs(reg = tmp)

#> Submitting 10 jobs in 10 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Setting seed to 1 ...

#> ### [bt]: Setting seed to 2 ...

#> ### [bt]: Setting seed to 3 ...

#> ### [bt]: Setting seed to 4 ...

#> ### [bt]: Setting seed to 5 ...

#> ### [bt]: Setting seed to 6 ...

#> ### [bt]: Setting seed to 7 ...

#> ### [bt]: Setting seed to 8 ...

#> ### [bt]: Setting seed to 9 ...

#> ### [bt]: Setting seed to 10 ...

waitForJobs(reg = tmp)

#> [1] TRUE

reduceResultsList(fun = sqrt, reg = tmp)

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 3
#> 
#> [[4]]
#> [1] 4
#> 
#> [[5]]
#> [1] 5
#> 
#> [[6]]
#> [1] 6
#> 
#> [[7]]
#> [1] 7
#> 
#> [[8]]
#> [1] 8
#> 
#> [[9]]
#> [1] 9
#> 
#> [[10]]
#> [1] 10
#> 

### Example 2 - reduceResultsDataTable
tmp = makeExperimentRegistry(file.dir = NA, make.default = FALSE)

#> Sourcing configuration file '/home/lang/.config/batchtools/config.R' ...

#> Created registry in '/tmp/batchtools-example/reg2' using cluster functions 'Interactive'

# add first problem
fun = function(job, data, n, mean, sd, ...) rnorm(n, mean = mean, sd = sd)
addProblem("rnorm", fun = fun, reg = tmp)

#> Adding problem 'rnorm'

# add second problem
fun = function(job, data, n, lambda, ...) rexp(n, rate = lambda)
addProblem("rexp", fun = fun, reg = tmp)

#> Adding problem 'rexp'

# add first algorithm
fun = function(instance, method, ...) if (method == "mean") mean(instance) else median(instance)
addAlgorithm("average", fun = fun, reg = tmp)

#> Adding algorithm 'average'

# add second algorithm
fun = function(instance, ...) sd(instance)
addAlgorithm("deviation", fun = fun, reg = tmp)

#> Adding algorithm 'deviation'

# define problem and algorithm designs
prob.designs = algo.designs = list()
prob.designs$rnorm = CJ(n = 100, mean = -1:1, sd = 1:5)
prob.designs$rexp = data.table(n = 100, lambda = 1:5)
algo.designs$average = data.table(method = c("mean", "median"))
algo.designs$deviation = data.table()

# add experiments and submit
addExperiments(prob.designs, algo.designs, reg = tmp)

#> Adding 30 experiments ('rnorm'[15] x 'average'[2] x repls[1]) ...

#> Adding 15 experiments ('rnorm'[15] x 'deviation'[1] x repls[1]) ...

#> Adding 10 experiments ('rexp'[5] x 'average'[2] x repls[1]) ...

#> Adding 5 experiments ('rexp'[5] x 'deviation'[1] x repls[1]) ...

submitJobs(reg = tmp)

#> Submitting 60 jobs in 60 chunks using cluster functions 'Interactive' ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 1 (seed = 4055) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 2 (seed = 4056) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 3 (seed = 4057) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 4 (seed = 4058) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 5 (seed = 4059) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 6 (seed = 4060) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 7 (seed = 4061) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 8 (seed = 4062) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 9 (seed = 4063) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 10 (seed = 4064) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 11 (seed = 4065) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 12 (seed = 4066) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 13 (seed = 4067) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 14 (seed = 4068) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 15 (seed = 4069) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 16 (seed = 4070) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 17 (seed = 4071) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 18 (seed = 4072) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 19 (seed = 4073) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 20 (seed = 4074) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 21 (seed = 4075) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 22 (seed = 4076) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 23 (seed = 4077) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 24 (seed = 4078) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 25 (seed = 4079) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 26 (seed = 4080) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 27 (seed = 4081) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 28 (seed = 4082) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 29 (seed = 4083) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rnorm' for job 30 (seed = 4084) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 31 (seed = 4085) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 32 (seed = 4086) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 33 (seed = 4087) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 34 (seed = 4088) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 35 (seed = 4089) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 36 (seed = 4090) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 37 (seed = 4091) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 38 (seed = 4092) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 39 (seed = 4093) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 40 (seed = 4094) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 41 (seed = 4095) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 42 (seed = 4096) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 43 (seed = 4097) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 44 (seed = 4098) ...

#> ### [bt]: Generating problem instance for problem 'rnorm' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rnorm' for job 45 (seed = 4099) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 46 (seed = 4100) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 47 (seed = 4101) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 48 (seed = 4102) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 49 (seed = 4103) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 50 (seed = 4104) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 51 (seed = 4105) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 52 (seed = 4106) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 53 (seed = 4107) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 54 (seed = 4108) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'average' on problem 'rexp' for job 55 (seed = 4109) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 56 (seed = 4110) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 57 (seed = 4111) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 58 (seed = 4112) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 59 (seed = 4113) ...

#> ### [bt]: Generating problem instance for problem 'rexp' ...

#> ### [bt]: Applying algorithm 'deviation' on problem 'rexp' for job 60 (seed = 4114) ...

# collect results and join them with problem and algorithm paramters
res = ijoin(
  getJobPars(reg = tmp),
  reduceResultsDataTable(reg = tmp, fun = function(x) list(res = x))
)
unwrap(res, sep = ".")

#>     job.id problem algorithm prob.pars.n prob.pars.mean prob.pars.sd
#>      <int>  <char>    <char>       <num>          <int>        <int>
#>  1:      1   rnorm   average         100             -1            1
#>  2:      2   rnorm   average         100             -1            1
#>  3:      3   rnorm   average         100             -1            2
#>  4:      4   rnorm   average         100             -1            2
#>  5:      5   rnorm   average         100             -1            3
#>  6:      6   rnorm   average         100             -1            3
#>  7:      7   rnorm   average         100             -1            4
#>  8:      8   rnorm   average         100             -1            4
#>  9:      9   rnorm   average         100             -1            5
#> 10:     10   rnorm   average         100             -1            5
#> 11:     11   rnorm   average         100              0            1
#> 12:     12   rnorm   average         100              0            1
#> 13:     13   rnorm   average         100              0            2
#> 14:     14   rnorm   average         100              0            2
#> 15:     15   rnorm   average         100              0            3
#> 16:     16   rnorm   average         100              0            3
#> 17:     17   rnorm   average         100              0            4
#> 18:     18   rnorm   average         100              0            4
#> 19:     19   rnorm   average         100              0            5
#> 20:     20   rnorm   average         100              0            5
#> 21:     21   rnorm   average         100              1            1
#> 22:     22   rnorm   average         100              1            1
#> 23:     23   rnorm   average         100              1            2
#> 24:     24   rnorm   average         100              1            2
#> 25:     25   rnorm   average         100              1            3
#> 26:     26   rnorm   average         100              1            3
#> 27:     27   rnorm   average         100              1            4
#> 28:     28   rnorm   average         100              1            4
#> 29:     29   rnorm   average         100              1            5
#> 30:     30   rnorm   average         100              1            5
#> 31:     31   rnorm deviation         100             -1            1
#> 32:     32   rnorm deviation         100             -1            2
#> 33:     33   rnorm deviation         100             -1            3
#> 34:     34   rnorm deviation         100             -1            4
#> 35:     35   rnorm deviation         100             -1            5
#> 36:     36   rnorm deviation         100              0            1
#> 37:     37   rnorm deviation         100              0            2
#> 38:     38   rnorm deviation         100              0            3
#> 39:     39   rnorm deviation         100              0            4
#> 40:     40   rnorm deviation         100              0            5
#> 41:     41   rnorm deviation         100              1            1
#> 42:     42   rnorm deviation         100              1            2
#> 43:     43   rnorm deviation         100              1            3
#> 44:     44   rnorm deviation         100              1            4
#> 45:     45   rnorm deviation         100              1            5
#> 46:     46    rexp   average         100             NA           NA
#> 47:     47    rexp   average         100             NA           NA
#> 48:     48    rexp   average         100             NA           NA
#> 49:     49    rexp   average         100             NA           NA
#> 50:     50    rexp   average         100             NA           NA
#> 51:     51    rexp   average         100             NA           NA
#> 52:     52    rexp   average         100             NA           NA
#> 53:     53    rexp   average         100             NA           NA
#> 54:     54    rexp   average         100             NA           NA
#> 55:     55    rexp   average         100             NA           NA
#> 56:     56    rexp deviation         100             NA           NA
#> 57:     57    rexp deviation         100             NA           NA
#> 58:     58    rexp deviation         100             NA           NA
#> 59:     59    rexp deviation         100             NA           NA
#> 60:     60    rexp deviation         100             NA           NA
#>     job.id problem algorithm prob.pars.n prob.pars.mean prob.pars.sd
#>     prob.pars.lambda algo.pars.method   result.res
#>                <int>           <char>        <num>
#>  1:               NA             mean -0.869992112
#>  2:               NA           median -1.025899375
#>  3:               NA             mean -0.975119673
#>  4:               NA           median -1.256491263
#>  5:               NA             mean -0.881804662
#>  6:               NA           median -1.518282095
#>  7:               NA             mean -1.103852846
#>  8:               NA           median -1.064450374
#>  9:               NA             mean -0.552667414
#> 10:               NA           median -0.887186301
#> 11:               NA             mean  0.005931588
#> 12:               NA           median -0.113783258
#> 13:               NA             mean -0.313519972
#> 14:               NA           median  0.388512978
#> 15:               NA             mean  0.196598803
#> 16:               NA           median  0.632184760
#> 17:               NA             mean -0.786698230
#> 18:               NA           median -0.486595008
#> 19:               NA             mean -0.556593643
#> 20:               NA           median -0.090044089
#> 21:               NA             mean  1.096183829
#> 22:               NA           median  0.923400543
#> 23:               NA             mean  0.871488642
#> 24:               NA           median  0.995850380
#> 25:               NA             mean  1.165503396
#> 26:               NA           median  1.040704016
#> 27:               NA             mean  1.225310112
#> 28:               NA           median  0.727732654
#> 29:               NA             mean  1.408755570
#> 30:               NA           median  1.315751178
#> 31:               NA               NA  0.954128756
#> 32:               NA               NA  2.109668878
#> 33:               NA               NA  3.547528683
#> 34:               NA               NA  3.652085923
#> 35:               NA               NA  5.026616864
#> 36:               NA               NA  1.049808162
#> 37:               NA               NA  2.073053899
#> 38:               NA               NA  2.756029245
#> 39:               NA               NA  3.596593906
#> 40:               NA               NA  5.592943503
#> 41:               NA               NA  1.095757731
#> 42:               NA               NA  1.860470749
#> 43:               NA               NA  3.114341794
#> 44:               NA               NA  3.980134036
#> 45:               NA               NA  4.988786861
#> 46:                1             mean  0.951403995
#> 47:                1           median  0.522297416
#> 48:                2             mean  0.559016998
#> 49:                2           median  0.344682942
#> 50:                3             mean  0.346089420
#> 51:                3           median  0.225886826
#> 52:                4             mean  0.206335677
#> 53:                4           median  0.146995310
#> 54:                5             mean  0.172466252
#> 55:                5           median  0.142389320
#> 56:                1               NA  1.167929414
#> 57:                2               NA  0.573476442
#> 58:                3               NA  0.416546472
#> 59:                4               NA  0.259982000
#> 60:                5               NA  0.218127087
#>     prob.pars.lambda algo.pars.method   result.res